unit test mock_requests
