--- floodviz/test/hydrograph_utils_tests.py
@@ -1,84 +1,0 @@
-import unittest
-import requests_mock
-from floodviz.hydrograph_utils import req_hydrodata, parse_hydrodata
-
-class TestReqHydroData(unittest.TestCase):
-
-  def setUp(self):
-
-    self.valid_data = [{"sourceInfo": {"siteName": "Black Hawk Creek at Hudson, IA", "siteCode": [{"value": "05463500", "network": "NWIS", "agencyCode": "USGS"}], "timeZoneInfo": {"defaultTimeZone": {"zoneOffset": "-06:00", "zoneAbbreviation": "CST"}, "daylightSavingsTimeZone": {"zoneOffset": "-05:00", "zoneAbbreviation": "CDT"}, "siteUsesDaylightSavingsTime": True}, "geoLocation": {"geogLocation": {"srs": "EPSG:4326", "latitude": 42.4077639, "longitude": -92.4632451}, "localSiteXY": []}, "note": [], "siteType": [], "siteProperty": [{"value": "ST", "name": "siteTypeCd"}, {"value": "07080205", "name": "hucCd"}, {"value": "19", "name": "stateCd"}, {"value": "19013", "name": "countyCd"}]}, "variable": {"variableCode": [{"value": "00060", "network": "NWIS", "vocabulary": "NWIS:UnitValues", "variableID": 45807197, "default": True}], "variableName": "Streamflow, ft&#179;/s", "variableDescription": "Discharge, cubic feet per second", "valueType": "Derived Value", "unit": {"unitCode": "ft3/s"}, "options": {"option": [{"name": "Statistic", "optionCode": "00000"}]}, "note": [], "noDataValue": -999999.0, "variableProperty": [], "oid": "45807197"}, "values": [{"value": [{"value": "446", "qualifiers": ["A"], "dateTime": "2008-05-20T00:00:00.000-05:00"}, {"value": "446", "qualifiers": ["A"], "dateTime": "2008-05-20T00:15:00.000-05:00"}, {"value": "446", "qualifiers": ["A"], "dateTime": "2008-05-20T00:30:00.000-05:00"}, {"value": "444", "qualifiers": ["A"], "dateTime": "2008-05-20T00:45:00.000-05:00"}, {"value": "446", "qualifiers": ["A"], "dateTime": "2008-05-20T01:00:00.000-05:00"}, {"value": "446", "qualifiers": ["A"], "dateTime": "2008-05-20T01:15:00.000-05:00"}, {"value": "444", "qualifiers": ["A"], "dateTime": "2008-05-20T01:30:00.000-05:00"}, {"value": "446", "qualifiers": ["A"], "dateTime": "2008-05-20T01:45:00.000-05:00"}, {"value": "446", "qualifiers": ["A"], "dateTime": "2008-05-20T02:00:00.000-05:00"}, {"value": "442", "qualifiers": ["A"], "dateTime": "2008-05-20T02:15:00.000-05:00"}, {"value": "442", "qualifiers": ["A"], "dateTime": "2008-05-20T02:30:00.000-05:00"}, {"value": "446", "qualifiers": ["A"], "dateTime": "2008-05-20T02:45:00.000-05:00"}, {"value": "444", "qualifiers": ["A"], "dateTime": "2008-05-20T03:00:00.000-05:00"}, {"value": "448", "qualifiers": ["A"], "dateTime": "2008-05-20T03:15:00.000-05:00"}, {"value": "444", "qualifiers": ["A"], "dateTime": "2008-05-20T03:30:00.000-05:00"}, {"value": "442", "qualifiers": ["A"], "dateTime": "2008-05-20T03:45:00.000-05:00"}, {"value": "442", "qualifiers": ["A"], "dateTime": "2008-05-20T04:00:00.000-05:00"}, {"value": "439", "qualifiers": ["A"], "dateTime": "2008-05-20T04:15:00.000-05:00"}, {"value": "442", "qualifiers": ["A"], "dateTime": "2008-05-20T04:30:00.000-05:00"}, {"value": "444", "qualifiers": ["A"], "dateTime": "2008-05-20T04:45:00.000-05:00"}, {"value": "442", "qualifiers": ["A"], "dateTime": "2008-05-20T05:00:00.000-05:00"}, {"value": "441", "qualifiers": ["A"], "dateTime": "2008-05-20T05:15:00.000-05:00"}, {"value": "441", "qualifiers": ["A"], "dateTime": "2008-05-20T05:30:00.000-05:00"}, {"value": "442", "qualifiers": ["A"], "dateTime": "2008-05-20T05:45:00.000-05:00"}, {"value": "441", "qualifiers": ["A"], "dateTime": "2008-05-20T06:00:00.000-05:00"}, {"value": "444", "qualifiers": ["A"], "dateTime": "2008-05-20T06:15:00.000-05:00"}, {"value": "444", "qualifiers": ["A"], "dateTime": "2008-05-20T06:30:00.000-05:00"}, {"value": "441", "qualifiers": ["A"], "dateTime": "2008-05-20T06:45:00.000-05:00"}, {"value": "442", "qualifiers": ["A"], "dateTime": "2008-05-20T07:00:00.000-05:00"}, {"value": "441", "qualifiers": ["A"], "dateTime": "2008-05-20T07:15:00.000-05:00"}, {"value": "442", "qualifiers": ["A"], "dateTime": "2008-05-20T07:30:00.000-05:00"}, {"value": "439", "qualifiers": ["A"], "dateTime": "2008-05-20T07:45:00.000-05:00"}, {"value": "439", "qualifiers": ["A"], "dateTime": "2008-05-20T08:00:00.000-05:00"}, {"value": "441", "qualifiers": ["A"], "dateTime": "2008-05-20T08:15:00.000-05:00"}, {"value": "439", "qualifiers": ["A"], "dateTime": "2008-05-20T08:30:00.000-05:00"}, {"value": "437", "qualifiers": ["A"], "dateTime": "2008-05-20T08:45:00.000-05:00"}, {"value": "437", "qualifiers": ["A"], "dateTime": "2008-05-20T09:00:00.000-05:00"}, {"value": "441", "qualifiers": ["A"], "dateTime": "2008-05-20T09:15:00.000-05:00"}, {"value": "437", "qualifiers": ["A"], "dateTime": "2008-05-20T09:30:00.000-05:00"}, {"value": "435", "qualifiers": ["A"], "dateTime": "2008-05-20T09:45:00.000-05:00"}, {"value": "441", "qualifiers": ["A"], "dateTime": "2008-05-20T10:00:00.000-05:00"}, {"value": "435", "qualifiers": ["A"], "dateTime": "2008-05-20T10:15:00.000-05:00"}, {"value": "435", "qualifiers": ["A"], "dateTime": "2008-05-20T10:30:00.000-05:00"}, {"value": "439", "qualifiers": ["A"], "dateTime": "2008-05-20T10:45:00.000-05:00"}, {"value": "437", "qualifiers": ["A"], "dateTime": "2008-05-20T11:00:00.000-05:00"}, {"value": "437", "qualifiers": ["A"], "dateTime": "2008-05-20T11:15:00.000-05:00"}, {"value": "437", "qualifiers": ["A"], "dateTime": "2008-05-20T11:30:00.000-05:00"}, {"value": "433", "qualifiers": ["A"], "dateTime": "2008-05-20T11:45:00.000-05:00"}, {"value": "433", "qualifiers": ["A"], "dateTime": "2008-05-20T12:00:00.000-05:00"}, {"value": "437", "qualifiers": ["A"], "dateTime": "2008-05-20T12:15:00.000-05:00"}, {"value": "433", "qualifiers": ["A"], "dateTime": "2008-05-20T12:30:00.000-05:00"}, {"value": "433", "qualifiers": ["A"], "dateTime": "2008-05-20T12:45:00.000-05:00"}, {"value": "435", "qualifiers": ["A"], "dateTime": "2008-05-20T13:00:00.000-05:00"}, {"value": "437", "qualifiers": ["A"], "dateTime": "2008-05-20T13:15:00.000-05:00"}, {"value": "437", "qualifiers": ["A"], "dateTime": "2008-05-20T13:30:00.000-05:00"}, {"value": "433", "qualifiers": ["A"], "dateTime": "2008-05-20T13:45:00.000-05:00"}, {"value": "433", "qualifiers": ["A"], "dateTime": "2008-05-20T14:00:00.000-05:00"}, {"value": "435", "qualifiers": ["A"], "dateTime": "2008-05-20T14:15:00.000-05:00"}, {"value": "433", "qualifiers": ["A"], "dateTime": "2008-05-20T14:30:00.000-05:00"}, {"value": "435", "qualifiers": ["A"], "dateTime": "2008-05-20T14:45:00.000-05:00"}, {"value": "432", "qualifiers": ["A"], "dateTime": "2008-05-20T15:00:00.000-05:00"}, {"value": "435", "qualifiers": ["A"], "dateTime": "2008-05-20T15:15:00.000-05:00"}, {"value": "432", "qualifiers": ["A"], "dateTime": "2008-05-20T15:30:00.000-05:00"}, {"value": "432", "qualifiers": ["A"], "dateTime": "2008-05-20T15:45:00.000-05:00"}, {"value": "432", "qualifiers": ["A"], "dateTime": "2008-05-20T16:00:00.000-05:00"}, {"value": "432", "qualifiers": ["A"], "dateTime": "2008-05-20T16:15:00.000-05:00"}, {"value": "433", "qualifiers": ["A"], "dateTime": "2008-05-20T16:30:00.000-05:00"}, {"value": "432", "qualifiers": ["A"], "dateTime": "2008-05-20T16:45:00.000-05:00"}, {"value": "428", "qualifiers": ["A"], "dateTime": "2008-05-20T17:00:00.000-05:00"}, {"value": "433", "qualifiers": ["A"], "dateTime": "2008-05-20T17:15:00.000-05:00"}, {"value": "432", "qualifiers": ["A"], "dateTime": "2008-05-20T17:30:00.000-05:00"}, {"value": "430", "qualifiers": ["A"], "dateTime": "2008-05-20T17:45:00.000-05:00"}, {"value": "428", "qualifiers": ["A"], "dateTime": "2008-05-20T18:00:00.000-05:00"}, {"value": "430", "qualifiers": ["A"], "dateTime": "2008-05-20T18:15:00.000-05:00"}, {"value": "426", "qualifiers": ["A"], "dateTime": "2008-05-20T18:30:00.000-05:00"}, {"value": "430", "qualifiers": ["A"], "dateTime": "2008-05-20T18:45:00.000-05:00"}, {"value": "430", "qualifiers": ["A"], "dateTime": "2008-05-20T19:00:00.000-05:00"}, {"value": "428", "qualifiers": ["A"], "dateTime": "2008-05-20T19:15:00.000-05:00"}, {"value": "426", "qualifiers": ["A"], "dateTime": "2008-05-20T19:30:00.000-05:00"}, {"value": "426", "qualifiers": ["A"], "dateTime": "2008-05-20T19:45:00.000-05:00"}, {"value": "426", "qualifiers": ["A"], "dateTime": "2008-05-20T20:00:00.000-05:00"}, {"value": "428", "qualifiers": ["A"], "dateTime": "2008-05-20T20:15:00.000-05:00"}, {"value": "424", "qualifiers": ["A"], "dateTime": "2008-05-20T20:30:00.000-05:00"}, {"value": "426", "qualifiers": ["A"], "dateTime": "2008-05-20T20:45:00.000-05:00"}, {"value": "424", "qualifiers": ["A"], "dateTime": "2008-05-20T21:00:00.000-05:00"}, {"value": "423", "qualifiers": ["A"], "dateTime": "2008-05-20T21:15:00.000-05:00"}, {"value": "421", "qualifiers": ["A"], "dateTime": "2008-05-20T21:30:00.000-05:00"}, {"value": "421", "qualifiers": ["A"], "dateTime": "2008-05-20T21:45:00.000-05:00"}, {"value": "421", "qualifiers": ["A"], "dateTime": "2008-05-20T22:00:00.000-05:00"}, {"value": "424", "qualifiers": ["A"], "dateTime": "2008-05-20T22:15:00.000-05:00"}, {"value": "423", "qualifiers": ["A"], "dateTime": "2008-05-20T22:30:00.000-05:00"}, {"value": "423", "qualifiers": ["A"], "dateTime": "2008-05-20T22:45:00.000-05:00"}, {"value": "423", "qualifiers": ["A"], "dateTime": "2008-05-20T23:00:00.000-05:00"}, {"value": "419", "qualifiers": ["A"], "dateTime": "2008-05-20T23:15:00.000-05:00"}, {"value": "419", "qualifiers": ["A"], "dateTime": "2008-05-20T23:30:00.000-05:00"}, {"value": "419", "qualifiers": ["A"], "dateTime": "2008-05-20T23:45:00.000-05:00"}, {"value": "421", "qualifiers": ["A"], "dateTime": "2008-05-21T00:00:00.000-05:00"}, {"value": "419", "qualifiers": ["A"], "dateTime": "2008-05-21T00:15:00.000-05:00"}, {"value": "421", "qualifiers": ["A"], "dateTime": "2008-05-21T00:30:00.000-05:00"}, {"value": "417", "qualifiers": ["A"], "dateTime": "2008-05-21T00:45:00.000-05:00"}, {"value": "419", "qualifiers": ["A"], "dateTime": "2008-05-21T01:00:00.000-05:00"}, {"value": "421", "qualifiers": ["A"], "dateTime": "2008-05-21T01:15:00.000-05:00"}, {"value": "417", "qualifiers": ["A"], "dateTime": "2008-05-21T01:30:00.000-05:00"}, {"value": "417", "qualifiers": ["A"], "dateTime": "2008-05-21T01:45:00.000-05:00"}, {"value": "417", "qualifiers": ["A"], "dateTime": "2008-05-21T02:00:00.000-05:00"}, {"value": "417", "qualifiers": ["A"], "dateTime": "2008-05-21T02:15:00.000-05:00"}, {"value": "417", "qualifiers": ["A"], "dateTime": "2008-05-21T02:30:00.000-05:00"}, {"value": "417", "qualifiers": ["A"], "dateTime": "2008-05-21T02:45:00.000-05:00"}, {"value": "415", "qualifiers": ["A"], "dateTime": "2008-05-21T03:00:00.000-05:00"}, {"value": "415", "qualifiers": ["A"], "dateTime": "2008-05-21T03:15:00.000-05:00"}, {"value": "417", "qualifiers": ["A"], "dateTime": "2008-05-21T03:30:00.000-05:00"}, {"value": "417", "qualifiers": ["A"], "dateTime": "2008-05-21T03:45:00.000-05:00"}, {"value": "414", "qualifiers": ["A"], "dateTime": "2008-05-21T04:00:00.000-05:00"}, {"value": "415", "qualifiers": ["A"], "dateTime": "2008-05-21T04:15:00.000-05:00"}, {"value": "415", "qualifiers": ["A"], "dateTime": "2008-05-21T04:30:00.000-05:00"}, {"value": "414", "qualifiers": ["A"], "dateTime": "2008-05-21T04:45:00.000-05:00"}, {"value": "415", "qualifiers": ["A"], "dateTime": "2008-05-21T05:00:00.000-05:00"}, {"value": "415", "qualifiers": ["A"], "dateTime": "2008-05-21T05:15:00.000-05:00"}, {"value": "415", "qualifiers": ["A"], "dateTime": "2008-05-21T05:30:00.000-05:00"}, {"value": "412", "qualifiers": ["A"], "dateTime": "2008-05-21T05:45:00.000-05:00"}, {"value": "415", "qualifiers": ["A"], "dateTime": "2008-05-21T06:00:00.000-05:00"}, {"value": "412", "qualifiers": ["A"], "dateTime": "2008-05-21T06:15:00.000-05:00"}, {"value": "412", "qualifiers": ["A"], "dateTime": "2008-05-21T06:30:00.000-05:00"}, {"value": "412", "qualifiers": ["A"], "dateTime": "2008-05-21T06:45:00.000-05:00"}, {"value": "412", "qualifiers": ["A"], "dateTime": "2008-05-21T07:00:00.000-05:00"}, {"value": "414", "qualifiers": ["A"], "dateTime": "2008-05-21T07:15:00.000-05:00"}, {"value": "412", "qualifiers": ["A"], "dateTime": "2008-05-21T07:30:00.000-05:00"}, {"value": "414", "qualifiers": ["A"], "dateTime": "2008-05-21T07:45:00.000-05:00"}, {"value": "414", "qualifiers": ["A"], "dateTime": "2008-05-21T08:00:00.000-05:00"}, {"value": "414", "qualifiers": ["A"], "dateTime": "2008-05-21T08:15:00.000-05:00"}, {"value": "412", "qualifiers": ["A"], "dateTime": "2008-05-21T08:30:00.000-05:00"}, {"value": "410", "qualifiers": ["A"], "dateTime": "2008-05-21T08:45:00.000-05:00"}, {"value": "410", "qualifiers": ["A"], "dateTime": "2008-05-21T09:00:00.000-05:00"}, {"value": "414", "qualifiers": ["A"], "dateTime": "2008-05-21T09:15:00.000-05:00"}, {"value": "412", "qualifiers": ["A"], "dateTime": "2008-05-21T09:30:00.000-05:00"}, {"value": "412", "qualifiers": ["A"], "dateTime": "2008-05-21T09:45:00.000-05:00"}, {"value": "410", "qualifiers": ["A"], "dateTime": "2008-05-21T10:00:00.000-05:00"}, {"value": "412", "qualifiers": ["A"], "dateTime": "2008-05-21T10:15:00.000-05:00"}, {"value": "412", "qualifiers": ["A"], "dateTime": "2008-05-21T10:30:00.000-05:00"}, {"value": "408", "qualifiers": ["A"], "dateTime": "2008-05-21T10:45:00.000-05:00"}, {"value": "412", "qualifiers": ["A"], "dateTime": "2008-05-21T11:00:00.000-05:00"}, {"value": "412", "qualifiers": ["A"], "dateTime": "2008-05-21T11:15:00.000-05:00"}, {"value": "410", "qualifiers": ["A"], "dateTime": "2008-05-21T11:30:00.000-05:00"}, {"value": "408", "qualifiers": ["A"], "dateTime": "2008-05-21T11:45:00.000-05:00"}, {"value": "408", "qualifiers": ["A"], "dateTime": "2008-05-21T12:00:00.000-05:00"}, {"value": "410", "qualifiers": ["A"], "dateTime": "2008-05-21T12:15:00.000-05:00"}, {"value": "408", "qualifiers": ["A"], "dateTime": "2008-05-21T12:30:00.000-05:00"}, {"value": "407", "qualifiers": ["A"], "dateTime": "2008-05-21T12:45:00.000-05:00"}, {"value": "410", "qualifiers": ["A"], "dateTime": "2008-05-21T13:00:00.000-05:00"}, {"value": "407", "qualifiers": ["A"], "dateTime": "2008-05-21T13:15:00.000-05:00"}, {"value": "407", "qualifiers": ["A"], "dateTime": "2008-05-21T13:30:00.000-05:00"}, {"value": "407", "qualifiers": ["A"], "dateTime": "2008-05-21T13:45:00.000-05:00"}, {"value": "408", "qualifiers": ["A"], "dateTime": "2008-05-21T14:00:00.000-05:00"}, {"value": "407", "qualifiers": ["A"], "dateTime": "2008-05-21T14:15:00.000-05:00"}, {"value": "407", "qualifiers": ["A"], "dateTime": "2008-05-21T14:30:00.000-05:00"}, {"value": "407", "qualifiers": ["A"], "dateTime": "2008-05-21T14:45:00.000-05:00"}, {"value": "407", "qualifiers": ["A"], "dateTime": "2008-05-21T15:00:00.000-05:00"}, {"value": "408", "qualifiers": ["A"], "dateTime": "2008-05-21T15:15:00.000-05:00"}, {"value": "407", "qualifiers": ["A"], "dateTime": "2008-05-21T15:30:00.000-05:00"}, {"value": "407", "qualifiers": ["A"], "dateTime": "2008-05-21T15:45:00.000-05:00"}, {"value": "407", "qualifiers": ["A"], "dateTime": "2008-05-21T16:00:00.000-05:00"}, {"value": "407", "qualifiers": ["A"], "dateTime": "2008-05-21T16:15:00.000-05:00"}, {"value": "407", "qualifiers": ["A"], "dateTime": "2008-05-21T16:30:00.000-05:00"}, {"value": "407", "qualifiers": ["A"], "dateTime": "2008-05-21T16:45:00.000-05:00"}, {"value": "407", "qualifiers": ["A"], "dateTime": "2008-05-21T17:00:00.000-05:00"}, {"value": "403", "qualifiers": ["A"], "dateTime": "2008-05-21T17:15:00.000-05:00"}, {"value": "405", "qualifiers": ["A"], "dateTime": "2008-05-21T17:30:00.000-05:00"}, {"value": "407", "qualifiers": ["A"], "dateTime": "2008-05-21T17:45:00.000-05:00"}, {"value": "407", "qualifiers": ["A"], "dateTime": "2008-05-21T18:00:00.000-05:00"}, {"value": "403", "qualifiers": ["A"], "dateTime": "2008-05-21T18:15:00.000-05:00"}, {"value": "403", "qualifiers": ["A"], "dateTime": "2008-05-21T18:30:00.000-05:00"}, {"value": "403", "qualifiers": ["A"], "dateTime": "2008-05-21T18:45:00.000-05:00"}, {"value": "405", "qualifiers": ["A"], "dateTime": "2008-05-21T19:00:00.000-05:00"}, {"value": "403", "qualifiers": ["A"], "dateTime": "2008-05-21T19:15:00.000-05:00"}, {"value": "403", "qualifiers": ["A"], "dateTime": "2008-05-21T19:30:00.000-05:00"}, {"value": "405", "qualifiers": ["A"], "dateTime": "2008-05-21T19:45:00.000-05:00"}, {"value": "405", "qualifiers": ["A"], "dateTime": "2008-05-21T20:00:00.000-05:00"}, {"value": "403", "qualifiers": ["A"], "dateTime": "2008-05-21T20:15:00.000-05:00"}, {"value": "405", "qualifiers": ["A"], "dateTime": "2008-05-21T20:30:00.000-05:00"}, {"value": "401", "qualifiers": ["A"], "dateTime": "2008-05-21T20:45:00.000-05:00"}, {"value": "401", "qualifiers": ["A"], "dateTime": "2008-05-21T21:00:00.000-05:00"}, {"value": "401", "qualifiers": ["A"], "dateTime": "2008-05-21T21:15:00.000-05:00"}, {"value": "403", "qualifiers": ["A"], "dateTime": "2008-05-21T21:30:00.000-05:00"}, {"value": "403", "qualifiers": ["A"], "dateTime": "2008-05-21T21:45:00.000-05:00"}, {"value": "403", "qualifiers": ["A"], "dateTime": "2008-05-21T22:00:00.000-05:00"}, {"value": "401", "qualifiers": ["A"], "dateTime": "2008-05-21T22:15:00.000-05:00"}, {"value": "403", "qualifiers": ["A"], "dateTime": "2008-05-21T22:30:00.000-05:00"}, {"value": "400", "qualifiers": ["A"], "dateTime": "2008-05-21T22:45:00.000-05:00"}, {"value": "403", "qualifiers": ["A"], "dateTime": "2008-05-21T23:00:00.000-05:00"}, {"value": "401", "qualifiers": ["A"], "dateTime": "2008-05-21T23:15:00.000-05:00"}, {"value": "400", "qualifiers": ["A"], "dateTime": "2008-05-21T23:30:00.000-05:00"}, {"value": "403", "qualifiers": ["A"], "dateTime": "2008-05-21T23:45:00.000-05:00"}, {"value": "403", "qualifiers": ["A"], "dateTime": "2008-05-22T00:00:00.000-05:00"}, {"value": "401", "qualifiers": ["A"], "dateTime": "2008-05-22T00:15:00.000-05:00"}, {"value": "401", "qualifiers": ["A"], "dateTime": "2008-05-22T00:30:00.000-05:00"}, {"value": "398", "qualifiers": ["A"], "dateTime": "2008-05-22T00:45:00.000-05:00"}, {"value": "398", "qualifiers": ["A"], "dateTime": "2008-05-22T01:00:00.000-05:00"}, {"value": "401", "qualifiers": ["A"], "dateTime": "2008-05-22T01:15:00.000-05:00"}, {"value": "401", "qualifiers": ["A"], "dateTime": "2008-05-22T01:30:00.000-05:00"}, {"value": "401", "qualifiers": ["A"], "dateTime": "2008-05-22T01:45:00.000-05:00"}, {"value": "398", "qualifiers": ["A"], "dateTime": "2008-05-22T02:00:00.000-05:00"}, {"value": "401", "qualifiers": ["A"], "dateTime": "2008-05-22T02:15:00.000-05:00"}, {"value": "398", "qualifiers": ["A"], "dateTime": "2008-05-22T02:30:00.000-05:00"}, {"value": "400", "qualifiers": ["A"], "dateTime": "2008-05-22T02:45:00.000-05:00"}, {"value": "401", "qualifiers": ["A"], "dateTime": "2008-05-22T03:00:00.000-05:00"}, {"value": "400", "qualifiers": ["A"], "dateTime": "2008-05-22T03:15:00.000-05:00"}, {"value": "400", "qualifiers": ["A"], "dateTime": "2008-05-22T03:30:00.000-05:00"}, {"value": "394", "qualifiers": ["A"], "dateTime": "2008-05-22T03:45:00.000-05:00"}, {"value": "400", "qualifiers": ["A"], "dateTime": "2008-05-22T04:00:00.000-05:00"}, {"value": "396", "qualifiers": ["A"], "dateTime": "2008-05-22T04:15:00.000-05:00"}, {"value": "398", "qualifiers": ["A"], "dateTime": "2008-05-22T04:30:00.000-05:00"}, {"value": "394", "qualifiers": ["A"], "dateTime": "2008-05-22T04:45:00.000-05:00"}, {"value": "398", "qualifiers": ["A"], "dateTime": "2008-05-22T05:00:00.000-05:00"}, {"value": "398", "qualifiers": ["A"], "dateTime": "2008-05-22T05:15:00.000-05:00"}, {"value": "396", "qualifiers": ["A"], "dateTime": "2008-05-22T05:30:00.000-05:00"}, {"value": "398", "qualifiers": ["A"], "dateTime": "2008-05-22T05:45:00.000-05:00"}, {"value": "394", "qualifiers": ["A"], "dateTime": "2008-05-22T06:00:00.000-05:00"}, {"value": "394", "qualifiers": ["A"], "dateTime": "2008-05-22T06:15:00.000-05:00"}, {"value": "396", "qualifiers": ["A"], "dateTime": "2008-05-22T06:30:00.000-05:00"}, {"value": "396", "qualifiers": ["A"], "dateTime": "2008-05-22T06:45:00.000-05:00"}, {"value": "394", "qualifiers": ["A"], "dateTime": "2008-05-22T07:00:00.000-05:00"}, {"value": "393", "qualifiers": ["A"], "dateTime": "2008-05-22T07:15:00.000-05:00"}, {"value": "391", "qualifiers": ["A"], "dateTime": "2008-05-22T07:30:00.000-05:00"}, {"value": "396", "qualifiers": ["A"], "dateTime": "2008-05-22T07:45:00.000-05:00"}, {"value": "393", "qualifiers": ["A"], "dateTime": "2008-05-22T08:00:00.000-05:00"}, {"value": "393", "qualifiers": ["A"], "dateTime": "2008-05-22T08:15:00.000-05:00"}, {"value": "394", "qualifiers": ["A"], "dateTime": "2008-05-22T08:30:00.000-05:00"}, {"value": "394", "qualifiers": ["A"], "dateTime": "2008-05-22T08:45:00.000-05:00"}, {"value": "393", "qualifiers": ["A"], "dateTime": "2008-05-22T09:00:00.000-05:00"}, {"value": "396", "qualifiers": ["A"], "dateTime": "2008-05-22T09:15:00.000-05:00"}, {"value": "393", "qualifiers": ["A"], "dateTime": "2008-05-22T09:30:00.000-05:00"}, {"value": "393", "qualifiers": ["A"], "dateTime": "2008-05-22T09:45:00.000-05:00"}, {"value": "394", "qualifiers": ["A"], "dateTime": "2008-05-22T10:00:00.000-05:00"}, {"value": "393", "qualifiers": ["A"], "dateTime": "2008-05-22T10:15:00.000-05:00"}, {"value": "393", "qualifiers": ["A"], "dateTime": "2008-05-22T10:30:00.000-05:00"}, {"value": "391", "qualifiers": ["A"], "dateTime": "2008-05-22T10:45:00.000-05:00"}, {"value": "393", "qualifiers": ["A"], "dateTime": "2008-05-22T11:00:00.000-05:00"}, {"value": "394", "qualifiers": ["A"], "dateTime": "2008-05-22T11:15:00.000-05:00"}, {"value": "391", "qualifiers": ["A"], "dateTime": "2008-05-22T11:30:00.000-05:00"}, {"value": "394", "qualifiers": ["A"], "dateTime": "2008-05-22T11:45:00.000-05:00"}, {"value": "393", "qualifiers": ["A"], "dateTime": "2008-05-22T12:00:00.000-05:00"}, {"value": "394", "qualifiers": ["A"], "dateTime": "2008-05-22T12:15:00.000-05:00"}, {"value": "393", "qualifiers": ["A"], "dateTime": "2008-05-22T12:30:00.000-05:00"}, {"value": "393", "qualifiers": ["A"], "dateTime": "2008-05-22T12:45:00.000-05:00"}, {"value": "393", "qualifiers": ["A"], "dateTime": "2008-05-22T13:00:00.000-05:00"}, {"value": "391", "qualifiers": ["A"], "dateTime": "2008-05-22T13:15:00.000-05:00"}, {"value": "391", "qualifiers": ["A"], "dateTime": "2008-05-22T13:30:00.000-05:00"}, {"value": "391", "qualifiers": ["A"], "dateTime": "2008-05-22T13:45:00.000-05:00"}, {"value": "391", "qualifiers": ["A"], "dateTime": "2008-05-22T14:00:00.000-05:00"}, {"value": "391", "qualifiers": ["A"], "dateTime": "2008-05-22T14:15:00.000-05:00"}, {"value": "389", "qualifiers": ["A"], "dateTime": "2008-05-22T14:30:00.000-05:00"}, {"value": "391", "qualifiers": ["A"], "dateTime": "2008-05-22T14:45:00.000-05:00"}, {"value": "388", "qualifiers": ["A"], "dateTime": "2008-05-22T15:00:00.000-05:00"}, {"value": "389", "qualifiers": ["A"], "dateTime": "2008-05-22T15:15:00.000-05:00"}, {"value": "391", "qualifiers": ["A"], "dateTime": "2008-05-22T15:30:00.000-05:00"}, {"value": "389", "qualifiers": ["A"], "dateTime": "2008-05-22T15:45:00.000-05:00"}, {"value": "388", "qualifiers": ["A"], "dateTime": "2008-05-22T16:00:00.000-05:00"}, {"value": "388", "qualifiers": ["A"], "dateTime": "2008-05-22T16:15:00.000-05:00"}, {"value": "391", "qualifiers": ["A"], "dateTime": "2008-05-22T16:30:00.000-05:00"}, {"value": "389", "qualifiers": ["A"], "dateTime": "2008-05-22T16:45:00.000-05:00"}, {"value": "389", "qualifiers": ["A"], "dateTime": "2008-05-22T17:00:00.000-05:00"}, {"value": "391", "qualifiers": ["A"], "dateTime": "2008-05-22T17:15:00.000-05:00"}, {"value": "391", "qualifiers": ["A"], "dateTime": "2008-05-22T17:30:00.000-05:00"}, {"value": "391", "qualifiers": ["A"], "dateTime": "2008-05-22T17:45:00.000-05:00"}, {"value": "389", "qualifiers": ["A"], "dateTime": "2008-05-22T18:00:00.000-05:00"}, {"value": "391", "qualifiers": ["A"], "dateTime": "2008-05-22T18:15:00.000-05:00"}, {"value": "389", "qualifiers": ["A"], "dateTime": "2008-05-22T18:30:00.000-05:00"}, {"value": "391", "qualifiers": ["A"], "dateTime": "2008-05-22T18:45:00.000-05:00"}, {"value": "389", "qualifiers": ["A"], "dateTime": "2008-05-22T19:00:00.000-05:00"}, {"value": "389", "qualifiers": ["A"], "dateTime": "2008-05-22T19:15:00.000-05:00"}, {"value": "388", "qualifiers": ["A"], "dateTime": "2008-05-22T19:30:00.000-05:00"}, {"value": "386", "qualifiers": ["A"], "dateTime": "2008-05-22T19:45:00.000-05:00"}, {"value": "389", "qualifiers": ["A"], "dateTime": "2008-05-22T20:00:00.000-05:00"}, {"value": "386", "qualifiers": ["A"], "dateTime": "2008-05-22T20:15:00.000-05:00"}, {"value": "389", "qualifiers": ["A"], "dateTime": "2008-05-22T20:30:00.000-05:00"}, {"value": "388", "qualifiers": ["A"], "dateTime": "2008-05-22T20:45:00.000-05:00"}, {"value": "384", "qualifiers": ["A"], "dateTime": "2008-05-22T21:00:00.000-05:00"}, {"value": "386", "qualifiers": ["A"], "dateTime": "2008-05-22T21:15:00.000-05:00"}, {"value": "388", "qualifiers": ["A"], "dateTime": "2008-05-22T21:30:00.000-05:00"}, {"value": "386", "qualifiers": ["A"], "dateTime": "2008-05-22T21:45:00.000-05:00"}, {"value": "384", "qualifiers": ["A"], "dateTime": "2008-05-22T22:00:00.000-05:00"}, {"value": "386", "qualifiers": ["A"], "dateTime": "2008-05-22T22:15:00.000-05:00"}, {"value": "384", "qualifiers": ["A"], "dateTime": "2008-05-22T22:30:00.000-05:00"}, {"value": "384", "qualifiers": ["A"], "dateTime": "2008-05-22T22:45:00.000-05:00"}, {"value": "388", "qualifiers": ["A"], "dateTime": "2008-05-22T23:00:00.000-05:00"}, {"value": "382", "qualifiers": ["A"], "dateTime": "2008-05-22T23:15:00.000-05:00"}, {"value": "384", "qualifiers": ["A"], "dateTime": "2008-05-22T23:30:00.000-05:00"}, {"value": "386", "qualifiers": ["A"], "dateTime": "2008-05-22T23:45:00.000-05:00"}, {"value": "384", "qualifiers": ["A"], "dateTime": "2008-05-23T00:00:00.000-05:00"}, {"value": "386", "qualifiers": ["A"], "dateTime": "2008-05-23T00:15:00.000-05:00"}, {"value": "382", "qualifiers": ["A"], "dateTime": "2008-05-23T00:30:00.000-05:00"}, {"value": "384", "qualifiers": ["A"], "dateTime": "2008-05-23T00:45:00.000-05:00"}, {"value": "386", "qualifiers": ["A"], "dateTime": "2008-05-23T01:00:00.000-05:00"}, {"value": "384", "qualifiers": ["A"], "dateTime": "2008-05-23T01:15:00.000-05:00"}, {"value": "384", "qualifiers": ["A"], "dateTime": "2008-05-23T01:30:00.000-05:00"}, {"value": "382", "qualifiers": ["A"], "dateTime": "2008-05-23T01:45:00.000-05:00"}, {"value": "381", "qualifiers": ["A"], "dateTime": "2008-05-23T02:00:00.000-05:00"}, {"value": "382", "qualifiers": ["A"], "dateTime": "2008-05-23T02:15:00.000-05:00"}, {"value": "381", "qualifiers": ["A"], "dateTime": "2008-05-23T02:30:00.000-05:00"}, {"value": "382", "qualifiers": ["A"], "dateTime": "2008-05-23T02:45:00.000-05:00"}, {"value": "381", "qualifiers": ["A"], "dateTime": "2008-05-23T03:00:00.000-05:00"}, {"value": "384", "qualifiers": ["A"], "dateTime": "2008-05-23T03:15:00.000-05:00"}, {"value": "384", "qualifiers": ["A"], "dateTime": "2008-05-23T03:30:00.000-05:00"}, {"value": "381", "qualifiers": ["A"], "dateTime": "2008-05-23T03:45:00.000-05:00"}, {"value": "382", "qualifiers": ["A"], "dateTime": "2008-05-23T04:00:00.000-05:00"}, {"value": "382", "qualifiers": ["A"], "dateTime": "2008-05-23T04:15:00.000-05:00"}, {"value": "381", "qualifiers": ["A"], "dateTime": "2008-05-23T04:30:00.000-05:00"}, {"value": "382", "qualifiers": ["A"], "dateTime": "2008-05-23T04:45:00.000-05:00"}, {"value": "379", "qualifiers": ["A"], "dateTime": "2008-05-23T05:00:00.000-05:00"}, {"value": "382", "qualifiers": ["A"], "dateTime": "2008-05-23T05:15:00.000-05:00"}, {"value": "382", "qualifiers": ["A"], "dateTime": "2008-05-23T05:30:00.000-05:00"}, {"value": "379", "qualifiers": ["A"], "dateTime": "2008-05-23T05:45:00.000-05:00"}, {"value": "377", "qualifiers": ["A"], "dateTime": "2008-05-23T06:00:00.000-05:00"}, {"value": "377", "qualifiers": ["A"], "dateTime": "2008-05-23T06:15:00.000-05:00"}, {"value": "379", "qualifiers": ["A"], "dateTime": "2008-05-23T06:30:00.000-05:00"}, {"value": "381", "qualifiers": ["A"], "dateTime": "2008-05-23T06:45:00.000-05:00"}, {"value": "377", "qualifiers": ["A"], "dateTime": "2008-05-23T07:00:00.000-05:00"}, {"value": "377", "qualifiers": ["A"], "dateTime": "2008-05-23T07:15:00.000-05:00"}, {"value": "379", "qualifiers": ["A"], "dateTime": "2008-05-23T07:30:00.000-05:00"}, {"value": "381", "qualifiers": ["A"], "dateTime": "2008-05-23T07:45:00.000-05:00"}, {"value": "379", "qualifiers": ["A"], "dateTime": "2008-05-23T08:00:00.000-05:00"}, {"value": "379", "qualifiers": ["A"], "dateTime": "2008-05-23T08:15:00.000-05:00"}, {"value": "381", "qualifiers": ["A"], "dateTime": "2008-05-23T08:30:00.000-05:00"}, {"value": "381", "qualifiers": ["A"], "dateTime": "2008-05-23T08:45:00.000-05:00"}, {"value": "381", "qualifiers": ["A"], "dateTime": "2008-05-23T09:00:00.000-05:00"}, {"value": "384", "qualifiers": ["A"], "dateTime": "2008-05-23T09:15:00.000-05:00"}, {"value": "382", "qualifiers": ["A"], "dateTime": "2008-05-23T09:30:00.000-05:00"}, {"value": "382", "qualifiers": ["A"], "dateTime": "2008-05-23T09:45:00.000-05:00"}, {"value": "384", "qualifiers": ["A"], "dateTime": "2008-05-23T10:00:00.000-05:00"}, {"value": "382", "qualifiers": ["A"], "dateTime": "2008-05-23T10:15:00.000-05:00"}, {"value": "382", "qualifiers": ["A"], "dateTime": "2008-05-23T10:30:00.000-05:00"}, {"value": "384", "qualifiers": ["A"], "dateTime": "2008-05-23T10:45:00.000-05:00"}, {"value": "388", "qualifiers": ["A"], "dateTime": "2008-05-23T11:00:00.000-05:00"}, {"value": "388", "qualifiers": ["A"], "dateTime": "2008-05-23T11:15:00.000-05:00"}, {"value": "389", "qualifiers": ["A"], "dateTime": "2008-05-23T11:30:00.000-05:00"}, {"value": "384", "qualifiers": ["A"], "dateTime": "2008-05-23T11:45:00.000-05:00"}, {"value": "384", "qualifiers": ["A"], "dateTime": "2008-05-23T12:00:00.000-05:00"}, {"value": "384", "qualifiers": ["A"], "dateTime": "2008-05-23T12:15:00.000-05:00"}, {"value": "384", "qualifiers": ["A"], "dateTime": "2008-05-23T12:30:00.000-05:00"}, {"value": "386", "qualifiers": ["A"], "dateTime": "2008-05-23T12:45:00.000-05:00"}, {"value": "384", "qualifiers": ["A"], "dateTime": "2008-05-23T13:00:00.000-05:00"}, {"value": "386", "qualifiers": ["A"], "dateTime": "2008-05-23T13:15:00.000-05:00"}, {"value": "384", "qualifiers": ["A"], "dateTime": "2008-05-23T13:30:00.000-05:00"}, {"value": "384", "qualifiers": ["A"], "dateTime": "2008-05-23T13:45:00.000-05:00"}, {"value": "382", "qualifiers": ["A"], "dateTime": "2008-05-23T14:00:00.000-05:00"}, {"value": "388", "qualifiers": ["A"], "dateTime": "2008-05-23T14:15:00.000-05:00"}, {"value": "384", "qualifiers": ["A"], "dateTime": "2008-05-23T14:30:00.000-05:00"}, {"value": "388", "qualifiers": ["A"], "dateTime": "2008-05-23T14:45:00.000-05:00"}, {"value": "381", "qualifiers": ["A"], "dateTime": "2008-05-23T15:00:00.000-05:00"}, {"value": "381", "qualifiers": ["A"], "dateTime": "2008-05-23T15:15:00.000-05:00"}, {"value": "382", "qualifiers": ["A"], "dateTime": "2008-05-23T15:30:00.000-05:00"}, {"value": "382", "qualifiers": ["A"], "dateTime": "2008-05-23T15:45:00.000-05:00"}, {"value": "384", "qualifiers": ["A"], "dateTime": "2008-05-23T16:00:00.000-05:00"}, {"value": "381", "qualifiers": ["A"], "dateTime": "2008-05-23T16:15:00.000-05:00"}, {"value": "381", "qualifiers": ["A"], "dateTime": "2008-05-23T16:30:00.000-05:00"}, {"value": "384", "qualifiers": ["A"], "dateTime": "2008-05-23T16:45:00.000-05:00"}, {"value": "382", "qualifiers": ["A"], "dateTime": "2008-05-23T17:00:00.000-05:00"}, {"value": "384", "qualifiers": ["A"], "dateTime": "2008-05-23T17:15:00.000-05:00"}, {"value": "382", "qualifiers": ["A"], "dateTime": "2008-05-23T17:30:00.000-05:00"}, {"value": "382", "qualifiers": ["A"], "dateTime": "2008-05-23T17:45:00.000-05:00"}, {"value": "386", "qualifiers": ["A"], "dateTime": "2008-05-23T18:00:00.000-05:00"}, {"value": "386", "qualifiers": ["A"], "dateTime": "2008-05-23T18:15:00.000-05:00"}, {"value": "382", "qualifiers": ["A"], "dateTime": "2008-05-23T18:30:00.000-05:00"}, {"value": "381", "qualifiers": ["A"], "dateTime": "2008-05-23T18:45:00.000-05:00"}, {"value": "384", "qualifiers": ["A"], "dateTime": "2008-05-23T19:00:00.000-05:00"}, {"value": "382", "qualifiers": ["A"], "dateTime": "2008-05-23T19:15:00.000-05:00"}, {"value": "386", "qualifiers": ["A"], "dateTime": "2008-05-23T19:30:00.000-05:00"}, {"value": "384", "qualifiers": ["A"], "dateTime": "2008-05-23T19:45:00.000-05:00"}, {"value": "384", "qualifiers": ["A"], "dateTime": "2008-05-23T20:00:00.000-05:00"}, {"value": "386", "qualifiers": ["A"], "dateTime": "2008-05-23T20:15:00.000-05:00"}, {"value": "386", "qualifiers": ["A"], "dateTime": "2008-05-23T20:30:00.000-05:00"}, {"value": "382", "qualifiers": ["A"], "dateTime": "2008-05-23T20:45:00.000-05:00"}, {"value": "382", "qualifiers": ["A"], "dateTime": "2008-05-23T21:00:00.000-05:00"}, {"value": "384", "qualifiers": ["A"], "dateTime": "2008-05-23T21:15:00.000-05:00"}, {"value": "382", "qualifiers": ["A"], "dateTime": "2008-05-23T21:30:00.000-05:00"}, {"value": "382", "qualifiers": ["A"], "dateTime": "2008-05-23T21:45:00.000-05:00"}, {"value": "384", "qualifiers": ["A"], "dateTime": "2008-05-23T22:00:00.000-05:00"}, {"value": "384", "qualifiers": ["A"], "dateTime": "2008-05-23T22:15:00.000-05:00"}, {"value": "384", "qualifiers": ["A"], "dateTime": "2008-05-23T22:30:00.000-05:00"}, {"value": "381", "qualifiers": ["A"], "dateTime": "2008-05-23T22:45:00.000-05:00"}, {"value": "384", "qualifiers": ["A"], "dateTime": "2008-05-23T23:00:00.000-05:00"}, {"value": "384", "qualifiers": ["A"], "dateTime": "2008-05-23T23:15:00.000-05:00"}, {"value": "381", "qualifiers": ["A"], "dateTime": "2008-05-23T23:30:00.000-05:00"}, {"value": "382", "qualifiers": ["A"], "dateTime": "2008-05-23T23:45:00.000-05:00"}, {"value": "382", "qualifiers": ["A"], "dateTime": "2008-05-24T00:00:00.000-05:00"}, {"value": "386", "qualifiers": ["A"], "dateTime": "2008-05-24T00:15:00.000-05:00"}, {"value": "384", "qualifiers": ["A"], "dateTime": "2008-05-24T00:30:00.000-05:00"}, {"value": "384", "qualifiers": ["A"], "dateTime": "2008-05-24T00:45:00.000-05:00"}, {"value": "381", "qualifiers": ["A"], "dateTime": "2008-05-24T01:00:00.000-05:00"}, {"value": "386", "qualifiers": ["A"], "dateTime": "2008-05-24T01:15:00.000-05:00"}, {"value": "384", "qualifiers": ["A"], "dateTime": "2008-05-24T01:30:00.000-05:00"}, {"value": "386", "qualifiers": ["A"], "dateTime": "2008-05-24T01:45:00.000-05:00"}, {"value": "382", "qualifiers": ["A"], "dateTime": "2008-05-24T02:00:00.000-05:00"}, {"value": "384", "qualifiers": ["A"], "dateTime": "2008-05-24T02:15:00.000-05:00"}, {"value": "382", "qualifiers": ["A"], "dateTime": "2008-05-24T02:30:00.000-05:00"}, {"value": "386", "qualifiers": ["A"], "dateTime": "2008-05-24T02:45:00.000-05:00"}, {"value": "384", "qualifiers": ["A"], "dateTime": "2008-05-24T03:00:00.000-05:00"}, {"value": "386", "qualifiers": ["A"], "dateTime": "2008-05-24T03:15:00.000-05:00"}, {"value": "382", "qualifiers": ["A"], "dateTime": "2008-05-24T03:30:00.000-05:00"}, {"value": "382", "qualifiers": ["A"], "dateTime": "2008-05-24T03:45:00.000-05:00"}, {"value": "386", "qualifiers": ["A"], "dateTime": "2008-05-24T04:00:00.000-05:00"}, {"value": "384", "qualifiers": ["A"], "dateTime": "2008-05-24T04:15:00.000-05:00"}, {"value": "384", "qualifiers": ["A"], "dateTime": "2008-05-24T04:30:00.000-05:00"}, {"value": "384", "qualifiers": ["A"], "dateTime": "2008-05-24T04:45:00.000-05:00"}, {"value": "381", "qualifiers": ["A"], "dateTime": "2008-05-24T05:00:00.000-05:00"}, {"value": "381", "qualifiers": ["A"], "dateTime": "2008-05-24T05:15:00.000-05:00"}, {"value": "386", "qualifiers": ["A"], "dateTime": "2008-05-24T05:30:00.000-05:00"}, {"value": "384", "qualifiers": ["A"], "dateTime": "2008-05-24T05:45:00.000-05:00"}, {"value": "384", "qualifiers": ["A"], "dateTime": "2008-05-24T06:00:00.000-05:00"}, {"value": "382", "qualifiers": ["A"], "dateTime": "2008-05-24T06:15:00.000-05:00"}, {"value": "384", "qualifiers": ["A"], "dateTime": "2008-05-24T06:30:00.000-05:00"}, {"value": "386", "qualifiers": ["A"], "dateTime": "2008-05-24T06:45:00.000-05:00"}, {"value": "382", "qualifiers": ["A"], "dateTime": "2008-05-24T07:00:00.000-05:00"}, {"value": "386", "qualifiers": ["A"], "dateTime": "2008-05-24T07:15:00.000-05:00"}, {"value": "384", "qualifiers": ["A"], "dateTime": "2008-05-24T07:30:00.000-05:00"}, {"value": "382", "qualifiers": ["A"], "dateTime": "2008-05-24T07:45:00.000-05:00"}, {"value": "384", "qualifiers": ["A"], "dateTime": "2008-05-24T08:00:00.000-05:00"}, {"value": "382", "qualifiers": ["A"], "dateTime": "2008-05-24T08:15:00.000-05:00"}, {"value": "384", "qualifiers": ["A"], "dateTime": "2008-05-24T08:30:00.000-05:00"}, {"value": "382", "qualifiers": ["A"], "dateTime": "2008-05-24T08:45:00.000-05:00"}, {"value": "379", "qualifiers": ["A"], "dateTime": "2008-05-24T09:00:00.000-05:00"}, {"value": "384", "qualifiers": ["A"], "dateTime": "2008-05-24T09:15:00.000-05:00"}, {"value": "384", "qualifiers": ["A"], "dateTime": "2008-05-24T09:30:00.000-05:00"}, {"value": "379", "qualifiers": ["A"], "dateTime": "2008-05-24T09:45:00.000-05:00"}, {"value": "382", "qualifiers": ["A"], "dateTime": "2008-05-24T10:00:00.000-05:00"}, {"value": "381", "qualifiers": ["A"], "dateTime": "2008-05-24T10:15:00.000-05:00"}, {"value": "379", "qualifiers": ["A"], "dateTime": "2008-05-24T10:30:00.000-05:00"}, {"value": "382", "qualifiers": ["A"], "dateTime": "2008-05-24T10:45:00.000-05:00"}, {"value": "379", "qualifiers": ["A"], "dateTime": "2008-05-24T11:00:00.000-05:00"}, {"value": "379", "qualifiers": ["A"], "dateTime": "2008-05-24T11:15:00.000-05:00"}, {"value": "382", "qualifiers": ["A"], "dateTime": "2008-05-24T11:30:00.000-05:00"}, {"value": "376", "qualifiers": ["A"], "dateTime": "2008-05-24T11:45:00.000-05:00"}, {"value": "377", "qualifiers": ["A"], "dateTime": "2008-05-24T12:00:00.000-05:00"}, {"value": "379", "qualifiers": ["A"], "dateTime": "2008-05-24T12:15:00.000-05:00"}, {"value": "377", "qualifiers": ["A"], "dateTime": "2008-05-24T12:30:00.000-05:00"}, {"value": "379", "qualifiers": ["A"], "dateTime": "2008-05-24T12:45:00.000-05:00"}, {"value": "377", "qualifiers": ["A"], "dateTime": "2008-05-24T13:00:00.000-05:00"}, {"value": "381", "qualifiers": ["A"], "dateTime": "2008-05-24T13:15:00.000-05:00"}, {"value": "376", "qualifiers": ["A"], "dateTime": "2008-05-24T13:30:00.000-05:00"}, {"value": "379", "qualifiers": ["A"], "dateTime": "2008-05-24T13:45:00.000-05:00"}, {"value": "379", "qualifiers": ["A"], "dateTime": "2008-05-24T14:00:00.000-05:00"}, {"value": "376", "qualifiers": ["A"], "dateTime": "2008-05-24T14:15:00.000-05:00"}, {"value": "377", "qualifiers": ["A"], "dateTime": "2008-05-24T14:30:00.000-05:00"}, {"value": "377", "qualifiers": ["A"], "dateTime": "2008-05-24T14:45:00.000-05:00"}, {"value": "374", "qualifiers": ["A"], "dateTime": "2008-05-24T15:00:00.000-05:00"}, {"value": "374", "qualifiers": ["A"], "dateTime": "2008-05-24T15:15:00.000-05:00"}, {"value": "377", "qualifiers": ["A"], "dateTime": "2008-05-24T15:30:00.000-05:00"}, {"value": "377", "qualifiers": ["A"], "dateTime": "2008-05-24T15:45:00.000-05:00"}, {"value": "374", "qualifiers": ["A"], "dateTime": "2008-05-24T16:00:00.000-05:00"}, {"value": "376", "qualifiers": ["A"], "dateTime": "2008-05-24T16:15:00.000-05:00"}, {"value": "376", "qualifiers": ["A"], "dateTime": "2008-05-24T16:30:00.000-05:00"}, {"value": "374", "qualifiers": ["A"], "dateTime": "2008-05-24T16:45:00.000-05:00"}, {"value": "376", "qualifiers": ["A"], "dateTime": "2008-05-24T17:00:00.000-05:00"}, {"value": "374", "qualifiers": ["A"], "dateTime": "2008-05-24T17:15:00.000-05:00"}, {"value": "376", "qualifiers": ["A"], "dateTime": "2008-05-24T17:30:00.000-05:00"}, {"value": "374", "qualifiers": ["A"], "dateTime": "2008-05-24T17:45:00.000-05:00"}, {"value": "372", "qualifiers": ["A"], "dateTime": "2008-05-24T18:00:00.000-05:00"}, {"value": "374", "qualifiers": ["A"], "dateTime": "2008-05-24T18:15:00.000-05:00"}, {"value": "374", "qualifiers": ["A"], "dateTime": "2008-05-24T18:30:00.000-05:00"}, {"value": "374", "qualifiers": ["A"], "dateTime": "2008-05-24T18:45:00.000-05:00"}, {"value": "374", "qualifiers": ["A"], "dateTime": "2008-05-24T19:00:00.000-05:00"}, {"value": "371", "qualifiers": ["A"], "dateTime": "2008-05-24T19:15:00.000-05:00"}, {"value": "374", "qualifiers": ["A"], "dateTime": "2008-05-24T19:30:00.000-05:00"}, {"value": "372", "qualifiers": ["A"], "dateTime": "2008-05-24T19:45:00.000-05:00"}, {"value": "374", "qualifiers": ["A"], "dateTime": "2008-05-24T20:00:00.000-05:00"}, {"value": "374", "qualifiers": ["A"], "dateTime": "2008-05-24T20:15:00.000-05:00"}, {"value": "371", "qualifiers": ["A"], "dateTime": "2008-05-24T20:30:00.000-05:00"}, {"value": "372", "qualifiers": ["A"], "dateTime": "2008-05-24T20:45:00.000-05:00"}, {"value": "374", "qualifiers": ["A"], "dateTime": "2008-05-24T21:00:00.000-05:00"}, {"value": "372", "qualifiers": ["A"], "dateTime": "2008-05-24T21:15:00.000-05:00"}, {"value": "374", "qualifiers": ["A"], "dateTime": "2008-05-24T21:30:00.000-05:00"}, {"value": "374", "qualifiers": ["A"], "dateTime": "2008-05-24T21:45:00.000-05:00"}, {"value": "374", "qualifiers": ["A"], "dateTime": "2008-05-24T22:00:00.000-05:00"}, {"value": "374", "qualifiers": ["A"], "dateTime": "2008-05-24T22:15:00.000-05:00"}, {"value": "374", "qualifiers": ["A"], "dateTime": "2008-05-24T22:30:00.000-05:00"}, {"value": "371", "qualifiers": ["A"], "dateTime": "2008-05-24T22:45:00.000-05:00"}, {"value": "374", "qualifiers": ["A"], "dateTime": "2008-05-24T23:00:00.000-05:00"}, {"value": "374", "qualifiers": ["A"], "dateTime": "2008-05-24T23:15:00.000-05:00"}, {"value": "371", "qualifiers": ["A"], "dateTime": "2008-05-24T23:30:00.000-05:00"}, {"value": "372", "qualifiers": ["A"], "dateTime": "2008-05-24T23:45:00.000-05:00"}, {"value": "372", "qualifiers": ["A"], "dateTime": "2008-05-25T00:00:00.000-05:00"}, {"value": "371", "qualifiers": ["A"], "dateTime": "2008-05-25T00:15:00.000-05:00"}, {"value": "371", "qualifiers": ["A"], "dateTime": "2008-05-25T00:30:00.000-05:00"}, {"value": "371", "qualifiers": ["A"], "dateTime": "2008-05-25T00:45:00.000-05:00"}, {"value": "371", "qualifiers": ["A"], "dateTime": "2008-05-25T01:00:00.000-05:00"}, {"value": "372", "qualifiers": ["A"], "dateTime": "2008-05-25T01:15:00.000-05:00"}, {"value": "371", "qualifiers": ["A"], "dateTime": "2008-05-25T01:30:00.000-05:00"}, {"value": "369", "qualifiers": ["A"], "dateTime": "2008-05-25T01:45:00.000-05:00"}, {"value": "371", "qualifiers": ["A"], "dateTime": "2008-05-25T02:00:00.000-05:00"}, {"value": "371", "qualifiers": ["A"], "dateTime": "2008-05-25T02:15:00.000-05:00"}, {"value": "371", "qualifiers": ["A"], "dateTime": "2008-05-25T02:30:00.000-05:00"}, {"value": "374", "qualifiers": ["A"], "dateTime": "2008-05-25T02:45:00.000-05:00"}, {"value": "371", "qualifiers": ["A"], "dateTime": "2008-05-25T03:00:00.000-05:00"}, {"value": "369", "qualifiers": ["A"], "dateTime": "2008-05-25T03:15:00.000-05:00"}, {"value": "369", "qualifiers": ["A"], "dateTime": "2008-05-25T03:30:00.000-05:00"}, {"value": "372", "qualifiers": ["A"], "dateTime": "2008-05-25T03:45:00.000-05:00"}, {"value": "369", "qualifiers": ["A"], "dateTime": "2008-05-25T04:00:00.000-05:00"}, {"value": "371", "qualifiers": ["A"], "dateTime": "2008-05-25T04:15:00.000-05:00"}, {"value": "371", "qualifiers": ["A"], "dateTime": "2008-05-25T04:30:00.000-05:00"}, {"value": "369", "qualifiers": ["A"], "dateTime": "2008-05-25T04:45:00.000-05:00"}, {"value": "372", "qualifiers": ["A"], "dateTime": "2008-05-25T05:00:00.000-05:00"}, {"value": "367", "qualifiers": ["A"], "dateTime": "2008-05-25T05:15:00.000-05:00"}, {"value": "371", "qualifiers": ["A"], "dateTime": "2008-05-25T05:30:00.000-05:00"}, {"value": "372", "qualifiers": ["A"], "dateTime": "2008-05-25T05:45:00.000-05:00"}, {"value": "367", "qualifiers": ["A"], "dateTime": "2008-05-25T06:00:00.000-05:00"}, {"value": "367", "qualifiers": ["A"], "dateTime": "2008-05-25T06:15:00.000-05:00"}, {"value": "369", "qualifiers": ["A"], "dateTime": "2008-05-25T06:30:00.000-05:00"}, {"value": "369", "qualifiers": ["A"], "dateTime": "2008-05-25T06:45:00.000-05:00"}, {"value": "369", "qualifiers": ["A"], "dateTime": "2008-05-25T07:00:00.000-05:00"}, {"value": "367", "qualifiers": ["A"], "dateTime": "2008-05-25T07:15:00.000-05:00"}, {"value": "367", "qualifiers": ["A"], "dateTime": "2008-05-25T07:30:00.000-05:00"}, {"value": "367", "qualifiers": ["A"], "dateTime": "2008-05-25T07:45:00.000-05:00"}, {"value": "369", "qualifiers": ["A"], "dateTime": "2008-05-25T08:00:00.000-05:00"}, {"value": "367", "qualifiers": ["A"], "dateTime": "2008-05-25T08:15:00.000-05:00"}, {"value": "369", "qualifiers": ["A"], "dateTime": "2008-05-25T08:30:00.000-05:00"}, {"value": "367", "qualifiers": ["A"], "dateTime": "2008-05-25T08:45:00.000-05:00"}, {"value": "366", "qualifiers": ["A"], "dateTime": "2008-05-25T09:00:00.000-05:00"}, {"value": "367", "qualifiers": ["A"], "dateTime": "2008-05-25T09:15:00.000-05:00"}, {"value": "369", "qualifiers": ["A"], "dateTime": "2008-05-25T09:30:00.000-05:00"}, {"value": "367", "qualifiers": ["A"], "dateTime": "2008-05-25T09:45:00.000-05:00"}, {"value": "369", "qualifiers": ["A"], "dateTime": "2008-05-25T10:00:00.000-05:00"}, {"value": "366", "qualifiers": ["A"], "dateTime": "2008-05-25T10:15:00.000-05:00"}, {"value": "367", "qualifiers": ["A"], "dateTime": "2008-05-25T10:30:00.000-05:00"}, {"value": "364", "qualifiers": ["A"], "dateTime": "2008-05-25T10:45:00.000-05:00"}, {"value": "371", "qualifiers": ["A"], "dateTime": "2008-05-25T11:00:00.000-05:00"}, {"value": "367", "qualifiers": ["A"], "dateTime": "2008-05-25T11:15:00.000-05:00"}, {"value": "367", "qualifiers": ["A"], "dateTime": "2008-05-25T11:30:00.000-05:00"}, {"value": "369", "qualifiers": ["A"], "dateTime": "2008-05-25T11:45:00.000-05:00"}, {"value": "369", "qualifiers": ["A"], "dateTime": "2008-05-25T12:00:00.000-05:00"}, {"value": "367", "qualifiers": ["A"], "dateTime": "2008-05-25T12:15:00.000-05:00"}, {"value": "367", "qualifiers": ["A"], "dateTime": "2008-05-25T12:30:00.000-05:00"}, {"value": "369", "qualifiers": ["A"], "dateTime": "2008-05-25T12:45:00.000-05:00"}, {"value": "369", "qualifiers": ["A"], "dateTime": "2008-05-25T13:00:00.000-05:00"}, {"value": "369", "qualifiers": ["A"], "dateTime": "2008-05-25T13:15:00.000-05:00"}, {"value": "369", "qualifiers": ["A"], "dateTime": "2008-05-25T13:30:00.000-05:00"}, {"value": "367", "qualifiers": ["A"], "dateTime": "2008-05-25T13:45:00.000-05:00"}, {"value": "369", "qualifiers": ["A"], "dateTime": "2008-05-25T14:00:00.000-05:00"}, {"value": "367", "qualifiers": ["A"], "dateTime": "2008-05-25T14:15:00.000-05:00"}, {"value": "364", "qualifiers": ["A"], "dateTime": "2008-05-25T14:30:00.000-05:00"}, {"value": "366", "qualifiers": ["A"], "dateTime": "2008-05-25T14:45:00.000-05:00"}, {"value": "366", "qualifiers": ["A"], "dateTime": "2008-05-25T15:00:00.000-05:00"}, {"value": "367", "qualifiers": ["A"], "dateTime": "2008-05-25T15:15:00.000-05:00"}, {"value": "366", "qualifiers": ["A"], "dateTime": "2008-05-25T15:30:00.000-05:00"}, {"value": "369", "qualifiers": ["A"], "dateTime": "2008-05-25T15:45:00.000-05:00"}, {"value": "369", "qualifiers": ["A"], "dateTime": "2008-05-25T16:00:00.000-05:00"}, {"value": "369", "qualifiers": ["A"], "dateTime": "2008-05-25T16:15:00.000-05:00"}, {"value": "364", "qualifiers": ["A"], "dateTime": "2008-05-25T16:30:00.000-05:00"}, {"value": "369", "qualifiers": ["A"], "dateTime": "2008-05-25T16:45:00.000-05:00"}, {"value": "369", "qualifiers": ["A"], "dateTime": "2008-05-25T17:00:00.000-05:00"}, {"value": "366", "qualifiers": ["A"], "dateTime": "2008-05-25T17:15:00.000-05:00"}, {"value": "366", "qualifiers": ["A"], "dateTime": "2008-05-25T17:30:00.000-05:00"}, {"value": "369", "qualifiers": ["A"], "dateTime": "2008-05-25T17:45:00.000-05:00"}, {"value": "367", "qualifiers": ["A"], "dateTime": "2008-05-25T18:00:00.000-05:00"}, {"value": "364", "qualifiers": ["A"], "dateTime": "2008-05-25T18:15:00.000-05:00"}, {"value": "367", "qualifiers": ["A"], "dateTime": "2008-05-25T18:30:00.000-05:00"}, {"value": "364", "qualifiers": ["A"], "dateTime": "2008-05-25T18:45:00.000-05:00"}, {"value": "362", "qualifiers": ["A"], "dateTime": "2008-05-25T19:00:00.000-05:00"}, {"value": "367", "qualifiers": ["A"], "dateTime": "2008-05-25T19:15:00.000-05:00"}, {"value": "367", "qualifiers": ["A"], "dateTime": "2008-05-25T19:30:00.000-05:00"}, {"value": "364", "qualifiers": ["A"], "dateTime": "2008-05-25T19:45:00.000-05:00"}, {"value": "366", "qualifiers": ["A"], "dateTime": "2008-05-25T20:00:00.000-05:00"}, {"value": "367", "qualifiers": ["A"], "dateTime": "2008-05-25T20:15:00.000-05:00"}, {"value": "367", "qualifiers": ["A"], "dateTime": "2008-05-25T20:30:00.000-05:00"}, {"value": "362", "qualifiers": ["A"], "dateTime": "2008-05-25T20:45:00.000-05:00"}, {"value": "366", "qualifiers": ["A"], "dateTime": "2008-05-25T21:00:00.000-05:00"}, {"value": "364", "qualifiers": ["A"], "dateTime": "2008-05-25T21:15:00.000-05:00"}, {"value": "367", "qualifiers": ["A"], "dateTime": "2008-05-25T21:30:00.000-05:00"}, {"value": "367", "qualifiers": ["A"], "dateTime": "2008-05-25T21:45:00.000-05:00"}, {"value": "367", "qualifiers": ["A"], "dateTime": "2008-05-25T22:00:00.000-05:00"}, {"value": "367", "qualifiers": ["A"], "dateTime": "2008-05-25T22:15:00.000-05:00"}, {"value": "366", "qualifiers": ["A"], "dateTime": "2008-05-25T22:30:00.000-05:00"}, {"value": "364", "qualifiers": ["A"], "dateTime": "2008-05-25T22:45:00.000-05:00"}, {"value": "364", "qualifiers": ["A"], "dateTime": "2008-05-25T23:00:00.000-05:00"}, {"value": "367", "qualifiers": ["A"], "dateTime": "2008-05-25T23:15:00.000-05:00"}, {"value": "367", "qualifiers": ["A"], "dateTime": "2008-05-25T23:30:00.000-05:00"}, {"value": "364", "qualifiers": ["A"], "dateTime": "2008-05-25T23:45:00.000-05:00"}, {"value": "361", "qualifiers": ["A"], "dateTime": "2008-05-26T00:00:00.000-05:00"}, {"value": "366", "qualifiers": ["A"], "dateTime": "2008-05-26T00:15:00.000-05:00"}, {"value": "367", "qualifiers": ["A"], "dateTime": "2008-05-26T00:30:00.000-05:00"}, {"value": "367", "qualifiers": ["A"], "dateTime": "2008-05-26T00:45:00.000-05:00"}, {"value": "364", "qualifiers": ["A"], "dateTime": "2008-05-26T01:00:00.000-05:00"}, {"value": "361", "qualifiers": ["A"], "dateTime": "2008-05-26T01:15:00.000-05:00"}, {"value": "364", "qualifiers": ["A"], "dateTime": "2008-05-26T01:30:00.000-05:00"}, {"value": "364", "qualifiers": ["A"], "dateTime": "2008-05-26T01:45:00.000-05:00"}, {"value": "362", "qualifiers": ["A"], "dateTime": "2008-05-26T02:00:00.000-05:00"}, {"value": "367", "qualifiers": ["A"], "dateTime": "2008-05-26T02:15:00.000-05:00"}, {"value": "367", "qualifiers": ["A"], "dateTime": "2008-05-26T02:30:00.000-05:00"}, {"value": "367", "qualifiers": ["A"], "dateTime": "2008-05-26T02:45:00.000-05:00"}, {"value": "366", "qualifiers": ["A"], "dateTime": "2008-05-26T03:00:00.000-05:00"}, {"value": "364", "qualifiers": ["A"], "dateTime": "2008-05-26T03:15:00.000-05:00"}, {"value": "364", "qualifiers": ["A"], "dateTime": "2008-05-26T03:30:00.000-05:00"}, {"value": "362", "qualifiers": ["A"], "dateTime": "2008-05-26T03:45:00.000-05:00"}, {"value": "367", "qualifiers": ["A"], "dateTime": "2008-05-26T04:00:00.000-05:00"}, {"value": "362", "qualifiers": ["A"], "dateTime": "2008-05-26T04:15:00.000-05:00"}, {"value": "366", "qualifiers": ["A"], "dateTime": "2008-05-26T04:30:00.000-05:00"}, {"value": "362", "qualifiers": ["A"], "dateTime": "2008-05-26T04:45:00.000-05:00"}, {"value": "366", "qualifiers": ["A"], "dateTime": "2008-05-26T05:00:00.000-05:00"}, {"value": "364", "qualifiers": ["A"], "dateTime": "2008-05-26T05:15:00.000-05:00"}, {"value": "364", "qualifiers": ["A"], "dateTime": "2008-05-26T05:30:00.000-05:00"}, {"value": "367", "qualifiers": ["A"], "dateTime": "2008-05-26T05:45:00.000-05:00"}, {"value": "366", "qualifiers": ["A"], "dateTime": "2008-05-26T06:00:00.000-05:00"}, {"value": "366", "qualifiers": ["A"], "dateTime": "2008-05-26T06:15:00.000-05:00"}, {"value": "366", "qualifiers": ["A"], "dateTime": "2008-05-26T06:30:00.000-05:00"}, {"value": "367", "qualifiers": ["A"], "dateTime": "2008-05-26T06:45:00.000-05:00"}, {"value": "366", "qualifiers": ["A"], "dateTime": "2008-05-26T07:00:00.000-05:00"}, {"value": "366", "qualifiers": ["A"], "dateTime": "2008-05-26T07:15:00.000-05:00"}, {"value": "367", "qualifiers": ["A"], "dateTime": "2008-05-26T07:30:00.000-05:00"}, {"value": "362", "qualifiers": ["A"], "dateTime": "2008-05-26T07:45:00.000-05:00"}, {"value": "367", "qualifiers": ["A"], "dateTime": "2008-05-26T08:00:00.000-05:00"}, {"value": "367", "qualifiers": ["A"], "dateTime": "2008-05-26T08:15:00.000-05:00"}, {"value": "364", "qualifiers": ["A"], "dateTime": "2008-05-26T08:30:00.000-05:00"}, {"value": "364", "qualifiers": ["A"], "dateTime": "2008-05-26T08:45:00.000-05:00"}, {"value": "362", "qualifiers": ["A"], "dateTime": "2008-05-26T09:00:00.000-05:00"}, {"value": "366", "qualifiers": ["A"], "dateTime": "2008-05-26T09:15:00.000-05:00"}, {"value": "367", "qualifiers": ["A"], "dateTime": "2008-05-26T09:30:00.000-05:00"}, {"value": "364", "qualifiers": ["A"], "dateTime": "2008-05-26T09:45:00.000-05:00"}, {"value": "362", "qualifiers": ["A"], "dateTime": "2008-05-26T10:00:00.000-05:00"}, {"value": "367", "qualifiers": ["A"], "dateTime": "2008-05-26T10:15:00.000-05:00"}, {"value": "367", "qualifiers": ["A"], "dateTime": "2008-05-26T10:30:00.000-05:00"}, {"value": "364", "qualifiers": ["A"], "dateTime": "2008-05-26T10:45:00.000-05:00"}, {"value": "366", "qualifiers": ["A"], "dateTime": "2008-05-26T11:00:00.000-05:00"}, {"value": "367", "qualifiers": ["A"], "dateTime": "2008-05-26T11:15:00.000-05:00"}, {"value": "362", "qualifiers": ["A"], "dateTime": "2008-05-26T11:30:00.000-05:00"}, {"value": "364", "qualifiers": ["A"], "dateTime": "2008-05-26T11:45:00.000-05:00"}, {"value": "364", "qualifiers": ["A"], "dateTime": "2008-05-26T12:00:00.000-05:00"}, {"value": "369", "qualifiers": ["A"], "dateTime": "2008-05-26T12:15:00.000-05:00"}, {"value": "362", "qualifiers": ["A"], "dateTime": "2008-05-26T12:30:00.000-05:00"}, {"value": "366", "qualifiers": ["A"], "dateTime": "2008-05-26T12:45:00.000-05:00"}, {"value": "362", "qualifiers": ["A"], "dateTime": "2008-05-26T13:00:00.000-05:00"}, {"value": "364", "qualifiers": ["A"], "dateTime": "2008-05-26T13:15:00.000-05:00"}, {"value": "364", "qualifiers": ["A"], "dateTime": "2008-05-26T13:30:00.000-05:00"}, {"value": "362", "qualifiers": ["A"], "dateTime": "2008-05-26T13:45:00.000-05:00"}, {"value": "364", "qualifiers": ["A"], "dateTime": "2008-05-26T14:00:00.000-05:00"}, {"value": "367", "qualifiers": ["A"], "dateTime": "2008-05-26T14:15:00.000-05:00"}, {"value": "366", "qualifiers": ["A"], "dateTime": "2008-05-26T14:30:00.000-05:00"}, {"value": "367", "qualifiers": ["A"], "dateTime": "2008-05-26T14:45:00.000-05:00"}, {"value": "366", "qualifiers": ["A"], "dateTime": "2008-05-26T15:00:00.000-05:00"}, {"value": "364", "qualifiers": ["A"], "dateTime": "2008-05-26T15:15:00.000-05:00"}, {"value": "366", "qualifiers": ["A"], "dateTime": "2008-05-26T15:30:00.000-05:00"}, {"value": "366", "qualifiers": ["A"], "dateTime": "2008-05-26T15:45:00.000-05:00"}, {"value": "366", "qualifiers": ["A"], "dateTime": "2008-05-26T16:00:00.000-05:00"}, {"value": "367", "qualifiers": ["A"], "dateTime": "2008-05-26T16:15:00.000-05:00"}, {"value": "366", "qualifiers": ["A"], "dateTime": "2008-05-26T16:30:00.000-05:00"}, {"value": "366", "qualifiers": ["A"], "dateTime": "2008-05-26T16:45:00.000-05:00"}, {"value": "362", "qualifiers": ["A"], "dateTime": "2008-05-26T17:00:00.000-05:00"}, {"value": "364", "qualifiers": ["A"], "dateTime": "2008-05-26T17:15:00.000-05:00"}, {"value": "367", "qualifiers": ["A"], "dateTime": "2008-05-26T17:30:00.000-05:00"}, {"value": "366", "qualifiers": ["A"], "dateTime": "2008-05-26T17:45:00.000-05:00"}, {"value": "362", "qualifiers": ["A"], "dateTime": "2008-05-26T18:00:00.000-05:00"}, {"value": "366", "qualifiers": ["A"], "dateTime": "2008-05-26T18:15:00.000-05:00"}, {"value": "362", "qualifiers": ["A"], "dateTime": "2008-05-26T18:30:00.000-05:00"}, {"value": "366", "qualifiers": ["A"], "dateTime": "2008-05-26T18:45:00.000-05:00"}, {"value": "361", "qualifiers": ["A"], "dateTime": "2008-05-26T19:00:00.000-05:00"}, {"value": "364", "qualifiers": ["A"], "dateTime": "2008-05-26T19:15:00.000-05:00"}, {"value": "361", "qualifiers": ["A"], "dateTime": "2008-05-26T19:30:00.000-05:00"}, {"value": "364", "qualifiers": ["A"], "dateTime": "2008-05-26T19:45:00.000-05:00"}, {"value": "361", "qualifiers": ["A"], "dateTime": "2008-05-26T20:00:00.000-05:00"}, {"value": "361", "qualifiers": ["A"], "dateTime": "2008-05-26T20:15:00.000-05:00"}, {"value": "361", "qualifiers": ["A"], "dateTime": "2008-05-26T20:30:00.000-05:00"}, {"value": "359", "qualifiers": ["A"], "dateTime": "2008-05-26T20:45:00.000-05:00"}, {"value": "357", "qualifiers": ["A"], "dateTime": "2008-05-26T21:00:00.000-05:00"}, {"value": "362", "qualifiers": ["A"], "dateTime": "2008-05-26T21:15:00.000-05:00"}, {"value": "357", "qualifiers": ["A"], "dateTime": "2008-05-26T21:30:00.000-05:00"}, {"value": "357", "qualifiers": ["A"], "dateTime": "2008-05-26T21:45:00.000-05:00"}, {"value": "357", "qualifiers": ["A"], "dateTime": "2008-05-26T22:00:00.000-05:00"}, {"value": "359", "qualifiers": ["A"], "dateTime": "2008-05-26T22:15:00.000-05:00"}, {"value": "359", "qualifiers": ["A"], "dateTime": "2008-05-26T22:30:00.000-05:00"}, {"value": "359", "qualifiers": ["A"], "dateTime": "2008-05-26T22:45:00.000-05:00"}, {"value": "356", "qualifiers": ["A"], "dateTime": "2008-05-26T23:00:00.000-05:00"}, {"value": "354", "qualifiers": ["A"], "dateTime": "2008-05-26T23:15:00.000-05:00"}, {"value": "356", "qualifiers": ["A"], "dateTime": "2008-05-26T23:30:00.000-05:00"}, {"value": "352", "qualifiers": ["A"], "dateTime": "2008-05-26T23:45:00.000-05:00"}, {"value": "352", "qualifiers": ["A"], "dateTime": "2008-05-27T00:00:00.000-05:00"}, {"value": "357", "qualifiers": ["A"], "dateTime": "2008-05-27T00:15:00.000-05:00"}, {"value": "357", "qualifiers": ["A"], "dateTime": "2008-05-27T00:30:00.000-05:00"}, {"value": "356", "qualifiers": ["A"], "dateTime": "2008-05-27T00:45:00.000-05:00"}, {"value": "357", "qualifiers": ["A"], "dateTime": "2008-05-27T01:00:00.000-05:00"}, {"value": "352", "qualifiers": ["A"], "dateTime": "2008-05-27T01:15:00.000-05:00"}, {"value": "351", "qualifiers": ["A"], "dateTime": "2008-05-27T01:30:00.000-05:00"}, {"value": "354", "qualifiers": ["A"], "dateTime": "2008-05-27T01:45:00.000-05:00"}, {"value": "352", "qualifiers": ["A"], "dateTime": "2008-05-27T02:00:00.000-05:00"}, {"value": "347", "qualifiers": ["A"], "dateTime": "2008-05-27T02:15:00.000-05:00"}, {"value": "352", "qualifiers": ["A"], "dateTime": "2008-05-27T02:30:00.000-05:00"}, {"value": "349", "qualifiers": ["A"], "dateTime": "2008-05-27T02:45:00.000-05:00"}, {"value": "349", "qualifiers": ["A"], "dateTime": "2008-05-27T03:00:00.000-05:00"}, {"value": "352", "qualifiers": ["A"], "dateTime": "2008-05-27T03:15:00.000-05:00"}, {"value": "347", "qualifiers": ["A"], "dateTime": "2008-05-27T03:30:00.000-05:00"}, {"value": "351", "qualifiers": ["A"], "dateTime": "2008-05-27T03:45:00.000-05:00"}, {"value": "349", "qualifiers": ["A"], "dateTime": "2008-05-27T04:00:00.000-05:00"}, {"value": "351", "qualifiers": ["A"], "dateTime": "2008-05-27T04:15:00.000-05:00"}, {"value": "347", "qualifiers": ["A"], "dateTime": "2008-05-27T04:30:00.000-05:00"}, {"value": "349", "qualifiers": ["A"], "dateTime": "2008-05-27T04:45:00.000-05:00"}, {"value": "351", "qualifiers": ["A"], "dateTime": "2008-05-27T05:00:00.000-05:00"}, {"value": "346", "qualifiers": ["A"], "dateTime": "2008-05-27T05:15:00.000-05:00"}, {"value": "347", "qualifiers": ["A"], "dateTime": "2008-05-27T05:30:00.000-05:00"}, {"value": "349", "qualifiers": ["A"], "dateTime": "2008-05-27T05:45:00.000-05:00"}, {"value": "346", "qualifiers": ["A"], "dateTime": "2008-05-27T06:00:00.000-05:00"}, {"value": "346", "qualifiers": ["A"], "dateTime": "2008-05-27T06:15:00.000-05:00"}, {"value": "347", "qualifiers": ["A"], "dateTime": "2008-05-27T06:30:00.000-05:00"}, {"value": "344", "qualifiers": ["A"], "dateTime": "2008-05-27T06:45:00.000-05:00"}, {"value": "344", "qualifiers": ["A"], "dateTime": "2008-05-27T07:00:00.000-05:00"}, {"value": "346", "qualifiers": ["A"], "dateTime": "2008-05-27T07:15:00.000-05:00"}, {"value": "344", "qualifiers": ["A"], "dateTime": "2008-05-27T07:30:00.000-05:00"}, {"value": "346", "qualifiers": ["A"], "dateTime": "2008-05-27T07:45:00.000-05:00"}, {"value": "343", "qualifiers": ["A"], "dateTime": "2008-05-27T08:00:00.000-05:00"}, {"value": "344", "qualifiers": ["A"], "dateTime": "2008-05-27T08:15:00.000-05:00"}, {"value": "346", "qualifiers": ["A"], "dateTime": "2008-05-27T08:30:00.000-05:00"}, {"value": "344", "qualifiers": ["A"], "dateTime": "2008-05-27T08:45:00.000-05:00"}, {"value": "343", "qualifiers": ["A"], "dateTime": "2008-05-27T09:00:00.000-05:00"}, {"value": "343", "qualifiers": ["A"], "dateTime": "2008-05-27T09:15:00.000-05:00"}, {"value": "343", "qualifiers": ["A"], "dateTime": "2008-05-27T09:30:00.000-05:00"}, {"value": "343", "qualifiers": ["A"], "dateTime": "2008-05-27T09:45:00.000-05:00"}, {"value": "341", "qualifiers": ["A"], "dateTime": "2008-05-27T10:00:00.000-05:00"}, {"value": "341", "qualifiers": ["A"], "dateTime": "2008-05-27T10:15:00.000-05:00"}, {"value": "344", "qualifiers": ["A"], "dateTime": "2008-05-27T10:30:00.000-05:00"}, {"value": "339", "qualifiers": ["A"], "dateTime": "2008-05-27T10:45:00.000-05:00"}, {"value": "339", "qualifiers": ["A"], "dateTime": "2008-05-27T11:00:00.000-05:00"}, {"value": "341", "qualifiers": ["A"], "dateTime": "2008-05-27T11:15:00.000-05:00"}, {"value": "344", "qualifiers": ["A"], "dateTime": "2008-05-27T11:30:00.000-05:00"}, {"value": "339", "qualifiers": ["A"], "dateTime": "2008-05-27T11:45:00.000-05:00"}, {"value": "338", "qualifiers": ["A"], "dateTime": "2008-05-27T12:00:00.000-05:00"}, {"value": "341", "qualifiers": ["A"], "dateTime": "2008-05-27T12:15:00.000-05:00"}, {"value": "341", "qualifiers": ["A"], "dateTime": "2008-05-27T12:30:00.000-05:00"}, {"value": "338", "qualifiers": ["A"], "dateTime": "2008-05-27T12:45:00.000-05:00"}, {"value": "341", "qualifiers": ["A"], "dateTime": "2008-05-27T13:00:00.000-05:00"}, {"value": "339", "qualifiers": ["A"], "dateTime": "2008-05-27T13:15:00.000-05:00"}, {"value": "339", "qualifiers": ["A"], "dateTime": "2008-05-27T13:30:00.000-05:00"}, {"value": "338", "qualifiers": ["A"], "dateTime": "2008-05-27T13:45:00.000-05:00"}, {"value": "338", "qualifiers": ["A"], "dateTime": "2008-05-27T14:00:00.000-05:00"}, {"value": "336", "qualifiers": ["A"], "dateTime": "2008-05-27T14:15:00.000-05:00"}, {"value": "339", "qualifiers": ["A"], "dateTime": "2008-05-27T14:30:00.000-05:00"}, {"value": "341", "qualifiers": ["A"], "dateTime": "2008-05-27T14:45:00.000-05:00"}, {"value": "339", "qualifiers": ["A"], "dateTime": "2008-05-27T15:00:00.000-05:00"}, {"value": "341", "qualifiers": ["A"], "dateTime": "2008-05-27T15:15:00.000-05:00"}, {"value": "341", "qualifiers": ["A"], "dateTime": "2008-05-27T15:30:00.000-05:00"}, {"value": "339", "qualifiers": ["A"], "dateTime": "2008-05-27T15:45:00.000-05:00"}, {"value": "338", "qualifiers": ["A"], "dateTime": "2008-05-27T16:00:00.000-05:00"}, {"value": "334", "qualifiers": ["A"], "dateTime": "2008-05-27T16:15:00.000-05:00"}, {"value": "339", "qualifiers": ["A"], "dateTime": "2008-05-27T16:30:00.000-05:00"}, {"value": "336", "qualifiers": ["A"], "dateTime": "2008-05-27T16:45:00.000-05:00"}, {"value": "336", "qualifiers": ["A"], "dateTime": "2008-05-27T17:00:00.000-05:00"}, {"value": "338", "qualifiers": ["A"], "dateTime": "2008-05-27T17:15:00.000-05:00"}, {"value": "336", "qualifiers": ["A"], "dateTime": "2008-05-27T17:30:00.000-05:00"}, {"value": "339", "qualifiers": ["A"], "dateTime": "2008-05-27T17:45:00.000-05:00"}, {"value": "338", "qualifiers": ["A"], "dateTime": "2008-05-27T18:00:00.000-05:00"}, {"value": "336", "qualifiers": ["A"], "dateTime": "2008-05-27T18:15:00.000-05:00"}, {"value": "334", "qualifiers": ["A"], "dateTime": "2008-05-27T18:30:00.000-05:00"}, {"value": "333", "qualifiers": ["A"], "dateTime": "2008-05-27T18:45:00.000-05:00"}, {"value": "334", "qualifiers": ["A"], "dateTime": "2008-05-27T19:00:00.000-05:00"}, {"value": "336", "qualifiers": ["A"], "dateTime": "2008-05-27T19:15:00.000-05:00"}, {"value": "334", "qualifiers": ["A"], "dateTime": "2008-05-27T19:30:00.000-05:00"}, {"value": "334", "qualifiers": ["A"], "dateTime": "2008-05-27T19:45:00.000-05:00"}, {"value": "334", "qualifiers": ["A"], "dateTime": "2008-05-27T20:00:00.000-05:00"}, {"value": "333", "qualifiers": ["A"], "dateTime": "2008-05-27T20:15:00.000-05:00"}, {"value": "334", "qualifiers": ["A"], "dateTime": "2008-05-27T20:30:00.000-05:00"}, {"value": "333", "qualifiers": ["A"], "dateTime": "2008-05-27T20:45:00.000-05:00"}, {"value": "336", "qualifiers": ["A"], "dateTime": "2008-05-27T21:00:00.000-05:00"}, {"value": "334", "qualifiers": ["A"], "dateTime": "2008-05-27T21:15:00.000-05:00"}, {"value": "338", "qualifiers": ["A"], "dateTime": "2008-05-27T21:30:00.000-05:00"}, {"value": "333", "qualifiers": ["A"], "dateTime": "2008-05-27T21:45:00.000-05:00"}, {"value": "331", "qualifiers": ["A"], "dateTime": "2008-05-27T22:00:00.000-05:00"}, {"value": "336", "qualifiers": ["A"], "dateTime": "2008-05-27T22:15:00.000-05:00"}, {"value": "334", "qualifiers": ["A"], "dateTime": "2008-05-27T22:30:00.000-05:00"}, {"value": "336", "qualifiers": ["A"], "dateTime": "2008-05-27T22:45:00.000-05:00"}, {"value": "333", "qualifiers": ["A"], "dateTime": "2008-05-27T23:00:00.000-05:00"}, {"value": "336", "qualifiers": ["A"], "dateTime": "2008-05-27T23:15:00.000-05:00"}, {"value": "336", "qualifiers": ["A"], "dateTime": "2008-05-27T23:30:00.000-05:00"}, {"value": "333", "qualifiers": ["A"], "dateTime": "2008-05-27T23:45:00.000-05:00"}, {"value": "333", "qualifiers": ["A"], "dateTime": "2008-05-28T00:00:00.000-05:00"}, {"value": "333", "qualifiers": ["A"], "dateTime": "2008-05-28T00:15:00.000-05:00"}, {"value": "333", "qualifiers": ["A"], "dateTime": "2008-05-28T00:30:00.000-05:00"}, {"value": "333", "qualifiers": ["A"], "dateTime": "2008-05-28T00:45:00.000-05:00"}, {"value": "331", "qualifiers": ["A"], "dateTime": "2008-05-28T01:00:00.000-05:00"}, {"value": "331", "qualifiers": ["A"], "dateTime": "2008-05-28T01:15:00.000-05:00"}, {"value": "333", "qualifiers": ["A"], "dateTime": "2008-05-28T01:30:00.000-05:00"}, {"value": "334", "qualifiers": ["A"], "dateTime": "2008-05-28T01:45:00.000-05:00"}, {"value": "333", "qualifiers": ["A"], "dateTime": "2008-05-28T02:00:00.000-05:00"}, {"value": "333", "qualifiers": ["A"], "dateTime": "2008-05-28T02:15:00.000-05:00"}, {"value": "333", "qualifiers": ["A"], "dateTime": "2008-05-28T02:30:00.000-05:00"}, {"value": "330", "qualifiers": ["A"], "dateTime": "2008-05-28T02:45:00.000-05:00"}, {"value": "331", "qualifiers": ["A"], "dateTime": "2008-05-28T03:00:00.000-05:00"}, {"value": "331", "qualifiers": ["A"], "dateTime": "2008-05-28T03:15:00.000-05:00"}, {"value": "331", "qualifiers": ["A"], "dateTime": "2008-05-28T03:30:00.000-05:00"}, {"value": "330", "qualifiers": ["A"], "dateTime": "2008-05-28T03:45:00.000-05:00"}, {"value": "330", "qualifiers": ["A"], "dateTime": "2008-05-28T04:00:00.000-05:00"}, {"value": "328", "qualifiers": ["A"], "dateTime": "2008-05-28T04:15:00.000-05:00"}, {"value": "328", "qualifiers": ["A"], "dateTime": "2008-05-28T04:30:00.000-05:00"}, {"value": "328", "qualifiers": ["A"], "dateTime": "2008-05-28T04:45:00.000-05:00"}, {"value": "328", "qualifiers": ["A"], "dateTime": "2008-05-28T05:00:00.000-05:00"}, {"value": "328", "qualifiers": ["A"], "dateTime": "2008-05-28T05:15:00.000-05:00"}, {"value": "328", "qualifiers": ["A"], "dateTime": "2008-05-28T05:30:00.000-05:00"}, {"value": "326", "qualifiers": ["A"], "dateTime": "2008-05-28T05:45:00.000-05:00"}, {"value": "330", "qualifiers": ["A"], "dateTime": "2008-05-28T06:00:00.000-05:00"}, {"value": "328", "qualifiers": ["A"], "dateTime": "2008-05-28T06:15:00.000-05:00"}, {"value": "325", "qualifiers": ["A"], "dateTime": "2008-05-28T06:30:00.000-05:00"}, {"value": "326", "qualifiers": ["A"], "dateTime": "2008-05-28T06:45:00.000-05:00"}, {"value": "330", "qualifiers": ["A"], "dateTime": "2008-05-28T07:00:00.000-05:00"}, {"value": "330", "qualifiers": ["A"], "dateTime": "2008-05-28T07:15:00.000-05:00"}, {"value": "325", "qualifiers": ["A"], "dateTime": "2008-05-28T07:30:00.000-05:00"}, {"value": "330", "qualifiers": ["A"], "dateTime": "2008-05-28T07:45:00.000-05:00"}, {"value": "326", "qualifiers": ["A"], "dateTime": "2008-05-28T08:00:00.000-05:00"}, {"value": "328", "qualifiers": ["A"], "dateTime": "2008-05-28T08:15:00.000-05:00"}, {"value": "326", "qualifiers": ["A"], "dateTime": "2008-05-28T08:30:00.000-05:00"}, {"value": "325", "qualifiers": ["A"], "dateTime": "2008-05-28T08:45:00.000-05:00"}, {"value": "326", "qualifiers": ["A"], "dateTime": "2008-05-28T09:00:00.000-05:00"}, {"value": "325", "qualifiers": ["A"], "dateTime": "2008-05-28T09:15:00.000-05:00"}, {"value": "330", "qualifiers": ["A"], "dateTime": "2008-05-28T09:30:00.000-05:00"}, {"value": "328", "qualifiers": ["A"], "dateTime": "2008-05-28T09:45:00.000-05:00"}, {"value": "330", "qualifiers": ["A"], "dateTime": "2008-05-28T10:00:00.000-05:00"}, {"value": "328", "qualifiers": ["A"], "dateTime": "2008-05-28T10:15:00.000-05:00"}, {"value": "330", "qualifiers": ["A"], "dateTime": "2008-05-28T10:30:00.000-05:00"}, {"value": "323", "qualifiers": ["A"], "dateTime": "2008-05-28T10:45:00.000-05:00"}, {"value": "328", "qualifiers": ["A"], "dateTime": "2008-05-28T11:00:00.000-05:00"}, {"value": "323", "qualifiers": ["A"], "dateTime": "2008-05-28T11:15:00.000-05:00"}, {"value": "323", "qualifiers": ["A"], "dateTime": "2008-05-28T11:30:00.000-05:00"}, {"value": "328", "qualifiers": ["A"], "dateTime": "2008-05-28T11:45:00.000-05:00"}, {"value": "325", "qualifiers": ["A"], "dateTime": "2008-05-28T12:00:00.000-05:00"}, {"value": "326", "qualifiers": ["A"], "dateTime": "2008-05-28T12:15:00.000-05:00"}, {"value": "323", "qualifiers": ["A"], "dateTime": "2008-05-28T12:30:00.000-05:00"}, {"value": "326", "qualifiers": ["A"], "dateTime": "2008-05-28T12:45:00.000-05:00"}, {"value": "328", "qualifiers": ["A"], "dateTime": "2008-05-28T13:00:00.000-05:00"}, {"value": "326", "qualifiers": ["A"], "dateTime": "2008-05-28T13:15:00.000-05:00"}, {"value": "325", "qualifiers": ["A"], "dateTime": "2008-05-28T13:30:00.000-05:00"}, {"value": "322", "qualifiers": ["A"], "dateTime": "2008-05-28T13:45:00.000-05:00"}, {"value": "326", "qualifiers": ["A"], "dateTime": "2008-05-28T14:00:00.000-05:00"}, {"value": "325", "qualifiers": ["A"], "dateTime": "2008-05-28T14:15:00.000-05:00"}, {"value": "325", "qualifiers": ["A"], "dateTime": "2008-05-28T14:30:00.000-05:00"}, {"value": "320", "qualifiers": ["A"], "dateTime": "2008-05-28T14:45:00.000-05:00"}, {"value": "323", "qualifiers": ["A"], "dateTime": "2008-05-28T15:00:00.000-05:00"}, {"value": "322", "qualifiers": ["A"], "dateTime": "2008-05-28T15:15:00.000-05:00"}, {"value": "322", "qualifiers": ["A"], "dateTime": "2008-05-28T15:30:00.000-05:00"}, {"value": "322", "qualifiers": ["A"], "dateTime": "2008-05-28T15:45:00.000-05:00"}, {"value": "320", "qualifiers": ["A"], "dateTime": "2008-05-28T16:00:00.000-05:00"}, {"value": "323", "qualifiers": ["A"], "dateTime": "2008-05-28T16:15:00.000-05:00"}, {"value": "323", "qualifiers": ["A"], "dateTime": "2008-05-28T16:30:00.000-05:00"}, {"value": "323", "qualifiers": ["A"], "dateTime": "2008-05-28T16:45:00.000-05:00"}, {"value": "322", "qualifiers": ["A"], "dateTime": "2008-05-28T17:00:00.000-05:00"}, {"value": "322", "qualifiers": ["A"], "dateTime": "2008-05-28T17:15:00.000-05:00"}, {"value": "323", "qualifiers": ["A"], "dateTime": "2008-05-28T17:30:00.000-05:00"}, {"value": "322", "qualifiers": ["A"], "dateTime": "2008-05-28T17:45:00.000-05:00"}, {"value": "325", "qualifiers": ["A"], "dateTime": "2008-05-28T18:00:00.000-05:00"}, {"value": "322", "qualifiers": ["A"], "dateTime": "2008-05-28T18:15:00.000-05:00"}, {"value": "322", "qualifiers": ["A"], "dateTime": "2008-05-28T18:30:00.000-05:00"}, {"value": "323", "qualifiers": ["A"], "dateTime": "2008-05-28T18:45:00.000-05:00"}, {"value": "322", "qualifiers": ["A"], "dateTime": "2008-05-28T19:00:00.000-05:00"}, {"value": "320", "qualifiers": ["A"], "dateTime": "2008-05-28T19:15:00.000-05:00"}, {"value": "319", "qualifiers": ["A"], "dateTime": "2008-05-28T19:30:00.000-05:00"}, {"value": "322", "qualifiers": ["A"], "dateTime": "2008-05-28T19:45:00.000-05:00"}, {"value": "319", "qualifiers": ["A"], "dateTime": "2008-05-28T20:00:00.000-05:00"}, {"value": "319", "qualifiers": ["A"], "dateTime": "2008-05-28T20:15:00.000-05:00"}, {"value": "319", "qualifiers": ["A"], "dateTime": "2008-05-28T20:30:00.000-05:00"}, {"value": "319", "qualifiers": ["A"], "dateTime": "2008-05-28T20:45:00.000-05:00"}, {"value": "319", "qualifiers": ["A"], "dateTime": "2008-05-28T21:00:00.000-05:00"}, {"value": "317", "qualifiers": ["A"], "dateTime": "2008-05-28T21:15:00.000-05:00"}, {"value": "322", "qualifiers": ["A"], "dateTime": "2008-05-28T21:30:00.000-05:00"}, {"value": "322", "qualifiers": ["A"], "dateTime": "2008-05-28T21:45:00.000-05:00"}, {"value": "320", "qualifiers": ["A"], "dateTime": "2008-05-28T22:00:00.000-05:00"}, {"value": "319", "qualifiers": ["A"], "dateTime": "2008-05-28T22:15:00.000-05:00"}, {"value": "319", "qualifiers": ["A"], "dateTime": "2008-05-28T22:30:00.000-05:00"}, {"value": "319", "qualifiers": ["A"], "dateTime": "2008-05-28T22:45:00.000-05:00"}, {"value": "320", "qualifiers": ["A"], "dateTime": "2008-05-28T23:00:00.000-05:00"}, {"value": "317", "qualifiers": ["A"], "dateTime": "2008-05-28T23:15:00.000-05:00"}, {"value": "319", "qualifiers": ["A"], "dateTime": "2008-05-28T23:30:00.000-05:00"}, {"value": "315", "qualifiers": ["A"], "dateTime": "2008-05-28T23:45:00.000-05:00"}, {"value": "315", "qualifiers": ["A"], "dateTime": "2008-05-29T00:00:00.000-05:00"}, {"value": "315", "qualifiers": ["A"], "dateTime": "2008-05-29T00:15:00.000-05:00"}, {"value": "320", "qualifiers": ["A"], "dateTime": "2008-05-29T00:30:00.000-05:00"}, {"value": "317", "qualifiers": ["A"], "dateTime": "2008-05-29T00:45:00.000-05:00"}, {"value": "315", "qualifiers": ["A"], "dateTime": "2008-05-29T01:00:00.000-05:00"}, {"value": "317", "qualifiers": ["A"], "dateTime": "2008-05-29T01:15:00.000-05:00"}, {"value": "319", "qualifiers": ["A"], "dateTime": "2008-05-29T01:30:00.000-05:00"}, {"value": "315", "qualifiers": ["A"], "dateTime": "2008-05-29T01:45:00.000-05:00"}, {"value": "315", "qualifiers": ["A"], "dateTime": "2008-05-29T02:00:00.000-05:00"}, {"value": "317", "qualifiers": ["A"], "dateTime": "2008-05-29T02:15:00.000-05:00"}, {"value": "320", "qualifiers": ["A"], "dateTime": "2008-05-29T02:30:00.000-05:00"}, {"value": "317", "qualifiers": ["A"], "dateTime": "2008-05-29T02:45:00.000-05:00"}, {"value": "317", "qualifiers": ["A"], "dateTime": "2008-05-29T03:00:00.000-05:00"}, {"value": "317", "qualifiers": ["A"], "dateTime": "2008-05-29T03:15:00.000-05:00"}, {"value": "315", "qualifiers": ["A"], "dateTime": "2008-05-29T03:30:00.000-05:00"}, {"value": "317", "qualifiers": ["A"], "dateTime": "2008-05-29T03:45:00.000-05:00"}, {"value": "320", "qualifiers": ["A"], "dateTime": "2008-05-29T04:00:00.000-05:00"}, {"value": "315", "qualifiers": ["A"], "dateTime": "2008-05-29T04:15:00.000-05:00"}, {"value": "315", "qualifiers": ["A"], "dateTime": "2008-05-29T04:30:00.000-05:00"}, {"value": "317", "qualifiers": ["A"], "dateTime": "2008-05-29T04:45:00.000-05:00"}, {"value": "315", "qualifiers": ["A"], "dateTime": "2008-05-29T05:00:00.000-05:00"}, {"value": "317", "qualifiers": ["A"], "dateTime": "2008-05-29T05:15:00.000-05:00"}, {"value": "314", "qualifiers": ["A"], "dateTime": "2008-05-29T05:30:00.000-05:00"}, {"value": "315", "qualifiers": ["A"], "dateTime": "2008-05-29T05:45:00.000-05:00"}, {"value": "315", "qualifiers": ["A"], "dateTime": "2008-05-29T06:00:00.000-05:00"}, {"value": "319", "qualifiers": ["A"], "dateTime": "2008-05-29T06:15:00.000-05:00"}, {"value": "319", "qualifiers": ["A"], "dateTime": "2008-05-29T06:30:00.000-05:00"}, {"value": "315", "qualifiers": ["A"], "dateTime": "2008-05-29T06:45:00.000-05:00"}, {"value": "319", "qualifiers": ["A"], "dateTime": "2008-05-29T07:00:00.000-05:00"}, {"value": "319", "qualifiers": ["A"], "dateTime": "2008-05-29T07:15:00.000-05:00"}, {"value": "322", "qualifiers": ["A"], "dateTime": "2008-05-29T07:30:00.000-05:00"}, {"value": "330", "qualifiers": ["A"], "dateTime": "2008-05-29T07:45:00.000-05:00"}, {"value": "330", "qualifiers": ["A"], "dateTime": "2008-05-29T08:00:00.000-05:00"}, {"value": "330", "qualifiers": ["A"], "dateTime": "2008-05-29T08:15:00.000-05:00"}, {"value": "331", "qualifiers": ["A"], "dateTime": "2008-05-29T08:30:00.000-05:00"}, {"value": "331", "qualifiers": ["A"], "dateTime": "2008-05-29T08:45:00.000-05:00"}, {"value": "331", "qualifiers": ["A"], "dateTime": "2008-05-29T09:00:00.000-05:00"}, {"value": "330", "qualifiers": ["A"], "dateTime": "2008-05-29T09:15:00.000-05:00"}, {"value": "330", "qualifiers": ["A"], "dateTime": "2008-05-29T09:30:00.000-05:00"}, {"value": "328", "qualifiers": ["A"], "dateTime": "2008-05-29T09:45:00.000-05:00"}, {"value": "328", "qualifiers": ["A"], "dateTime": "2008-05-29T10:00:00.000-05:00"}, {"value": "328", "qualifiers": ["A"], "dateTime": "2008-05-29T10:15:00.000-05:00"}, {"value": "328", "qualifiers": ["A"], "dateTime": "2008-05-29T10:30:00.000-05:00"}, {"value": "325", "qualifiers": ["A"], "dateTime": "2008-05-29T10:45:00.000-05:00"}, {"value": "330", "qualifiers": ["A"], "dateTime": "2008-05-29T11:00:00.000-05:00"}, {"value": "325", "qualifiers": ["A"], "dateTime": "2008-05-29T11:15:00.000-05:00"}, {"value": "328", "qualifiers": ["A"], "dateTime": "2008-05-29T11:30:00.000-05:00"}, {"value": "331", "qualifiers": ["A"], "dateTime": "2008-05-29T11:45:00.000-05:00"}, {"value": "331", "qualifiers": ["A"], "dateTime": "2008-05-29T12:00:00.000-05:00"}, {"value": "334", "qualifiers": ["A"], "dateTime": "2008-05-29T12:15:00.000-05:00"}, {"value": "334", "qualifiers": ["A"], "dateTime": "2008-05-29T12:30:00.000-05:00"}, {"value": "333", "qualifiers": ["A"], "dateTime": "2008-05-29T12:45:00.000-05:00"}, {"value": "336", "qualifiers": ["A"], "dateTime": "2008-05-29T13:00:00.000-05:00"}, {"value": "339", "qualifiers": ["A"], "dateTime": "2008-05-29T13:15:00.000-05:00"}, {"value": "338", "qualifiers": ["A"], "dateTime": "2008-05-29T13:30:00.000-05:00"}, {"value": "338", "qualifiers": ["A"], "dateTime": "2008-05-29T13:45:00.000-05:00"}, {"value": "343", "qualifiers": ["A"], "dateTime": "2008-05-29T14:00:00.000-05:00"}, {"value": "339", "qualifiers": ["A"], "dateTime": "2008-05-29T14:15:00.000-05:00"}, {"value": "343", "qualifiers": ["A"], "dateTime": "2008-05-29T14:30:00.000-05:00"}, {"value": "341", "qualifiers": ["A"], "dateTime": "2008-05-29T14:45:00.000-05:00"}, {"value": "344", "qualifiers": ["A"], "dateTime": "2008-05-29T15:00:00.000-05:00"}, {"value": "339", "qualifiers": ["A"], "dateTime": "2008-05-29T15:15:00.000-05:00"}, {"value": "346", "qualifiers": ["A"], "dateTime": "2008-05-29T15:30:00.000-05:00"}, {"value": "347", "qualifiers": ["A"], "dateTime": "2008-05-29T15:45:00.000-05:00"}, {"value": "343", "qualifiers": ["A"], "dateTime": "2008-05-29T16:00:00.000-05:00"}, {"value": "346", "qualifiers": ["A"], "dateTime": "2008-05-29T16:15:00.000-05:00"}, {"value": "346", "qualifiers": ["A"], "dateTime": "2008-05-29T16:30:00.000-05:00"}, {"value": "347", "qualifiers": ["A"], "dateTime": "2008-05-29T16:45:00.000-05:00"}, {"value": "351", "qualifiers": ["A"], "dateTime": "2008-05-29T17:00:00.000-05:00"}, {"value": "351", "qualifiers": ["A"], "dateTime": "2008-05-29T17:15:00.000-05:00"}, {"value": "354", "qualifiers": ["A"], "dateTime": "2008-05-29T17:30:00.000-05:00"}, {"value": "359", "qualifiers": ["A"], "dateTime": "2008-05-29T17:45:00.000-05:00"}, {"value": "357", "qualifiers": ["A"], "dateTime": "2008-05-29T18:00:00.000-05:00"}, {"value": "357", "qualifiers": ["A"], "dateTime": "2008-05-29T18:15:00.000-05:00"}, {"value": "361", "qualifiers": ["A"], "dateTime": "2008-05-29T18:30:00.000-05:00"}, {"value": "361", "qualifiers": ["A"], "dateTime": "2008-05-29T18:45:00.000-05:00"}, {"value": "361", "qualifiers": ["A"], "dateTime": "2008-05-29T19:00:00.000-05:00"}, {"value": "362", "qualifiers": ["A"], "dateTime": "2008-05-29T19:15:00.000-05:00"}, {"value": "367", "qualifiers": ["A"], "dateTime": "2008-05-29T19:30:00.000-05:00"}, {"value": "366", "qualifiers": ["A"], "dateTime": "2008-05-29T19:45:00.000-05:00"}, {"value": "372", "qualifiers": ["A"], "dateTime": "2008-05-29T20:00:00.000-05:00"}, {"value": "372", "qualifiers": ["A"], "dateTime": "2008-05-29T20:15:00.000-05:00"}, {"value": "374", "qualifiers": ["A"], "dateTime": "2008-05-29T20:30:00.000-05:00"}, {"value": "374", "qualifiers": ["A"], "dateTime": "2008-05-29T20:45:00.000-05:00"}, {"value": "377", "qualifiers": ["A"], "dateTime": "2008-05-29T21:00:00.000-05:00"}, {"value": "381", "qualifiers": ["A"], "dateTime": "2008-05-29T21:15:00.000-05:00"}, {"value": "381", "qualifiers": ["A"], "dateTime": "2008-05-29T21:30:00.000-05:00"}, {"value": "382", "qualifiers": ["A"], "dateTime": "2008-05-29T21:45:00.000-05:00"}, {"value": "391", "qualifiers": ["A"], "dateTime": "2008-05-29T22:00:00.000-05:00"}, {"value": "391", "qualifiers": ["A"], "dateTime": "2008-05-29T22:15:00.000-05:00"}, {"value": "398", "qualifiers": ["A"], "dateTime": "2008-05-29T22:30:00.000-05:00"}, {"value": "398", "qualifiers": ["A"], "dateTime": "2008-05-29T22:45:00.000-05:00"}, {"value": "407", "qualifiers": ["A"], "dateTime": "2008-05-29T23:00:00.000-05:00"}, {"value": "412", "qualifiers": ["A"], "dateTime": "2008-05-29T23:15:00.000-05:00"}, {"value": "417", "qualifiers": ["A"], "dateTime": "2008-05-29T23:30:00.000-05:00"}, {"value": "428", "qualifiers": ["A"], "dateTime": "2008-05-29T23:45:00.000-05:00"}, {"value": "437", "qualifiers": ["A"], "dateTime": "2008-05-30T00:00:00.000-05:00"}, {"value": "444", "qualifiers": ["A"], "dateTime": "2008-05-30T00:15:00.000-05:00"}, {"value": "459", "qualifiers": ["A"], "dateTime": "2008-05-30T00:30:00.000-05:00"}, {"value": "472", "qualifiers": ["A"], "dateTime": "2008-05-30T00:45:00.000-05:00"}, {"value": "493", "qualifiers": ["A"], "dateTime": "2008-05-30T01:00:00.000-05:00"}, {"value": "504", "qualifiers": ["A"], "dateTime": "2008-05-30T01:15:00.000-05:00"}, {"value": "526", "qualifiers": ["A"], "dateTime": "2008-05-30T01:30:00.000-05:00"}, {"value": "546", "qualifiers": ["A"], "dateTime": "2008-05-30T01:45:00.000-05:00"}, {"value": "583", "qualifiers": ["A"], "dateTime": "2008-05-30T02:00:00.000-05:00"}, {"value": "629", "qualifiers": ["A"], "dateTime": "2008-05-30T02:15:00.000-05:00"}, {"value": "663", "qualifiers": ["A"], "dateTime": "2008-05-30T02:30:00.000-05:00"}, {"value": "697", "qualifiers": ["A"], "dateTime": "2008-05-30T02:45:00.000-05:00"}, {"value": "717", "qualifiers": ["A"], "dateTime": "2008-05-30T03:00:00.000-05:00"}, {"value": "733", "qualifiers": ["A"], "dateTime": "2008-05-30T03:15:00.000-05:00"}, {"value": "758", "qualifiers": ["A"], "dateTime": "2008-05-30T03:30:00.000-05:00"}, {"value": "782", "qualifiers": ["A"], "dateTime": "2008-05-30T03:45:00.000-05:00"}, {"value": "806", "qualifiers": ["A"], "dateTime": "2008-05-30T04:00:00.000-05:00"}, {"value": "844", "qualifiers": ["A"], "dateTime": "2008-05-30T04:15:00.000-05:00"}, {"value": "877", "qualifiers": ["A"], "dateTime": "2008-05-30T04:30:00.000-05:00"}, {"value": "915", "qualifiers": ["A"], "dateTime": "2008-05-30T04:45:00.000-05:00"}, {"value": "948", "qualifiers": ["A"], "dateTime": "2008-05-30T05:00:00.000-05:00"}, {"value": "995", "qualifiers": ["A"], "dateTime": "2008-05-30T05:15:00.000-05:00"}, {"value": "1040", "qualifiers": ["A"], "dateTime": "2008-05-30T05:30:00.000-05:00"}, {"value": "1080", "qualifiers": ["A"], "dateTime": "2008-05-30T05:45:00.000-05:00"}, {"value": "1110", "qualifiers": ["A"], "dateTime": "2008-05-30T06:00:00.000-05:00"}, {"value": "1150", "qualifiers": ["A"], "dateTime": "2008-05-30T06:15:00.000-05:00"}, {"value": "1190", "qualifiers": ["A"], "dateTime": "2008-05-30T06:30:00.000-05:00"}, {"value": "1220", "qualifiers": ["A"], "dateTime": "2008-05-30T06:45:00.000-05:00"}, {"value": "1260", "qualifiers": ["A"], "dateTime": "2008-05-30T07:00:00.000-05:00"}, {"value": "1300", "qualifiers": ["A"], "dateTime": "2008-05-30T07:15:00.000-05:00"}, {"value": "1330", "qualifiers": ["A"], "dateTime": "2008-05-30T07:30:00.000-05:00"}, {"value": "1360", "qualifiers": ["A"], "dateTime": "2008-05-30T07:45:00.000-05:00"}, {"value": "1380", "qualifiers": ["A"], "dateTime": "2008-05-30T08:00:00.000-05:00"}, {"value": "1400", "qualifiers": ["A"], "dateTime": "2008-05-30T08:15:00.000-05:00"}, {"value": "1450", "qualifiers": ["A"], "dateTime": "2008-05-30T08:30:00.000-05:00"}, {"value": "1470", "qualifiers": ["A"], "dateTime": "2008-05-30T08:45:00.000-05:00"}, {"value": "1490", "qualifiers": ["A"], "dateTime": "2008-05-30T09:00:00.000-05:00"}, {"value": "1490", "qualifiers": ["A"], "dateTime": "2008-05-30T09:15:00.000-05:00"}, {"value": "1530", "qualifiers": ["A"], "dateTime": "2008-05-30T09:30:00.000-05:00"}, {"value": "1550", "qualifiers": ["A"], "dateTime": "2008-05-30T09:45:00.000-05:00"}, {"value": "1570", "qualifiers": ["A"], "dateTime": "2008-05-30T10:00:00.000-05:00"}, {"value": "1600", "qualifiers": ["A"], "dateTime": "2008-05-30T10:15:00.000-05:00"}, {"value": "1620", "qualifiers": ["A"], "dateTime": "2008-05-30T10:30:00.000-05:00"}, {"value": "1640", "qualifiers": ["A"], "dateTime": "2008-05-30T10:45:00.000-05:00"}, {"value": "1680", "qualifiers": ["A"], "dateTime": "2008-05-30T11:00:00.000-05:00"}, {"value": "1700", "qualifiers": ["A"], "dateTime": "2008-05-30T11:15:00.000-05:00"}, {"value": "1710", "qualifiers": ["A"], "dateTime": "2008-05-30T11:30:00.000-05:00"}, {"value": "1760", "qualifiers": ["A"], "dateTime": "2008-05-30T11:45:00.000-05:00"}, {"value": "1780", "qualifiers": ["A"], "dateTime": "2008-05-30T12:00:00.000-05:00"}, {"value": "1800", "qualifiers": ["A"], "dateTime": "2008-05-30T12:15:00.000-05:00"}, {"value": "1820", "qualifiers": ["A"], "dateTime": "2008-05-30T12:30:00.000-05:00"}, {"value": "1850", "qualifiers": ["A"], "dateTime": "2008-05-30T12:45:00.000-05:00"}, {"value": "1880", "qualifiers": ["A"], "dateTime": "2008-05-30T13:00:00.000-05:00"}, {"value": "1920", "qualifiers": ["A"], "dateTime": "2008-05-30T13:15:00.000-05:00"}, {"value": "1930", "qualifiers": ["A"], "dateTime": "2008-05-30T13:30:00.000-05:00"}, {"value": "1950", "qualifiers": ["A"], "dateTime": "2008-05-30T13:45:00.000-05:00"}, {"value": "1970", "qualifiers": ["A"], "dateTime": "2008-05-30T14:00:00.000-05:00"}, {"value": "2000", "qualifiers": ["A"], "dateTime": "2008-05-30T14:15:00.000-05:00"}, {"value": "2010", "qualifiers": ["A"], "dateTime": "2008-05-30T14:30:00.000-05:00"}, {"value": "2050", "qualifiers": ["A"], "dateTime": "2008-05-30T14:45:00.000-05:00"}, {"value": "2070", "qualifiers": ["A"], "dateTime": "2008-05-30T15:00:00.000-05:00"}, {"value": "2090", "qualifiers": ["A"], "dateTime": "2008-05-30T15:15:00.000-05:00"}, {"value": "2130", "qualifiers": ["A"], "dateTime": "2008-05-30T15:30:00.000-05:00"}, {"value": "2150", "qualifiers": ["A"], "dateTime": "2008-05-30T15:45:00.000-05:00"}, {"value": "2190", "qualifiers": ["A"], "dateTime": "2008-05-30T16:00:00.000-05:00"}, {"value": "2210", "qualifiers": ["A"], "dateTime": "2008-05-30T16:15:00.000-05:00"}, {"value": "2250", "qualifiers": ["A"], "dateTime": "2008-05-30T16:30:00.000-05:00"}, {"value": "2290", "qualifiers": ["A"], "dateTime": "2008-05-30T16:45:00.000-05:00"}, {"value": "2320", "qualifiers": ["A"], "dateTime": "2008-05-30T17:00:00.000-05:00"}, {"value": "2370", "qualifiers": ["A"], "dateTime": "2008-05-30T17:15:00.000-05:00"}, {"value": "2430", "qualifiers": ["A"], "dateTime": "2008-05-30T17:30:00.000-05:00"}, {"value": "2470", "qualifiers": ["A"], "dateTime": "2008-05-30T17:45:00.000-05:00"}, {"value": "2540", "qualifiers": ["A"], "dateTime": "2008-05-30T18:00:00.000-05:00"}, {"value": "2590", "qualifiers": ["A"], "dateTime": "2008-05-30T18:15:00.000-05:00"}, {"value": "2640", "qualifiers": ["A"], "dateTime": "2008-05-30T18:30:00.000-05:00"}, {"value": "2730", "qualifiers": ["A"], "dateTime": "2008-05-30T18:45:00.000-05:00"}, {"value": "2770", "qualifiers": ["A"], "dateTime": "2008-05-30T19:00:00.000-05:00"}, {"value": "2870", "qualifiers": ["A"], "dateTime": "2008-05-30T19:15:00.000-05:00"}, {"value": "2920", "qualifiers": ["A"], "dateTime": "2008-05-30T19:30:00.000-05:00"}, {"value": "3000", "qualifiers": ["A"], "dateTime": "2008-05-30T19:45:00.000-05:00"}, {"value": "3140", "qualifiers": ["A"], "dateTime": "2008-05-30T20:00:00.000-05:00"}, {"value": "3200", "qualifiers": ["A"], "dateTime": "2008-05-30T20:15:00.000-05:00"}, {"value": "3320", "qualifiers": ["A"], "dateTime": "2008-05-30T20:30:00.000-05:00"}, {"value": "3380", "qualifiers": ["A"], "dateTime": "2008-05-30T20:45:00.000-05:00"}, {"value": "3480", "qualifiers": ["A"], "dateTime": "2008-05-30T21:00:00.000-05:00"}, {"value": "3530", "qualifiers": ["A"], "dateTime": "2008-05-30T21:15:00.000-05:00"}, {"value": "3650", "qualifiers": ["A"], "dateTime": "2008-05-30T21:30:00.000-05:00"}, {"value": "3720", "qualifiers": ["A"], "dateTime": "2008-05-30T21:45:00.000-05:00"}, {"value": "3870", "qualifiers": ["A"], "dateTime": "2008-05-30T22:00:00.000-05:00"}, {"value": "3930", "qualifiers": ["A"], "dateTime": "2008-05-30T22:15:00.000-05:00"}, {"value": "4030", "qualifiers": ["A"], "dateTime": "2008-05-30T22:30:00.000-05:00"}, {"value": "4110", "qualifiers": ["A"], "dateTime": "2008-05-30T22:45:00.000-05:00"}, {"value": "4210", "qualifiers": ["A"], "dateTime": "2008-05-30T23:00:00.000-05:00"}, {"value": "4230", "qualifiers": ["A"], "dateTime": "2008-05-30T23:15:00.000-05:00"}, {"value": "4330", "qualifiers": ["A"], "dateTime": "2008-05-30T23:30:00.000-05:00"}, {"value": "4360", "qualifiers": ["A"], "dateTime": "2008-05-30T23:45:00.000-05:00"}, {"value": "4440", "qualifiers": ["A"], "dateTime": "2008-05-31T00:00:00.000-05:00"}, {"value": "4550", "qualifiers": ["A"], "dateTime": "2008-05-31T00:15:00.000-05:00"}, {"value": "4550", "qualifiers": ["A"], "dateTime": "2008-05-31T00:30:00.000-05:00"}, {"value": "4550", "qualifiers": ["A"], "dateTime": "2008-05-31T00:45:00.000-05:00"}, {"value": "4660", "qualifiers": ["A"], "dateTime": "2008-05-31T01:00:00.000-05:00"}, {"value": "4640", "qualifiers": ["A"], "dateTime": "2008-05-31T01:15:00.000-05:00"}, {"value": "4720", "qualifiers": ["A"], "dateTime": "2008-05-31T01:30:00.000-05:00"}, {"value": "4720", "qualifiers": ["A"], "dateTime": "2008-05-31T01:45:00.000-05:00"}, {"value": "4720", "qualifiers": ["A"], "dateTime": "2008-05-31T02:00:00.000-05:00"}, {"value": "4770", "qualifiers": ["A"], "dateTime": "2008-05-31T02:15:00.000-05:00"}, {"value": "4790", "qualifiers": ["A"], "dateTime": "2008-05-31T02:30:00.000-05:00"}, {"value": "4820", "qualifiers": ["A"], "dateTime": "2008-05-31T02:45:00.000-05:00"}, {"value": "4820", "qualifiers": ["A"], "dateTime": "2008-05-31T03:00:00.000-05:00"}, {"value": "4820", "qualifiers": ["A"], "dateTime": "2008-05-31T03:15:00.000-05:00"}, {"value": "4820", "qualifiers": ["A"], "dateTime": "2008-05-31T03:30:00.000-05:00"}, {"value": "4770", "qualifiers": ["A"], "dateTime": "2008-05-31T03:45:00.000-05:00"}, {"value": "4820", "qualifiers": ["A"], "dateTime": "2008-05-31T04:00:00.000-05:00"}, {"value": "4790", "qualifiers": ["A"], "dateTime": "2008-05-31T04:15:00.000-05:00"}, {"value": "4820", "qualifiers": ["A"], "dateTime": "2008-05-31T04:30:00.000-05:00"}, {"value": "4770", "qualifiers": ["A"], "dateTime": "2008-05-31T04:45:00.000-05:00"}, {"value": "4770", "qualifiers": ["A"], "dateTime": "2008-05-31T05:00:00.000-05:00"}, {"value": "4750", "qualifiers": ["A"], "dateTime": "2008-05-31T05:15:00.000-05:00"}, {"value": "4700", "qualifiers": ["A"], "dateTime": "2008-05-31T05:30:00.000-05:00"}, {"value": "4700", "qualifiers": ["A"], "dateTime": "2008-05-31T05:45:00.000-05:00"}, {"value": "4680", "qualifiers": ["A"], "dateTime": "2008-05-31T06:00:00.000-05:00"}, {"value": "4660", "qualifiers": ["A"], "dateTime": "2008-05-31T06:15:00.000-05:00"}, {"value": "4660", "qualifiers": ["A"], "dateTime": "2008-05-31T06:30:00.000-05:00"}, {"value": "4570", "qualifiers": ["A"], "dateTime": "2008-05-31T06:45:00.000-05:00"}, {"value": "4570", "qualifiers": ["A"], "dateTime": "2008-05-31T07:00:00.000-05:00"}, {"value": "4590", "qualifiers": ["A"], "dateTime": "2008-05-31T07:15:00.000-05:00"}, {"value": "4570", "qualifiers": ["A"], "dateTime": "2008-05-31T07:30:00.000-05:00"}, {"value": "4530", "qualifiers": ["A"], "dateTime": "2008-05-31T07:45:00.000-05:00"}, {"value": "4530", "qualifiers": ["A"], "dateTime": "2008-05-31T08:00:00.000-05:00"}, {"value": "4480", "qualifiers": ["A"], "dateTime": "2008-05-31T08:15:00.000-05:00"}, {"value": "4460", "qualifiers": ["A"], "dateTime": "2008-05-31T08:30:00.000-05:00"}, {"value": "4360", "qualifiers": ["A"], "dateTime": "2008-05-31T08:45:00.000-05:00"}, {"value": "4420", "qualifiers": ["A"], "dateTime": "2008-05-31T09:00:00.000-05:00"}, {"value": "4380", "qualifiers": ["A"], "dateTime": "2008-05-31T09:15:00.000-05:00"}, {"value": "4380", "qualifiers": ["A"], "dateTime": "2008-05-31T09:30:00.000-05:00"}, {"value": "4330", "qualifiers": ["A"], "dateTime": "2008-05-31T09:45:00.000-05:00"}, {"value": "4310", "qualifiers": ["A"], "dateTime": "2008-05-31T10:00:00.000-05:00"}, {"value": "4290", "qualifiers": ["A"], "dateTime": "2008-05-31T10:15:00.000-05:00"}, {"value": "4210", "qualifiers": ["A"], "dateTime": "2008-05-31T10:30:00.000-05:00"}, {"value": "4210", "qualifiers": ["A"], "dateTime": "2008-05-31T10:45:00.000-05:00"}, {"value": "4210", "qualifiers": ["A"], "dateTime": "2008-05-31T11:00:00.000-05:00"}, {"value": "4210", "qualifiers": ["A"], "dateTime": "2008-05-31T11:15:00.000-05:00"}, {"value": "4150", "qualifiers": ["A"], "dateTime": "2008-05-31T11:30:00.000-05:00"}, {"value": "4130", "qualifiers": ["A"], "dateTime": "2008-05-31T11:45:00.000-05:00"}, {"value": "4130", "qualifiers": ["A"], "dateTime": "2008-05-31T12:00:00.000-05:00"}, {"value": "4110", "qualifiers": ["A"], "dateTime": "2008-05-31T12:15:00.000-05:00"}, {"value": "4070", "qualifiers": ["A"], "dateTime": "2008-05-31T12:30:00.000-05:00"}, {"value": "4030", "qualifiers": ["A"], "dateTime": "2008-05-31T12:45:00.000-05:00"}, {"value": "4030", "qualifiers": ["A"], "dateTime": "2008-05-31T13:00:00.000-05:00"}, {"value": "4010", "qualifiers": ["A"], "dateTime": "2008-05-31T13:15:00.000-05:00"}, {"value": "4010", "qualifiers": ["A"], "dateTime": "2008-05-31T13:30:00.000-05:00"}, {"value": "3930", "qualifiers": ["A"], "dateTime": "2008-05-31T13:45:00.000-05:00"}, {"value": "3910", "qualifiers": ["A"], "dateTime": "2008-05-31T14:00:00.000-05:00"}, {"value": "3930", "qualifiers": ["A"], "dateTime": "2008-05-31T14:15:00.000-05:00"}, {"value": "3860", "qualifiers": ["A"], "dateTime": "2008-05-31T14:30:00.000-05:00"}, {"value": "3840", "qualifiers": ["A"], "dateTime": "2008-05-31T14:45:00.000-05:00"}, {"value": "3890", "qualifiers": ["A"], "dateTime": "2008-05-31T15:00:00.000-05:00"}, {"value": "3870", "qualifiers": ["A"], "dateTime": "2008-05-31T15:15:00.000-05:00"}, {"value": "3840", "qualifiers": ["A"], "dateTime": "2008-05-31T15:30:00.000-05:00"}, {"value": "3800", "qualifiers": ["A"], "dateTime": "2008-05-31T15:45:00.000-05:00"}, {"value": "3780", "qualifiers": ["A"], "dateTime": "2008-05-31T16:00:00.000-05:00"}, {"value": "3770", "qualifiers": ["A"], "dateTime": "2008-05-31T16:15:00.000-05:00"}, {"value": "3750", "qualifiers": ["A"], "dateTime": "2008-05-31T16:30:00.000-05:00"}, {"value": "3770", "qualifiers": ["A"], "dateTime": "2008-05-31T16:45:00.000-05:00"}, {"value": "3690", "qualifiers": ["A"], "dateTime": "2008-05-31T17:00:00.000-05:00"}, {"value": "3770", "qualifiers": ["A"], "dateTime": "2008-05-31T17:15:00.000-05:00"}, {"value": "3700", "qualifiers": ["A"], "dateTime": "2008-05-31T17:30:00.000-05:00"}, {"value": "3740", "qualifiers": ["A"], "dateTime": "2008-05-31T17:45:00.000-05:00"}, {"value": "3670", "qualifiers": ["A"], "dateTime": "2008-05-31T18:00:00.000-05:00"}, {"value": "3670", "qualifiers": ["A"], "dateTime": "2008-05-31T18:15:00.000-05:00"}, {"value": "3700", "qualifiers": ["A"], "dateTime": "2008-05-31T18:30:00.000-05:00"}, {"value": "3620", "qualifiers": ["A"], "dateTime": "2008-05-31T18:45:00.000-05:00"}, {"value": "3610", "qualifiers": ["A"], "dateTime": "2008-05-31T19:00:00.000-05:00"}, {"value": "3610", "qualifiers": ["A"], "dateTime": "2008-05-31T19:15:00.000-05:00"}, {"value": "3610", "qualifiers": ["A"], "dateTime": "2008-05-31T19:30:00.000-05:00"}, {"value": "3610", "qualifiers": ["A"], "dateTime": "2008-05-31T19:45:00.000-05:00"}, {"value": "3620", "qualifiers": ["A"], "dateTime": "2008-05-31T20:00:00.000-05:00"}, {"value": "3590", "qualifiers": ["A"], "dateTime": "2008-05-31T20:15:00.000-05:00"}, {"value": "3620", "qualifiers": ["A"], "dateTime": "2008-05-31T20:30:00.000-05:00"}, {"value": "3610", "qualifiers": ["A"], "dateTime": "2008-05-31T20:45:00.000-05:00"}, {"value": "3620", "qualifiers": ["A"], "dateTime": "2008-05-31T21:00:00.000-05:00"}, {"value": "3620", "qualifiers": ["A"], "dateTime": "2008-05-31T21:15:00.000-05:00"}, {"value": "3590", "qualifiers": ["A"], "dateTime": "2008-05-31T21:30:00.000-05:00"}, {"value": "3580", "qualifiers": ["A"], "dateTime": "2008-05-31T21:45:00.000-05:00"}, {"value": "3580", "qualifiers": ["A"], "dateTime": "2008-05-31T22:00:00.000-05:00"}, {"value": "3560", "qualifiers": ["A"], "dateTime": "2008-05-31T22:15:00.000-05:00"}, {"value": "3560", "qualifiers": ["A"], "dateTime": "2008-05-31T22:30:00.000-05:00"}, {"value": "3540", "qualifiers": ["A"], "dateTime": "2008-05-31T22:45:00.000-05:00"}, {"value": "3560", "qualifiers": ["A"], "dateTime": "2008-05-31T23:00:00.000-05:00"}, {"value": "3540", "qualifiers": ["A"], "dateTime": "2008-05-31T23:15:00.000-05:00"}, {"value": "3510", "qualifiers": ["A"], "dateTime": "2008-05-31T23:30:00.000-05:00"}, {"value": "3540", "qualifiers": ["A"], "dateTime": "2008-05-31T23:45:00.000-05:00"}, {"value": "3530", "qualifiers": ["A"], "dateTime": "2008-06-01T00:00:00.000-05:00"}, {"value": "3510", "qualifiers": ["A"], "dateTime": "2008-06-01T00:15:00.000-05:00"}, {"value": "3530", "qualifiers": ["A"], "dateTime": "2008-06-01T00:30:00.000-05:00"}, {"value": "3470", "qualifiers": ["A"], "dateTime": "2008-06-01T00:45:00.000-05:00"}, {"value": "3500", "qualifiers": ["A"], "dateTime": "2008-06-01T01:00:00.000-05:00"}, {"value": "3500", "qualifiers": ["A"], "dateTime": "2008-06-01T01:15:00.000-05:00"}, {"value": "3510", "qualifiers": ["A"], "dateTime": "2008-06-01T01:30:00.000-05:00"}, {"value": "3480", "qualifiers": ["A"], "dateTime": "2008-06-01T01:45:00.000-05:00"}, {"value": "3480", "qualifiers": ["A"], "dateTime": "2008-06-01T02:00:00.000-05:00"}, {"value": "3480", "qualifiers": ["A"], "dateTime": "2008-06-01T02:15:00.000-05:00"}, {"value": "3500", "qualifiers": ["A"], "dateTime": "2008-06-01T02:30:00.000-05:00"}, {"value": "3500", "qualifiers": ["A"], "dateTime": "2008-06-01T02:45:00.000-05:00"}, {"value": "3500", "qualifiers": ["A"], "dateTime": "2008-06-01T03:00:00.000-05:00"}, {"value": "3420", "qualifiers": ["A"], "dateTime": "2008-06-01T03:15:00.000-05:00"}, {"value": "3410", "qualifiers": ["A"], "dateTime": "2008-06-01T03:30:00.000-05:00"}, {"value": "3420", "qualifiers": ["A"], "dateTime": "2008-06-01T03:45:00.000-05:00"}, {"value": "3410", "qualifiers": ["A"], "dateTime": "2008-06-01T04:00:00.000-05:00"}, {"value": "3450", "qualifiers": ["A"], "dateTime": "2008-06-01T04:15:00.000-05:00"}, {"value": "3440", "qualifiers": ["A"], "dateTime": "2008-06-01T04:30:00.000-05:00"}, {"value": "3420", "qualifiers": ["A"], "dateTime": "2008-06-01T04:45:00.000-05:00"}, {"value": "3390", "qualifiers": ["A"], "dateTime": "2008-06-01T05:00:00.000-05:00"}, {"value": "3420", "qualifiers": ["A"], "dateTime": "2008-06-01T05:15:00.000-05:00"}, {"value": "3440", "qualifiers": ["A"], "dateTime": "2008-06-01T05:30:00.000-05:00"}, {"value": "3420", "qualifiers": ["A"], "dateTime": "2008-06-01T05:45:00.000-05:00"}, {"value": "3380", "qualifiers": ["A"], "dateTime": "2008-06-01T06:00:00.000-05:00"}, {"value": "3380", "qualifiers": ["A"], "dateTime": "2008-06-01T06:15:00.000-05:00"}, {"value": "3320", "qualifiers": ["A"], "dateTime": "2008-06-01T06:30:00.000-05:00"}, {"value": "3360", "qualifiers": ["A"], "dateTime": "2008-06-01T06:45:00.000-05:00"}, {"value": "3300", "qualifiers": ["A"], "dateTime": "2008-06-01T07:00:00.000-05:00"}, {"value": "3320", "qualifiers": ["A"], "dateTime": "2008-06-01T07:15:00.000-05:00"}, {"value": "3300", "qualifiers": ["A"], "dateTime": "2008-06-01T07:30:00.000-05:00"}, {"value": "3290", "qualifiers": ["A"], "dateTime": "2008-06-01T07:45:00.000-05:00"}, {"value": "3270", "qualifiers": ["A"], "dateTime": "2008-06-01T08:00:00.000-05:00"}, {"value": "3270", "qualifiers": ["A"], "dateTime": "2008-06-01T08:15:00.000-05:00"}, {"value": "3290", "qualifiers": ["A"], "dateTime": "2008-06-01T08:30:00.000-05:00"}, {"value": "3230", "qualifiers": ["A"], "dateTime": "2008-06-01T08:45:00.000-05:00"}, {"value": "3240", "qualifiers": ["A"], "dateTime": "2008-06-01T09:00:00.000-05:00"}, {"value": "3240", "qualifiers": ["A"], "dateTime": "2008-06-01T09:15:00.000-05:00"}, {"value": "3180", "qualifiers": ["A"], "dateTime": "2008-06-01T09:30:00.000-05:00"}, {"value": "3210", "qualifiers": ["A"], "dateTime": "2008-06-01T09:45:00.000-05:00"}, {"value": "3170", "qualifiers": ["A"], "dateTime": "2008-06-01T10:00:00.000-05:00"}, {"value": "3160", "qualifiers": ["A"], "dateTime": "2008-06-01T10:15:00.000-05:00"}, {"value": "3130", "qualifiers": ["A"], "dateTime": "2008-06-01T10:30:00.000-05:00"}, {"value": "3130", "qualifiers": ["A"], "dateTime": "2008-06-01T10:45:00.000-05:00"}, {"value": "3110", "qualifiers": ["A"], "dateTime": "2008-06-01T11:00:00.000-05:00"}, {"value": "3130", "qualifiers": ["A"], "dateTime": "2008-06-01T11:15:00.000-05:00"}, {"value": "3090", "qualifiers": ["A"], "dateTime": "2008-06-01T11:30:00.000-05:00"}, {"value": "3060", "qualifiers": ["A"], "dateTime": "2008-06-01T11:45:00.000-05:00"}, {"value": "3040", "qualifiers": ["A"], "dateTime": "2008-06-01T12:00:00.000-05:00"}, {"value": "3030", "qualifiers": ["A"], "dateTime": "2008-06-01T12:15:00.000-05:00"}, {"value": "3020", "qualifiers": ["A"], "dateTime": "2008-06-01T12:30:00.000-05:00"}, {"value": "3000", "qualifiers": ["A"], "dateTime": "2008-06-01T12:45:00.000-05:00"}, {"value": "2930", "qualifiers": ["A"], "dateTime": "2008-06-01T13:00:00.000-05:00"}, {"value": "2950", "qualifiers": ["A"], "dateTime": "2008-06-01T13:15:00.000-05:00"}, {"value": "2920", "qualifiers": ["A"], "dateTime": "2008-06-01T13:30:00.000-05:00"}, {"value": "2910", "qualifiers": ["A"], "dateTime": "2008-06-01T13:45:00.000-05:00"}, {"value": "2880", "qualifiers": ["A"], "dateTime": "2008-06-01T14:00:00.000-05:00"}, {"value": "2880", "qualifiers": ["A"], "dateTime": "2008-06-01T14:15:00.000-05:00"}, {"value": "2860", "qualifiers": ["A"], "dateTime": "2008-06-01T14:30:00.000-05:00"}, {"value": "2810", "qualifiers": ["A"], "dateTime": "2008-06-01T14:45:00.000-05:00"}, {"value": "2830", "qualifiers": ["A"], "dateTime": "2008-06-01T15:00:00.000-05:00"}, {"value": "2800", "qualifiers": ["A"], "dateTime": "2008-06-01T15:15:00.000-05:00"}, {"value": "2760", "qualifiers": ["A"], "dateTime": "2008-06-01T15:30:00.000-05:00"}, {"value": "2770", "qualifiers": ["A"], "dateTime": "2008-06-01T15:45:00.000-05:00"}, {"value": "2750", "qualifiers": ["A"], "dateTime": "2008-06-01T16:00:00.000-05:00"}, {"value": "2730", "qualifiers": ["A"], "dateTime": "2008-06-01T16:15:00.000-05:00"}, {"value": "2690", "qualifiers": ["A"], "dateTime": "2008-06-01T16:30:00.000-05:00"}, {"value": "2700", "qualifiers": ["A"], "dateTime": "2008-06-01T16:45:00.000-05:00"}, {"value": "2680", "qualifiers": ["A"], "dateTime": "2008-06-01T17:00:00.000-05:00"}, {"value": "2660", "qualifiers": ["A"], "dateTime": "2008-06-01T17:15:00.000-05:00"}, {"value": "2630", "qualifiers": ["A"], "dateTime": "2008-06-01T17:30:00.000-05:00"}, {"value": "2610", "qualifiers": ["A"], "dateTime": "2008-06-01T17:45:00.000-05:00"}, {"value": "2600", "qualifiers": ["A"], "dateTime": "2008-06-01T18:00:00.000-05:00"}, {"value": "2590", "qualifiers": ["A"], "dateTime": "2008-06-01T18:15:00.000-05:00"}, {"value": "2560", "qualifiers": ["A"], "dateTime": "2008-06-01T18:30:00.000-05:00"}, {"value": "2550", "qualifiers": ["A"], "dateTime": "2008-06-01T18:45:00.000-05:00"}, {"value": "2510", "qualifiers": ["A"], "dateTime": "2008-06-01T19:00:00.000-05:00"}, {"value": "2500", "qualifiers": ["A"], "dateTime": "2008-06-01T19:15:00.000-05:00"}, {"value": "2470", "qualifiers": ["A"], "dateTime": "2008-06-01T19:30:00.000-05:00"}, {"value": "2440", "qualifiers": ["A"], "dateTime": "2008-06-01T19:45:00.000-05:00"}, {"value": "2440", "qualifiers": ["A"], "dateTime": "2008-06-01T20:00:00.000-05:00"}, {"value": "2390", "qualifiers": ["A"], "dateTime": "2008-06-01T20:15:00.000-05:00"}, {"value": "2380", "qualifiers": ["A"], "dateTime": "2008-06-01T20:30:00.000-05:00"}, {"value": "2370", "qualifiers": ["A"], "dateTime": "2008-06-01T20:45:00.000-05:00"}, {"value": "2340", "qualifiers": ["A"], "dateTime": "2008-06-01T21:00:00.000-05:00"}, {"value": "2330", "qualifiers": ["A"], "dateTime": "2008-06-01T21:15:00.000-05:00"}, {"value": "2310", "qualifiers": ["A"], "dateTime": "2008-06-01T21:30:00.000-05:00"}, {"value": "2290", "qualifiers": ["A"], "dateTime": "2008-06-01T21:45:00.000-05:00"}, {"value": "2260", "qualifiers": ["A"], "dateTime": "2008-06-01T22:00:00.000-05:00"}, {"value": "2230", "qualifiers": ["A"], "dateTime": "2008-06-01T22:15:00.000-05:00"}, {"value": "2220", "qualifiers": ["A"], "dateTime": "2008-06-01T22:30:00.000-05:00"}, {"value": "2200", "qualifiers": ["A"], "dateTime": "2008-06-01T22:45:00.000-05:00"}, {"value": "2180", "qualifiers": ["A"], "dateTime": "2008-06-01T23:00:00.000-05:00"}, {"value": "2180", "qualifiers": ["A"], "dateTime": "2008-06-01T23:15:00.000-05:00"}, {"value": "2170", "qualifiers": ["A"], "dateTime": "2008-06-01T23:30:00.000-05:00"}, {"value": "2150", "qualifiers": ["A"], "dateTime": "2008-06-01T23:45:00.000-05:00"}, {"value": "2140", "qualifiers": ["A"], "dateTime": "2008-06-02T00:00:00.000-05:00"}, {"value": "2080", "qualifiers": ["A"], "dateTime": "2008-06-02T00:15:00.000-05:00"}, {"value": "2090", "qualifiers": ["A"], "dateTime": "2008-06-02T00:30:00.000-05:00"}, {"value": "2060", "qualifiers": ["A"], "dateTime": "2008-06-02T00:45:00.000-05:00"}, {"value": "2070", "qualifiers": ["A"], "dateTime": "2008-06-02T01:00:00.000-05:00"}, {"value": "2060", "qualifiers": ["A"], "dateTime": "2008-06-02T01:15:00.000-05:00"}, {"value": "2040", "qualifiers": ["A"], "dateTime": "2008-06-02T01:30:00.000-05:00"}, {"value": "2030", "qualifiers": ["A"], "dateTime": "2008-06-02T01:45:00.000-05:00"}, {"value": "2000", "qualifiers": ["A"], "dateTime": "2008-06-02T02:00:00.000-05:00"}, {"value": "1990", "qualifiers": ["A"], "dateTime": "2008-06-02T02:15:00.000-05:00"}, {"value": "1970", "qualifiers": ["A"], "dateTime": "2008-06-02T02:30:00.000-05:00"}, {"value": "1960", "qualifiers": ["A"], "dateTime": "2008-06-02T02:45:00.000-05:00"}, {"value": "1950", "qualifiers": ["A"], "dateTime": "2008-06-02T03:00:00.000-05:00"}, {"value": "1940", "qualifiers": ["A"], "dateTime": "2008-06-02T03:15:00.000-05:00"}, {"value": "1920", "qualifiers": ["A"], "dateTime": "2008-06-02T03:30:00.000-05:00"}, {"value": "1900", "qualifiers": ["A"], "dateTime": "2008-06-02T03:45:00.000-05:00"}, {"value": "1890", "qualifiers": ["A"], "dateTime": "2008-06-02T04:00:00.000-05:00"}, {"value": "1870", "qualifiers": ["A"], "dateTime": "2008-06-02T04:15:00.000-05:00"}, {"value": "1850", "qualifiers": ["A"], "dateTime": "2008-06-02T04:30:00.000-05:00"}, {"value": "1840", "qualifiers": ["A"], "dateTime": "2008-06-02T04:45:00.000-05:00"}, {"value": "1820", "qualifiers": ["A"], "dateTime": "2008-06-02T05:00:00.000-05:00"}, {"value": "1810", "qualifiers": ["A"], "dateTime": "2008-06-02T05:15:00.000-05:00"}, {"value": "1800", "qualifiers": ["A"], "dateTime": "2008-06-02T05:30:00.000-05:00"}, {"value": "1780", "qualifiers": ["A"], "dateTime": "2008-06-02T05:45:00.000-05:00"}, {"value": "1780", "qualifiers": ["A"], "dateTime": "2008-06-02T06:00:00.000-05:00"}, {"value": "1760", "qualifiers": ["A"], "dateTime": "2008-06-02T06:15:00.000-05:00"}, {"value": "1750", "qualifiers": ["A"], "dateTime": "2008-06-02T06:30:00.000-05:00"}, {"value": "1740", "qualifiers": ["A"], "dateTime": "2008-06-02T06:45:00.000-05:00"}, {"value": "1710", "qualifiers": ["A"], "dateTime": "2008-06-02T07:00:00.000-05:00"}, {"value": "1700", "qualifiers": ["A"], "dateTime": "2008-06-02T07:15:00.000-05:00"}, {"value": "1690", "qualifiers": ["A"], "dateTime": "2008-06-02T07:30:00.000-05:00"}, {"value": "1680", "qualifiers": ["A"], "dateTime": "2008-06-02T07:45:00.000-05:00"}, {"value": "1670", "qualifiers": ["A"], "dateTime": "2008-06-02T08:00:00.000-05:00"}, {"value": "1650", "qualifiers": ["A"], "dateTime": "2008-06-02T08:15:00.000-05:00"}, {"value": "1640", "qualifiers": ["A"], "dateTime": "2008-06-02T08:30:00.000-05:00"}, {"value": "1630", "qualifiers": ["A"], "dateTime": "2008-06-02T08:45:00.000-05:00"}, {"value": "1620", "qualifiers": ["A"], "dateTime": "2008-06-02T09:00:00.000-05:00"}, {"value": "1610", "qualifiers": ["A"], "dateTime": "2008-06-02T09:15:00.000-05:00"}, {"value": "1590", "qualifiers": ["A"], "dateTime": "2008-06-02T09:30:00.000-05:00"}, {"value": "1580", "qualifiers": ["A"], "dateTime": "2008-06-02T09:45:00.000-05:00"}, {"value": "1570", "qualifiers": ["A"], "dateTime": "2008-06-02T10:00:00.000-05:00"}, {"value": "1560", "qualifiers": ["A"], "dateTime": "2008-06-02T10:15:00.000-05:00"}, {"value": "1550", "qualifiers": ["A"], "dateTime": "2008-06-02T10:30:00.000-05:00"}, {"value": "1540", "qualifiers": ["A"], "dateTime": "2008-06-02T10:45:00.000-05:00"}, {"value": "1530", "qualifiers": ["A"], "dateTime": "2008-06-02T11:00:00.000-05:00"}, {"value": "1520", "qualifiers": ["A"], "dateTime": "2008-06-02T11:15:00.000-05:00"}, {"value": "1510", "qualifiers": ["A"], "dateTime": "2008-06-02T11:30:00.000-05:00"}, {"value": "1490", "qualifiers": ["A"], "dateTime": "2008-06-02T11:45:00.000-05:00"}, {"value": "1490", "qualifiers": ["A"], "dateTime": "2008-06-02T12:00:00.000-05:00"}, {"value": "1480", "qualifiers": ["A"], "dateTime": "2008-06-02T12:15:00.000-05:00"}, {"value": "1470", "qualifiers": ["A"], "dateTime": "2008-06-02T12:30:00.000-05:00"}, {"value": "1450", "qualifiers": ["A"], "dateTime": "2008-06-02T12:45:00.000-05:00"}, {"value": "1450", "qualifiers": ["A"], "dateTime": "2008-06-02T13:00:00.000-05:00"}, {"value": "1440", "qualifiers": ["A"], "dateTime": "2008-06-02T13:15:00.000-05:00"}, {"value": "1430", "qualifiers": ["A"], "dateTime": "2008-06-02T13:30:00.000-05:00"}, {"value": "1420", "qualifiers": ["A"], "dateTime": "2008-06-02T13:45:00.000-05:00"}, {"value": "1420", "qualifiers": ["A"], "dateTime": "2008-06-02T14:00:00.000-05:00"}, {"value": "1400", "qualifiers": ["A"], "dateTime": "2008-06-02T14:15:00.000-05:00"}, {"value": "1390", "qualifiers": ["A"], "dateTime": "2008-06-02T14:30:00.000-05:00"}, {"value": "1390", "qualifiers": ["A"], "dateTime": "2008-06-02T14:45:00.000-05:00"}, {"value": "1380", "qualifiers": ["A"], "dateTime": "2008-06-02T15:00:00.000-05:00"}, {"value": "1370", "qualifiers": ["A"], "dateTime": "2008-06-02T15:15:00.000-05:00"}, {"value": "1370", "qualifiers": ["A"], "dateTime": "2008-06-02T15:30:00.000-05:00"}, {"value": "1360", "qualifiers": ["A"], "dateTime": "2008-06-02T15:45:00.000-05:00"}, {"value": "1350", "qualifiers": ["A"], "dateTime": "2008-06-02T16:00:00.000-05:00"}, {"value": "1340", "qualifiers": ["A"], "dateTime": "2008-06-02T16:15:00.000-05:00"}, {"value": "1340", "qualifiers": ["A"], "dateTime": "2008-06-02T16:30:00.000-05:00"}, {"value": "1330", "qualifiers": ["A"], "dateTime": "2008-06-02T16:45:00.000-05:00"}, {"value": "1320", "qualifiers": ["A"], "dateTime": "2008-06-02T17:00:00.000-05:00"}, {"value": "1320", "qualifiers": ["A"], "dateTime": "2008-06-02T17:15:00.000-05:00"}, {"value": "1300", "qualifiers": ["A"], "dateTime": "2008-06-02T17:30:00.000-05:00"}, {"value": "1300", "qualifiers": ["A"], "dateTime": "2008-06-02T17:45:00.000-05:00"}, {"value": "1290", "qualifiers": ["A"], "dateTime": "2008-06-02T18:00:00.000-05:00"}, {"value": "1290", "qualifiers": ["A"], "dateTime": "2008-06-02T18:15:00.000-05:00"}, {"value": "1280", "qualifiers": ["A"], "dateTime": "2008-06-02T18:30:00.000-05:00"}, {"value": "1270", "qualifiers": ["A"], "dateTime": "2008-06-02T18:45:00.000-05:00"}, {"value": "1260", "qualifiers": ["A"], "dateTime": "2008-06-02T19:00:00.000-05:00"}, {"value": "1260", "qualifiers": ["A"], "dateTime": "2008-06-02T19:15:00.000-05:00"}, {"value": "1260", "qualifiers": ["A"], "dateTime": "2008-06-02T19:30:00.000-05:00"}, {"value": "1250", "qualifiers": ["A"], "dateTime": "2008-06-02T19:45:00.000-05:00"}, {"value": "1240", "qualifiers": ["A"], "dateTime": "2008-06-02T20:00:00.000-05:00"}, {"value": "1240", "qualifiers": ["A"], "dateTime": "2008-06-02T20:15:00.000-05:00"}, {"value": "1230", "qualifiers": ["A"], "dateTime": "2008-06-02T20:30:00.000-05:00"}, {"value": "1220", "qualifiers": ["A"], "dateTime": "2008-06-02T20:45:00.000-05:00"}, {"value": "1220", "qualifiers": ["A"], "dateTime": "2008-06-02T21:00:00.000-05:00"}, {"value": "1220", "qualifiers": ["A"], "dateTime": "2008-06-02T21:15:00.000-05:00"}, {"value": "1210", "qualifiers": ["A"], "dateTime": "2008-06-02T21:30:00.000-05:00"}, {"value": "1200", "qualifiers": ["A"], "dateTime": "2008-06-02T21:45:00.000-05:00"}, {"value": "1200", "qualifiers": ["A"], "dateTime": "2008-06-02T22:00:00.000-05:00"}, {"value": "1200", "qualifiers": ["A"], "dateTime": "2008-06-02T22:15:00.000-05:00"}, {"value": "1190", "qualifiers": ["A"], "dateTime": "2008-06-02T22:30:00.000-05:00"}, {"value": "1190", "qualifiers": ["A"], "dateTime": "2008-06-02T22:45:00.000-05:00"}, {"value": "1180", "qualifiers": ["A"], "dateTime": "2008-06-02T23:00:00.000-05:00"}, {"value": "1170", "qualifiers": ["A"], "dateTime": "2008-06-02T23:15:00.000-05:00"}, {"value": "1180", "qualifiers": ["A"], "dateTime": "2008-06-02T23:30:00.000-05:00"}, {"value": "1170", "qualifiers": ["A"], "dateTime": "2008-06-02T23:45:00.000-05:00"}, {"value": "1160", "qualifiers": ["A"], "dateTime": "2008-06-03T00:00:00.000-05:00"}, {"value": "1160", "qualifiers": ["A"], "dateTime": "2008-06-03T00:15:00.000-05:00"}, {"value": "1160", "qualifiers": ["A"], "dateTime": "2008-06-03T00:30:00.000-05:00"}, {"value": "1160", "qualifiers": ["A"], "dateTime": "2008-06-03T00:45:00.000-05:00"}, {"value": "1150", "qualifiers": ["A"], "dateTime": "2008-06-03T01:00:00.000-05:00"}, {"value": "1150", "qualifiers": ["A"], "dateTime": "2008-06-03T01:15:00.000-05:00"}, {"value": "1150", "qualifiers": ["A"], "dateTime": "2008-06-03T01:30:00.000-05:00"}, {"value": "1150", "qualifiers": ["A"], "dateTime": "2008-06-03T01:45:00.000-05:00"}, {"value": "1140", "qualifiers": ["A"], "dateTime": "2008-06-03T02:00:00.000-05:00"}, {"value": "1140", "qualifiers": ["A"], "dateTime": "2008-06-03T02:15:00.000-05:00"}, {"value": "1130", "qualifiers": ["A"], "dateTime": "2008-06-03T02:30:00.000-05:00"}, {"value": "1140", "qualifiers": ["A"], "dateTime": "2008-06-03T02:45:00.000-05:00"}, {"value": "1130", "qualifiers": ["A"], "dateTime": "2008-06-03T03:00:00.000-05:00"}, {"value": "1120", "qualifiers": ["A"], "dateTime": "2008-06-03T03:15:00.000-05:00"}, {"value": "1130", "qualifiers": ["A"], "dateTime": "2008-06-03T03:30:00.000-05:00"}, {"value": "1130", "qualifiers": ["A"], "dateTime": "2008-06-03T03:45:00.000-05:00"}, {"value": "1120", "qualifiers": ["A"], "dateTime": "2008-06-03T04:00:00.000-05:00"}, {"value": "1120", "qualifiers": ["A"], "dateTime": "2008-06-03T04:15:00.000-05:00"}, {"value": "1110", "qualifiers": ["A"], "dateTime": "2008-06-03T04:30:00.000-05:00"}, {"value": "1110", "qualifiers": ["A"], "dateTime": "2008-06-03T04:45:00.000-05:00"}, {"value": "1110", "qualifiers": ["A"], "dateTime": "2008-06-03T05:00:00.000-05:00"}, {"value": "1100", "qualifiers": ["A"], "dateTime": "2008-06-03T05:15:00.000-05:00"}, {"value": "1100", "qualifiers": ["A"], "dateTime": "2008-06-03T05:30:00.000-05:00"}, {"value": "1100", "qualifiers": ["A"], "dateTime": "2008-06-03T05:45:00.000-05:00"}, {"value": "1100", "qualifiers": ["A"], "dateTime": "2008-06-03T06:00:00.000-05:00"}, {"value": "1100", "qualifiers": ["A"], "dateTime": "2008-06-03T06:15:00.000-05:00"}, {"value": "1100", "qualifiers": ["A"], "dateTime": "2008-06-03T06:30:00.000-05:00"}, {"value": "1100", "qualifiers": ["A"], "dateTime": "2008-06-03T06:45:00.000-05:00"}, {"value": "1110", "qualifiers": ["A"], "dateTime": "2008-06-03T07:00:00.000-05:00"}, {"value": "1110", "qualifiers": ["A"], "dateTime": "2008-06-03T07:15:00.000-05:00"}, {"value": "1110", "qualifiers": ["A"], "dateTime": "2008-06-03T07:30:00.000-05:00"}, {"value": "1120", "qualifiers": ["A"], "dateTime": "2008-06-03T07:45:00.000-05:00"}, {"value": "1120", "qualifiers": ["A"], "dateTime": "2008-06-03T08:00:00.000-05:00"}, {"value": "1130", "qualifiers": ["A"], "dateTime": "2008-06-03T08:15:00.000-05:00"}, {"value": "1130", "qualifiers": ["A"], "dateTime": "2008-06-03T08:30:00.000-05:00"}, {"value": "1130", "qualifiers": ["A"], "dateTime": "2008-06-03T08:45:00.000-05:00"}, {"value": "1130", "qualifiers": ["A"], "dateTime": "2008-06-03T09:00:00.000-05:00"}, {"value": "1130", "qualifiers": ["A"], "dateTime": "2008-06-03T09:15:00.000-05:00"}, {"value": "1130", "qualifiers": ["A"], "dateTime": "2008-06-03T09:30:00.000-05:00"}, {"value": "1140", "qualifiers": ["A"], "dateTime": "2008-06-03T09:45:00.000-05:00"}, {"value": "1140", "qualifiers": ["A"], "dateTime": "2008-06-03T10:00:00.000-05:00"}, {"value": "1150", "qualifiers": ["A"], "dateTime": "2008-06-03T10:15:00.000-05:00"}, {"value": "1150", "qualifiers": ["A"], "dateTime": "2008-06-03T10:30:00.000-05:00"}, {"value": "1150", "qualifiers": ["A"], "dateTime": "2008-06-03T10:45:00.000-05:00"}, {"value": "1160", "qualifiers": ["A"], "dateTime": "2008-06-03T11:00:00.000-05:00"}, {"value": "1160", "qualifiers": ["A"], "dateTime": "2008-06-03T11:15:00.000-05:00"}, {"value": "1160", "qualifiers": ["A"], "dateTime": "2008-06-03T11:30:00.000-05:00"}, {"value": "1160", "qualifiers": ["A"], "dateTime": "2008-06-03T11:45:00.000-05:00"}, {"value": "1170", "qualifiers": ["A"], "dateTime": "2008-06-03T12:00:00.000-05:00"}, {"value": "1170", "qualifiers": ["A"], "dateTime": "2008-06-03T12:15:00.000-05:00"}, {"value": "1170", "qualifiers": ["A"], "dateTime": "2008-06-03T12:30:00.000-05:00"}, {"value": "1180", "qualifiers": ["A"], "dateTime": "2008-06-03T12:45:00.000-05:00"}, {"value": "1190", "qualifiers": ["A"], "dateTime": "2008-06-03T13:00:00.000-05:00"}, {"value": "1180", "qualifiers": ["A"], "dateTime": "2008-06-03T13:15:00.000-05:00"}, {"value": "1190", "qualifiers": ["A"], "dateTime": "2008-06-03T13:30:00.000-05:00"}, {"value": "1190", "qualifiers": ["A"], "dateTime": "2008-06-03T13:45:00.000-05:00"}, {"value": "1190", "qualifiers": ["A"], "dateTime": "2008-06-03T14:00:00.000-05:00"}, {"value": "1200", "qualifiers": ["A"], "dateTime": "2008-06-03T14:15:00.000-05:00"}, {"value": "1200", "qualifiers": ["A"], "dateTime": "2008-06-03T14:30:00.000-05:00"}, {"value": "1200", "qualifiers": ["A"], "dateTime": "2008-06-03T14:45:00.000-05:00"}, {"value": "1210", "qualifiers": ["A"], "dateTime": "2008-06-03T15:00:00.000-05:00"}, {"value": "1210", "qualifiers": ["A"], "dateTime": "2008-06-03T15:15:00.000-05:00"}, {"value": "1210", "qualifiers": ["A"], "dateTime": "2008-06-03T15:30:00.000-05:00"}, {"value": "1220", "qualifiers": ["A"], "dateTime": "2008-06-03T15:45:00.000-05:00"}, {"value": "1220", "qualifiers": ["A"], "dateTime": "2008-06-03T16:00:00.000-05:00"}, {"value": "1220", "qualifiers": ["A"], "dateTime": "2008-06-03T16:15:00.000-05:00"}, {"value": "1230", "qualifiers": ["A"], "dateTime": "2008-06-03T16:30:00.000-05:00"}, {"value": "1230", "qualifiers": ["A"], "dateTime": "2008-06-03T16:45:00.000-05:00"}, {"value": "1240", "qualifiers": ["A"], "dateTime": "2008-06-03T17:00:00.000-05:00"}, {"value": "1240", "qualifiers": ["A"], "dateTime": "2008-06-03T17:15:00.000-05:00"}, {"value": "1250", "qualifiers": ["A"], "dateTime": "2008-06-03T17:30:00.000-05:00"}, {"value": "1240", "qualifiers": ["A"], "dateTime": "2008-06-03T17:45:00.000-05:00"}, {"value": "1260", "qualifiers": ["A"], "dateTime": "2008-06-03T18:00:00.000-05:00"}, {"value": "1260", "qualifiers": ["A"], "dateTime": "2008-06-03T18:15:00.000-05:00"}, {"value": "1260", "qualifiers": ["A"], "dateTime": "2008-06-03T18:30:00.000-05:00"}, {"value": "1260", "qualifiers": ["A"], "dateTime": "2008-06-03T18:45:00.000-05:00"}, {"value": "1270", "qualifiers": ["A"], "dateTime": "2008-06-03T19:00:00.000-05:00"}, {"value": "1270", "qualifiers": ["A"], "dateTime": "2008-06-03T19:15:00.000-05:00"}, {"value": "1280", "qualifiers": ["A"], "dateTime": "2008-06-03T19:30:00.000-05:00"}, {"value": "1280", "qualifiers": ["A"], "dateTime": "2008-06-03T19:45:00.000-05:00"}, {"value": "1280", "qualifiers": ["A"], "dateTime": "2008-06-03T20:00:00.000-05:00"}, {"value": "1290", "qualifiers": ["A"], "dateTime": "2008-06-03T20:15:00.000-05:00"}, {"value": "1290", "qualifiers": ["A"], "dateTime": "2008-06-03T20:30:00.000-05:00"}, {"value": "1300", "qualifiers": ["A"], "dateTime": "2008-06-03T20:45:00.000-05:00"}, {"value": "1300", "qualifiers": ["A"], "dateTime": "2008-06-03T21:00:00.000-05:00"}, {"value": "1300", "qualifiers": ["A"], "dateTime": "2008-06-03T21:15:00.000-05:00"}, {"value": "1300", "qualifiers": ["A"], "dateTime": "2008-06-03T21:30:00.000-05:00"}, {"value": "1320", "qualifiers": ["A"], "dateTime": "2008-06-03T21:45:00.000-05:00"}, {"value": "1320", "qualifiers": ["A"], "dateTime": "2008-06-03T22:00:00.000-05:00"}, {"value": "1320", "qualifiers": ["A"], "dateTime": "2008-06-03T22:15:00.000-05:00"}, {"value": "1320", "qualifiers": ["A"], "dateTime": "2008-06-03T22:30:00.000-05:00"}, {"value": "1320", "qualifiers": ["A"], "dateTime": "2008-06-03T22:45:00.000-05:00"}, {"value": "1330", "qualifiers": ["A"], "dateTime": "2008-06-03T23:00:00.000-05:00"}, {"value": "1340", "qualifiers": ["A"], "dateTime": "2008-06-03T23:15:00.000-05:00"}, {"value": "1340", "qualifiers": ["A"], "dateTime": "2008-06-03T23:30:00.000-05:00"}, {"value": "1350", "qualifiers": ["A"], "dateTime": "2008-06-03T23:45:00.000-05:00"}, {"value": "1350", "qualifiers": ["A"], "dateTime": "2008-06-04T00:00:00.000-05:00"}, {"value": "1350", "qualifiers": ["A"], "dateTime": "2008-06-04T00:15:00.000-05:00"}, {"value": "1350", "qualifiers": ["A"], "dateTime": "2008-06-04T00:30:00.000-05:00"}, {"value": "1360", "qualifiers": ["A"], "dateTime": "2008-06-04T00:45:00.000-05:00"}, {"value": "1360", "qualifiers": ["A"], "dateTime": "2008-06-04T01:00:00.000-05:00"}, {"value": "1370", "qualifiers": ["A"], "dateTime": "2008-06-04T01:15:00.000-05:00"}, {"value": "1370", "qualifiers": ["A"], "dateTime": "2008-06-04T01:30:00.000-05:00"}, {"value": "1370", "qualifiers": ["A"], "dateTime": "2008-06-04T01:45:00.000-05:00"}, {"value": "1380", "qualifiers": ["A"], "dateTime": "2008-06-04T02:00:00.000-05:00"}, {"value": "1390", "qualifiers": ["A"], "dateTime": "2008-06-04T02:15:00.000-05:00"}, {"value": "1390", "qualifiers": ["A"], "dateTime": "2008-06-04T02:30:00.000-05:00"}, {"value": "1390", "qualifiers": ["A"], "dateTime": "2008-06-04T02:45:00.000-05:00"}, {"value": "1390", "qualifiers": ["A"], "dateTime": "2008-06-04T03:00:00.000-05:00"}, {"value": "1390", "qualifiers": ["A"], "dateTime": "2008-06-04T03:15:00.000-05:00"}, {"value": "1400", "qualifiers": ["A"], "dateTime": "2008-06-04T03:30:00.000-05:00"}, {"value": "1390", "qualifiers": ["A"], "dateTime": "2008-06-04T03:45:00.000-05:00"}, {"value": "1410", "qualifiers": ["A"], "dateTime": "2008-06-04T04:00:00.000-05:00"}, {"value": "1410", "qualifiers": ["A"], "dateTime": "2008-06-04T04:15:00.000-05:00"}, {"value": "1420", "qualifiers": ["A"], "dateTime": "2008-06-04T04:30:00.000-05:00"}, {"value": "1410", "qualifiers": ["A"], "dateTime": "2008-06-04T04:45:00.000-05:00"}, {"value": "1420", "qualifiers": ["A"], "dateTime": "2008-06-04T05:00:00.000-05:00"}, {"value": "1420", "qualifiers": ["A"], "dateTime": "2008-06-04T05:15:00.000-05:00"}, {"value": "1420", "qualifiers": ["A"], "dateTime": "2008-06-04T05:30:00.000-05:00"}, {"value": "1430", "qualifiers": ["A"], "dateTime": "2008-06-04T05:45:00.000-05:00"}, {"value": "1420", "qualifiers": ["A"], "dateTime": "2008-06-04T06:00:00.000-05:00"}, {"value": "1430", "qualifiers": ["A"], "dateTime": "2008-06-04T06:15:00.000-05:00"}, {"value": "1440", "qualifiers": ["A"], "dateTime": "2008-06-04T06:30:00.000-05:00"}, {"value": "1440", "qualifiers": ["A"], "dateTime": "2008-06-04T06:45:00.000-05:00"}, {"value": "1440", "qualifiers": ["A"], "dateTime": "2008-06-04T07:00:00.000-05:00"}, {"value": "1440", "qualifiers": ["A"], "dateTime": "2008-06-04T07:15:00.000-05:00"}, {"value": "1450", "qualifiers": ["A"], "dateTime": "2008-06-04T07:30:00.000-05:00"}, {"value": "1450", "qualifiers": ["A"], "dateTime": "2008-06-04T07:45:00.000-05:00"}, {"value": "1450", "qualifiers": ["A"], "dateTime": "2008-06-04T08:00:00.000-05:00"}, {"value": "1450", "qualifiers": ["A"], "dateTime": "2008-06-04T08:15:00.000-05:00"}, {"value": "1450", "qualifiers": ["A"], "dateTime": "2008-06-04T08:30:00.000-05:00"}, {"value": "1460", "qualifiers": ["A"], "dateTime": "2008-06-04T08:45:00.000-05:00"}, {"value": "1450", "qualifiers": ["A"], "dateTime": "2008-06-04T09:00:00.000-05:00"}, {"value": "1450", "qualifiers": ["A"], "dateTime": "2008-06-04T09:15:00.000-05:00"}, {"value": "1460", "qualifiers": ["A"], "dateTime": "2008-06-04T09:30:00.000-05:00"}, {"value": "1460", "qualifiers": ["A"], "dateTime": "2008-06-04T09:45:00.000-05:00"}, {"value": "1460", "qualifiers": ["A"], "dateTime": "2008-06-04T10:00:00.000-05:00"}, {"value": "1470", "qualifiers": ["A"], "dateTime": "2008-06-04T10:15:00.000-05:00"}, {"value": "1470", "qualifiers": ["A"], "dateTime": "2008-06-04T10:30:00.000-05:00"}, {"value": "1470", "qualifiers": ["A"], "dateTime": "2008-06-04T10:45:00.000-05:00"}, {"value": "1470", "qualifiers": ["A"], "dateTime": "2008-06-04T11:00:00.000-05:00"}, {"value": "1470", "qualifiers": ["A"], "dateTime": "2008-06-04T11:15:00.000-05:00"}, {"value": "1470", "qualifiers": ["A"], "dateTime": "2008-06-04T11:30:00.000-05:00"}, {"value": "1470", "qualifiers": ["A"], "dateTime": "2008-06-04T11:45:00.000-05:00"}, {"value": "1470", "qualifiers": ["A"], "dateTime": "2008-06-04T12:00:00.000-05:00"}, {"value": "1470", "qualifiers": ["A"], "dateTime": "2008-06-04T12:15:00.000-05:00"}, {"value": "1480", "qualifiers": ["A"], "dateTime": "2008-06-04T12:30:00.000-05:00"}, {"value": "1470", "qualifiers": ["A"], "dateTime": "2008-06-04T12:45:00.000-05:00"}, {"value": "1480", "qualifiers": ["A"], "dateTime": "2008-06-04T13:00:00.000-05:00"}, {"value": "1480", "qualifiers": ["A"], "dateTime": "2008-06-04T13:15:00.000-05:00"}, {"value": "1470", "qualifiers": ["A"], "dateTime": "2008-06-04T13:30:00.000-05:00"}, {"value": "1480", "qualifiers": ["A"], "dateTime": "2008-06-04T13:45:00.000-05:00"}, {"value": "1480", "qualifiers": ["A"], "dateTime": "2008-06-04T14:00:00.000-05:00"}, {"value": "1480", "qualifiers": ["A"], "dateTime": "2008-06-04T14:15:00.000-05:00"}, {"value": "1480", "qualifiers": ["A"], "dateTime": "2008-06-04T14:30:00.000-05:00"}, {"value": "1480", "qualifiers": ["A"], "dateTime": "2008-06-04T14:45:00.000-05:00"}, {"value": "1480", "qualifiers": ["A"], "dateTime": "2008-06-04T15:00:00.000-05:00"}, {"value": "1480", "qualifiers": ["A"], "dateTime": "2008-06-04T15:15:00.000-05:00"}, {"value": "1480", "qualifiers": ["A"], "dateTime": "2008-06-04T15:30:00.000-05:00"}, {"value": "1480", "qualifiers": ["A"], "dateTime": "2008-06-04T15:45:00.000-05:00"}, {"value": "1480", "qualifiers": ["A"], "dateTime": "2008-06-04T16:00:00.000-05:00"}, {"value": "1480", "qualifiers": ["A"], "dateTime": "2008-06-04T16:15:00.000-05:00"}, {"value": "1480", "qualifiers": ["A"], "dateTime": "2008-06-04T16:30:00.000-05:00"}, {"value": "1480", "qualifiers": ["A"], "dateTime": "2008-06-04T16:45:00.000-05:00"}, {"value": "1480", "qualifiers": ["A"], "dateTime": "2008-06-04T17:00:00.000-05:00"}, {"value": "1480", "qualifiers": ["A"], "dateTime": "2008-06-04T17:15:00.000-05:00"}, {"value": "1480", "qualifiers": ["A"], "dateTime": "2008-06-04T17:30:00.000-05:00"}, {"value": "1470", "qualifiers": ["A"], "dateTime": "2008-06-04T17:45:00.000-05:00"}, {"value": "1480", "qualifiers": ["A"], "dateTime": "2008-06-04T18:00:00.000-05:00"}, {"value": "1480", "qualifiers": ["A"], "dateTime": "2008-06-04T18:15:00.000-05:00"}, {"value": "1480", "qualifiers": ["A"], "dateTime": "2008-06-04T18:30:00.000-05:00"}, {"value": "1480", "qualifiers": ["A"], "dateTime": "2008-06-04T18:45:00.000-05:00"}, {"value": "1480", "qualifiers": ["A"], "dateTime": "2008-06-04T19:00:00.000-05:00"}, {"value": "1470", "qualifiers": ["A"], "dateTime": "2008-06-04T19:15:00.000-05:00"}, {"value": "1470", "qualifiers": ["A"], "dateTime": "2008-06-04T19:30:00.000-05:00"}, {"value": "1470", "qualifiers": ["A"], "dateTime": "2008-06-04T19:45:00.000-05:00"}, {"value": "1470", "qualifiers": ["A"], "dateTime": "2008-06-04T20:00:00.000-05:00"}, {"value": "1460", "qualifiers": ["A"], "dateTime": "2008-06-04T20:15:00.000-05:00"}, {"value": "1460", "qualifiers": ["A"], "dateTime": "2008-06-04T20:30:00.000-05:00"}, {"value": "1450", "qualifiers": ["A"], "dateTime": "2008-06-04T20:45:00.000-05:00"}, {"value": "1450", "qualifiers": ["A"], "dateTime": "2008-06-04T21:00:00.000-05:00"}, {"value": "1450", "qualifiers": ["A"], "dateTime": "2008-06-04T21:15:00.000-05:00"}, {"value": "1450", "qualifiers": ["A"], "dateTime": "2008-06-04T21:30:00.000-05:00"}, {"value": "1450", "qualifiers": ["A"], "dateTime": "2008-06-04T21:45:00.000-05:00"}, {"value": "1440", "qualifiers": ["A"], "dateTime": "2008-06-04T22:00:00.000-05:00"}, {"value": "1450", "qualifiers": ["A"], "dateTime": "2008-06-04T22:15:00.000-05:00"}, {"value": "1440", "qualifiers": ["A"], "dateTime": "2008-06-04T22:30:00.000-05:00"}, {"value": "1440", "qualifiers": ["A"], "dateTime": "2008-06-04T22:45:00.000-05:00"}, {"value": "1430", "qualifiers": ["A"], "dateTime": "2008-06-04T23:00:00.000-05:00"}, {"value": "1430", "qualifiers": ["A"], "dateTime": "2008-06-04T23:15:00.000-05:00"}, {"value": "1420", "qualifiers": ["A"], "dateTime": "2008-06-04T23:30:00.000-05:00"}, {"value": "1420", "qualifiers": ["A"], "dateTime": "2008-06-04T23:45:00.000-05:00"}, {"value": "1420", "qualifiers": ["A"], "dateTime": "2008-06-05T00:00:00.000-05:00"}, {"value": "1420", "qualifiers": ["A"], "dateTime": "2008-06-05T00:15:00.000-05:00"}, {"value": "1410", "qualifiers": ["A"], "dateTime": "2008-06-05T00:30:00.000-05:00"}, {"value": "1410", "qualifiers": ["A"], "dateTime": "2008-06-05T00:45:00.000-05:00"}, {"value": "1400", "qualifiers": ["A"], "dateTime": "2008-06-05T01:00:00.000-05:00"}, {"value": "1390", "qualifiers": ["A"], "dateTime": "2008-06-05T01:15:00.000-05:00"}, {"value": "1390", "qualifiers": ["A"], "dateTime": "2008-06-05T01:30:00.000-05:00"}, {"value": "1400", "qualifiers": ["A"], "dateTime": "2008-06-05T01:45:00.000-05:00"}, {"value": "1390", "qualifiers": ["A"], "dateTime": "2008-06-05T02:00:00.000-05:00"}, {"value": "1390", "qualifiers": ["A"], "dateTime": "2008-06-05T02:15:00.000-05:00"}, {"value": "1390", "qualifiers": ["A"], "dateTime": "2008-06-05T02:30:00.000-05:00"}, {"value": "1420", "qualifiers": ["A"], "dateTime": "2008-06-05T02:45:00.000-05:00"}, {"value": "1420", "qualifiers": ["A"], "dateTime": "2008-06-05T03:00:00.000-05:00"}, {"value": "1430", "qualifiers": ["A"], "dateTime": "2008-06-05T03:15:00.000-05:00"}, {"value": "1430", "qualifiers": ["A"], "dateTime": "2008-06-05T03:30:00.000-05:00"}, {"value": "1430", "qualifiers": ["A"], "dateTime": "2008-06-05T03:45:00.000-05:00"}, {"value": "1420", "qualifiers": ["A"], "dateTime": "2008-06-05T04:00:00.000-05:00"}, {"value": "1440", "qualifiers": ["A"], "dateTime": "2008-06-05T04:15:00.000-05:00"}, {"value": "1430", "qualifiers": ["A"], "dateTime": "2008-06-05T04:30:00.000-05:00"}, {"value": "1440", "qualifiers": ["A"], "dateTime": "2008-06-05T04:45:00.000-05:00"}, {"value": "1450", "qualifiers": ["A"], "dateTime": "2008-06-05T05:00:00.000-05:00"}, {"value": "1460", "qualifiers": ["A"], "dateTime": "2008-06-05T05:15:00.000-05:00"}, {"value": "1460", "qualifiers": ["A"], "dateTime": "2008-06-05T05:30:00.000-05:00"}, {"value": "1470", "qualifiers": ["A"], "dateTime": "2008-06-05T05:45:00.000-05:00"}, {"value": "1480", "qualifiers": ["A"], "dateTime": "2008-06-05T06:00:00.000-05:00"}, {"value": "1480", "qualifiers": ["A"], "dateTime": "2008-06-05T06:15:00.000-05:00"}, {"value": "1500", "qualifiers": ["A"], "dateTime": "2008-06-05T06:30:00.000-05:00"}, {"value": "1510", "qualifiers": ["A"], "dateTime": "2008-06-05T06:45:00.000-05:00"}, {"value": "1510", "qualifiers": ["A"], "dateTime": "2008-06-05T07:00:00.000-05:00"}, {"value": "1520", "qualifiers": ["A"], "dateTime": "2008-06-05T07:15:00.000-05:00"}, {"value": "1520", "qualifiers": ["A"], "dateTime": "2008-06-05T07:30:00.000-05:00"}, {"value": "1540", "qualifiers": ["A"], "dateTime": "2008-06-05T07:45:00.000-05:00"}, {"value": "1540", "qualifiers": ["A"], "dateTime": "2008-06-05T08:00:00.000-05:00"}, {"value": "1550", "qualifiers": ["A"], "dateTime": "2008-06-05T08:15:00.000-05:00"}, {"value": "1570", "qualifiers": ["A"], "dateTime": "2008-06-05T08:30:00.000-05:00"}, {"value": "1580", "qualifiers": ["A"], "dateTime": "2008-06-05T08:45:00.000-05:00"}, {"value": "1590", "qualifiers": ["A"], "dateTime": "2008-06-05T09:00:00.000-05:00"}, {"value": "1600", "qualifiers": ["A"], "dateTime": "2008-06-05T09:15:00.000-05:00"}, {"value": "1610", "qualifiers": ["A"], "dateTime": "2008-06-05T09:30:00.000-05:00"}, {"value": "1620", "qualifiers": ["A"], "dateTime": "2008-06-05T09:45:00.000-05:00"}, {"value": "1630", "qualifiers": ["A"], "dateTime": "2008-06-05T10:00:00.000-05:00"}, {"value": "1630", "qualifiers": ["A"], "dateTime": "2008-06-05T10:15:00.000-05:00"}, {"value": "1640", "qualifiers": ["A"], "dateTime": "2008-06-05T10:30:00.000-05:00"}, {"value": "1660", "qualifiers": ["A"], "dateTime": "2008-06-05T10:45:00.000-05:00"}, {"value": "1680", "qualifiers": ["A"], "dateTime": "2008-06-05T11:00:00.000-05:00"}, {"value": "1680", "qualifiers": ["A"], "dateTime": "2008-06-05T11:15:00.000-05:00"}, {"value": "1690", "qualifiers": ["A"], "dateTime": "2008-06-05T11:30:00.000-05:00"}, {"value": "1700", "qualifiers": ["A"], "dateTime": "2008-06-05T11:45:00.000-05:00"}, {"value": "1720", "qualifiers": ["A"], "dateTime": "2008-06-05T12:00:00.000-05:00"}, {"value": "1710", "qualifiers": ["A"], "dateTime": "2008-06-05T12:15:00.000-05:00"}, {"value": "1730", "qualifiers": ["A"], "dateTime": "2008-06-05T12:30:00.000-05:00"}, {"value": "1740", "qualifiers": ["A"], "dateTime": "2008-06-05T12:45:00.000-05:00"}, {"value": "1740", "qualifiers": ["A"], "dateTime": "2008-06-05T13:00:00.000-05:00"}, {"value": "1750", "qualifiers": ["A"], "dateTime": "2008-06-05T13:15:00.000-05:00"}, {"value": "1760", "qualifiers": ["A"], "dateTime": "2008-06-05T13:30:00.000-05:00"}, {"value": "1760", "qualifiers": ["A"], "dateTime": "2008-06-05T13:45:00.000-05:00"}, {"value": "1760", "qualifiers": ["A"], "dateTime": "2008-06-05T14:00:00.000-05:00"}, {"value": "1780", "qualifiers": ["A"], "dateTime": "2008-06-05T14:15:00.000-05:00"}, {"value": "1780", "qualifiers": ["A"], "dateTime": "2008-06-05T14:30:00.000-05:00"}, {"value": "1790", "qualifiers": ["A"], "dateTime": "2008-06-05T14:45:00.000-05:00"}, {"value": "1780", "qualifiers": ["A"], "dateTime": "2008-06-05T15:00:00.000-05:00"}, {"value": "1790", "qualifiers": ["A"], "dateTime": "2008-06-05T15:15:00.000-05:00"}, {"value": "1800", "qualifiers": ["A"], "dateTime": "2008-06-05T15:30:00.000-05:00"}, {"value": "1800", "qualifiers": ["A"], "dateTime": "2008-06-05T15:45:00.000-05:00"}, {"value": "1800", "qualifiers": ["A"], "dateTime": "2008-06-05T16:00:00.000-05:00"}, {"value": "1790", "qualifiers": ["A"], "dateTime": "2008-06-05T16:15:00.000-05:00"}, {"value": "1820", "qualifiers": ["A"], "dateTime": "2008-06-05T16:30:00.000-05:00"}, {"value": "1810", "qualifiers": ["A"], "dateTime": "2008-06-05T16:45:00.000-05:00"}, {"value": "1820", "qualifiers": ["A"], "dateTime": "2008-06-05T17:00:00.000-05:00"}, {"value": "1820", "qualifiers": ["A"], "dateTime": "2008-06-05T17:15:00.000-05:00"}, {"value": "1830", "qualifiers": ["A"], "dateTime": "2008-06-05T17:30:00.000-05:00"}, {"value": "1830", "qualifiers": ["A"], "dateTime": "2008-06-05T17:45:00.000-05:00"}, {"value": "1830", "qualifiers": ["A"], "dateTime": "2008-06-05T18:00:00.000-05:00"}, {"value": "1840", "qualifiers": ["A"], "dateTime": "2008-06-05T18:15:00.000-05:00"}, {"value": "1850", "qualifiers": ["A"], "dateTime": "2008-06-05T18:30:00.000-05:00"}, {"value": "1850", "qualifiers": ["A"], "dateTime": "2008-06-05T18:45:00.000-05:00"}, {"value": "1840", "qualifiers": ["A"], "dateTime": "2008-06-05T19:00:00.000-05:00"}, {"value": "1850", "qualifiers": ["A"], "dateTime": "2008-06-05T19:15:00.000-05:00"}, {"value": "1850", "qualifiers": ["A"], "dateTime": "2008-06-05T19:30:00.000-05:00"}, {"value": "1850", "qualifiers": ["A"], "dateTime": "2008-06-05T19:45:00.000-05:00"}, {"value": "1870", "qualifiers": ["A"], "dateTime": "2008-06-05T20:00:00.000-05:00"}, {"value": "1880", "qualifiers": ["A"], "dateTime": "2008-06-05T20:15:00.000-05:00"}, {"value": "1880", "qualifiers": ["A"], "dateTime": "2008-06-05T20:30:00.000-05:00"}, {"value": "1880", "qualifiers": ["A"], "dateTime": "2008-06-05T20:45:00.000-05:00"}, {"value": "1860", "qualifiers": ["A"], "dateTime": "2008-06-05T21:00:00.000-05:00"}, {"value": "1870", "qualifiers": ["A"], "dateTime": "2008-06-05T21:15:00.000-05:00"}, {"value": "1890", "qualifiers": ["A"], "dateTime": "2008-06-05T21:30:00.000-05:00"}, {"value": "1900", "qualifiers": ["A"], "dateTime": "2008-06-05T21:45:00.000-05:00"}, {"value": "1890", "qualifiers": ["A"], "dateTime": "2008-06-05T22:00:00.000-05:00"}, {"value": "1890", "qualifiers": ["A"], "dateTime": "2008-06-05T22:15:00.000-05:00"}, {"value": "1900", "qualifiers": ["A"], "dateTime": "2008-06-05T22:30:00.000-05:00"}, {"value": "1910", "qualifiers": ["A"], "dateTime": "2008-06-05T22:45:00.000-05:00"}, {"value": "1910", "qualifiers": ["A"], "dateTime": "2008-06-05T23:00:00.000-05:00"}, {"value": "1910", "qualifiers": ["A"], "dateTime": "2008-06-05T23:15:00.000-05:00"}, {"value": "1900", "qualifiers": ["A"], "dateTime": "2008-06-05T23:30:00.000-05:00"}, {"value": "1920", "qualifiers": ["A"], "dateTime": "2008-06-05T23:45:00.000-05:00"}, {"value": "1920", "qualifiers": ["A"], "dateTime": "2008-06-06T00:00:00.000-05:00"}, {"value": "1920", "qualifiers": ["A"], "dateTime": "2008-06-06T00:15:00.000-05:00"}, {"value": "1920", "qualifiers": ["A"], "dateTime": "2008-06-06T00:30:00.000-05:00"}, {"value": "1920", "qualifiers": ["A"], "dateTime": "2008-06-06T00:45:00.000-05:00"}, {"value": "1930", "qualifiers": ["A"], "dateTime": "2008-06-06T01:00:00.000-05:00"}, {"value": "1930", "qualifiers": ["A"], "dateTime": "2008-06-06T01:15:00.000-05:00"}, {"value": "1920", "qualifiers": ["A"], "dateTime": "2008-06-06T01:30:00.000-05:00"}, {"value": "1930", "qualifiers": ["A"], "dateTime": "2008-06-06T01:45:00.000-05:00"}, {"value": "1930", "qualifiers": ["A"], "dateTime": "2008-06-06T02:00:00.000-05:00"}, {"value": "1940", "qualifiers": ["A"], "dateTime": "2008-06-06T02:15:00.000-05:00"}, {"value": "1940", "qualifiers": ["A"], "dateTime": "2008-06-06T02:30:00.000-05:00"}, {"value": "1940", "qualifiers": ["A"], "dateTime": "2008-06-06T02:45:00.000-05:00"}, {"value": "1940", "qualifiers": ["A"], "dateTime": "2008-06-06T03:00:00.000-05:00"}, {"value": "1950", "qualifiers": ["A"], "dateTime": "2008-06-06T03:15:00.000-05:00"}, {"value": "1950", "qualifiers": ["A"], "dateTime": "2008-06-06T03:30:00.000-05:00"}, {"value": "1950", "qualifiers": ["A"], "dateTime": "2008-06-06T03:45:00.000-05:00"}, {"value": "1950", "qualifiers": ["A"], "dateTime": "2008-06-06T04:00:00.000-05:00"}, {"value": "1950", "qualifiers": ["A"], "dateTime": "2008-06-06T04:15:00.000-05:00"}, {"value": "1960", "qualifiers": ["A"], "dateTime": "2008-06-06T04:30:00.000-05:00"}, {"value": "1960", "qualifiers": ["A"], "dateTime": "2008-06-06T04:45:00.000-05:00"}, {"value": "1960", "qualifiers": ["A"], "dateTime": "2008-06-06T05:00:00.000-05:00"}, {"value": "1950", "qualifiers": ["A"], "dateTime": "2008-06-06T05:15:00.000-05:00"}, {"value": "1950", "qualifiers": ["A"], "dateTime": "2008-06-06T05:30:00.000-05:00"}, {"value": "1960", "qualifiers": ["A"], "dateTime": "2008-06-06T05:45:00.000-05:00"}, {"value": "1950", "qualifiers": ["A"], "dateTime": "2008-06-06T06:00:00.000-05:00"}, {"value": "1970", "qualifiers": ["A"], "dateTime": "2008-06-06T06:15:00.000-05:00"}, {"value": "1960", "qualifiers": ["A"], "dateTime": "2008-06-06T06:30:00.000-05:00"}, {"value": "1950", "qualifiers": ["A"], "dateTime": "2008-06-06T06:45:00.000-05:00"}, {"value": "1950", "qualifiers": ["A"], "dateTime": "2008-06-06T07:00:00.000-05:00"}, {"value": "1950", "qualifiers": ["A"], "dateTime": "2008-06-06T07:15:00.000-05:00"}, {"value": "1950", "qualifiers": ["A"], "dateTime": "2008-06-06T07:30:00.000-05:00"}, {"value": "1960", "qualifiers": ["A"], "dateTime": "2008-06-06T07:45:00.000-05:00"}, {"value": "1950", "qualifiers": ["A"], "dateTime": "2008-06-06T08:00:00.000-05:00"}, {"value": "1950", "qualifiers": ["A"], "dateTime": "2008-06-06T08:15:00.000-05:00"}, {"value": "1950", "qualifiers": ["A"], "dateTime": "2008-06-06T08:30:00.000-05:00"}, {"value": "1970", "qualifiers": ["A"], "dateTime": "2008-06-06T08:45:00.000-05:00"}, {"value": "1960", "qualifiers": ["A"], "dateTime": "2008-06-06T09:00:00.000-05:00"}, {"value": "1950", "qualifiers": ["A"], "dateTime": "2008-06-06T09:15:00.000-05:00"}, {"value": "1960", "qualifiers": ["A"], "dateTime": "2008-06-06T09:30:00.000-05:00"}, {"value": "1970", "qualifiers": ["A"], "dateTime": "2008-06-06T09:45:00.000-05:00"}, {"value": "1940", "qualifiers": ["A"], "dateTime": "2008-06-06T10:00:00.000-05:00"}, {"value": "1960", "qualifiers": ["A"], "dateTime": "2008-06-06T10:15:00.000-05:00"}, {"value": "1950", "qualifiers": ["A"], "dateTime": "2008-06-06T10:30:00.000-05:00"}, {"value": "1950", "qualifiers": ["A"], "dateTime": "2008-06-06T10:45:00.000-05:00"}, {"value": "1970", "qualifiers": ["A"], "dateTime": "2008-06-06T11:00:00.000-05:00"}, {"value": "1960", "qualifiers": ["A"], "dateTime": "2008-06-06T11:15:00.000-05:00"}, {"value": "1960", "qualifiers": ["A"], "dateTime": "2008-06-06T11:30:00.000-05:00"}, {"value": "1950", "qualifiers": ["A"], "dateTime": "2008-06-06T11:45:00.000-05:00"}, {"value": "1970", "qualifiers": ["A"], "dateTime": "2008-06-06T12:00:00.000-05:00"}, {"value": "1950", "qualifiers": ["A"], "dateTime": "2008-06-06T12:15:00.000-05:00"}, {"value": "1960", "qualifiers": ["A"], "dateTime": "2008-06-06T12:30:00.000-05:00"}, {"value": "1960", "qualifiers": ["A"], "dateTime": "2008-06-06T12:45:00.000-05:00"}, {"value": "1960", "qualifiers": ["A"], "dateTime": "2008-06-06T13:00:00.000-05:00"}, {"value": "1970", "qualifiers": ["A"], "dateTime": "2008-06-06T13:15:00.000-05:00"}, {"value": "1970", "qualifiers": ["A"], "dateTime": "2008-06-06T13:30:00.000-05:00"}, {"value": "1950", "qualifiers": ["A"], "dateTime": "2008-06-06T13:45:00.000-05:00"}, {"value": "1970", "qualifiers": ["A"], "dateTime": "2008-06-06T14:00:00.000-05:00"}, {"value": "1970", "qualifiers": ["A"], "dateTime": "2008-06-06T14:15:00.000-05:00"}, {"value": "1970", "qualifiers": ["A"], "dateTime": "2008-06-06T14:30:00.000-05:00"}, {"value": "1970", "qualifiers": ["A"], "dateTime": "2008-06-06T14:45:00.000-05:00"}, {"value": "1970", "qualifiers": ["A"], "dateTime": "2008-06-06T15:00:00.000-05:00"}, {"value": "1980", "qualifiers": ["A"], "dateTime": "2008-06-06T15:15:00.000-05:00"}, {"value": "1980", "qualifiers": ["A"], "dateTime": "2008-06-06T15:30:00.000-05:00"}, {"value": "1990", "qualifiers": ["A"], "dateTime": "2008-06-06T15:45:00.000-05:00"}, {"value": "1970", "qualifiers": ["A"], "dateTime": "2008-06-06T16:00:00.000-05:00"}, {"value": "1980", "qualifiers": ["A"], "dateTime": "2008-06-06T16:15:00.000-05:00"}, {"value": "1990", "qualifiers": ["A"], "dateTime": "2008-06-06T16:30:00.000-05:00"}, {"value": "1970", "qualifiers": ["A"], "dateTime": "2008-06-06T16:45:00.000-05:00"}, {"value": "1990", "qualifiers": ["A"], "dateTime": "2008-06-06T17:00:00.000-05:00"}, {"value": "1990", "qualifiers": ["A"], "dateTime": "2008-06-06T17:15:00.000-05:00"}, {"value": "1990", "qualifiers": ["A"], "dateTime": "2008-06-06T17:30:00.000-05:00"}, {"value": "2000", "qualifiers": ["A"], "dateTime": "2008-06-06T17:45:00.000-05:00"}, {"value": "1990", "qualifiers": ["A"], "dateTime": "2008-06-06T18:00:00.000-05:00"}, {"value": "2000", "qualifiers": ["A"], "dateTime": "2008-06-06T18:15:00.000-05:00"}, {"value": "2000", "qualifiers": ["A"], "dateTime": "2008-06-06T18:30:00.000-05:00"}, {"value": "2000", "qualifiers": ["A"], "dateTime": "2008-06-06T18:45:00.000-05:00"}, {"value": "2010", "qualifiers": ["A"], "dateTime": "2008-06-06T19:00:00.000-05:00"}, {"value": "2020", "qualifiers": ["A"], "dateTime": "2008-06-06T19:15:00.000-05:00"}, {"value": "2020", "qualifiers": ["A"], "dateTime": "2008-06-06T19:30:00.000-05:00"}, {"value": "2020", "qualifiers": ["A"], "dateTime": "2008-06-06T19:45:00.000-05:00"}, {"value": "2020", "qualifiers": ["A"], "dateTime": "2008-06-06T20:00:00.000-05:00"}, {"value": "2030", "qualifiers": ["A"], "dateTime": "2008-06-06T20:15:00.000-05:00"}, {"value": "2030", "qualifiers": ["A"], "dateTime": "2008-06-06T20:30:00.000-05:00"}, {"value": "2030", "qualifiers": ["A"], "dateTime": "2008-06-06T20:45:00.000-05:00"}, {"value": "2040", "qualifiers": ["A"], "dateTime": "2008-06-06T21:00:00.000-05:00"}, {"value": "2040", "qualifiers": ["A"], "dateTime": "2008-06-06T21:15:00.000-05:00"}, {"value": "2040", "qualifiers": ["A"], "dateTime": "2008-06-06T21:30:00.000-05:00"}, {"value": "2050", "qualifiers": ["A"], "dateTime": "2008-06-06T21:45:00.000-05:00"}, {"value": "2060", "qualifiers": ["A"], "dateTime": "2008-06-06T22:00:00.000-05:00"}, {"value": "2060", "qualifiers": ["A"], "dateTime": "2008-06-06T22:15:00.000-05:00"}, {"value": "2060", "qualifiers": ["A"], "dateTime": "2008-06-06T22:30:00.000-05:00"}, {"value": "2070", "qualifiers": ["A"], "dateTime": "2008-06-06T22:45:00.000-05:00"}, {"value": "2060", "qualifiers": ["A"], "dateTime": "2008-06-06T23:00:00.000-05:00"}, {"value": "2050", "qualifiers": ["A"], "dateTime": "2008-06-06T23:15:00.000-05:00"}, {"value": "2070", "qualifiers": ["A"], "dateTime": "2008-06-06T23:30:00.000-05:00"}, {"value": "2070", "qualifiers": ["A"], "dateTime": "2008-06-06T23:45:00.000-05:00"}, {"value": "2070", "qualifiers": ["A"], "dateTime": "2008-06-07T00:00:00.000-05:00"}, {"value": "2070", "qualifiers": ["A"], "dateTime": "2008-06-07T00:15:00.000-05:00"}, {"value": "2090", "qualifiers": ["A"], "dateTime": "2008-06-07T00:30:00.000-05:00"}, {"value": "2090", "qualifiers": ["A"], "dateTime": "2008-06-07T00:45:00.000-05:00"}, {"value": "2100", "qualifiers": ["A"], "dateTime": "2008-06-07T01:00:00.000-05:00"}, {"value": "2070", "qualifiers": ["A"], "dateTime": "2008-06-07T01:15:00.000-05:00"}, {"value": "2110", "qualifiers": ["A"], "dateTime": "2008-06-07T01:30:00.000-05:00"}, {"value": "2110", "qualifiers": ["A"], "dateTime": "2008-06-07T01:45:00.000-05:00"}, {"value": "2110", "qualifiers": ["A"], "dateTime": "2008-06-07T02:00:00.000-05:00"}, {"value": "2130", "qualifiers": ["A"], "dateTime": "2008-06-07T02:15:00.000-05:00"}, {"value": "2120", "qualifiers": ["A"], "dateTime": "2008-06-07T02:30:00.000-05:00"}, {"value": "2120", "qualifiers": ["A"], "dateTime": "2008-06-07T02:45:00.000-05:00"}, {"value": "2130", "qualifiers": ["A"], "dateTime": "2008-06-07T03:00:00.000-05:00"}, {"value": "2130", "qualifiers": ["A"], "dateTime": "2008-06-07T03:15:00.000-05:00"}, {"value": "2130", "qualifiers": ["A"], "dateTime": "2008-06-07T03:30:00.000-05:00"}, {"value": "2140", "qualifiers": ["A"], "dateTime": "2008-06-07T03:45:00.000-05:00"}, {"value": "2150", "qualifiers": ["A"], "dateTime": "2008-06-07T04:00:00.000-05:00"}, {"value": "2150", "qualifiers": ["A"], "dateTime": "2008-06-07T04:15:00.000-05:00"}, {"value": "2160", "qualifiers": ["A"], "dateTime": "2008-06-07T04:30:00.000-05:00"}, {"value": "2150", "qualifiers": ["A"], "dateTime": "2008-06-07T04:45:00.000-05:00"}, {"value": "2150", "qualifiers": ["A"], "dateTime": "2008-06-07T05:00:00.000-05:00"}, {"value": "2150", "qualifiers": ["A"], "dateTime": "2008-06-07T05:15:00.000-05:00"}, {"value": "2170", "qualifiers": ["A"], "dateTime": "2008-06-07T05:30:00.000-05:00"}, {"value": "2170", "qualifiers": ["A"], "dateTime": "2008-06-07T05:45:00.000-05:00"}, {"value": "2180", "qualifiers": ["A"], "dateTime": "2008-06-07T06:00:00.000-05:00"}, {"value": "2180", "qualifiers": ["A"], "dateTime": "2008-06-07T06:15:00.000-05:00"}, {"value": "2180", "qualifiers": ["A"], "dateTime": "2008-06-07T06:30:00.000-05:00"}, {"value": "2180", "qualifiers": ["A"], "dateTime": "2008-06-07T06:45:00.000-05:00"}, {"value": "2190", "qualifiers": ["A"], "dateTime": "2008-06-07T07:00:00.000-05:00"}, {"value": "2200", "qualifiers": ["A"], "dateTime": "2008-06-07T07:15:00.000-05:00"}, {"value": "2200", "qualifiers": ["A"], "dateTime": "2008-06-07T07:30:00.000-05:00"}, {"value": "2200", "qualifiers": ["A"], "dateTime": "2008-06-07T07:45:00.000-05:00"}, {"value": "2200", "qualifiers": ["A"], "dateTime": "2008-06-07T08:00:00.000-05:00"}, {"value": "2210", "qualifiers": ["A"], "dateTime": "2008-06-07T08:15:00.000-05:00"}, {"value": "2220", "qualifiers": ["A"], "dateTime": "2008-06-07T08:30:00.000-05:00"}, {"value": "2230", "qualifiers": ["A"], "dateTime": "2008-06-07T08:45:00.000-05:00"}, {"value": "2240", "qualifiers": ["A"], "dateTime": "2008-06-07T09:00:00.000-05:00"}, {"value": "2240", "qualifiers": ["A"], "dateTime": "2008-06-07T09:15:00.000-05:00"}, {"value": "2230", "qualifiers": ["A"], "dateTime": "2008-06-07T09:30:00.000-05:00"}, {"value": "2230", "qualifiers": ["A"], "dateTime": "2008-06-07T09:45:00.000-05:00"}, {"value": "2230", "qualifiers": ["A"], "dateTime": "2008-06-07T10:00:00.000-05:00"}, {"value": "2260", "qualifiers": ["A"], "dateTime": "2008-06-07T10:15:00.000-05:00"}, {"value": "2260", "qualifiers": ["A"], "dateTime": "2008-06-07T10:30:00.000-05:00"}, {"value": "2240", "qualifiers": ["A"], "dateTime": "2008-06-07T10:45:00.000-05:00"}, {"value": "2260", "qualifiers": ["A"], "dateTime": "2008-06-07T11:00:00.000-05:00"}, {"value": "2280", "qualifiers": ["A"], "dateTime": "2008-06-07T11:15:00.000-05:00"}, {"value": "2260", "qualifiers": ["A"], "dateTime": "2008-06-07T11:30:00.000-05:00"}, {"value": "2290", "qualifiers": ["A"], "dateTime": "2008-06-07T11:45:00.000-05:00"}, {"value": "2270", "qualifiers": ["A"], "dateTime": "2008-06-07T12:00:00.000-05:00"}, {"value": "2280", "qualifiers": ["A"], "dateTime": "2008-06-07T12:15:00.000-05:00"}, {"value": "2290", "qualifiers": ["A"], "dateTime": "2008-06-07T12:30:00.000-05:00"}, {"value": "2290", "qualifiers": ["A"], "dateTime": "2008-06-07T12:45:00.000-05:00"}, {"value": "2290", "qualifiers": ["A"], "dateTime": "2008-06-07T13:00:00.000-05:00"}, {"value": "2310", "qualifiers": ["A"], "dateTime": "2008-06-07T13:15:00.000-05:00"}, {"value": "2300", "qualifiers": ["A"], "dateTime": "2008-06-07T13:30:00.000-05:00"}, {"value": "2320", "qualifiers": ["A"], "dateTime": "2008-06-07T13:45:00.000-05:00"}, {"value": "2320", "qualifiers": ["A"], "dateTime": "2008-06-07T14:00:00.000-05:00"}, {"value": "2320", "qualifiers": ["A"], "dateTime": "2008-06-07T14:15:00.000-05:00"}, {"value": "2340", "qualifiers": ["A"], "dateTime": "2008-06-07T14:30:00.000-05:00"}, {"value": "2340", "qualifiers": ["A"], "dateTime": "2008-06-07T14:45:00.000-05:00"}, {"value": "2330", "qualifiers": ["A"], "dateTime": "2008-06-07T15:00:00.000-05:00"}, {"value": "2340", "qualifiers": ["A"], "dateTime": "2008-06-07T15:15:00.000-05:00"}, {"value": "2350", "qualifiers": ["A"], "dateTime": "2008-06-07T15:30:00.000-05:00"}, {"value": "2350", "qualifiers": ["A"], "dateTime": "2008-06-07T15:45:00.000-05:00"}, {"value": "2370", "qualifiers": ["A"], "dateTime": "2008-06-07T16:00:00.000-05:00"}, {"value": "2380", "qualifiers": ["A"], "dateTime": "2008-06-07T16:15:00.000-05:00"}, {"value": "2380", "qualifiers": ["A"], "dateTime": "2008-06-07T16:30:00.000-05:00"}, {"value": "2380", "qualifiers": ["A"], "dateTime": "2008-06-07T16:45:00.000-05:00"}, {"value": "2400", "qualifiers": ["A"], "dateTime": "2008-06-07T17:00:00.000-05:00"}, {"value": "2410", "qualifiers": ["A"], "dateTime": "2008-06-07T17:15:00.000-05:00"}, {"value": "2410", "qualifiers": ["A"], "dateTime": "2008-06-07T17:30:00.000-05:00"}, {"value": "2410", "qualifiers": ["A"], "dateTime": "2008-06-07T17:45:00.000-05:00"}, {"value": "2420", "qualifiers": ["A"], "dateTime": "2008-06-07T18:00:00.000-05:00"}, {"value": "2400", "qualifiers": ["A"], "dateTime": "2008-06-07T18:15:00.000-05:00"}, {"value": "2400", "qualifiers": ["A"], "dateTime": "2008-06-07T18:30:00.000-05:00"}, {"value": "2410", "qualifiers": ["A"], "dateTime": "2008-06-07T18:45:00.000-05:00"}, {"value": "2430", "qualifiers": ["A"], "dateTime": "2008-06-07T19:00:00.000-05:00"}, {"value": "2420", "qualifiers": ["A"], "dateTime": "2008-06-07T19:15:00.000-05:00"}, {"value": "2430", "qualifiers": ["A"], "dateTime": "2008-06-07T19:30:00.000-05:00"}, {"value": "2430", "qualifiers": ["A"], "dateTime": "2008-06-07T19:45:00.000-05:00"}, {"value": "2430", "qualifiers": ["A"], "dateTime": "2008-06-07T20:00:00.000-05:00"}, {"value": "2430", "qualifiers": ["A"], "dateTime": "2008-06-07T20:15:00.000-05:00"}, {"value": "2450", "qualifiers": ["A"], "dateTime": "2008-06-07T20:30:00.000-05:00"}, {"value": "2450", "qualifiers": ["A"], "dateTime": "2008-06-07T20:45:00.000-05:00"}, {"value": "2450", "qualifiers": ["A"], "dateTime": "2008-06-07T21:00:00.000-05:00"}, {"value": "2460", "qualifiers": ["A"], "dateTime": "2008-06-07T21:15:00.000-05:00"}, {"value": "2470", "qualifiers": ["A"], "dateTime": "2008-06-07T21:30:00.000-05:00"}, {"value": "2450", "qualifiers": ["A"], "dateTime": "2008-06-07T21:45:00.000-05:00"}, {"value": "2480", "qualifiers": ["A"], "dateTime": "2008-06-07T22:00:00.000-05:00"}, {"value": "2510", "qualifiers": ["A"], "dateTime": "2008-06-07T22:15:00.000-05:00"}, {"value": "2510", "qualifiers": ["A"], "dateTime": "2008-06-07T22:30:00.000-05:00"}, {"value": "2500", "qualifiers": ["A"], "dateTime": "2008-06-07T22:45:00.000-05:00"}, {"value": "2530", "qualifiers": ["A"], "dateTime": "2008-06-07T23:00:00.000-05:00"}, {"value": "2520", "qualifiers": ["A"], "dateTime": "2008-06-07T23:15:00.000-05:00"}, {"value": "2580", "qualifiers": ["A"], "dateTime": "2008-06-07T23:30:00.000-05:00"}, {"value": "2580", "qualifiers": ["A"], "dateTime": "2008-06-07T23:45:00.000-05:00"}, {"value": "2600", "qualifiers": ["A"], "dateTime": "2008-06-08T00:00:00.000-05:00"}, {"value": "2600", "qualifiers": ["A"], "dateTime": "2008-06-08T00:15:00.000-05:00"}, {"value": "2650", "qualifiers": ["A"], "dateTime": "2008-06-08T00:30:00.000-05:00"}, {"value": "2630", "qualifiers": ["A"], "dateTime": "2008-06-08T00:45:00.000-05:00"}, {"value": "2670", "qualifiers": ["A"], "dateTime": "2008-06-08T01:00:00.000-05:00"}, {"value": "2660", "qualifiers": ["A"], "dateTime": "2008-06-08T01:15:00.000-05:00"}, {"value": "2670", "qualifiers": ["A"], "dateTime": "2008-06-08T01:30:00.000-05:00"}, {"value": "2660", "qualifiers": ["A"], "dateTime": "2008-06-08T01:45:00.000-05:00"}, {"value": "2660", "qualifiers": ["A"], "dateTime": "2008-06-08T02:00:00.000-05:00"}, {"value": "2670", "qualifiers": ["A"], "dateTime": "2008-06-08T02:15:00.000-05:00"}, {"value": "2680", "qualifiers": ["A"], "dateTime": "2008-06-08T02:30:00.000-05:00"}, {"value": "2690", "qualifiers": ["A"], "dateTime": "2008-06-08T02:45:00.000-05:00"}, {"value": "2670", "qualifiers": ["A"], "dateTime": "2008-06-08T03:00:00.000-05:00"}, {"value": "2680", "qualifiers": ["A"], "dateTime": "2008-06-08T03:15:00.000-05:00"}, {"value": "2700", "qualifiers": ["A"], "dateTime": "2008-06-08T03:30:00.000-05:00"}, {"value": "2710", "qualifiers": ["A"], "dateTime": "2008-06-08T03:45:00.000-05:00"}, {"value": "2720", "qualifiers": ["A"], "dateTime": "2008-06-08T04:00:00.000-05:00"}, {"value": "2700", "qualifiers": ["A"], "dateTime": "2008-06-08T04:15:00.000-05:00"}, {"value": "2710", "qualifiers": ["A"], "dateTime": "2008-06-08T04:30:00.000-05:00"}, {"value": "2720", "qualifiers": ["A"], "dateTime": "2008-06-08T04:45:00.000-05:00"}, {"value": "2730", "qualifiers": ["A"], "dateTime": "2008-06-08T05:00:00.000-05:00"}, {"value": "2730", "qualifiers": ["A"], "dateTime": "2008-06-08T05:15:00.000-05:00"}, {"value": "2760", "qualifiers": ["A"], "dateTime": "2008-06-08T05:30:00.000-05:00"}, {"value": "2750", "qualifiers": ["A"], "dateTime": "2008-06-08T05:45:00.000-05:00"}, {"value": "2740", "qualifiers": ["A"], "dateTime": "2008-06-08T06:00:00.000-05:00"}, {"value": "2730", "qualifiers": ["A"], "dateTime": "2008-06-08T06:15:00.000-05:00"}, {"value": "2760", "qualifiers": ["A"], "dateTime": "2008-06-08T06:30:00.000-05:00"}, {"value": "2760", "qualifiers": ["A"], "dateTime": "2008-06-08T06:45:00.000-05:00"}, {"value": "2760", "qualifiers": ["A"], "dateTime": "2008-06-08T07:00:00.000-05:00"}, {"value": "2800", "qualifiers": ["A"], "dateTime": "2008-06-08T07:15:00.000-05:00"}, {"value": "2860", "qualifiers": ["A"], "dateTime": "2008-06-08T07:30:00.000-05:00"}, {"value": "2870", "qualifiers": ["A"], "dateTime": "2008-06-08T07:45:00.000-05:00"}, {"value": "2910", "qualifiers": ["A"], "dateTime": "2008-06-08T08:00:00.000-05:00"}, {"value": "3000", "qualifiers": ["A"], "dateTime": "2008-06-08T08:15:00.000-05:00"}, {"value": "3030", "qualifiers": ["A"], "dateTime": "2008-06-08T08:30:00.000-05:00"}, {"value": "3040", "qualifiers": ["A"], "dateTime": "2008-06-08T08:45:00.000-05:00"}, {"value": "3110", "qualifiers": ["A"], "dateTime": "2008-06-08T09:00:00.000-05:00"}, {"value": "3160", "qualifiers": ["A"], "dateTime": "2008-06-08T09:15:00.000-05:00"}, {"value": "3140", "qualifiers": ["A"], "dateTime": "2008-06-08T09:30:00.000-05:00"}, {"value": "3200", "qualifiers": ["A"], "dateTime": "2008-06-08T09:45:00.000-05:00"}, {"value": "3260", "qualifiers": ["A"], "dateTime": "2008-06-08T10:00:00.000-05:00"}, {"value": "3230", "qualifiers": ["A"], "dateTime": "2008-06-08T10:15:00.000-05:00"}, {"value": "3320", "qualifiers": ["A"], "dateTime": "2008-06-08T10:30:00.000-05:00"}, {"value": "3330", "qualifiers": ["A"], "dateTime": "2008-06-08T10:45:00.000-05:00"}, {"value": "3360", "qualifiers": ["A"], "dateTime": "2008-06-08T11:00:00.000-05:00"}, {"value": "3390", "qualifiers": ["A"], "dateTime": "2008-06-08T11:15:00.000-05:00"}, {"value": "3480", "qualifiers": ["A"], "dateTime": "2008-06-08T11:30:00.000-05:00"}, {"value": "3540", "qualifiers": ["A"], "dateTime": "2008-06-08T11:45:00.000-05:00"}, {"value": "3650", "qualifiers": ["A"], "dateTime": "2008-06-08T12:00:00.000-05:00"}, {"value": "3740", "qualifiers": ["A"], "dateTime": "2008-06-08T12:15:00.000-05:00"}, {"value": "3820", "qualifiers": ["A"], "dateTime": "2008-06-08T12:30:00.000-05:00"}, {"value": "3950", "qualifiers": ["A"], "dateTime": "2008-06-08T12:45:00.000-05:00"}, {"value": "4050", "qualifiers": ["A"], "dateTime": "2008-06-08T13:00:00.000-05:00"}, {"value": "4150", "qualifiers": ["A"], "dateTime": "2008-06-08T13:15:00.000-05:00"}, {"value": "4230", "qualifiers": ["A"], "dateTime": "2008-06-08T13:30:00.000-05:00"}, {"value": "4440", "qualifiers": ["A"], "dateTime": "2008-06-08T13:45:00.000-05:00"}, {"value": "4500", "qualifiers": ["A"], "dateTime": "2008-06-08T14:00:00.000-05:00"}, {"value": "4640", "qualifiers": ["A"], "dateTime": "2008-06-08T14:15:00.000-05:00"}, {"value": "4720", "qualifiers": ["A"], "dateTime": "2008-06-08T14:30:00.000-05:00"}, {"value": "4770", "qualifiers": ["A"], "dateTime": "2008-06-08T14:45:00.000-05:00"}, {"value": "4770", "qualifiers": ["A"], "dateTime": "2008-06-08T15:00:00.000-05:00"}, {"value": "4820", "qualifiers": ["A"], "dateTime": "2008-06-08T15:15:00.000-05:00"}, {"value": "4840", "qualifiers": ["A"], "dateTime": "2008-06-08T15:30:00.000-05:00"}, {"value": "4950", "qualifiers": ["A"], "dateTime": "2008-06-08T15:45:00.000-05:00"}, {"value": "4950", "qualifiers": ["A"], "dateTime": "2008-06-08T16:00:00.000-05:00"}, {"value": "4910", "qualifiers": ["A"], "dateTime": "2008-06-08T16:15:00.000-05:00"}, {"value": "5050", "qualifiers": ["A"], "dateTime": "2008-06-08T16:30:00.000-05:00"}, {"value": "5020", "qualifiers": ["A"], "dateTime": "2008-06-08T16:45:00.000-05:00"}, {"value": "5070", "qualifiers": ["A"], "dateTime": "2008-06-08T17:00:00.000-05:00"}, {"value": "5100", "qualifiers": ["A"], "dateTime": "2008-06-08T17:15:00.000-05:00"}, {"value": "5170", "qualifiers": ["A"], "dateTime": "2008-06-08T17:30:00.000-05:00"}, {"value": "5310", "qualifiers": ["A"], "dateTime": "2008-06-08T17:45:00.000-05:00"}, {"value": "5460", "qualifiers": ["A"], "dateTime": "2008-06-08T18:00:00.000-05:00"}, {"value": "5460", "qualifiers": ["A"], "dateTime": "2008-06-08T18:15:00.000-05:00"}, {"value": "5570", "qualifiers": ["A"], "dateTime": "2008-06-08T18:30:00.000-05:00"}, {"value": "5770", "qualifiers": ["A"], "dateTime": "2008-06-08T18:45:00.000-05:00"}, {"value": "5930", "qualifiers": ["A"], "dateTime": "2008-06-08T19:00:00.000-05:00"}, {"value": "6210", "qualifiers": ["A"], "dateTime": "2008-06-08T19:15:00.000-05:00"}, {"value": "6320", "qualifiers": ["A"], "dateTime": "2008-06-08T19:30:00.000-05:00"}, {"value": "6760", "qualifiers": ["A"], "dateTime": "2008-06-08T19:45:00.000-05:00"}, {"value": "7160", "qualifiers": ["A"], "dateTime": "2008-06-08T20:00:00.000-05:00"}, {"value": "7390", "qualifiers": ["A"], "dateTime": "2008-06-08T20:15:00.000-05:00"}, {"value": "7680", "qualifiers": ["A"], "dateTime": "2008-06-08T20:30:00.000-05:00"}, {"value": "8270", "qualifiers": ["A"], "dateTime": "2008-06-08T20:45:00.000-05:00"}, {"value": "8520", "qualifiers": ["A"], "dateTime": "2008-06-08T21:00:00.000-05:00"}, {"value": "8920", "qualifiers": ["A"], "dateTime": "2008-06-08T21:15:00.000-05:00"}, {"value": "9430", "qualifiers": ["A"], "dateTime": "2008-06-08T21:30:00.000-05:00"}, {"value": "9710", "qualifiers": ["A"], "dateTime": "2008-06-08T21:45:00.000-05:00"}, {"value": "10000", "qualifiers": ["A"], "dateTime": "2008-06-08T22:00:00.000-05:00"}, {"value": "10200", "qualifiers": ["A"], "dateTime": "2008-06-08T22:15:00.000-05:00"}, {"value": "10500", "qualifiers": ["A"], "dateTime": "2008-06-08T22:30:00.000-05:00"}, {"value": "10700", "qualifiers": ["A"], "dateTime": "2008-06-08T22:45:00.000-05:00"}, {"value": "10900", "qualifiers": ["A"], "dateTime": "2008-06-08T23:00:00.000-05:00"}, {"value": "11100", "qualifiers": ["A"], "dateTime": "2008-06-08T23:15:00.000-05:00"}, {"value": "11200", "qualifiers": ["A"], "dateTime": "2008-06-08T23:30:00.000-05:00"}, {"value": "11400", "qualifiers": ["A"], "dateTime": "2008-06-08T23:45:00.000-05:00"}, {"value": "11300", "qualifiers": ["A"], "dateTime": "2008-06-09T00:00:00.000-05:00"}, {"value": "11300", "qualifiers": ["A"], "dateTime": "2008-06-09T00:15:00.000-05:00"}, {"value": "11600", "qualifiers": ["A"], "dateTime": "2008-06-09T00:30:00.000-05:00"}, {"value": "11500", "qualifiers": ["A"], "dateTime": "2008-06-09T00:45:00.000-05:00"}, {"value": "11500", "qualifiers": ["A"], "dateTime": "2008-06-09T01:00:00.000-05:00"}, {"value": "11700", "qualifiers": ["A"], "dateTime": "2008-06-09T01:15:00.000-05:00"}, {"value": "11600", "qualifiers": ["A"], "dateTime": "2008-06-09T01:30:00.000-05:00"}, {"value": "11600", "qualifiers": ["A"], "dateTime": "2008-06-09T01:45:00.000-05:00"}, {"value": "11500", "qualifiers": ["A"], "dateTime": "2008-06-09T02:00:00.000-05:00"}, {"value": "11600", "qualifiers": ["A"], "dateTime": "2008-06-09T02:15:00.000-05:00"}, {"value": "11400", "qualifiers": ["A"], "dateTime": "2008-06-09T02:30:00.000-05:00"}, {"value": "11400", "qualifiers": ["A"], "dateTime": "2008-06-09T02:45:00.000-05:00"}, {"value": "11400", "qualifiers": ["A"], "dateTime": "2008-06-09T03:00:00.000-05:00"}, {"value": "11300", "qualifiers": ["A"], "dateTime": "2008-06-09T03:15:00.000-05:00"}, {"value": "11300", "qualifiers": ["A"], "dateTime": "2008-06-09T03:30:00.000-05:00"}, {"value": "11000", "qualifiers": ["A"], "dateTime": "2008-06-09T03:45:00.000-05:00"}, {"value": "11100", "qualifiers": ["A"], "dateTime": "2008-06-09T04:00:00.000-05:00"}, {"value": "11000", "qualifiers": ["A"], "dateTime": "2008-06-09T04:15:00.000-05:00"}, {"value": "10800", "qualifiers": ["A"], "dateTime": "2008-06-09T04:30:00.000-05:00"}, {"value": "10800", "qualifiers": ["A"], "dateTime": "2008-06-09T04:45:00.000-05:00"}, {"value": "10800", "qualifiers": ["A"], "dateTime": "2008-06-09T05:00:00.000-05:00"}, {"value": "10600", "qualifiers": ["A"], "dateTime": "2008-06-09T05:15:00.000-05:00"}, {"value": "10700", "qualifiers": ["A"], "dateTime": "2008-06-09T05:30:00.000-05:00"}, {"value": "10500", "qualifiers": ["A"], "dateTime": "2008-06-09T05:45:00.000-05:00"}, {"value": "10300", "qualifiers": ["A"], "dateTime": "2008-06-09T06:00:00.000-05:00"}, {"value": "10300", "qualifiers": ["A"], "dateTime": "2008-06-09T06:15:00.000-05:00"}, {"value": "10200", "qualifiers": ["A"], "dateTime": "2008-06-09T06:30:00.000-05:00"}, {"value": "10200", "qualifiers": ["A"], "dateTime": "2008-06-09T06:45:00.000-05:00"}, {"value": "10100", "qualifiers": ["A"], "dateTime": "2008-06-09T07:00:00.000-05:00"}, {"value": "9990", "qualifiers": ["A"], "dateTime": "2008-06-09T07:15:00.000-05:00"}, {"value": "9870", "qualifiers": ["A"], "dateTime": "2008-06-09T07:30:00.000-05:00"}, {"value": "9870", "qualifiers": ["A"], "dateTime": "2008-06-09T07:45:00.000-05:00"}, {"value": "9590", "qualifiers": ["A"], "dateTime": "2008-06-09T08:00:00.000-05:00"}, {"value": "9630", "qualifiers": ["A"], "dateTime": "2008-06-09T08:15:00.000-05:00"}, {"value": "9590", "qualifiers": ["A"], "dateTime": "2008-06-09T08:30:00.000-05:00"}, {"value": "9430", "qualifiers": ["A"], "dateTime": "2008-06-09T08:45:00.000-05:00"}, {"value": "9510", "qualifiers": ["A"], "dateTime": "2008-06-09T09:00:00.000-05:00"}, {"value": "9310", "qualifiers": ["A"], "dateTime": "2008-06-09T09:15:00.000-05:00"}, {"value": "9310", "qualifiers": ["A"], "dateTime": "2008-06-09T09:30:00.000-05:00"}, {"value": "9230", "qualifiers": ["A"], "dateTime": "2008-06-09T09:45:00.000-05:00"}, {"value": "9040", "qualifiers": ["A"], "dateTime": "2008-06-09T10:00:00.000-05:00"}, {"value": "9040", "qualifiers": ["A"], "dateTime": "2008-06-09T10:15:00.000-05:00"}, {"value": "8960", "qualifiers": ["A"], "dateTime": "2008-06-09T10:30:00.000-05:00"}, {"value": "8850", "qualifiers": ["A"], "dateTime": "2008-06-09T10:45:00.000-05:00"}, {"value": "8810", "qualifiers": ["A"], "dateTime": "2008-06-09T11:00:00.000-05:00"}, {"value": "8850", "qualifiers": ["A"], "dateTime": "2008-06-09T11:15:00.000-05:00"}, {"value": "8740", "qualifiers": ["A"], "dateTime": "2008-06-09T11:30:00.000-05:00"}, {"value": "8660", "qualifiers": ["A"], "dateTime": "2008-06-09T11:45:00.000-05:00"}, {"value": "8450", "qualifiers": ["A"], "dateTime": "2008-06-09T12:00:00.000-05:00"}, {"value": "8480", "qualifiers": ["A"], "dateTime": "2008-06-09T12:15:00.000-05:00"}, {"value": "8450", "qualifiers": ["A"], "dateTime": "2008-06-09T12:30:00.000-05:00"}, {"value": "8300", "qualifiers": ["A"], "dateTime": "2008-06-09T12:45:00.000-05:00"}, {"value": "8300", "qualifiers": ["A"], "dateTime": "2008-06-09T13:00:00.000-05:00"}, {"value": "8160", "qualifiers": ["A"], "dateTime": "2008-06-09T13:15:00.000-05:00"}, {"value": "8230", "qualifiers": ["A"], "dateTime": "2008-06-09T13:30:00.000-05:00"}, {"value": "8090", "qualifiers": ["A"], "dateTime": "2008-06-09T13:45:00.000-05:00"}, {"value": "8090", "qualifiers": ["A"], "dateTime": "2008-06-09T14:00:00.000-05:00"}, {"value": "7990", "qualifiers": ["A"], "dateTime": "2008-06-09T14:15:00.000-05:00"}, {"value": "7850", "qualifiers": ["A"], "dateTime": "2008-06-09T14:30:00.000-05:00"}, {"value": "7820", "qualifiers": ["A"], "dateTime": "2008-06-09T14:45:00.000-05:00"}, {"value": "7780", "qualifiers": ["A"], "dateTime": "2008-06-09T15:00:00.000-05:00"}, {"value": "7850", "qualifiers": ["A"], "dateTime": "2008-06-09T15:15:00.000-05:00"}, {"value": "7620", "qualifiers": ["A"], "dateTime": "2008-06-09T15:30:00.000-05:00"}, {"value": "7650", "qualifiers": ["A"], "dateTime": "2008-06-09T15:45:00.000-05:00"}, {"value": "7580", "qualifiers": ["A"], "dateTime": "2008-06-09T16:00:00.000-05:00"}, {"value": "7450", "qualifiers": ["A"], "dateTime": "2008-06-09T16:15:00.000-05:00"}, {"value": "7480", "qualifiers": ["A"], "dateTime": "2008-06-09T16:30:00.000-05:00"}, {"value": "7450", "qualifiers": ["A"], "dateTime": "2008-06-09T16:45:00.000-05:00"}, {"value": "7350", "qualifiers": ["A"], "dateTime": "2008-06-09T17:00:00.000-05:00"}, {"value": "7260", "qualifiers": ["A"], "dateTime": "2008-06-09T17:15:00.000-05:00"}, {"value": "7260", "qualifiers": ["A"], "dateTime": "2008-06-09T17:30:00.000-05:00"}, {"value": "7260", "qualifiers": ["A"], "dateTime": "2008-06-09T17:45:00.000-05:00"}, {"value": "7130", "qualifiers": ["A"], "dateTime": "2008-06-09T18:00:00.000-05:00"}, {"value": "6980", "qualifiers": ["A"], "dateTime": "2008-06-09T18:15:00.000-05:00"}, {"value": "7010", "qualifiers": ["A"], "dateTime": "2008-06-09T18:30:00.000-05:00"}, {"value": "7010", "qualifiers": ["A"], "dateTime": "2008-06-09T18:45:00.000-05:00"}, {"value": "6950", "qualifiers": ["A"], "dateTime": "2008-06-09T19:00:00.000-05:00"}, {"value": "6820", "qualifiers": ["A"], "dateTime": "2008-06-09T19:15:00.000-05:00"}, {"value": "6760", "qualifiers": ["A"], "dateTime": "2008-06-09T19:30:00.000-05:00"}, {"value": "6730", "qualifiers": ["A"], "dateTime": "2008-06-09T19:45:00.000-05:00"}, {"value": "6670", "qualifiers": ["A"], "dateTime": "2008-06-09T20:00:00.000-05:00"}, {"value": "6610", "qualifiers": ["A"], "dateTime": "2008-06-09T20:15:00.000-05:00"}, {"value": "6610", "qualifiers": ["A"], "dateTime": "2008-06-09T20:30:00.000-05:00"}, {"value": "6640", "qualifiers": ["A"], "dateTime": "2008-06-09T20:45:00.000-05:00"}, {"value": "6470", "qualifiers": ["A"], "dateTime": "2008-06-09T21:00:00.000-05:00"}, {"value": "6550", "qualifiers": ["A"], "dateTime": "2008-06-09T21:15:00.000-05:00"}, {"value": "6500", "qualifiers": ["A"], "dateTime": "2008-06-09T21:30:00.000-05:00"}, {"value": "6350", "qualifiers": ["A"], "dateTime": "2008-06-09T21:45:00.000-05:00"}, {"value": "6410", "qualifiers": ["A"], "dateTime": "2008-06-09T22:00:00.000-05:00"}, {"value": "6270", "qualifiers": ["A"], "dateTime": "2008-06-09T22:15:00.000-05:00"}, {"value": "6300", "qualifiers": ["A"], "dateTime": "2008-06-09T22:30:00.000-05:00"}, {"value": "6180", "qualifiers": ["A"], "dateTime": "2008-06-09T22:45:00.000-05:00"}, {"value": "6210", "qualifiers": ["A"], "dateTime": "2008-06-09T23:00:00.000-05:00"}, {"value": "6210", "qualifiers": ["A"], "dateTime": "2008-06-09T23:15:00.000-05:00"}, {"value": "6130", "qualifiers": ["A"], "dateTime": "2008-06-09T23:30:00.000-05:00"}, {"value": "6020", "qualifiers": ["A"], "dateTime": "2008-06-09T23:45:00.000-05:00"}, {"value": "6020", "qualifiers": ["A"], "dateTime": "2008-06-10T00:00:00.000-05:00"}, {"value": "5910", "qualifiers": ["A"], "dateTime": "2008-06-10T00:15:00.000-05:00"}, {"value": "6020", "qualifiers": ["A"], "dateTime": "2008-06-10T00:30:00.000-05:00"}, {"value": "5910", "qualifiers": ["A"], "dateTime": "2008-06-10T00:45:00.000-05:00"}, {"value": "5850", "qualifiers": ["A"], "dateTime": "2008-06-10T01:00:00.000-05:00"}, {"value": "5800", "qualifiers": ["A"], "dateTime": "2008-06-10T01:15:00.000-05:00"}, {"value": "5850", "qualifiers": ["A"], "dateTime": "2008-06-10T01:30:00.000-05:00"}, {"value": "5770", "qualifiers": ["A"], "dateTime": "2008-06-10T01:45:00.000-05:00"}, {"value": "5720", "qualifiers": ["A"], "dateTime": "2008-06-10T02:00:00.000-05:00"}, {"value": "5700", "qualifiers": ["A"], "dateTime": "2008-06-10T02:15:00.000-05:00"}, {"value": "5700", "qualifiers": ["A"], "dateTime": "2008-06-10T02:30:00.000-05:00"}, {"value": "5670", "qualifiers": ["A"], "dateTime": "2008-06-10T02:45:00.000-05:00"}, {"value": "5510", "qualifiers": ["A"], "dateTime": "2008-06-10T03:00:00.000-05:00"}, {"value": "5570", "qualifiers": ["A"], "dateTime": "2008-06-10T03:15:00.000-05:00"}, {"value": "5510", "qualifiers": ["A"], "dateTime": "2008-06-10T03:30:00.000-05:00"}, {"value": "5540", "qualifiers": ["A"], "dateTime": "2008-06-10T03:45:00.000-05:00"}, {"value": "5440", "qualifiers": ["A"], "dateTime": "2008-06-10T04:00:00.000-05:00"}, {"value": "5490", "qualifiers": ["A"], "dateTime": "2008-06-10T04:15:00.000-05:00"}, {"value": "5440", "qualifiers": ["A"], "dateTime": "2008-06-10T04:30:00.000-05:00"}, {"value": "5360", "qualifiers": ["A"], "dateTime": "2008-06-10T04:45:00.000-05:00"}, {"value": "5360", "qualifiers": ["A"], "dateTime": "2008-06-10T05:00:00.000-05:00"}, {"value": "5240", "qualifiers": ["A"], "dateTime": "2008-06-10T05:15:00.000-05:00"}, {"value": "5220", "qualifiers": ["A"], "dateTime": "2008-06-10T05:30:00.000-05:00"}, {"value": "5260", "qualifiers": ["A"], "dateTime": "2008-06-10T05:45:00.000-05:00"}, {"value": "5190", "qualifiers": ["A"], "dateTime": "2008-06-10T06:00:00.000-05:00"}, {"value": "5120", "qualifiers": ["A"], "dateTime": "2008-06-10T06:15:00.000-05:00"}, {"value": "5140", "qualifiers": ["A"], "dateTime": "2008-06-10T06:30:00.000-05:00"}, {"value": "5050", "qualifiers": ["A"], "dateTime": "2008-06-10T06:45:00.000-05:00"}, {"value": "5100", "qualifiers": ["A"], "dateTime": "2008-06-10T07:00:00.000-05:00"}, {"value": "5070", "qualifiers": ["A"], "dateTime": "2008-06-10T07:15:00.000-05:00"}, {"value": "5070", "qualifiers": ["A"], "dateTime": "2008-06-10T07:30:00.000-05:00"}, {"value": "4930", "qualifiers": ["A"], "dateTime": "2008-06-10T07:45:00.000-05:00"}, {"value": "4950", "qualifiers": ["A"], "dateTime": "2008-06-10T08:00:00.000-05:00"}, {"value": "4860", "qualifiers": ["A"], "dateTime": "2008-06-10T08:15:00.000-05:00"}, {"value": "4910", "qualifiers": ["A"], "dateTime": "2008-06-10T08:30:00.000-05:00"}, {"value": "4840", "qualifiers": ["A"], "dateTime": "2008-06-10T08:45:00.000-05:00"}, {"value": "4750", "qualifiers": ["A"], "dateTime": "2008-06-10T09:00:00.000-05:00"}, {"value": "4750", "qualifiers": ["A"], "dateTime": "2008-06-10T09:15:00.000-05:00"}, {"value": "4750", "qualifiers": ["A"], "dateTime": "2008-06-10T09:30:00.000-05:00"}, {"value": "4680", "qualifiers": ["A"], "dateTime": "2008-06-10T09:45:00.000-05:00"}, {"value": "4660", "qualifiers": ["A"], "dateTime": "2008-06-10T10:00:00.000-05:00"}, {"value": "4640", "qualifiers": ["A"], "dateTime": "2008-06-10T10:15:00.000-05:00"}, {"value": "4640", "qualifiers": ["A"], "dateTime": "2008-06-10T10:30:00.000-05:00"}, {"value": "4530", "qualifiers": ["A"], "dateTime": "2008-06-10T10:45:00.000-05:00"}, {"value": "4590", "qualifiers": ["A"], "dateTime": "2008-06-10T11:00:00.000-05:00"}, {"value": "4480", "qualifiers": ["A"], "dateTime": "2008-06-10T11:15:00.000-05:00"}, {"value": "4460", "qualifiers": ["A"], "dateTime": "2008-06-10T11:30:00.000-05:00"}, {"value": "4460", "qualifiers": ["A"], "dateTime": "2008-06-10T11:45:00.000-05:00"}, {"value": "4360", "qualifiers": ["A"], "dateTime": "2008-06-10T12:00:00.000-05:00"}, {"value": "4400", "qualifiers": ["A"], "dateTime": "2008-06-10T12:15:00.000-05:00"}, {"value": "4400", "qualifiers": ["A"], "dateTime": "2008-06-10T12:30:00.000-05:00"}, {"value": "4360", "qualifiers": ["A"], "dateTime": "2008-06-10T12:45:00.000-05:00"}, {"value": "4330", "qualifiers": ["A"], "dateTime": "2008-06-10T13:00:00.000-05:00"}, {"value": "4230", "qualifiers": ["A"], "dateTime": "2008-06-10T13:15:00.000-05:00"}, {"value": "4170", "qualifiers": ["A"], "dateTime": "2008-06-10T13:30:00.000-05:00"}, {"value": "4170", "qualifiers": ["A"], "dateTime": "2008-06-10T13:45:00.000-05:00"}, {"value": "4170", "qualifiers": ["A"], "dateTime": "2008-06-10T14:00:00.000-05:00"}, {"value": "4090", "qualifiers": ["A"], "dateTime": "2008-06-10T14:15:00.000-05:00"}, {"value": "4050", "qualifiers": ["A"], "dateTime": "2008-06-10T14:30:00.000-05:00"}, {"value": "4070", "qualifiers": ["A"], "dateTime": "2008-06-10T14:45:00.000-05:00"}, {"value": "4030", "qualifiers": ["A"], "dateTime": "2008-06-10T15:00:00.000-05:00"}, {"value": "4030", "qualifiers": ["A"], "dateTime": "2008-06-10T15:15:00.000-05:00"}, {"value": "3930", "qualifiers": ["A"], "dateTime": "2008-06-10T15:30:00.000-05:00"}, {"value": "3890", "qualifiers": ["A"], "dateTime": "2008-06-10T15:45:00.000-05:00"}, {"value": "3910", "qualifiers": ["A"], "dateTime": "2008-06-10T16:00:00.000-05:00"}, {"value": "3890", "qualifiers": ["A"], "dateTime": "2008-06-10T16:15:00.000-05:00"}, {"value": "3840", "qualifiers": ["A"], "dateTime": "2008-06-10T16:30:00.000-05:00"}, {"value": "3840", "qualifiers": ["A"], "dateTime": "2008-06-10T16:45:00.000-05:00"}, {"value": "3750", "qualifiers": ["A"], "dateTime": "2008-06-10T17:00:00.000-05:00"}, {"value": "3690", "qualifiers": ["A"], "dateTime": "2008-06-10T17:15:00.000-05:00"}, {"value": "3740", "qualifiers": ["A"], "dateTime": "2008-06-10T17:30:00.000-05:00"}, {"value": "3690", "qualifiers": ["A"], "dateTime": "2008-06-10T17:45:00.000-05:00"}, {"value": "3670", "qualifiers": ["A"], "dateTime": "2008-06-10T18:00:00.000-05:00"}, {"value": "3620", "qualifiers": ["A"], "dateTime": "2008-06-10T18:15:00.000-05:00"}, {"value": "3540", "qualifiers": ["A"], "dateTime": "2008-06-10T18:30:00.000-05:00"}, {"value": "3540", "qualifiers": ["A"], "dateTime": "2008-06-10T18:45:00.000-05:00"}, {"value": "3560", "qualifiers": ["A"], "dateTime": "2008-06-10T19:00:00.000-05:00"}, {"value": "3540", "qualifiers": ["A"], "dateTime": "2008-06-10T19:15:00.000-05:00"}, {"value": "3510", "qualifiers": ["A"], "dateTime": "2008-06-10T19:30:00.000-05:00"}, {"value": "3500", "qualifiers": ["A"], "dateTime": "2008-06-10T19:45:00.000-05:00"}, {"value": "3470", "qualifiers": ["A"], "dateTime": "2008-06-10T20:00:00.000-05:00"}, {"value": "3440", "qualifiers": ["A"], "dateTime": "2008-06-10T20:15:00.000-05:00"}, {"value": "3410", "qualifiers": ["A"], "dateTime": "2008-06-10T20:30:00.000-05:00"}, {"value": "3360", "qualifiers": ["A"], "dateTime": "2008-06-10T20:45:00.000-05:00"}, {"value": "3330", "qualifiers": ["A"], "dateTime": "2008-06-10T21:00:00.000-05:00"}, {"value": "3320", "qualifiers": ["A"], "dateTime": "2008-06-10T21:15:00.000-05:00"}, {"value": "3290", "qualifiers": ["A"], "dateTime": "2008-06-10T21:30:00.000-05:00"}, {"value": "3200", "qualifiers": ["A"], "dateTime": "2008-06-10T21:45:00.000-05:00"}, {"value": "3200", "qualifiers": ["A"], "dateTime": "2008-06-10T22:00:00.000-05:00"}, {"value": "3200", "qualifiers": ["A"], "dateTime": "2008-06-10T22:15:00.000-05:00"}, {"value": "3160", "qualifiers": ["A"], "dateTime": "2008-06-10T22:30:00.000-05:00"}, {"value": "3140", "qualifiers": ["A"], "dateTime": "2008-06-10T22:45:00.000-05:00"}, {"value": "3110", "qualifiers": ["A"], "dateTime": "2008-06-10T23:00:00.000-05:00"}, {"value": "3100", "qualifiers": ["A"], "dateTime": "2008-06-10T23:15:00.000-05:00"}, {"value": "3090", "qualifiers": ["A"], "dateTime": "2008-06-10T23:30:00.000-05:00"}, {"value": "3060", "qualifiers": ["A"], "dateTime": "2008-06-10T23:45:00.000-05:00"}, {"value": "3000", "qualifiers": ["A"], "dateTime": "2008-06-11T00:00:00.000-05:00"}, {"value": "2970", "qualifiers": ["A"], "dateTime": "2008-06-11T00:15:00.000-05:00"}, {"value": "2960", "qualifiers": ["A"], "dateTime": "2008-06-11T00:30:00.000-05:00"}, {"value": "2920", "qualifiers": ["A"], "dateTime": "2008-06-11T00:45:00.000-05:00"}, {"value": "2900", "qualifiers": ["A"], "dateTime": "2008-06-11T01:00:00.000-05:00"}, {"value": "2860", "qualifiers": ["A"], "dateTime": "2008-06-11T01:15:00.000-05:00"}, {"value": "2870", "qualifiers": ["A"], "dateTime": "2008-06-11T01:30:00.000-05:00"}, {"value": "2850", "qualifiers": ["A"], "dateTime": "2008-06-11T01:45:00.000-05:00"}, {"value": "2850", "qualifiers": ["A"], "dateTime": "2008-06-11T02:00:00.000-05:00"}, {"value": "2810", "qualifiers": ["A"], "dateTime": "2008-06-11T02:15:00.000-05:00"}, {"value": "2810", "qualifiers": ["A"], "dateTime": "2008-06-11T02:30:00.000-05:00"}, {"value": "2790", "qualifiers": ["A"], "dateTime": "2008-06-11T02:45:00.000-05:00"}, {"value": "2760", "qualifiers": ["A"], "dateTime": "2008-06-11T03:00:00.000-05:00"}, {"value": "2730", "qualifiers": ["A"], "dateTime": "2008-06-11T03:15:00.000-05:00"}, {"value": "2720", "qualifiers": ["A"], "dateTime": "2008-06-11T03:30:00.000-05:00"}, {"value": "2700", "qualifiers": ["A"], "dateTime": "2008-06-11T03:45:00.000-05:00"}, {"value": "2690", "qualifiers": ["A"], "dateTime": "2008-06-11T04:00:00.000-05:00"}, {"value": "2670", "qualifiers": ["A"], "dateTime": "2008-06-11T04:15:00.000-05:00"}, {"value": "2640", "qualifiers": ["A"], "dateTime": "2008-06-11T04:30:00.000-05:00"}, {"value": "2660", "qualifiers": ["A"], "dateTime": "2008-06-11T04:45:00.000-05:00"}, {"value": "2610", "qualifiers": ["A"], "dateTime": "2008-06-11T05:00:00.000-05:00"}, {"value": "2600", "qualifiers": ["A"], "dateTime": "2008-06-11T05:15:00.000-05:00"}, {"value": "2580", "qualifiers": ["A"], "dateTime": "2008-06-11T05:30:00.000-05:00"}, {"value": "2560", "qualifiers": ["A"], "dateTime": "2008-06-11T05:45:00.000-05:00"}, {"value": "2530", "qualifiers": ["A"], "dateTime": "2008-06-11T06:00:00.000-05:00"}, {"value": "2530", "qualifiers": ["A"], "dateTime": "2008-06-11T06:15:00.000-05:00"}, {"value": "2530", "qualifiers": ["A"], "dateTime": "2008-06-11T06:30:00.000-05:00"}, {"value": "2500", "qualifiers": ["A"], "dateTime": "2008-06-11T06:45:00.000-05:00"}, {"value": "2480", "qualifiers": ["A"], "dateTime": "2008-06-11T07:00:00.000-05:00"}, {"value": "2420", "qualifiers": ["A"], "dateTime": "2008-06-11T07:15:00.000-05:00"}, {"value": "2430", "qualifiers": ["A"], "dateTime": "2008-06-11T07:30:00.000-05:00"}, {"value": "2390", "qualifiers": ["A"], "dateTime": "2008-06-11T07:45:00.000-05:00"}, {"value": "2380", "qualifiers": ["A"], "dateTime": "2008-06-11T08:00:00.000-05:00"}, {"value": "2400", "qualifiers": ["A"], "dateTime": "2008-06-11T08:15:00.000-05:00"}, {"value": "2350", "qualifiers": ["A"], "dateTime": "2008-06-11T08:30:00.000-05:00"}, {"value": "2340", "qualifiers": ["A"], "dateTime": "2008-06-11T08:45:00.000-05:00"}, {"value": "2340", "qualifiers": ["A"], "dateTime": "2008-06-11T09:00:00.000-05:00"}, {"value": "2330", "qualifiers": ["A"], "dateTime": "2008-06-11T09:15:00.000-05:00"}, {"value": "2310", "qualifiers": ["A"], "dateTime": "2008-06-11T09:30:00.000-05:00"}, {"value": "2270", "qualifiers": ["A"], "dateTime": "2008-06-11T09:45:00.000-05:00"}, {"value": "2260", "qualifiers": ["A"], "dateTime": "2008-06-11T10:00:00.000-05:00"}, {"value": "2260", "qualifiers": ["A"], "dateTime": "2008-06-11T10:15:00.000-05:00"}, {"value": "2230", "qualifiers": ["A"], "dateTime": "2008-06-11T10:30:00.000-05:00"}, {"value": "2180", "qualifiers": ["A"], "dateTime": "2008-06-11T10:45:00.000-05:00"}, {"value": "2220", "qualifiers": ["A"], "dateTime": "2008-06-11T11:00:00.000-05:00"}, {"value": "2200", "qualifiers": ["A"], "dateTime": "2008-06-11T11:15:00.000-05:00"}, {"value": "2170", "qualifiers": ["A"], "dateTime": "2008-06-11T11:30:00.000-05:00"}, {"value": "2170", "qualifiers": ["A"], "dateTime": "2008-06-11T11:45:00.000-05:00"}, {"value": "2160", "qualifiers": ["A"], "dateTime": "2008-06-11T12:00:00.000-05:00"}, {"value": "2150", "qualifiers": ["A"], "dateTime": "2008-06-11T12:15:00.000-05:00"}, {"value": "2140", "qualifiers": ["A"], "dateTime": "2008-06-11T12:30:00.000-05:00"}, {"value": "2130", "qualifiers": ["A"], "dateTime": "2008-06-11T12:45:00.000-05:00"}, {"value": "2130", "qualifiers": ["A"], "dateTime": "2008-06-11T13:00:00.000-05:00"}, {"value": "2110", "qualifiers": ["A"], "dateTime": "2008-06-11T13:15:00.000-05:00"}, {"value": "2090", "qualifiers": ["A"], "dateTime": "2008-06-11T13:30:00.000-05:00"}, {"value": "2090", "qualifiers": ["A"], "dateTime": "2008-06-11T13:45:00.000-05:00"}, {"value": "2070", "qualifiers": ["A"], "dateTime": "2008-06-11T14:00:00.000-05:00"}, {"value": "2050", "qualifiers": ["A"], "dateTime": "2008-06-11T14:15:00.000-05:00"}, {"value": "2060", "qualifiers": ["A"], "dateTime": "2008-06-11T14:30:00.000-05:00"}, {"value": "2060", "qualifiers": ["A"], "dateTime": "2008-06-11T14:45:00.000-05:00"}, {"value": "2020", "qualifiers": ["A"], "dateTime": "2008-06-11T15:00:00.000-05:00"}, {"value": "2020", "qualifiers": ["A"], "dateTime": "2008-06-11T15:15:00.000-05:00"}, {"value": "2020", "qualifiers": ["A"], "dateTime": "2008-06-11T15:30:00.000-05:00"}, {"value": "2000", "qualifiers": ["A"], "dateTime": "2008-06-11T15:45:00.000-05:00"}, {"value": "1970", "qualifiers": ["A"], "dateTime": "2008-06-11T16:00:00.000-05:00"}, {"value": "1970", "qualifiers": ["A"], "dateTime": "2008-06-11T16:15:00.000-05:00"}, {"value": "1970", "qualifiers": ["A"], "dateTime": "2008-06-11T16:30:00.000-05:00"}, {"value": "1950", "qualifiers": ["A"], "dateTime": "2008-06-11T16:45:00.000-05:00"}, {"value": "1940", "qualifiers": ["A"], "dateTime": "2008-06-11T17:00:00.000-05:00"}, {"value": "1940", "qualifiers": ["A"], "dateTime": "2008-06-11T17:15:00.000-05:00"}, {"value": "1940", "qualifiers": ["A"], "dateTime": "2008-06-11T17:30:00.000-05:00"}, {"value": "1910", "qualifiers": ["A"], "dateTime": "2008-06-11T17:45:00.000-05:00"}, {"value": "1910", "qualifiers": ["A"], "dateTime": "2008-06-11T18:00:00.000-05:00"}, {"value": "1910", "qualifiers": ["A"], "dateTime": "2008-06-11T18:15:00.000-05:00"}, {"value": "1880", "qualifiers": ["A"], "dateTime": "2008-06-11T18:30:00.000-05:00"}, {"value": "1880", "qualifiers": ["A"], "dateTime": "2008-06-11T18:45:00.000-05:00"}, {"value": "1880", "qualifiers": ["A"], "dateTime": "2008-06-11T19:00:00.000-05:00"}, {"value": "1860", "qualifiers": ["A"], "dateTime": "2008-06-11T19:15:00.000-05:00"}, {"value": "1860", "qualifiers": ["A"], "dateTime": "2008-06-11T19:30:00.000-05:00"}, {"value": "1840", "qualifiers": ["A"], "dateTime": "2008-06-11T19:45:00.000-05:00"}, {"value": "1830", "qualifiers": ["A"], "dateTime": "2008-06-11T20:00:00.000-05:00"}, {"value": "1820", "qualifiers": ["A"], "dateTime": "2008-06-11T20:15:00.000-05:00"}, {"value": "1810", "qualifiers": ["A"], "dateTime": "2008-06-11T20:30:00.000-05:00"}, {"value": "1800", "qualifiers": ["A"], "dateTime": "2008-06-11T20:45:00.000-05:00"}, {"value": "1790", "qualifiers": ["A"], "dateTime": "2008-06-11T21:00:00.000-05:00"}, {"value": "1790", "qualifiers": ["A"], "dateTime": "2008-06-11T21:15:00.000-05:00"}, {"value": "1750", "qualifiers": ["A"], "dateTime": "2008-06-11T21:30:00.000-05:00"}, {"value": "1780", "qualifiers": ["A"], "dateTime": "2008-06-11T21:45:00.000-05:00"}, {"value": "1770", "qualifiers": ["A"], "dateTime": "2008-06-11T22:00:00.000-05:00"}, {"value": "1740", "qualifiers": ["A"], "dateTime": "2008-06-11T22:15:00.000-05:00"}, {"value": "1750", "qualifiers": ["A"], "dateTime": "2008-06-11T22:30:00.000-05:00"}, {"value": "1740", "qualifiers": ["A"], "dateTime": "2008-06-11T22:45:00.000-05:00"}, {"value": "1720", "qualifiers": ["A"], "dateTime": "2008-06-11T23:00:00.000-05:00"}, {"value": "1720", "qualifiers": ["A"], "dateTime": "2008-06-11T23:15:00.000-05:00"}, {"value": "1700", "qualifiers": ["A"], "dateTime": "2008-06-11T23:30:00.000-05:00"}, {"value": "1700", "qualifiers": ["A"], "dateTime": "2008-06-11T23:45:00.000-05:00"}, {"value": "1680", "qualifiers": ["A"], "dateTime": "2008-06-12T00:00:00.000-05:00"}, {"value": "1690", "qualifiers": ["A"], "dateTime": "2008-06-12T00:15:00.000-05:00"}, {"value": "1680", "qualifiers": ["A"], "dateTime": "2008-06-12T00:30:00.000-05:00"}, {"value": "1680", "qualifiers": ["A"], "dateTime": "2008-06-12T00:45:00.000-05:00"}, {"value": "1670", "qualifiers": ["A"], "dateTime": "2008-06-12T01:00:00.000-05:00"}, {"value": "1690", "qualifiers": ["A"], "dateTime": "2008-06-12T01:15:00.000-05:00"}, {"value": "1700", "qualifiers": ["A"], "dateTime": "2008-06-12T01:30:00.000-05:00"}, {"value": "1730", "qualifiers": ["A"], "dateTime": "2008-06-12T01:45:00.000-05:00"}, {"value": "1770", "qualifiers": ["A"], "dateTime": "2008-06-12T02:00:00.000-05:00"}, {"value": "1790", "qualifiers": ["A"], "dateTime": "2008-06-12T02:15:00.000-05:00"}, {"value": "1790", "qualifiers": ["A"], "dateTime": "2008-06-12T02:30:00.000-05:00"}, {"value": "1820", "qualifiers": ["A"], "dateTime": "2008-06-12T02:45:00.000-05:00"}, {"value": "1840", "qualifiers": ["A"], "dateTime": "2008-06-12T03:00:00.000-05:00"}, {"value": "1850", "qualifiers": ["A"], "dateTime": "2008-06-12T03:15:00.000-05:00"}, {"value": "1880", "qualifiers": ["A"], "dateTime": "2008-06-12T03:30:00.000-05:00"}, {"value": "1880", "qualifiers": ["A"], "dateTime": "2008-06-12T03:45:00.000-05:00"}, {"value": "1910", "qualifiers": ["A"], "dateTime": "2008-06-12T04:00:00.000-05:00"}, {"value": "1930", "qualifiers": ["A"], "dateTime": "2008-06-12T04:15:00.000-05:00"}, {"value": "1920", "qualifiers": ["A"], "dateTime": "2008-06-12T04:30:00.000-05:00"}, {"value": "1950", "qualifiers": ["A"], "dateTime": "2008-06-12T04:45:00.000-05:00"}, {"value": "1950", "qualifiers": ["A"], "dateTime": "2008-06-12T05:00:00.000-05:00"}, {"value": "1990", "qualifiers": ["A"], "dateTime": "2008-06-12T05:15:00.000-05:00"}, {"value": "2010", "qualifiers": ["A"], "dateTime": "2008-06-12T05:30:00.000-05:00"}, {"value": "2010", "qualifiers": ["A"], "dateTime": "2008-06-12T05:45:00.000-05:00"}, {"value": "2030", "qualifiers": ["A"], "dateTime": "2008-06-12T06:00:00.000-05:00"}, {"value": "2060", "qualifiers": ["A"], "dateTime": "2008-06-12T06:15:00.000-05:00"}, {"value": "2100", "qualifiers": ["A"], "dateTime": "2008-06-12T06:30:00.000-05:00"}, {"value": "2100", "qualifiers": ["A"], "dateTime": "2008-06-12T06:45:00.000-05:00"}, {"value": "2150", "qualifiers": ["A"], "dateTime": "2008-06-12T07:00:00.000-05:00"}, {"value": "2180", "qualifiers": ["A"], "dateTime": "2008-06-12T07:15:00.000-05:00"}, {"value": "2210", "qualifiers": ["A"], "dateTime": "2008-06-12T07:30:00.000-05:00"}, {"value": "2270", "qualifiers": ["A"], "dateTime": "2008-06-12T07:45:00.000-05:00"}, {"value": "2320", "qualifiers": ["A"], "dateTime": "2008-06-12T08:00:00.000-05:00"}, {"value": "2380", "qualifiers": ["A"], "dateTime": "2008-06-12T08:15:00.000-05:00"}, {"value": "2440", "qualifiers": ["A"], "dateTime": "2008-06-12T08:30:00.000-05:00"}, {"value": "2480", "qualifiers": ["A"], "dateTime": "2008-06-12T08:45:00.000-05:00"}, {"value": "2550", "qualifiers": ["A"], "dateTime": "2008-06-12T09:00:00.000-05:00"}, {"value": "2610", "qualifiers": ["A"], "dateTime": "2008-06-12T09:15:00.000-05:00"}, {"value": "2640", "qualifiers": ["A"], "dateTime": "2008-06-12T09:30:00.000-05:00"}, {"value": "2690", "qualifiers": ["A"], "dateTime": "2008-06-12T09:45:00.000-05:00"}, {"value": "2730", "qualifiers": ["A"], "dateTime": "2008-06-12T10:00:00.000-05:00"}, {"value": "2790", "qualifiers": ["A"], "dateTime": "2008-06-12T10:15:00.000-05:00"}, {"value": "2820", "qualifiers": ["A"], "dateTime": "2008-06-12T10:30:00.000-05:00"}, {"value": "2880", "qualifiers": ["A"], "dateTime": "2008-06-12T10:45:00.000-05:00"}, {"value": "2910", "qualifiers": ["A"], "dateTime": "2008-06-12T11:00:00.000-05:00"}, {"value": "2960", "qualifiers": ["A"], "dateTime": "2008-06-12T11:15:00.000-05:00"}, {"value": "3000", "qualifiers": ["A"], "dateTime": "2008-06-12T11:30:00.000-05:00"}, {"value": "3100", "qualifiers": ["A"], "dateTime": "2008-06-12T11:45:00.000-05:00"}, {"value": "3130", "qualifiers": ["A"], "dateTime": "2008-06-12T12:00:00.000-05:00"}, {"value": "3210", "qualifiers": ["A"], "dateTime": "2008-06-12T12:15:00.000-05:00"}, {"value": "3260", "qualifiers": ["A"], "dateTime": "2008-06-12T12:30:00.000-05:00"}, {"value": "3290", "qualifiers": ["A"], "dateTime": "2008-06-12T12:45:00.000-05:00"}, {"value": "3330", "qualifiers": ["A"], "dateTime": "2008-06-12T13:00:00.000-05:00"}, {"value": "3390", "qualifiers": ["A"], "dateTime": "2008-06-12T13:15:00.000-05:00"}, {"value": "3470", "qualifiers": ["A"], "dateTime": "2008-06-12T13:30:00.000-05:00"}, {"value": "3450", "qualifiers": ["A"], "dateTime": "2008-06-12T13:45:00.000-05:00"}, {"value": "3500", "qualifiers": ["A"], "dateTime": "2008-06-12T14:00:00.000-05:00"}, {"value": "3580", "qualifiers": ["A"], "dateTime": "2008-06-12T14:15:00.000-05:00"}, {"value": "3560", "qualifiers": ["A"], "dateTime": "2008-06-12T14:30:00.000-05:00"}, {"value": "3640", "qualifiers": ["A"], "dateTime": "2008-06-12T14:45:00.000-05:00"}, {"value": "3700", "qualifiers": ["A"], "dateTime": "2008-06-12T15:00:00.000-05:00"}, {"value": "3720", "qualifiers": ["A"], "dateTime": "2008-06-12T15:15:00.000-05:00"}, {"value": "3800", "qualifiers": ["A"], "dateTime": "2008-06-12T15:30:00.000-05:00"}, {"value": "3780", "qualifiers": ["A"], "dateTime": "2008-06-12T15:45:00.000-05:00"}, {"value": "3870", "qualifiers": ["A"], "dateTime": "2008-06-12T16:00:00.000-05:00"}, {"value": "3950", "qualifiers": ["A"], "dateTime": "2008-06-12T16:15:00.000-05:00"}, {"value": "3930", "qualifiers": ["A"], "dateTime": "2008-06-12T16:30:00.000-05:00"}, {"value": "4090", "qualifiers": ["A"], "dateTime": "2008-06-12T16:45:00.000-05:00"}, {"value": "4110", "qualifiers": ["A"], "dateTime": "2008-06-12T17:00:00.000-05:00"}, {"value": "4150", "qualifiers": ["A"], "dateTime": "2008-06-12T17:15:00.000-05:00"}, {"value": "4230", "qualifiers": ["A"], "dateTime": "2008-06-12T17:30:00.000-05:00"}, {"value": "4330", "qualifiers": ["A"], "dateTime": "2008-06-12T17:45:00.000-05:00"}, {"value": "4330", "qualifiers": ["A"], "dateTime": "2008-06-12T18:00:00.000-05:00"}, {"value": "4420", "qualifiers": ["A"], "dateTime": "2008-06-12T18:15:00.000-05:00"}, {"value": "4570", "qualifiers": ["A"], "dateTime": "2008-06-12T18:30:00.000-05:00"}, {"value": "4590", "qualifiers": ["A"], "dateTime": "2008-06-12T18:45:00.000-05:00"}, {"value": "4700", "qualifiers": ["A"], "dateTime": "2008-06-12T19:00:00.000-05:00"}, {"value": "4790", "qualifiers": ["A"], "dateTime": "2008-06-12T19:15:00.000-05:00"}, {"value": "4910", "qualifiers": ["A"], "dateTime": "2008-06-12T19:30:00.000-05:00"}, {"value": "5000", "qualifiers": ["A"], "dateTime": "2008-06-12T19:45:00.000-05:00"}, {"value": "5050", "qualifiers": ["A"], "dateTime": "2008-06-12T20:00:00.000-05:00"}, {"value": "5240", "qualifiers": ["A"], "dateTime": "2008-06-12T20:15:00.000-05:00"}, {"value": "5240", "qualifiers": ["A"], "dateTime": "2008-06-12T20:30:00.000-05:00"}, {"value": "5440", "qualifiers": ["A"], "dateTime": "2008-06-12T20:45:00.000-05:00"}, {"value": "5640", "qualifiers": ["A"], "dateTime": "2008-06-12T21:00:00.000-05:00"}, {"value": "5930", "qualifiers": ["A"], "dateTime": "2008-06-12T21:15:00.000-05:00"}, {"value": "6180", "qualifiers": ["A"], "dateTime": "2008-06-12T21:30:00.000-05:00"}, {"value": "6580", "qualifiers": ["A"], "dateTime": "2008-06-12T21:45:00.000-05:00"}, {"value": "7100", "qualifiers": ["A"], "dateTime": "2008-06-12T22:00:00.000-05:00"}, {"value": "7450", "qualifiers": ["A"], "dateTime": "2008-06-12T22:15:00.000-05:00"}, {"value": "8060", "qualifiers": ["A"], "dateTime": "2008-06-12T22:30:00.000-05:00"}, {"value": "8300", "qualifiers": ["A"], "dateTime": "2008-06-12T22:45:00.000-05:00"}, {"value": "8780", "qualifiers": ["A"], "dateTime": "2008-06-12T23:00:00.000-05:00"}, {"value": "9230", "qualifiers": ["A"], "dateTime": "2008-06-12T23:15:00.000-05:00"}, {"value": "9670", "qualifiers": ["A"], "dateTime": "2008-06-12T23:30:00.000-05:00"}, {"value": "9990", "qualifiers": ["A"], "dateTime": "2008-06-12T23:45:00.000-05:00"}, {"value": "10300", "qualifiers": ["A"], "dateTime": "2008-06-13T00:00:00.000-05:00"}, {"value": "10600", "qualifiers": ["A"], "dateTime": "2008-06-13T00:15:00.000-05:00"}, {"value": "10900", "qualifiers": ["A"], "dateTime": "2008-06-13T00:30:00.000-05:00"}, {"value": "11100", "qualifiers": ["A"], "dateTime": "2008-06-13T00:45:00.000-05:00"}, {"value": "11400", "qualifiers": ["A"], "dateTime": "2008-06-13T01:00:00.000-05:00"}, {"value": "11400", "qualifiers": ["A"], "dateTime": "2008-06-13T01:15:00.000-05:00"}, {"value": "11400", "qualifiers": ["A"], "dateTime": "2008-06-13T01:30:00.000-05:00"}, {"value": "11700", "qualifiers": ["A"], "dateTime": "2008-06-13T01:45:00.000-05:00"}, {"value": "11600", "qualifiers": ["A"], "dateTime": "2008-06-13T02:00:00.000-05:00"}, {"value": "11700", "qualifiers": ["A"], "dateTime": "2008-06-13T02:15:00.000-05:00"}, {"value": "11600", "qualifiers": ["A"], "dateTime": "2008-06-13T02:30:00.000-05:00"}, {"value": "11800", "qualifiers": ["A"], "dateTime": "2008-06-13T02:45:00.000-05:00"}, {"value": "11600", "qualifiers": ["A"], "dateTime": "2008-06-13T03:00:00.000-05:00"}, {"value": "11600", "qualifiers": ["A"], "dateTime": "2008-06-13T03:15:00.000-05:00"}, {"value": "11600", "qualifiers": ["A"], "dateTime": "2008-06-13T03:30:00.000-05:00"}, {"value": "11300", "qualifiers": ["A"], "dateTime": "2008-06-13T03:45:00.000-05:00"}, {"value": "11300", "qualifiers": ["A"], "dateTime": "2008-06-13T04:00:00.000-05:00"}, {"value": "11300", "qualifiers": ["A"], "dateTime": "2008-06-13T04:15:00.000-05:00"}, {"value": "11200", "qualifiers": ["A"], "dateTime": "2008-06-13T04:30:00.000-05:00"}, {"value": "11200", "qualifiers": ["A"], "dateTime": "2008-06-13T04:45:00.000-05:00"}, {"value": "10900", "qualifiers": ["A"], "dateTime": "2008-06-13T05:00:00.000-05:00"}, {"value": "10900", "qualifiers": ["A"], "dateTime": "2008-06-13T05:15:00.000-05:00"}, {"value": "10800", "qualifiers": ["A"], "dateTime": "2008-06-13T05:30:00.000-05:00"}, {"value": "10700", "qualifiers": ["A"], "dateTime": "2008-06-13T05:45:00.000-05:00"}, {"value": "10600", "qualifiers": ["A"], "dateTime": "2008-06-13T06:00:00.000-05:00"}, {"value": "10500", "qualifiers": ["A"], "dateTime": "2008-06-13T06:15:00.000-05:00"}, {"value": "10200", "qualifiers": ["A"], "dateTime": "2008-06-13T06:30:00.000-05:00"}, {"value": "10200", "qualifiers": ["A"], "dateTime": "2008-06-13T06:45:00.000-05:00"}, {"value": "9990", "qualifiers": ["A"], "dateTime": "2008-06-13T07:00:00.000-05:00"}, {"value": "9990", "qualifiers": ["A"], "dateTime": "2008-06-13T07:15:00.000-05:00"}, {"value": "9710", "qualifiers": ["A"], "dateTime": "2008-06-13T07:30:00.000-05:00"}, {"value": "9550", "qualifiers": ["A"], "dateTime": "2008-06-13T07:45:00.000-05:00"}, {"value": "9430", "qualifiers": ["A"], "dateTime": "2008-06-13T08:00:00.000-05:00"}, {"value": "9430", "qualifiers": ["A"], "dateTime": "2008-06-13T08:15:00.000-05:00"}, {"value": "9390", "qualifiers": ["A"], "dateTime": "2008-06-13T08:30:00.000-05:00"}, {"value": "9350", "qualifiers": ["A"], "dateTime": "2008-06-13T08:45:00.000-05:00"}, {"value": "9150", "qualifiers": ["A"], "dateTime": "2008-06-13T09:00:00.000-05:00"}, {"value": "9040", "qualifiers": ["A"], "dateTime": "2008-06-13T09:15:00.000-05:00"}, {"value": "9000", "qualifiers": ["A"], "dateTime": "2008-06-13T09:30:00.000-05:00"}, {"value": "8890", "qualifiers": ["A"], "dateTime": "2008-06-13T09:45:00.000-05:00"}, {"value": "8740", "qualifiers": ["A"], "dateTime": "2008-06-13T10:00:00.000-05:00"}, {"value": "8660", "qualifiers": ["A"], "dateTime": "2008-06-13T10:15:00.000-05:00"}, {"value": "8520", "qualifiers": ["A"], "dateTime": "2008-06-13T10:30:00.000-05:00"}, {"value": "8520", "qualifiers": ["A"], "dateTime": "2008-06-13T10:45:00.000-05:00"}, {"value": "8370", "qualifiers": ["A"], "dateTime": "2008-06-13T11:00:00.000-05:00"}, {"value": "8300", "qualifiers": ["A"], "dateTime": "2008-06-13T11:15:00.000-05:00"}, {"value": "8300", "qualifiers": ["A"], "dateTime": "2008-06-13T11:30:00.000-05:00"}, {"value": "8230", "qualifiers": ["A"], "dateTime": "2008-06-13T11:45:00.000-05:00"}, {"value": "8060", "qualifiers": ["A"], "dateTime": "2008-06-13T12:00:00.000-05:00"}, {"value": "8020", "qualifiers": ["A"], "dateTime": "2008-06-13T12:15:00.000-05:00"}, {"value": "7950", "qualifiers": ["A"], "dateTime": "2008-06-13T12:30:00.000-05:00"}, {"value": "7880", "qualifiers": ["A"], "dateTime": "2008-06-13T12:45:00.000-05:00"}, {"value": "7680", "qualifiers": ["A"], "dateTime": "2008-06-13T13:00:00.000-05:00"}, {"value": "7750", "qualifiers": ["A"], "dateTime": "2008-06-13T13:15:00.000-05:00"}, {"value": "7720", "qualifiers": ["A"], "dateTime": "2008-06-13T13:30:00.000-05:00"}, {"value": "7550", "qualifiers": ["A"], "dateTime": "2008-06-13T13:45:00.000-05:00"}, {"value": "7480", "qualifiers": ["A"], "dateTime": "2008-06-13T14:00:00.000-05:00"}, {"value": "7450", "qualifiers": ["A"], "dateTime": "2008-06-13T14:15:00.000-05:00"}, {"value": "7320", "qualifiers": ["A"], "dateTime": "2008-06-13T14:30:00.000-05:00"}, {"value": "7350", "qualifiers": ["A"], "dateTime": "2008-06-13T14:45:00.000-05:00"}, {"value": "7230", "qualifiers": ["A"], "dateTime": "2008-06-13T15:00:00.000-05:00"}, {"value": "7160", "qualifiers": ["A"], "dateTime": "2008-06-13T15:15:00.000-05:00"}, {"value": "7070", "qualifiers": ["A"], "dateTime": "2008-06-13T15:30:00.000-05:00"}, {"value": "7070", "qualifiers": ["A"], "dateTime": "2008-06-13T15:45:00.000-05:00"}, {"value": "7070", "qualifiers": ["A"], "dateTime": "2008-06-13T16:00:00.000-05:00"}, {"value": "6910", "qualifiers": ["A"], "dateTime": "2008-06-13T16:15:00.000-05:00"}, {"value": "6910", "qualifiers": ["A"], "dateTime": "2008-06-13T16:30:00.000-05:00"}, {"value": "6760", "qualifiers": ["A"], "dateTime": "2008-06-13T16:45:00.000-05:00"}, {"value": "6820", "qualifiers": ["A"], "dateTime": "2008-06-13T17:00:00.000-05:00"}, {"value": "6700", "qualifiers": ["A"], "dateTime": "2008-06-13T17:15:00.000-05:00"}, {"value": "6550", "qualifiers": ["A"], "dateTime": "2008-06-13T17:30:00.000-05:00"}, {"value": "6580", "qualifiers": ["A"], "dateTime": "2008-06-13T17:45:00.000-05:00"}, {"value": "6550", "qualifiers": ["A"], "dateTime": "2008-06-13T18:00:00.000-05:00"}, {"value": "6530", "qualifiers": ["A"], "dateTime": "2008-06-13T18:15:00.000-05:00"}, {"value": "6470", "qualifiers": ["A"], "dateTime": "2008-06-13T18:30:00.000-05:00"}, {"value": "6440", "qualifiers": ["A"], "dateTime": "2008-06-13T18:45:00.000-05:00"}, {"value": "6470", "qualifiers": ["A"], "dateTime": "2008-06-13T19:00:00.000-05:00"}, {"value": "6350", "qualifiers": ["A"], "dateTime": "2008-06-13T19:15:00.000-05:00"}, {"value": "6320", "qualifiers": ["A"], "dateTime": "2008-06-13T19:30:00.000-05:00"}, {"value": "6210", "qualifiers": ["A"], "dateTime": "2008-06-13T19:45:00.000-05:00"}, {"value": "6240", "qualifiers": ["A"], "dateTime": "2008-06-13T20:00:00.000-05:00"}, {"value": "6150", "qualifiers": ["A"], "dateTime": "2008-06-13T20:15:00.000-05:00"}, {"value": "6130", "qualifiers": ["A"], "dateTime": "2008-06-13T20:30:00.000-05:00"}, {"value": "6040", "qualifiers": ["A"], "dateTime": "2008-06-13T20:45:00.000-05:00"}, {"value": "6040", "qualifiers": ["A"], "dateTime": "2008-06-13T21:00:00.000-05:00"}, {"value": "5910", "qualifiers": ["A"], "dateTime": "2008-06-13T21:15:00.000-05:00"}, {"value": "5910", "qualifiers": ["A"], "dateTime": "2008-06-13T21:30:00.000-05:00"}, {"value": "5880", "qualifiers": ["A"], "dateTime": "2008-06-13T21:45:00.000-05:00"}, {"value": "5880", "qualifiers": ["A"], "dateTime": "2008-06-13T22:00:00.000-05:00"}, {"value": "5800", "qualifiers": ["A"], "dateTime": "2008-06-13T22:15:00.000-05:00"}, {"value": "5750", "qualifiers": ["A"], "dateTime": "2008-06-13T22:30:00.000-05:00"}, {"value": "5700", "qualifiers": ["A"], "dateTime": "2008-06-13T22:45:00.000-05:00"}, {"value": "5570", "qualifiers": ["A"], "dateTime": "2008-06-13T23:00:00.000-05:00"}, {"value": "5590", "qualifiers": ["A"], "dateTime": "2008-06-13T23:15:00.000-05:00"}, {"value": "5510", "qualifiers": ["A"], "dateTime": "2008-06-13T23:30:00.000-05:00"}, {"value": "5510", "qualifiers": ["A"], "dateTime": "2008-06-13T23:45:00.000-05:00"}, {"value": "5440", "qualifiers": ["A"], "dateTime": "2008-06-14T00:00:00.000-05:00"}, {"value": "5460", "qualifiers": ["A"], "dateTime": "2008-06-14T00:15:00.000-05:00"}, {"value": "5390", "qualifiers": ["A"], "dateTime": "2008-06-14T00:30:00.000-05:00"}, {"value": "5340", "qualifiers": ["A"], "dateTime": "2008-06-14T00:45:00.000-05:00"}, {"value": "5310", "qualifiers": ["A"], "dateTime": "2008-06-14T01:00:00.000-05:00"}, {"value": "5310", "qualifiers": ["A"], "dateTime": "2008-06-14T01:15:00.000-05:00"}, {"value": "5260", "qualifiers": ["A"], "dateTime": "2008-06-14T01:30:00.000-05:00"}, {"value": "5240", "qualifiers": ["A"], "dateTime": "2008-06-14T01:45:00.000-05:00"}, {"value": "5190", "qualifiers": ["A"], "dateTime": "2008-06-14T02:00:00.000-05:00"}, {"value": "5100", "qualifiers": ["A"], "dateTime": "2008-06-14T02:15:00.000-05:00"}, {"value": "5100", "qualifiers": ["A"], "dateTime": "2008-06-14T02:30:00.000-05:00"}, {"value": "4980", "qualifiers": ["A"], "dateTime": "2008-06-14T02:45:00.000-05:00"}, {"value": "5000", "qualifiers": ["A"], "dateTime": "2008-06-14T03:00:00.000-05:00"}, {"value": "4980", "qualifiers": ["A"], "dateTime": "2008-06-14T03:15:00.000-05:00"}, {"value": "4930", "qualifiers": ["A"], "dateTime": "2008-06-14T03:30:00.000-05:00"}, {"value": "4860", "qualifiers": ["A"], "dateTime": "2008-06-14T03:45:00.000-05:00"}, {"value": "4770", "qualifiers": ["A"], "dateTime": "2008-06-14T04:00:00.000-05:00"}, {"value": "4840", "qualifiers": ["A"], "dateTime": "2008-06-14T04:15:00.000-05:00"}, {"value": "4840", "qualifiers": ["A"], "dateTime": "2008-06-14T04:30:00.000-05:00"}, {"value": "4750", "qualifiers": ["A"], "dateTime": "2008-06-14T04:45:00.000-05:00"}, {"value": "4720", "qualifiers": ["A"], "dateTime": "2008-06-14T05:00:00.000-05:00"}, {"value": "4660", "qualifiers": ["A"], "dateTime": "2008-06-14T05:15:00.000-05:00"}, {"value": "4640", "qualifiers": ["A"], "dateTime": "2008-06-14T05:30:00.000-05:00"}, {"value": "4610", "qualifiers": ["A"], "dateTime": "2008-06-14T05:45:00.000-05:00"}, {"value": "4550", "qualifiers": ["A"], "dateTime": "2008-06-14T06:00:00.000-05:00"}, {"value": "4500", "qualifiers": ["A"], "dateTime": "2008-06-14T06:15:00.000-05:00"}, {"value": "4500", "qualifiers": ["A"], "dateTime": "2008-06-14T06:30:00.000-05:00"}, {"value": "4480", "qualifiers": ["A"], "dateTime": "2008-06-14T06:45:00.000-05:00"}, {"value": "4440", "qualifiers": ["A"], "dateTime": "2008-06-14T07:00:00.000-05:00"}, {"value": "4400", "qualifiers": ["A"], "dateTime": "2008-06-14T07:15:00.000-05:00"}, {"value": "4400", "qualifiers": ["A"], "dateTime": "2008-06-14T07:30:00.000-05:00"}, {"value": "4290", "qualifiers": ["A"], "dateTime": "2008-06-14T07:45:00.000-05:00"}, {"value": "4330", "qualifiers": ["A"], "dateTime": "2008-06-14T08:00:00.000-05:00"}, {"value": "4270", "qualifiers": ["A"], "dateTime": "2008-06-14T08:15:00.000-05:00"}, {"value": "4270", "qualifiers": ["A"], "dateTime": "2008-06-14T08:30:00.000-05:00"}, {"value": "4190", "qualifiers": ["A"], "dateTime": "2008-06-14T08:45:00.000-05:00"}, {"value": "4130", "qualifiers": ["A"], "dateTime": "2008-06-14T09:00:00.000-05:00"}, {"value": "4090", "qualifiers": ["A"], "dateTime": "2008-06-14T09:15:00.000-05:00"}, {"value": "4110", "qualifiers": ["A"], "dateTime": "2008-06-14T09:30:00.000-05:00"}, {"value": "4050", "qualifiers": ["A"], "dateTime": "2008-06-14T09:45:00.000-05:00"}, {"value": "3990", "qualifiers": ["A"], "dateTime": "2008-06-14T10:00:00.000-05:00"}, {"value": "4010", "qualifiers": ["A"], "dateTime": "2008-06-14T10:15:00.000-05:00"}, {"value": "3930", "qualifiers": ["A"], "dateTime": "2008-06-14T10:30:00.000-05:00"}, {"value": "3890", "qualifiers": ["A"], "dateTime": "2008-06-14T10:45:00.000-05:00"}, {"value": "3910", "qualifiers": ["A"], "dateTime": "2008-06-14T11:00:00.000-05:00"}, {"value": "3860", "qualifiers": ["A"], "dateTime": "2008-06-14T11:15:00.000-05:00"}, {"value": "3800", "qualifiers": ["A"], "dateTime": "2008-06-14T11:30:00.000-05:00"}, {"value": "3770", "qualifiers": ["A"], "dateTime": "2008-06-14T11:45:00.000-05:00"}, {"value": "3750", "qualifiers": ["A"], "dateTime": "2008-06-14T12:00:00.000-05:00"}, {"value": "3700", "qualifiers": ["A"], "dateTime": "2008-06-14T12:15:00.000-05:00"}, {"value": "3670", "qualifiers": ["A"], "dateTime": "2008-06-14T12:30:00.000-05:00"}, {"value": "3640", "qualifiers": ["A"], "dateTime": "2008-06-14T12:45:00.000-05:00"}, {"value": "3620", "qualifiers": ["A"], "dateTime": "2008-06-14T13:00:00.000-05:00"}, {"value": "3620", "qualifiers": ["A"], "dateTime": "2008-06-14T13:15:00.000-05:00"}, {"value": "3560", "qualifiers": ["A"], "dateTime": "2008-06-14T13:30:00.000-05:00"}, {"value": "3560", "qualifiers": ["A"], "dateTime": "2008-06-14T13:45:00.000-05:00"}, {"value": "3500", "qualifiers": ["A"], "dateTime": "2008-06-14T14:00:00.000-05:00"}, {"value": "3480", "qualifiers": ["A"], "dateTime": "2008-06-14T14:15:00.000-05:00"}, {"value": "3450", "qualifiers": ["A"], "dateTime": "2008-06-14T14:30:00.000-05:00"}, {"value": "3440", "qualifiers": ["A"], "dateTime": "2008-06-14T14:45:00.000-05:00"}, {"value": "3390", "qualifiers": ["A"], "dateTime": "2008-06-14T15:00:00.000-05:00"}, {"value": "3390", "qualifiers": ["A"], "dateTime": "2008-06-14T15:15:00.000-05:00"}, {"value": "3330", "qualifiers": ["A"], "dateTime": "2008-06-14T15:30:00.000-05:00"}, {"value": "3330", "qualifiers": ["A"], "dateTime": "2008-06-14T15:45:00.000-05:00"}, {"value": "3300", "qualifiers": ["A"], "dateTime": "2008-06-14T16:00:00.000-05:00"}, {"value": "3290", "qualifiers": ["A"], "dateTime": "2008-06-14T16:15:00.000-05:00"}, {"value": "3210", "qualifiers": ["A"], "dateTime": "2008-06-14T16:30:00.000-05:00"}, {"value": "3200", "qualifiers": ["A"], "dateTime": "2008-06-14T16:45:00.000-05:00"}, {"value": "3160", "qualifiers": ["A"], "dateTime": "2008-06-14T17:00:00.000-05:00"}, {"value": "3100", "qualifiers": ["A"], "dateTime": "2008-06-14T17:15:00.000-05:00"}, {"value": "3090", "qualifiers": ["A"], "dateTime": "2008-06-14T17:30:00.000-05:00"}, {"value": "3070", "qualifiers": ["A"], "dateTime": "2008-06-14T17:45:00.000-05:00"}, {"value": "3030", "qualifiers": ["A"], "dateTime": "2008-06-14T18:00:00.000-05:00"}, {"value": "2970", "qualifiers": ["A"], "dateTime": "2008-06-14T18:15:00.000-05:00"}, {"value": "2950", "qualifiers": ["A"], "dateTime": "2008-06-14T18:30:00.000-05:00"}, {"value": "2930", "qualifiers": ["A"], "dateTime": "2008-06-14T18:45:00.000-05:00"}, {"value": "2910", "qualifiers": ["A"], "dateTime": "2008-06-14T19:00:00.000-05:00"}, {"value": "2880", "qualifiers": ["A"], "dateTime": "2008-06-14T19:15:00.000-05:00"}, {"value": "2880", "qualifiers": ["A"], "dateTime": "2008-06-14T19:30:00.000-05:00"}, {"value": "2850", "qualifiers": ["A"], "dateTime": "2008-06-14T19:45:00.000-05:00"}, {"value": "2800", "qualifiers": ["A"], "dateTime": "2008-06-14T20:00:00.000-05:00"}, {"value": "2800", "qualifiers": ["A"], "dateTime": "2008-06-14T20:15:00.000-05:00"}, {"value": "2750", "qualifiers": ["A"], "dateTime": "2008-06-14T20:30:00.000-05:00"}, {"value": "2750", "qualifiers": ["A"], "dateTime": "2008-06-14T20:45:00.000-05:00"}, {"value": "2720", "qualifiers": ["A"], "dateTime": "2008-06-14T21:00:00.000-05:00"}, {"value": "2710", "qualifiers": ["A"], "dateTime": "2008-06-14T21:15:00.000-05:00"}, {"value": "2680", "qualifiers": ["A"], "dateTime": "2008-06-14T21:30:00.000-05:00"}, {"value": "2650", "qualifiers": ["A"], "dateTime": "2008-06-14T21:45:00.000-05:00"}, {"value": "2640", "qualifiers": ["A"], "dateTime": "2008-06-14T22:00:00.000-05:00"}, {"value": "2600", "qualifiers": ["A"], "dateTime": "2008-06-14T22:15:00.000-05:00"}, {"value": "2590", "qualifiers": ["A"], "dateTime": "2008-06-14T22:30:00.000-05:00"}, {"value": "2560", "qualifiers": ["A"], "dateTime": "2008-06-14T22:45:00.000-05:00"}, {"value": "2530", "qualifiers": ["A"], "dateTime": "2008-06-14T23:00:00.000-05:00"}, {"value": "2510", "qualifiers": ["A"], "dateTime": "2008-06-14T23:15:00.000-05:00"}, {"value": "2470", "qualifiers": ["A"], "dateTime": "2008-06-14T23:30:00.000-05:00"}, {"value": "2470", "qualifiers": ["A"], "dateTime": "2008-06-14T23:45:00.000-05:00"}, {"value": "2440", "qualifiers": ["A"], "dateTime": "2008-06-15T00:00:00.000-05:00"}, {"value": "2420", "qualifiers": ["A"], "dateTime": "2008-06-15T00:15:00.000-05:00"}, {"value": "2390", "qualifiers": ["A"], "dateTime": "2008-06-15T00:30:00.000-05:00"}, {"value": "2390", "qualifiers": ["A"], "dateTime": "2008-06-15T00:45:00.000-05:00"}, {"value": "2350", "qualifiers": ["A"], "dateTime": "2008-06-15T01:00:00.000-05:00"}, {"value": "2340", "qualifiers": ["A"], "dateTime": "2008-06-15T01:15:00.000-05:00"}, {"value": "2320", "qualifiers": ["A"], "dateTime": "2008-06-15T01:30:00.000-05:00"}, {"value": "2280", "qualifiers": ["A"], "dateTime": "2008-06-15T01:45:00.000-05:00"}, {"value": "2270", "qualifiers": ["A"], "dateTime": "2008-06-15T02:00:00.000-05:00"}, {"value": "2240", "qualifiers": ["A"], "dateTime": "2008-06-15T02:15:00.000-05:00"}, {"value": "2250", "qualifiers": ["A"], "dateTime": "2008-06-15T02:30:00.000-05:00"}, {"value": "2220", "qualifiers": ["A"], "dateTime": "2008-06-15T02:45:00.000-05:00"}, {"value": "2200", "qualifiers": ["A"], "dateTime": "2008-06-15T03:00:00.000-05:00"}, {"value": "2200", "qualifiers": ["A"], "dateTime": "2008-06-15T03:15:00.000-05:00"}, {"value": "2180", "qualifiers": ["A"], "dateTime": "2008-06-15T03:30:00.000-05:00"}, {"value": "2230", "qualifiers": ["A"], "dateTime": "2008-06-15T03:45:00.000-05:00"}, {"value": "2220", "qualifiers": ["A"], "dateTime": "2008-06-15T04:00:00.000-05:00"}, {"value": "2220", "qualifiers": ["A"], "dateTime": "2008-06-15T04:15:00.000-05:00"}, {"value": "2200", "qualifiers": ["A"], "dateTime": "2008-06-15T04:30:00.000-05:00"}, {"value": "2190", "qualifiers": ["A"], "dateTime": "2008-06-15T04:45:00.000-05:00"}, {"value": "2200", "qualifiers": ["A"], "dateTime": "2008-06-15T05:00:00.000-05:00"}, {"value": "2180", "qualifiers": ["A"], "dateTime": "2008-06-15T05:15:00.000-05:00"}, {"value": "2180", "qualifiers": ["A"], "dateTime": "2008-06-15T05:30:00.000-05:00"}, {"value": "2160", "qualifiers": ["A"], "dateTime": "2008-06-15T05:45:00.000-05:00"}, {"value": "2170", "qualifiers": ["A"], "dateTime": "2008-06-15T06:00:00.000-05:00"}, {"value": "2150", "qualifiers": ["A"], "dateTime": "2008-06-15T06:15:00.000-05:00"}, {"value": "2150", "qualifiers": ["A"], "dateTime": "2008-06-15T06:30:00.000-05:00"}, {"value": "2120", "qualifiers": ["A"], "dateTime": "2008-06-15T06:45:00.000-05:00"}, {"value": "2130", "qualifiers": ["A"], "dateTime": "2008-06-15T07:00:00.000-05:00"}, {"value": "2110", "qualifiers": ["A"], "dateTime": "2008-06-15T07:15:00.000-05:00"}, {"value": "2100", "qualifiers": ["A"], "dateTime": "2008-06-15T07:30:00.000-05:00"}, {"value": "2090", "qualifiers": ["A"], "dateTime": "2008-06-15T07:45:00.000-05:00"}, {"value": "2090", "qualifiers": ["A"], "dateTime": "2008-06-15T08:00:00.000-05:00"}, {"value": "2080", "qualifiers": ["A"], "dateTime": "2008-06-15T08:15:00.000-05:00"}, {"value": "2060", "qualifiers": ["A"], "dateTime": "2008-06-15T08:30:00.000-05:00"}, {"value": "2060", "qualifiers": ["A"], "dateTime": "2008-06-15T08:45:00.000-05:00"}, {"value": "2060", "qualifiers": ["A"], "dateTime": "2008-06-15T09:00:00.000-05:00"}, {"value": "2040", "qualifiers": ["A"], "dateTime": "2008-06-15T09:15:00.000-05:00"}, {"value": "2040", "qualifiers": ["A"], "dateTime": "2008-06-15T09:30:00.000-05:00"}, {"value": "2040", "qualifiers": ["A"], "dateTime": "2008-06-15T09:45:00.000-05:00"}, {"value": "2020", "qualifiers": ["A"], "dateTime": "2008-06-15T10:00:00.000-05:00"}, {"value": "2040", "qualifiers": ["A"], "dateTime": "2008-06-15T10:15:00.000-05:00"}, {"value": "2010", "qualifiers": ["A"], "dateTime": "2008-06-15T10:30:00.000-05:00"}, {"value": "2020", "qualifiers": ["A"], "dateTime": "2008-06-15T10:45:00.000-05:00"}, {"value": "2020", "qualifiers": ["A"], "dateTime": "2008-06-15T11:00:00.000-05:00"}, {"value": "2020", "qualifiers": ["A"], "dateTime": "2008-06-15T11:15:00.000-05:00"}, {"value": "2000", "qualifiers": ["A"], "dateTime": "2008-06-15T11:30:00.000-05:00"}, {"value": "2000", "qualifiers": ["A"], "dateTime": "2008-06-15T11:45:00.000-05:00"}, {"value": "2000", "qualifiers": ["A"], "dateTime": "2008-06-15T12:00:00.000-05:00"}, {"value": "2000", "qualifiers": ["A"], "dateTime": "2008-06-15T12:15:00.000-05:00"}, {"value": "2000", "qualifiers": ["A"], "dateTime": "2008-06-15T12:30:00.000-05:00"}, {"value": "1990", "qualifiers": ["A"], "dateTime": "2008-06-15T12:45:00.000-05:00"}, {"value": "1990", "qualifiers": ["A"], "dateTime": "2008-06-15T13:00:00.000-05:00"}, {"value": "1980", "qualifiers": ["A"], "dateTime": "2008-06-15T13:15:00.000-05:00"}, {"value": "1990", "qualifiers": ["A"], "dateTime": "2008-06-15T13:30:00.000-05:00"}, {"value": "1980", "qualifiers": ["A"], "dateTime": "2008-06-15T13:45:00.000-05:00"}, {"value": "1990", "qualifiers": ["A"], "dateTime": "2008-06-15T14:00:00.000-05:00"}, {"value": "1970", "qualifiers": ["A"], "dateTime": "2008-06-15T14:15:00.000-05:00"}, {"value": "1970", "qualifiers": ["A"], "dateTime": "2008-06-15T14:30:00.000-05:00"}, {"value": "1970", "qualifiers": ["A"], "dateTime": "2008-06-15T14:45:00.000-05:00"}, {"value": "1970", "qualifiers": ["A"], "dateTime": "2008-06-15T15:00:00.000-05:00"}, {"value": "1960", "qualifiers": ["A"], "dateTime": "2008-06-15T15:15:00.000-05:00"}, {"value": "1970", "qualifiers": ["A"], "dateTime": "2008-06-15T15:30:00.000-05:00"}, {"value": "1970", "qualifiers": ["A"], "dateTime": "2008-06-15T15:45:00.000-05:00"}, {"value": "1970", "qualifiers": ["A"], "dateTime": "2008-06-15T16:00:00.000-05:00"}, {"value": "1980", "qualifiers": ["A"], "dateTime": "2008-06-15T16:15:00.000-05:00"}, {"value": "1970", "qualifiers": ["A"], "dateTime": "2008-06-15T16:30:00.000-05:00"}, {"value": "1970", "qualifiers": ["A"], "dateTime": "2008-06-15T16:45:00.000-05:00"}, {"value": "1960", "qualifiers": ["A"], "dateTime": "2008-06-15T17:00:00.000-05:00"}, {"value": "1970", "qualifiers": ["A"], "dateTime": "2008-06-15T17:15:00.000-05:00"}, {"value": "1980", "qualifiers": ["A"], "dateTime": "2008-06-15T17:30:00.000-05:00"}, {"value": "1970", "qualifiers": ["A"], "dateTime": "2008-06-15T17:45:00.000-05:00"}, {"value": "1980", "qualifiers": ["A"], "dateTime": "2008-06-15T18:00:00.000-05:00"}, {"value": "1980", "qualifiers": ["A"], "dateTime": "2008-06-15T18:15:00.000-05:00"}, {"value": "1980", "qualifiers": ["A"], "dateTime": "2008-06-15T18:30:00.000-05:00"}, {"value": "1990", "qualifiers": ["A"], "dateTime": "2008-06-15T18:45:00.000-05:00"}, {"value": "2000", "qualifiers": ["A"], "dateTime": "2008-06-15T19:00:00.000-05:00"}, {"value": "2000", "qualifiers": ["A"], "dateTime": "2008-06-15T19:15:00.000-05:00"}, {"value": "2000", "qualifiers": ["A"], "dateTime": "2008-06-15T19:30:00.000-05:00"}, {"value": "2000", "qualifiers": ["A"], "dateTime": "2008-06-15T19:45:00.000-05:00"}, {"value": "2010", "qualifiers": ["A"], "dateTime": "2008-06-15T20:00:00.000-05:00"}, {"value": "2020", "qualifiers": ["A"], "dateTime": "2008-06-15T20:15:00.000-05:00"}, {"value": "2020", "qualifiers": ["A"], "dateTime": "2008-06-15T20:30:00.000-05:00"}, {"value": "2020", "qualifiers": ["A"], "dateTime": "2008-06-15T20:45:00.000-05:00"}, {"value": "2030", "qualifiers": ["A"], "dateTime": "2008-06-15T21:00:00.000-05:00"}, {"value": "2040", "qualifiers": ["A"], "dateTime": "2008-06-15T21:15:00.000-05:00"}, {"value": "2040", "qualifiers": ["A"], "dateTime": "2008-06-15T21:30:00.000-05:00"}, {"value": "2040", "qualifiers": ["A"], "dateTime": "2008-06-15T21:45:00.000-05:00"}, {"value": "2050", "qualifiers": ["A"], "dateTime": "2008-06-15T22:00:00.000-05:00"}, {"value": "2050", "qualifiers": ["A"], "dateTime": "2008-06-15T22:15:00.000-05:00"}, {"value": "2060", "qualifiers": ["A"], "dateTime": "2008-06-15T22:30:00.000-05:00"}, {"value": "2070", "qualifiers": ["A"], "dateTime": "2008-06-15T22:45:00.000-05:00"}, {"value": "2070", "qualifiers": ["A"], "dateTime": "2008-06-15T23:00:00.000-05:00"}, {"value": "2070", "qualifiers": ["A"], "dateTime": "2008-06-15T23:15:00.000-05:00"}, {"value": "2080", "qualifiers": ["A"], "dateTime": "2008-06-15T23:30:00.000-05:00"}, {"value": "2080", "qualifiers": ["A"], "dateTime": "2008-06-15T23:45:00.000-05:00"}, {"value": "2100", "qualifiers": ["A"], "dateTime": "2008-06-16T00:00:00.000-05:00"}, {"value": "2090", "qualifiers": ["A"], "dateTime": "2008-06-16T00:15:00.000-05:00"}, {"value": "2110", "qualifiers": ["A"], "dateTime": "2008-06-16T00:30:00.000-05:00"}, {"value": "2100", "qualifiers": ["A"], "dateTime": "2008-06-16T00:45:00.000-05:00"}, {"value": "2100", "qualifiers": ["A"], "dateTime": "2008-06-16T01:00:00.000-05:00"}, {"value": "2100", "qualifiers": ["A"], "dateTime": "2008-06-16T01:15:00.000-05:00"}, {"value": "2110", "qualifiers": ["A"], "dateTime": "2008-06-16T01:30:00.000-05:00"}, {"value": "2110", "qualifiers": ["A"], "dateTime": "2008-06-16T01:45:00.000-05:00"}, {"value": "2110", "qualifiers": ["A"], "dateTime": "2008-06-16T02:00:00.000-05:00"}, {"value": "2110", "qualifiers": ["A"], "dateTime": "2008-06-16T02:15:00.000-05:00"}, {"value": "2120", "qualifiers": ["A"], "dateTime": "2008-06-16T02:30:00.000-05:00"}, {"value": "2110", "qualifiers": ["A"], "dateTime": "2008-06-16T02:45:00.000-05:00"}, {"value": "2110", "qualifiers": ["A"], "dateTime": "2008-06-16T03:00:00.000-05:00"}, {"value": "2110", "qualifiers": ["A"], "dateTime": "2008-06-16T03:15:00.000-05:00"}, {"value": "2110", "qualifiers": ["A"], "dateTime": "2008-06-16T03:30:00.000-05:00"}, {"value": "2130", "qualifiers": ["A"], "dateTime": "2008-06-16T03:45:00.000-05:00"}, {"value": "2130", "qualifiers": ["A"], "dateTime": "2008-06-16T04:00:00.000-05:00"}, {"value": "2130", "qualifiers": ["A"], "dateTime": "2008-06-16T04:15:00.000-05:00"}, {"value": "2120", "qualifiers": ["A"], "dateTime": "2008-06-16T04:30:00.000-05:00"}, {"value": "2130", "qualifiers": ["A"], "dateTime": "2008-06-16T04:45:00.000-05:00"}, {"value": "2130", "qualifiers": ["A"], "dateTime": "2008-06-16T05:00:00.000-05:00"}, {"value": "2130", "qualifiers": ["A"], "dateTime": "2008-06-16T05:15:00.000-05:00"}, {"value": "2130", "qualifiers": ["A"], "dateTime": "2008-06-16T05:30:00.000-05:00"}, {"value": "2130", "qualifiers": ["A"], "dateTime": "2008-06-16T05:45:00.000-05:00"}, {"value": "2130", "qualifiers": ["A"], "dateTime": "2008-06-16T06:00:00.000-05:00"}, {"value": "2130", "qualifiers": ["A"], "dateTime": "2008-06-16T06:15:00.000-05:00"}, {"value": "2120", "qualifiers": ["A"], "dateTime": "2008-06-16T06:30:00.000-05:00"}, {"value": "2130", "qualifiers": ["A"], "dateTime": "2008-06-16T06:45:00.000-05:00"}, {"value": "2130", "qualifiers": ["A"], "dateTime": "2008-06-16T07:00:00.000-05:00"}, {"value": "2130", "qualifiers": ["A"], "dateTime": "2008-06-16T07:15:00.000-05:00"}, {"value": "2120", "qualifiers": ["A"], "dateTime": "2008-06-16T07:30:00.000-05:00"}, {"value": "2130", "qualifiers": ["A"], "dateTime": "2008-06-16T07:45:00.000-05:00"}, {"value": "2130", "qualifiers": ["A"], "dateTime": "2008-06-16T08:00:00.000-05:00"}, {"value": "2120", "qualifiers": ["A"], "dateTime": "2008-06-16T08:15:00.000-05:00"}, {"value": "2120", "qualifiers": ["A"], "dateTime": "2008-06-16T08:30:00.000-05:00"}, {"value": "2130", "qualifiers": ["A"], "dateTime": "2008-06-16T08:45:00.000-05:00"}, {"value": "2110", "qualifiers": ["A"], "dateTime": "2008-06-16T09:00:00.000-05:00"}, {"value": "2130", "qualifiers": ["A"], "dateTime": "2008-06-16T09:15:00.000-05:00"}, {"value": "2130", "qualifiers": ["A"], "dateTime": "2008-06-16T09:30:00.000-05:00"}, {"value": "2110", "qualifiers": ["A"], "dateTime": "2008-06-16T09:45:00.000-05:00"}, {"value": "2120", "qualifiers": ["A"], "dateTime": "2008-06-16T10:00:00.000-05:00"}, {"value": "2130", "qualifiers": ["A"], "dateTime": "2008-06-16T10:15:00.000-05:00"}, {"value": "2110", "qualifiers": ["A"], "dateTime": "2008-06-16T10:30:00.000-05:00"}, {"value": "2120", "qualifiers": ["A"], "dateTime": "2008-06-16T10:45:00.000-05:00"}, {"value": "2110", "qualifiers": ["A"], "dateTime": "2008-06-16T11:00:00.000-05:00"}, {"value": "2100", "qualifiers": ["A"], "dateTime": "2008-06-16T11:15:00.000-05:00"}, {"value": "2100", "qualifiers": ["A"], "dateTime": "2008-06-16T11:30:00.000-05:00"}, {"value": "2110", "qualifiers": ["A"], "dateTime": "2008-06-16T11:45:00.000-05:00"}, {"value": "2100", "qualifiers": ["A"], "dateTime": "2008-06-16T12:00:00.000-05:00"}, {"value": "2090", "qualifiers": ["A"], "dateTime": "2008-06-16T12:15:00.000-05:00"}, {"value": "2100", "qualifiers": ["A"], "dateTime": "2008-06-16T12:30:00.000-05:00"}, {"value": "2090", "qualifiers": ["A"], "dateTime": "2008-06-16T12:45:00.000-05:00"}, {"value": "2070", "qualifiers": ["A"], "dateTime": "2008-06-16T13:00:00.000-05:00"}, {"value": "2090", "qualifiers": ["A"], "dateTime": "2008-06-16T13:15:00.000-05:00"}, {"value": "2070", "qualifiers": ["A"], "dateTime": "2008-06-16T13:30:00.000-05:00"}, {"value": "2070", "qualifiers": ["A"], "dateTime": "2008-06-16T13:45:00.000-05:00"}, {"value": "2070", "qualifiers": ["A"], "dateTime": "2008-06-16T14:00:00.000-05:00"}, {"value": "2070", "qualifiers": ["A"], "dateTime": "2008-06-16T14:15:00.000-05:00"}, {"value": "2070", "qualifiers": ["A"], "dateTime": "2008-06-16T14:30:00.000-05:00"}, {"value": "2040", "qualifiers": ["A"], "dateTime": "2008-06-16T14:45:00.000-05:00"}, {"value": "2060", "qualifiers": ["A"], "dateTime": "2008-06-16T15:00:00.000-05:00"}, {"value": "2050", "qualifiers": ["A"], "dateTime": "2008-06-16T15:15:00.000-05:00"}, {"value": "2040", "qualifiers": ["A"], "dateTime": "2008-06-16T15:30:00.000-05:00"}, {"value": "2030", "qualifiers": ["A"], "dateTime": "2008-06-16T15:45:00.000-05:00"}, {"value": "2040", "qualifiers": ["A"], "dateTime": "2008-06-16T16:00:00.000-05:00"}, {"value": "2030", "qualifiers": ["A"], "dateTime": "2008-06-16T16:15:00.000-05:00"}, {"value": "2030", "qualifiers": ["A"], "dateTime": "2008-06-16T16:30:00.000-05:00"}, {"value": "2020", "qualifiers": ["A"], "dateTime": "2008-06-16T16:45:00.000-05:00"}, {"value": "2000", "qualifiers": ["A"], "dateTime": "2008-06-16T17:00:00.000-05:00"}, {"value": "2000", "qualifiers": ["A"], "dateTime": "2008-06-16T17:15:00.000-05:00"}, {"value": "1990", "qualifiers": ["A"], "dateTime": "2008-06-16T17:30:00.000-05:00"}, {"value": "1980", "qualifiers": ["A"], "dateTime": "2008-06-16T17:45:00.000-05:00"}, {"value": "1980", "qualifiers": ["A"], "dateTime": "2008-06-16T18:00:00.000-05:00"}, {"value": "1970", "qualifiers": ["A"], "dateTime": "2008-06-16T18:15:00.000-05:00"}, {"value": "1950", "qualifiers": ["A"], "dateTime": "2008-06-16T18:30:00.000-05:00"}, {"value": "1950", "qualifiers": ["A"], "dateTime": "2008-06-16T18:45:00.000-05:00"}, {"value": "1950", "qualifiers": ["A"], "dateTime": "2008-06-16T19:00:00.000-05:00"}, {"value": "1940", "qualifiers": ["A"], "dateTime": "2008-06-16T19:15:00.000-05:00"}, {"value": "1920", "qualifiers": ["A"], "dateTime": "2008-06-16T19:30:00.000-05:00"}, {"value": "1920", "qualifiers": ["A"], "dateTime": "2008-06-16T19:45:00.000-05:00"}, {"value": "1910", "qualifiers": ["A"], "dateTime": "2008-06-16T20:00:00.000-05:00"}, {"value": "1910", "qualifiers": ["A"], "dateTime": "2008-06-16T20:15:00.000-05:00"}, {"value": "1900", "qualifiers": ["A"], "dateTime": "2008-06-16T20:30:00.000-05:00"}, {"value": "1880", "qualifiers": ["A"], "dateTime": "2008-06-16T20:45:00.000-05:00"}, {"value": "1880", "qualifiers": ["A"], "dateTime": "2008-06-16T21:00:00.000-05:00"}, {"value": "1870", "qualifiers": ["A"], "dateTime": "2008-06-16T21:15:00.000-05:00"}, {"value": "1860", "qualifiers": ["A"], "dateTime": "2008-06-16T21:30:00.000-05:00"}, {"value": "1850", "qualifiers": ["A"], "dateTime": "2008-06-16T21:45:00.000-05:00"}, {"value": "1830", "qualifiers": ["A"], "dateTime": "2008-06-16T22:00:00.000-05:00"}, {"value": "1830", "qualifiers": ["A"], "dateTime": "2008-06-16T22:15:00.000-05:00"}, {"value": "1820", "qualifiers": ["A"], "dateTime": "2008-06-16T22:30:00.000-05:00"}, {"value": "1810", "qualifiers": ["A"], "dateTime": "2008-06-16T22:45:00.000-05:00"}, {"value": "1800", "qualifiers": ["A"], "dateTime": "2008-06-16T23:00:00.000-05:00"}, {"value": "1780", "qualifiers": ["A"], "dateTime": "2008-06-16T23:15:00.000-05:00"}, {"value": "1770", "qualifiers": ["A"], "dateTime": "2008-06-16T23:30:00.000-05:00"}, {"value": "1760", "qualifiers": ["A"], "dateTime": "2008-06-16T23:45:00.000-05:00"}, {"value": "1760", "qualifiers": ["A"], "dateTime": "2008-06-17T00:00:00.000-05:00"}, {"value": "1740", "qualifiers": ["A"], "dateTime": "2008-06-17T00:15:00.000-05:00"}, {"value": "1730", "qualifiers": ["A"], "dateTime": "2008-06-17T00:30:00.000-05:00"}, {"value": "1710", "qualifiers": ["A"], "dateTime": "2008-06-17T00:45:00.000-05:00"}, {"value": "1710", "qualifiers": ["A"], "dateTime": "2008-06-17T01:00:00.000-05:00"}, {"value": "1690", "qualifiers": ["A"], "dateTime": "2008-06-17T01:15:00.000-05:00"}, {"value": "1680", "qualifiers": ["A"], "dateTime": "2008-06-17T01:30:00.000-05:00"}, {"value": "1680", "qualifiers": ["A"], "dateTime": "2008-06-17T01:45:00.000-05:00"}, {"value": "1660", "qualifiers": ["A"], "dateTime": "2008-06-17T02:00:00.000-05:00"}, {"value": "1660", "qualifiers": ["A"], "dateTime": "2008-06-17T02:15:00.000-05:00"}, {"value": "1640", "qualifiers": ["A"], "dateTime": "2008-06-17T02:30:00.000-05:00"}, {"value": "1630", "qualifiers": ["A"], "dateTime": "2008-06-17T02:45:00.000-05:00"}, {"value": "1620", "qualifiers": ["A"], "dateTime": "2008-06-17T03:00:00.000-05:00"}, {"value": "1610", "qualifiers": ["A"], "dateTime": "2008-06-17T03:15:00.000-05:00"}, {"value": "1610", "qualifiers": ["A"], "dateTime": "2008-06-17T03:30:00.000-05:00"}, {"value": "1590", "qualifiers": ["A"], "dateTime": "2008-06-17T03:45:00.000-05:00"}, {"value": "1580", "qualifiers": ["A"], "dateTime": "2008-06-17T04:00:00.000-05:00"}, {"value": "1570", "qualifiers": ["A"], "dateTime": "2008-06-17T04:15:00.000-05:00"}, {"value": "1560", "qualifiers": ["A"], "dateTime": "2008-06-17T04:30:00.000-05:00"}, {"value": "1560", "qualifiers": ["A"], "dateTime": "2008-06-17T04:45:00.000-05:00"}, {"value": "1550", "qualifiers": ["A"], "dateTime": "2008-06-17T05:00:00.000-05:00"}, {"value": "1530", "qualifiers": ["A"], "dateTime": "2008-06-17T05:15:00.000-05:00"}, {"value": "1520", "qualifiers": ["A"], "dateTime": "2008-06-17T05:30:00.000-05:00"}, {"value": "1520", "qualifiers": ["A"], "dateTime": "2008-06-17T05:45:00.000-05:00"}, {"value": "1500", "qualifiers": ["A"], "dateTime": "2008-06-17T06:00:00.000-05:00"}, {"value": "1500", "qualifiers": ["A"], "dateTime": "2008-06-17T06:15:00.000-05:00"}, {"value": "1490", "qualifiers": ["A"], "dateTime": "2008-06-17T06:30:00.000-05:00"}, {"value": "1480", "qualifiers": ["A"], "dateTime": "2008-06-17T06:45:00.000-05:00"}, {"value": "1470", "qualifiers": ["A"], "dateTime": "2008-06-17T07:00:00.000-05:00"}, {"value": "1470", "qualifiers": ["A"], "dateTime": "2008-06-17T07:15:00.000-05:00"}, {"value": "1460", "qualifiers": ["A"], "dateTime": "2008-06-17T07:30:00.000-05:00"}, {"value": "1450", "qualifiers": ["A"], "dateTime": "2008-06-17T07:45:00.000-05:00"}, {"value": "1440", "qualifiers": ["A"], "dateTime": "2008-06-17T08:00:00.000-05:00"}, {"value": "1440", "qualifiers": ["A"], "dateTime": "2008-06-17T08:15:00.000-05:00"}, {"value": "1440", "qualifiers": ["A"], "dateTime": "2008-06-17T08:30:00.000-05:00"}, {"value": "1420", "qualifiers": ["A"], "dateTime": "2008-06-17T08:45:00.000-05:00"}, {"value": "1410", "qualifiers": ["A"], "dateTime": "2008-06-17T09:00:00.000-05:00"}, {"value": "1400", "qualifiers": ["A"], "dateTime": "2008-06-17T09:15:00.000-05:00"}, {"value": "1400", "qualifiers": ["A"], "dateTime": "2008-06-17T09:30:00.000-05:00"}, {"value": "1390", "qualifiers": ["A"], "dateTime": "2008-06-17T09:45:00.000-05:00"}, {"value": "1380", "qualifiers": ["A"], "dateTime": "2008-06-17T10:00:00.000-05:00"}, {"value": "1380", "qualifiers": ["A"], "dateTime": "2008-06-17T10:15:00.000-05:00"}, {"value": "1380", "qualifiers": ["A"], "dateTime": "2008-06-17T10:30:00.000-05:00"}, {"value": "1360", "qualifiers": ["A"], "dateTime": "2008-06-17T10:45:00.000-05:00"}, {"value": "1370", "qualifiers": ["A"], "dateTime": "2008-06-17T11:00:00.000-05:00"}, {"value": "1360", "qualifiers": ["A"], "dateTime": "2008-06-17T11:15:00.000-05:00"}, {"value": "1350", "qualifiers": ["A"], "dateTime": "2008-06-17T11:30:00.000-05:00"}, {"value": "1340", "qualifiers": ["A"], "dateTime": "2008-06-17T11:45:00.000-05:00"}, {"value": "1330", "qualifiers": ["A"], "dateTime": "2008-06-17T12:00:00.000-05:00"}, {"value": "1320", "qualifiers": ["A"], "dateTime": "2008-06-17T12:15:00.000-05:00"}, {"value": "1330", "qualifiers": ["A"], "dateTime": "2008-06-17T12:30:00.000-05:00"}, {"value": "1320", "qualifiers": ["A"], "dateTime": "2008-06-17T12:45:00.000-05:00"}, {"value": "1310", "qualifiers": ["A"], "dateTime": "2008-06-17T13:00:00.000-05:00"}, {"value": "1310", "qualifiers": ["A"], "dateTime": "2008-06-17T13:15:00.000-05:00"}, {"value": "1300", "qualifiers": ["A"], "dateTime": "2008-06-17T13:30:00.000-05:00"}, {"value": "1300", "qualifiers": ["A"], "dateTime": "2008-06-17T13:45:00.000-05:00"}, {"value": "1290", "qualifiers": ["A"], "dateTime": "2008-06-17T14:00:00.000-05:00"}, {"value": "1280", "qualifiers": ["A"], "dateTime": "2008-06-17T14:15:00.000-05:00"}, {"value": "1280", "qualifiers": ["A"], "dateTime": "2008-06-17T14:30:00.000-05:00"}, {"value": "1280", "qualifiers": ["A"], "dateTime": "2008-06-17T14:45:00.000-05:00"}, {"value": "1270", "qualifiers": ["A"], "dateTime": "2008-06-17T15:00:00.000-05:00"}, {"value": "1260", "qualifiers": ["A"], "dateTime": "2008-06-17T15:15:00.000-05:00"}, {"value": "1260", "qualifiers": ["A"], "dateTime": "2008-06-17T15:30:00.000-05:00"}, {"value": "1260", "qualifiers": ["A"], "dateTime": "2008-06-17T15:45:00.000-05:00"}, {"value": "1250", "qualifiers": ["A"], "dateTime": "2008-06-17T16:00:00.000-05:00"}, {"value": "1240", "qualifiers": ["A"], "dateTime": "2008-06-17T16:15:00.000-05:00"}, {"value": "1230", "qualifiers": ["A"], "dateTime": "2008-06-17T16:30:00.000-05:00"}, {"value": "1230", "qualifiers": ["A"], "dateTime": "2008-06-17T16:45:00.000-05:00"}, {"value": "1220", "qualifiers": ["A"], "dateTime": "2008-06-17T17:00:00.000-05:00"}, {"value": "1220", "qualifiers": ["A"], "dateTime": "2008-06-17T17:15:00.000-05:00"}, {"value": "1220", "qualifiers": ["A"], "dateTime": "2008-06-17T17:30:00.000-05:00"}, {"value": "1210", "qualifiers": ["A"], "dateTime": "2008-06-17T17:45:00.000-05:00"}, {"value": "1210", "qualifiers": ["A"], "dateTime": "2008-06-17T18:00:00.000-05:00"}, {"value": "1210", "qualifiers": ["A"], "dateTime": "2008-06-17T18:15:00.000-05:00"}, {"value": "1210", "qualifiers": ["A"], "dateTime": "2008-06-17T18:30:00.000-05:00"}, {"value": "1200", "qualifiers": ["A"], "dateTime": "2008-06-17T18:45:00.000-05:00"}, {"value": "1190", "qualifiers": ["A"], "dateTime": "2008-06-17T19:00:00.000-05:00"}, {"value": "1190", "qualifiers": ["A"], "dateTime": "2008-06-17T19:15:00.000-05:00"}, {"value": "1190", "qualifiers": ["A"], "dateTime": "2008-06-17T19:30:00.000-05:00"}, {"value": "1190", "qualifiers": ["A"], "dateTime": "2008-06-17T19:45:00.000-05:00"}, {"value": "1180", "qualifiers": ["A"], "dateTime": "2008-06-17T20:00:00.000-05:00"}, {"value": "1180", "qualifiers": ["A"], "dateTime": "2008-06-17T20:15:00.000-05:00"}, {"value": "1170", "qualifiers": ["A"], "dateTime": "2008-06-17T20:30:00.000-05:00"}, {"value": "1170", "qualifiers": ["A"], "dateTime": "2008-06-17T20:45:00.000-05:00"}, {"value": "1170", "qualifiers": ["A"], "dateTime": "2008-06-17T21:00:00.000-05:00"}, {"value": "1160", "qualifiers": ["A"], "dateTime": "2008-06-17T21:15:00.000-05:00"}, {"value": "1160", "qualifiers": ["A"], "dateTime": "2008-06-17T21:30:00.000-05:00"}, {"value": "1160", "qualifiers": ["A"], "dateTime": "2008-06-17T21:45:00.000-05:00"}, {"value": "1160", "qualifiers": ["A"], "dateTime": "2008-06-17T22:00:00.000-05:00"}, {"value": "1160", "qualifiers": ["A"], "dateTime": "2008-06-17T22:15:00.000-05:00"}, {"value": "1150", "qualifiers": ["A"], "dateTime": "2008-06-17T22:30:00.000-05:00"}, {"value": "1150", "qualifiers": ["A"], "dateTime": "2008-06-17T22:45:00.000-05:00"}, {"value": "1150", "qualifiers": ["A"], "dateTime": "2008-06-17T23:00:00.000-05:00"}, {"value": "1140", "qualifiers": ["A"], "dateTime": "2008-06-17T23:15:00.000-05:00"}, {"value": "1140", "qualifiers": ["A"], "dateTime": "2008-06-17T23:30:00.000-05:00"}, {"value": "1140", "qualifiers": ["A"], "dateTime": "2008-06-17T23:45:00.000-05:00"}, {"value": "1130", "qualifiers": ["A"], "dateTime": "2008-06-18T00:00:00.000-05:00"}, {"value": "1140", "qualifiers": ["A"], "dateTime": "2008-06-18T00:15:00.000-05:00"}, {"value": "1130", "qualifiers": ["A"], "dateTime": "2008-06-18T00:30:00.000-05:00"}, {"value": "1130", "qualifiers": ["A"], "dateTime": "2008-06-18T00:45:00.000-05:00"}, {"value": "1120", "qualifiers": ["A"], "dateTime": "2008-06-18T01:00:00.000-05:00"}, {"value": "1130", "qualifiers": ["A"], "dateTime": "2008-06-18T01:15:00.000-05:00"}, {"value": "1130", "qualifiers": ["A"], "dateTime": "2008-06-18T01:30:00.000-05:00"}, {"value": "1120", "qualifiers": ["A"], "dateTime": "2008-06-18T01:45:00.000-05:00"}, {"value": "1120", "qualifiers": ["A"], "dateTime": "2008-06-18T02:00:00.000-05:00"}, {"value": "1120", "qualifiers": ["A"], "dateTime": "2008-06-18T02:15:00.000-05:00"}, {"value": "1120", "qualifiers": ["A"], "dateTime": "2008-06-18T02:30:00.000-05:00"}, {"value": "1110", "qualifiers": ["A"], "dateTime": "2008-06-18T02:45:00.000-05:00"}, {"value": "1110", "qualifiers": ["A"], "dateTime": "2008-06-18T03:00:00.000-05:00"}, {"value": "1110", "qualifiers": ["A"], "dateTime": "2008-06-18T03:15:00.000-05:00"}, {"value": "1100", "qualifiers": ["A"], "dateTime": "2008-06-18T03:30:00.000-05:00"}, {"value": "1100", "qualifiers": ["A"], "dateTime": "2008-06-18T03:45:00.000-05:00"}, {"value": "1100", "qualifiers": ["A"], "dateTime": "2008-06-18T04:00:00.000-05:00"}, {"value": "1100", "qualifiers": ["A"], "dateTime": "2008-06-18T04:15:00.000-05:00"}, {"value": "1100", "qualifiers": ["A"], "dateTime": "2008-06-18T04:30:00.000-05:00"}, {"value": "1100", "qualifiers": ["A"], "dateTime": "2008-06-18T04:45:00.000-05:00"}, {"value": "1100", "qualifiers": ["A"], "dateTime": "2008-06-18T05:00:00.000-05:00"}, {"value": "1090", "qualifiers": ["A"], "dateTime": "2008-06-18T05:15:00.000-05:00"}, {"value": "1090", "qualifiers": ["A"], "dateTime": "2008-06-18T05:30:00.000-05:00"}, {"value": "1090", "qualifiers": ["A"], "dateTime": "2008-06-18T05:45:00.000-05:00"}, {"value": "1080", "qualifiers": ["A"], "dateTime": "2008-06-18T06:00:00.000-05:00"}, {"value": "1080", "qualifiers": ["A"], "dateTime": "2008-06-18T06:15:00.000-05:00"}, {"value": "1080", "qualifiers": ["A"], "dateTime": "2008-06-18T06:30:00.000-05:00"}, {"value": "1080", "qualifiers": ["A"], "dateTime": "2008-06-18T06:45:00.000-05:00"}, {"value": "1080", "qualifiers": ["A"], "dateTime": "2008-06-18T07:00:00.000-05:00"}, {"value": "1080", "qualifiers": ["A"], "dateTime": "2008-06-18T07:15:00.000-05:00"}, {"value": "1080", "qualifiers": ["A"], "dateTime": "2008-06-18T07:30:00.000-05:00"}, {"value": "1070", "qualifiers": ["A"], "dateTime": "2008-06-18T07:45:00.000-05:00"}, {"value": "1070", "qualifiers": ["A"], "dateTime": "2008-06-18T08:00:00.000-05:00"}, {"value": "1070", "qualifiers": ["A"], "dateTime": "2008-06-18T08:15:00.000-05:00"}, {"value": "1070", "qualifiers": ["A"], "dateTime": "2008-06-18T08:30:00.000-05:00"}, {"value": "1060", "qualifiers": ["A"], "dateTime": "2008-06-18T08:45:00.000-05:00"}, {"value": "1060", "qualifiers": ["A"], "dateTime": "2008-06-18T09:00:00.000-05:00"}, {"value": "1060", "qualifiers": ["A"], "dateTime": "2008-06-18T09:15:00.000-05:00"}, {"value": "1060", "qualifiers": ["A"], "dateTime": "2008-06-18T09:30:00.000-05:00"}, {"value": "1060", "qualifiers": ["A"], "dateTime": "2008-06-18T09:45:00.000-05:00"}, {"value": "1060", "qualifiers": ["A"], "dateTime": "2008-06-18T10:00:00.000-05:00"}, {"value": "1060", "qualifiers": ["A"], "dateTime": "2008-06-18T10:15:00.000-05:00"}, {"value": "1060", "qualifiers": ["A"], "dateTime": "2008-06-18T10:30:00.000-05:00"}, {"value": "1060", "qualifiers": ["A"], "dateTime": "2008-06-18T10:45:00.000-05:00"}, {"value": "1050", "qualifiers": ["A"], "dateTime": "2008-06-18T11:00:00.000-05:00"}, {"value": "1050", "qualifiers": ["A"], "dateTime": "2008-06-18T11:15:00.000-05:00"}, {"value": "1050", "qualifiers": ["A"], "dateTime": "2008-06-18T11:30:00.000-05:00"}, {"value": "1040", "qualifiers": ["A"], "dateTime": "2008-06-18T11:45:00.000-05:00"}, {"value": "1050", "qualifiers": ["A"], "dateTime": "2008-06-18T12:00:00.000-05:00"}, {"value": "1040", "qualifiers": ["A"], "dateTime": "2008-06-18T12:15:00.000-05:00"}, {"value": "1040", "qualifiers": ["A"], "dateTime": "2008-06-18T12:30:00.000-05:00"}, {"value": "1040", "qualifiers": ["A"], "dateTime": "2008-06-18T12:45:00.000-05:00"}, {"value": "1040", "qualifiers": ["A"], "dateTime": "2008-06-18T13:00:00.000-05:00"}, {"value": "1040", "qualifiers": ["A"], "dateTime": "2008-06-18T13:15:00.000-05:00"}, {"value": "1030", "qualifiers": ["A"], "dateTime": "2008-06-18T13:30:00.000-05:00"}, {"value": "1030", "qualifiers": ["A"], "dateTime": "2008-06-18T13:45:00.000-05:00"}, {"value": "1030", "qualifiers": ["A"], "dateTime": "2008-06-18T14:00:00.000-05:00"}, {"value": "1030", "qualifiers": ["A"], "dateTime": "2008-06-18T14:15:00.000-05:00"}, {"value": "1030", "qualifiers": ["A"], "dateTime": "2008-06-18T14:30:00.000-05:00"}, {"value": "1020", "qualifiers": ["A"], "dateTime": "2008-06-18T14:45:00.000-05:00"}, {"value": "1020", "qualifiers": ["A"], "dateTime": "2008-06-18T15:00:00.000-05:00"}, {"value": "1020", "qualifiers": ["A"], "dateTime": "2008-06-18T15:15:00.000-05:00"}, {"value": "1020", "qualifiers": ["A"], "dateTime": "2008-06-18T15:30:00.000-05:00"}, {"value": "1020", "qualifiers": ["A"], "dateTime": "2008-06-18T15:45:00.000-05:00"}, {"value": "1020", "qualifiers": ["A"], "dateTime": "2008-06-18T16:00:00.000-05:00"}, {"value": "1010", "qualifiers": ["A"], "dateTime": "2008-06-18T16:15:00.000-05:00"}, {"value": "1010", "qualifiers": ["A"], "dateTime": "2008-06-18T16:30:00.000-05:00"}, {"value": "1010", "qualifiers": ["A"], "dateTime": "2008-06-18T16:45:00.000-05:00"}, {"value": "1010", "qualifiers": ["A"], "dateTime": "2008-06-18T17:00:00.000-05:00"}, {"value": "1010", "qualifiers": ["A"], "dateTime": "2008-06-18T17:15:00.000-05:00"}, {"value": "1010", "qualifiers": ["A"], "dateTime": "2008-06-18T17:30:00.000-05:00"}, {"value": "1000", "qualifiers": ["A"], "dateTime": "2008-06-18T17:45:00.000-05:00"}, {"value": "1000", "qualifiers": ["A"], "dateTime": "2008-06-18T18:00:00.000-05:00"}, {"value": "998", "qualifiers": ["A"], "dateTime": "2008-06-18T18:15:00.000-05:00"}, {"value": "998", "qualifiers": ["A"], "dateTime": "2008-06-18T18:30:00.000-05:00"}, {"value": "1000", "qualifiers": ["A"], "dateTime": "2008-06-18T18:45:00.000-05:00"}, {"value": "998", "qualifiers": ["A"], "dateTime": "2008-06-18T19:00:00.000-05:00"}, {"value": "998", "qualifiers": ["A"], "dateTime": "2008-06-18T19:15:00.000-05:00"}, {"value": "995", "qualifiers": ["A"], "dateTime": "2008-06-18T19:30:00.000-05:00"}, {"value": "995", "qualifiers": ["A"], "dateTime": "2008-06-18T19:45:00.000-05:00"}, {"value": "992", "qualifiers": ["A"], "dateTime": "2008-06-18T20:00:00.000-05:00"}, {"value": "990", "qualifiers": ["A"], "dateTime": "2008-06-18T20:15:00.000-05:00"}, {"value": "992", "qualifiers": ["A"], "dateTime": "2008-06-18T20:30:00.000-05:00"}, {"value": "990", "qualifiers": ["A"], "dateTime": "2008-06-18T20:45:00.000-05:00"}, {"value": "984", "qualifiers": ["A"], "dateTime": "2008-06-18T21:00:00.000-05:00"}, {"value": "984", "qualifiers": ["A"], "dateTime": "2008-06-18T21:15:00.000-05:00"}, {"value": "987", "qualifiers": ["A"], "dateTime": "2008-06-18T21:30:00.000-05:00"}, {"value": "981", "qualifiers": ["A"], "dateTime": "2008-06-18T21:45:00.000-05:00"}, {"value": "978", "qualifiers": ["A"], "dateTime": "2008-06-18T22:00:00.000-05:00"}, {"value": "978", "qualifiers": ["A"], "dateTime": "2008-06-18T22:15:00.000-05:00"}, {"value": "978", "qualifiers": ["A"], "dateTime": "2008-06-18T22:30:00.000-05:00"}, {"value": "973", "qualifiers": ["A"], "dateTime": "2008-06-18T22:45:00.000-05:00"}, {"value": "973", "qualifiers": ["A"], "dateTime": "2008-06-18T23:00:00.000-05:00"}, {"value": "976", "qualifiers": ["A"], "dateTime": "2008-06-18T23:15:00.000-05:00"}, {"value": "973", "qualifiers": ["A"], "dateTime": "2008-06-18T23:30:00.000-05:00"}, {"value": "967", "qualifiers": ["A"], "dateTime": "2008-06-18T23:45:00.000-05:00"}, {"value": "967", "qualifiers": ["A"], "dateTime": "2008-06-19T00:00:00.000-05:00"}, {"value": "965", "qualifiers": ["A"], "dateTime": "2008-06-19T00:15:00.000-05:00"}, {"value": "967", "qualifiers": ["A"], "dateTime": "2008-06-19T00:30:00.000-05:00"}, {"value": "962", "qualifiers": ["A"], "dateTime": "2008-06-19T00:45:00.000-05:00"}, {"value": "962", "qualifiers": ["A"], "dateTime": "2008-06-19T01:00:00.000-05:00"}, {"value": "959", "qualifiers": ["A"], "dateTime": "2008-06-19T01:15:00.000-05:00"}, {"value": "965", "qualifiers": ["A"], "dateTime": "2008-06-19T01:30:00.000-05:00"}, {"value": "962", "qualifiers": ["A"], "dateTime": "2008-06-19T01:45:00.000-05:00"}, {"value": "954", "qualifiers": ["A"], "dateTime": "2008-06-19T02:00:00.000-05:00"}, {"value": "954", "qualifiers": ["A"], "dateTime": "2008-06-19T02:15:00.000-05:00"}, {"value": "956", "qualifiers": ["A"], "dateTime": "2008-06-19T02:30:00.000-05:00"}, {"value": "954", "qualifiers": ["A"], "dateTime": "2008-06-19T02:45:00.000-05:00"}, {"value": "954", "qualifiers": ["A"], "dateTime": "2008-06-19T03:00:00.000-05:00"}, {"value": "954", "qualifiers": ["A"], "dateTime": "2008-06-19T03:15:00.000-05:00"}, {"value": "954", "qualifiers": ["A"], "dateTime": "2008-06-19T03:30:00.000-05:00"}, {"value": "948", "qualifiers": ["A"], "dateTime": "2008-06-19T03:45:00.000-05:00"}, {"value": "954", "qualifiers": ["A"], "dateTime": "2008-06-19T04:00:00.000-05:00"}, {"value": "948", "qualifiers": ["A"], "dateTime": "2008-06-19T04:15:00.000-05:00"}, {"value": "945", "qualifiers": ["A"], "dateTime": "2008-06-19T04:30:00.000-05:00"}, {"value": "948", "qualifiers": ["A"], "dateTime": "2008-06-19T04:45:00.000-05:00"}, {"value": "943", "qualifiers": ["A"], "dateTime": "2008-06-19T05:00:00.000-05:00"}, {"value": "945", "qualifiers": ["A"], "dateTime": "2008-06-19T05:15:00.000-05:00"}, {"value": "937", "qualifiers": ["A"], "dateTime": "2008-06-19T05:30:00.000-05:00"}, {"value": "937", "qualifiers": ["A"], "dateTime": "2008-06-19T05:45:00.000-05:00"}, {"value": "937", "qualifiers": ["A"], "dateTime": "2008-06-19T06:00:00.000-05:00"}, {"value": "940", "qualifiers": ["A"], "dateTime": "2008-06-19T06:15:00.000-05:00"}, {"value": "932", "qualifiers": ["A"], "dateTime": "2008-06-19T06:30:00.000-05:00"}, {"value": "932", "qualifiers": ["A"], "dateTime": "2008-06-19T06:45:00.000-05:00"}, {"value": "937", "qualifiers": ["A"], "dateTime": "2008-06-19T07:00:00.000-05:00"}, {"value": "932", "qualifiers": ["A"], "dateTime": "2008-06-19T07:15:00.000-05:00"}, {"value": "929", "qualifiers": ["A"], "dateTime": "2008-06-19T07:30:00.000-05:00"}, {"value": "926", "qualifiers": ["A"], "dateTime": "2008-06-19T07:45:00.000-05:00"}, {"value": "932", "qualifiers": ["A"], "dateTime": "2008-06-19T08:00:00.000-05:00"}, {"value": "929", "qualifiers": ["A"], "dateTime": "2008-06-19T08:15:00.000-05:00"}, {"value": "929", "qualifiers": ["A"], "dateTime": "2008-06-19T08:30:00.000-05:00"}, {"value": "923", "qualifiers": ["A"], "dateTime": "2008-06-19T08:45:00.000-05:00"}, {"value": "926", "qualifiers": ["A"], "dateTime": "2008-06-19T09:00:00.000-05:00"}, {"value": "918", "qualifiers": ["A"], "dateTime": "2008-06-19T09:15:00.000-05:00"}, {"value": "923", "qualifiers": ["A"], "dateTime": "2008-06-19T09:30:00.000-05:00"}, {"value": "921", "qualifiers": ["A"], "dateTime": "2008-06-19T09:45:00.000-05:00"}, {"value": "921", "qualifiers": ["A"], "dateTime": "2008-06-19T10:00:00.000-05:00"}, {"value": "913", "qualifiers": ["A"], "dateTime": "2008-06-19T10:15:00.000-05:00"}, {"value": "913", "qualifiers": ["A"], "dateTime": "2008-06-19T10:30:00.000-05:00"}, {"value": "918", "qualifiers": ["A"], "dateTime": "2008-06-19T10:45:00.000-05:00"}, {"value": "913", "qualifiers": ["A"], "dateTime": "2008-06-19T11:00:00.000-05:00"}, {"value": "910", "qualifiers": ["A"], "dateTime": "2008-06-19T11:15:00.000-05:00"}, {"value": "910", "qualifiers": ["A"], "dateTime": "2008-06-19T11:30:00.000-05:00"}, {"value": "910", "qualifiers": ["A"], "dateTime": "2008-06-19T11:45:00.000-05:00"}, {"value": "915", "qualifiers": ["A"], "dateTime": "2008-06-19T12:00:00.000-05:00"}, {"value": "910", "qualifiers": ["A"], "dateTime": "2008-06-19T12:15:00.000-05:00"}, {"value": "910", "qualifiers": ["A"], "dateTime": "2008-06-19T12:30:00.000-05:00"}, {"value": "913", "qualifiers": ["A"], "dateTime": "2008-06-19T12:45:00.000-05:00"}, {"value": "905", "qualifiers": ["A"], "dateTime": "2008-06-19T13:00:00.000-05:00"}, {"value": "905", "qualifiers": ["A"], "dateTime": "2008-06-19T13:15:00.000-05:00"}, {"value": "907", "qualifiers": ["A"], "dateTime": "2008-06-19T13:30:00.000-05:00"}, {"value": "902", "qualifiers": ["A"], "dateTime": "2008-06-19T13:45:00.000-05:00"}, {"value": "905", "qualifiers": ["A"], "dateTime": "2008-06-19T14:00:00.000-05:00"}, {"value": "902", "qualifiers": ["A"], "dateTime": "2008-06-19T14:15:00.000-05:00"}, {"value": "902", "qualifiers": ["A"], "dateTime": "2008-06-19T14:30:00.000-05:00"}, {"value": "902", "qualifiers": ["A"], "dateTime": "2008-06-19T14:45:00.000-05:00"}, {"value": "902", "qualifiers": ["A"], "dateTime": "2008-06-19T15:00:00.000-05:00"}, {"value": "894", "qualifiers": ["A"], "dateTime": "2008-06-19T15:15:00.000-05:00"}, {"value": "902", "qualifiers": ["A"], "dateTime": "2008-06-19T15:30:00.000-05:00"}, {"value": "892", "qualifiers": ["A"], "dateTime": "2008-06-19T15:45:00.000-05:00"}, {"value": "899", "qualifiers": ["A"], "dateTime": "2008-06-19T16:00:00.000-05:00"}, {"value": "892", "qualifiers": ["A"], "dateTime": "2008-06-19T16:15:00.000-05:00"}, {"value": "894", "qualifiers": ["A"], "dateTime": "2008-06-19T16:30:00.000-05:00"}, {"value": "886", "qualifiers": ["A"], "dateTime": "2008-06-19T16:45:00.000-05:00"}, {"value": "892", "qualifiers": ["A"], "dateTime": "2008-06-19T17:00:00.000-05:00"}, {"value": "889", "qualifiers": ["A"], "dateTime": "2008-06-19T17:15:00.000-05:00"}, {"value": "889", "qualifiers": ["A"], "dateTime": "2008-06-19T17:30:00.000-05:00"}, {"value": "884", "qualifiers": ["A"], "dateTime": "2008-06-19T17:45:00.000-05:00"}, {"value": "881", "qualifiers": ["A"], "dateTime": "2008-06-19T18:00:00.000-05:00"}, {"value": "881", "qualifiers": ["A"], "dateTime": "2008-06-19T18:15:00.000-05:00"}, {"value": "884", "qualifiers": ["A"], "dateTime": "2008-06-19T18:30:00.000-05:00"}, {"value": "879", "qualifiers": ["A"], "dateTime": "2008-06-19T18:45:00.000-05:00"}, {"value": "881", "qualifiers": ["A"], "dateTime": "2008-06-19T19:00:00.000-05:00"}, {"value": "884", "qualifiers": ["A"], "dateTime": "2008-06-19T19:15:00.000-05:00"}, {"value": "877", "qualifiers": ["A"], "dateTime": "2008-06-19T19:30:00.000-05:00"}, {"value": "881", "qualifiers": ["A"], "dateTime": "2008-06-19T19:45:00.000-05:00"}, {"value": "874", "qualifiers": ["A"], "dateTime": "2008-06-19T20:00:00.000-05:00"}, {"value": "874", "qualifiers": ["A"], "dateTime": "2008-06-19T20:15:00.000-05:00"}, {"value": "874", "qualifiers": ["A"], "dateTime": "2008-06-19T20:30:00.000-05:00"}, {"value": "874", "qualifiers": ["A"], "dateTime": "2008-06-19T20:45:00.000-05:00"}, {"value": "877", "qualifiers": ["A"], "dateTime": "2008-06-19T21:00:00.000-05:00"}, {"value": "869", "qualifiers": ["A"], "dateTime": "2008-06-19T21:15:00.000-05:00"}, {"value": "872", "qualifiers": ["A"], "dateTime": "2008-06-19T21:30:00.000-05:00"}, {"value": "867", "qualifiers": ["A"], "dateTime": "2008-06-19T21:45:00.000-05:00"}, {"value": "869", "qualifiers": ["A"], "dateTime": "2008-06-19T22:00:00.000-05:00"}, {"value": "869", "qualifiers": ["A"], "dateTime": "2008-06-19T22:15:00.000-05:00"}, {"value": "867", "qualifiers": ["A"], "dateTime": "2008-06-19T22:30:00.000-05:00"}, {"value": "867", "qualifiers": ["A"], "dateTime": "2008-06-19T22:45:00.000-05:00"}, {"value": "867", "qualifiers": ["A"], "dateTime": "2008-06-19T23:00:00.000-05:00"}, {"value": "864", "qualifiers": ["A"], "dateTime": "2008-06-19T23:15:00.000-05:00"}, {"value": "864", "qualifiers": ["A"], "dateTime": "2008-06-19T23:30:00.000-05:00"}, {"value": "859", "qualifiers": ["A"], "dateTime": "2008-06-19T23:45:00.000-05:00"}, {"value": "862", "qualifiers": ["A"], "dateTime": "2008-06-20T00:00:00.000-05:00"}, {"value": "862", "qualifiers": ["A"], "dateTime": "2008-06-20T00:15:00.000-05:00"}, {"value": "862", "qualifiers": ["A"], "dateTime": "2008-06-20T00:30:00.000-05:00"}, {"value": "857", "qualifiers": ["A"], "dateTime": "2008-06-20T00:45:00.000-05:00"}, {"value": "857", "qualifiers": ["A"], "dateTime": "2008-06-20T01:00:00.000-05:00"}, {"value": "854", "qualifiers": ["A"], "dateTime": "2008-06-20T01:15:00.000-05:00"}, {"value": "859", "qualifiers": ["A"], "dateTime": "2008-06-20T01:30:00.000-05:00"}, {"value": "852", "qualifiers": ["A"], "dateTime": "2008-06-20T01:45:00.000-05:00"}, {"value": "857", "qualifiers": ["A"], "dateTime": "2008-06-20T02:00:00.000-05:00"}, {"value": "854", "qualifiers": ["A"], "dateTime": "2008-06-20T02:15:00.000-05:00"}, {"value": "854", "qualifiers": ["A"], "dateTime": "2008-06-20T02:30:00.000-05:00"}, {"value": "849", "qualifiers": ["A"], "dateTime": "2008-06-20T02:45:00.000-05:00"}, {"value": "854", "qualifiers": ["A"], "dateTime": "2008-06-20T03:00:00.000-05:00"}, {"value": "849", "qualifiers": ["A"], "dateTime": "2008-06-20T03:15:00.000-05:00"}, {"value": "849", "qualifiers": ["A"], "dateTime": "2008-06-20T03:30:00.000-05:00"}, {"value": "849", "qualifiers": ["A"], "dateTime": "2008-06-20T03:45:00.000-05:00"}, {"value": "849", "qualifiers": ["A"], "dateTime": "2008-06-20T04:00:00.000-05:00"}, {"value": "847", "qualifiers": ["A"], "dateTime": "2008-06-20T04:15:00.000-05:00"}, {"value": "847", "qualifiers": ["A"], "dateTime": "2008-06-20T04:30:00.000-05:00"}, {"value": "847", "qualifiers": ["A"], "dateTime": "2008-06-20T04:45:00.000-05:00"}, {"value": "847", "qualifiers": ["A"], "dateTime": "2008-06-20T05:00:00.000-05:00"}, {"value": "842", "qualifiers": ["A"], "dateTime": "2008-06-20T05:15:00.000-05:00"}, {"value": "842", "qualifiers": ["A"], "dateTime": "2008-06-20T05:30:00.000-05:00"}, {"value": "842", "qualifiers": ["A"], "dateTime": "2008-06-20T05:45:00.000-05:00"}, {"value": "840", "qualifiers": ["A"], "dateTime": "2008-06-20T06:00:00.000-05:00"}, {"value": "840", "qualifiers": ["A"], "dateTime": "2008-06-20T06:15:00.000-05:00"}, {"value": "840", "qualifiers": ["A"], "dateTime": "2008-06-20T06:30:00.000-05:00"}, {"value": "837", "qualifiers": ["A"], "dateTime": "2008-06-20T06:45:00.000-05:00"}, {"value": "840", "qualifiers": ["A"], "dateTime": "2008-06-20T07:00:00.000-05:00"}, {"value": "835", "qualifiers": ["A"], "dateTime": "2008-06-20T07:15:00.000-05:00"}, {"value": "837", "qualifiers": ["A"], "dateTime": "2008-06-20T07:30:00.000-05:00"}, {"value": "832", "qualifiers": ["A"], "dateTime": "2008-06-20T07:45:00.000-05:00"}, {"value": "837", "qualifiers": ["A"], "dateTime": "2008-06-20T08:00:00.000-05:00"}, {"value": "837", "qualifiers": ["A"], "dateTime": "2008-06-20T08:15:00.000-05:00"}, {"value": "832", "qualifiers": ["A"], "dateTime": "2008-06-20T08:30:00.000-05:00"}, {"value": "835", "qualifiers": ["A"], "dateTime": "2008-06-20T08:45:00.000-05:00"}, {"value": "835", "qualifiers": ["A"], "dateTime": "2008-06-20T09:00:00.000-05:00"}, {"value": "827", "qualifiers": ["A"], "dateTime": "2008-06-20T09:15:00.000-05:00"}, {"value": "827", "qualifiers": ["A"], "dateTime": "2008-06-20T09:30:00.000-05:00"}, {"value": "832", "qualifiers": ["A"], "dateTime": "2008-06-20T09:45:00.000-05:00"}, {"value": "830", "qualifiers": ["A"], "dateTime": "2008-06-20T10:00:00.000-05:00"}, {"value": "830", "qualifiers": ["A"], "dateTime": "2008-06-20T10:15:00.000-05:00"}, {"value": "830", "qualifiers": ["A"], "dateTime": "2008-06-20T10:30:00.000-05:00"}, {"value": "825", "qualifiers": ["A"], "dateTime": "2008-06-20T10:45:00.000-05:00"}, {"value": "827", "qualifiers": ["A"], "dateTime": "2008-06-20T11:00:00.000-05:00"}, {"value": "827", "qualifiers": ["A"], "dateTime": "2008-06-20T11:15:00.000-05:00"}, {"value": "827", "qualifiers": ["A"], "dateTime": "2008-06-20T11:30:00.000-05:00"}, {"value": "825", "qualifiers": ["A"], "dateTime": "2008-06-20T11:45:00.000-05:00"}, {"value": "823", "qualifiers": ["A"], "dateTime": "2008-06-20T12:00:00.000-05:00"}, {"value": "823", "qualifiers": ["A"], "dateTime": "2008-06-20T12:15:00.000-05:00"}, {"value": "823", "qualifiers": ["A"], "dateTime": "2008-06-20T12:30:00.000-05:00"}, {"value": "823", "qualifiers": ["A"], "dateTime": "2008-06-20T12:45:00.000-05:00"}, {"value": "815", "qualifiers": ["A"], "dateTime": "2008-06-20T13:00:00.000-05:00"}, {"value": "823", "qualifiers": ["A"], "dateTime": "2008-06-20T13:15:00.000-05:00"}, {"value": "815", "qualifiers": ["A"], "dateTime": "2008-06-20T13:30:00.000-05:00"}, {"value": "820", "qualifiers": ["A"], "dateTime": "2008-06-20T13:45:00.000-05:00"}, {"value": "820", "qualifiers": ["A"], "dateTime": "2008-06-20T14:00:00.000-05:00"}, {"value": "820", "qualifiers": ["A"], "dateTime": "2008-06-20T14:15:00.000-05:00"}, {"value": "820", "qualifiers": ["A"], "dateTime": "2008-06-20T14:30:00.000-05:00"}, {"value": "813", "qualifiers": ["A"], "dateTime": "2008-06-20T14:45:00.000-05:00"}, {"value": "813", "qualifiers": ["A"], "dateTime": "2008-06-20T15:00:00.000-05:00"}, {"value": "815", "qualifiers": ["A"], "dateTime": "2008-06-20T15:15:00.000-05:00"}, {"value": "806", "qualifiers": ["A"], "dateTime": "2008-06-20T15:30:00.000-05:00"}, {"value": "813", "qualifiers": ["A"], "dateTime": "2008-06-20T15:45:00.000-05:00"}, {"value": "808", "qualifiers": ["A"], "dateTime": "2008-06-20T16:00:00.000-05:00"}, {"value": "808", "qualifiers": ["A"], "dateTime": "2008-06-20T16:15:00.000-05:00"}, {"value": "810", "qualifiers": ["A"], "dateTime": "2008-06-20T16:30:00.000-05:00"}, {"value": "803", "qualifiers": ["A"], "dateTime": "2008-06-20T16:45:00.000-05:00"}, {"value": "808", "qualifiers": ["A"], "dateTime": "2008-06-20T17:00:00.000-05:00"}, {"value": "801", "qualifiers": ["A"], "dateTime": "2008-06-20T17:15:00.000-05:00"}, {"value": "806", "qualifiers": ["A"], "dateTime": "2008-06-20T17:30:00.000-05:00"}, {"value": "801", "qualifiers": ["A"], "dateTime": "2008-06-20T17:45:00.000-05:00"}, {"value": "803", "qualifiers": ["A"], "dateTime": "2008-06-20T18:00:00.000-05:00"}, {"value": "799", "qualifiers": ["A"], "dateTime": "2008-06-20T18:15:00.000-05:00"}, {"value": "799", "qualifiers": ["A"], "dateTime": "2008-06-20T18:30:00.000-05:00"}, {"value": "803", "qualifiers": ["A"], "dateTime": "2008-06-20T18:45:00.000-05:00"}, {"value": "796", "qualifiers": ["A"], "dateTime": "2008-06-20T19:00:00.000-05:00"}, {"value": "801", "qualifiers": ["A"], "dateTime": "2008-06-20T19:15:00.000-05:00"}, {"value": "799", "qualifiers": ["A"], "dateTime": "2008-06-20T19:30:00.000-05:00"}, {"value": "796", "qualifiers": ["A"], "dateTime": "2008-06-20T19:45:00.000-05:00"}, {"value": "799", "qualifiers": ["A"], "dateTime": "2008-06-20T20:00:00.000-05:00"}, {"value": "794", "qualifiers": ["A"], "dateTime": "2008-06-20T20:15:00.000-05:00"}, {"value": "799", "qualifiers": ["A"], "dateTime": "2008-06-20T20:30:00.000-05:00"}, {"value": "796", "qualifiers": ["A"], "dateTime": "2008-06-20T20:45:00.000-05:00"}, {"value": "794", "qualifiers": ["A"], "dateTime": "2008-06-20T21:00:00.000-05:00"}, {"value": "789", "qualifiers": ["A"], "dateTime": "2008-06-20T21:15:00.000-05:00"}, {"value": "789", "qualifiers": ["A"], "dateTime": "2008-06-20T21:30:00.000-05:00"}, {"value": "789", "qualifiers": ["A"], "dateTime": "2008-06-20T21:45:00.000-05:00"}, {"value": "789", "qualifiers": ["A"], "dateTime": "2008-06-20T22:00:00.000-05:00"}, {"value": "789", "qualifiers": ["A"], "dateTime": "2008-06-20T22:15:00.000-05:00"}, {"value": "791", "qualifiers": ["A"], "dateTime": "2008-06-20T22:30:00.000-05:00"}, {"value": "784", "qualifiers": ["A"], "dateTime": "2008-06-20T22:45:00.000-05:00"}, {"value": "787", "qualifiers": ["A"], "dateTime": "2008-06-20T23:00:00.000-05:00"}, {"value": "784", "qualifiers": ["A"], "dateTime": "2008-06-20T23:15:00.000-05:00"}, {"value": "782", "qualifiers": ["A"], "dateTime": "2008-06-20T23:30:00.000-05:00"}, {"value": "780", "qualifiers": ["A"], "dateTime": "2008-06-20T23:45:00.000-05:00"}, {"value": "782", "qualifiers": ["A"], "dateTime": "2008-06-21T00:00:00.000-05:00"}, {"value": "780", "qualifiers": ["A"], "dateTime": "2008-06-21T00:15:00.000-05:00"}, {"value": "775", "qualifiers": ["A"], "dateTime": "2008-06-21T00:30:00.000-05:00"}, {"value": "782", "qualifiers": ["A"], "dateTime": "2008-06-21T00:45:00.000-05:00"}, {"value": "780", "qualifiers": ["A"], "dateTime": "2008-06-21T01:00:00.000-05:00"}, {"value": "780", "qualifiers": ["A"], "dateTime": "2008-06-21T01:15:00.000-05:00"}, {"value": "775", "qualifiers": ["A"], "dateTime": "2008-06-21T01:30:00.000-05:00"}, {"value": "782", "qualifiers": ["A"], "dateTime": "2008-06-21T01:45:00.000-05:00"}, {"value": "780", "qualifiers": ["A"], "dateTime": "2008-06-21T02:00:00.000-05:00"}, {"value": "782", "qualifiers": ["A"], "dateTime": "2008-06-21T02:15:00.000-05:00"}, {"value": "782", "qualifiers": ["A"], "dateTime": "2008-06-21T02:30:00.000-05:00"}, {"value": "780", "qualifiers": ["A"], "dateTime": "2008-06-21T02:45:00.000-05:00"}, {"value": "775", "qualifiers": ["A"], "dateTime": "2008-06-21T03:00:00.000-05:00"}, {"value": "772", "qualifiers": ["A"], "dateTime": "2008-06-21T03:15:00.000-05:00"}, {"value": "775", "qualifiers": ["A"], "dateTime": "2008-06-21T03:30:00.000-05:00"}, {"value": "772", "qualifiers": ["A"], "dateTime": "2008-06-21T03:45:00.000-05:00"}, {"value": "775", "qualifiers": ["A"], "dateTime": "2008-06-21T04:00:00.000-05:00"}, {"value": "770", "qualifiers": ["A"], "dateTime": "2008-06-21T04:15:00.000-05:00"}, {"value": "772", "qualifiers": ["A"], "dateTime": "2008-06-21T04:30:00.000-05:00"}, {"value": "768", "qualifiers": ["A"], "dateTime": "2008-06-21T04:45:00.000-05:00"}, {"value": "768", "qualifiers": ["A"], "dateTime": "2008-06-21T05:00:00.000-05:00"}, {"value": "772", "qualifiers": ["A"], "dateTime": "2008-06-21T05:15:00.000-05:00"}, {"value": "770", "qualifiers": ["A"], "dateTime": "2008-06-21T05:30:00.000-05:00"}, {"value": "772", "qualifiers": ["A"], "dateTime": "2008-06-21T05:45:00.000-05:00"}, {"value": "768", "qualifiers": ["A"], "dateTime": "2008-06-21T06:00:00.000-05:00"}, {"value": "763", "qualifiers": ["A"], "dateTime": "2008-06-21T06:15:00.000-05:00"}, {"value": "768", "qualifiers": ["A"], "dateTime": "2008-06-21T06:30:00.000-05:00"}, {"value": "765", "qualifiers": ["A"], "dateTime": "2008-06-21T06:45:00.000-05:00"}, {"value": "768", "qualifiers": ["A"], "dateTime": "2008-06-21T07:00:00.000-05:00"}, {"value": "763", "qualifiers": ["A"], "dateTime": "2008-06-21T07:15:00.000-05:00"}, {"value": "765", "qualifiers": ["A"], "dateTime": "2008-06-21T07:30:00.000-05:00"}, {"value": "765", "qualifiers": ["A"], "dateTime": "2008-06-21T07:45:00.000-05:00"}, {"value": "765", "qualifiers": ["A"], "dateTime": "2008-06-21T08:00:00.000-05:00"}, {"value": "765", "qualifiers": ["A"], "dateTime": "2008-06-21T08:15:00.000-05:00"}, {"value": "763", "qualifiers": ["A"], "dateTime": "2008-06-21T08:30:00.000-05:00"}, {"value": "763", "qualifiers": ["A"], "dateTime": "2008-06-21T08:45:00.000-05:00"}, {"value": "756", "qualifiers": ["A"], "dateTime": "2008-06-21T09:00:00.000-05:00"}, {"value": "761", "qualifiers": ["A"], "dateTime": "2008-06-21T09:15:00.000-05:00"}, {"value": "761", "qualifiers": ["A"], "dateTime": "2008-06-21T09:30:00.000-05:00"}, {"value": "754", "qualifiers": ["A"], "dateTime": "2008-06-21T09:45:00.000-05:00"}, {"value": "761", "qualifiers": ["A"], "dateTime": "2008-06-21T10:00:00.000-05:00"}, {"value": "754", "qualifiers": ["A"], "dateTime": "2008-06-21T10:15:00.000-05:00"}, {"value": "751", "qualifiers": ["A"], "dateTime": "2008-06-21T10:30:00.000-05:00"}, {"value": "754", "qualifiers": ["A"], "dateTime": "2008-06-21T10:45:00.000-05:00"}, {"value": "758", "qualifiers": ["A"], "dateTime": "2008-06-21T11:00:00.000-05:00"}, {"value": "751", "qualifiers": ["A"], "dateTime": "2008-06-21T11:15:00.000-05:00"}, {"value": "751", "qualifiers": ["A"], "dateTime": "2008-06-21T11:30:00.000-05:00"}, {"value": "754", "qualifiers": ["A"], "dateTime": "2008-06-21T11:45:00.000-05:00"}, {"value": "756", "qualifiers": ["A"], "dateTime": "2008-06-21T12:00:00.000-05:00"}, {"value": "751", "qualifiers": ["A"], "dateTime": "2008-06-21T12:15:00.000-05:00"}, {"value": "756", "qualifiers": ["A"], "dateTime": "2008-06-21T12:30:00.000-05:00"}, {"value": "749", "qualifiers": ["A"], "dateTime": "2008-06-21T12:45:00.000-05:00"}, {"value": "747", "qualifiers": ["A"], "dateTime": "2008-06-21T13:00:00.000-05:00"}, {"value": "749", "qualifiers": ["A"], "dateTime": "2008-06-21T13:15:00.000-05:00"}, {"value": "747", "qualifiers": ["A"], "dateTime": "2008-06-21T13:30:00.000-05:00"}, {"value": "747", "qualifiers": ["A"], "dateTime": "2008-06-21T13:45:00.000-05:00"}, {"value": "749", "qualifiers": ["A"], "dateTime": "2008-06-21T14:00:00.000-05:00"}, {"value": "740", "qualifiers": ["A"], "dateTime": "2008-06-21T14:15:00.000-05:00"}, {"value": "749", "qualifiers": ["A"], "dateTime": "2008-06-21T14:30:00.000-05:00"}, {"value": "747", "qualifiers": ["A"], "dateTime": "2008-06-21T14:45:00.000-05:00"}, {"value": "747", "qualifiers": ["A"], "dateTime": "2008-06-21T15:00:00.000-05:00"}, {"value": "744", "qualifiers": ["A"], "dateTime": "2008-06-21T15:15:00.000-05:00"}, {"value": "744", "qualifiers": ["A"], "dateTime": "2008-06-21T15:30:00.000-05:00"}, {"value": "740", "qualifiers": ["A"], "dateTime": "2008-06-21T15:45:00.000-05:00"}, {"value": "744", "qualifiers": ["A"], "dateTime": "2008-06-21T16:00:00.000-05:00"}, {"value": "740", "qualifiers": ["A"], "dateTime": "2008-06-21T16:15:00.000-05:00"}, {"value": "735", "qualifiers": ["A"], "dateTime": "2008-06-21T16:30:00.000-05:00"}, {"value": "740", "qualifiers": ["A"], "dateTime": "2008-06-21T16:45:00.000-05:00"}, {"value": "740", "qualifiers": ["A"], "dateTime": "2008-06-21T17:00:00.000-05:00"}, {"value": "735", "qualifiers": ["A"], "dateTime": "2008-06-21T17:15:00.000-05:00"}, {"value": "735", "qualifiers": ["A"], "dateTime": "2008-06-21T17:30:00.000-05:00"}, {"value": "733", "qualifiers": ["A"], "dateTime": "2008-06-21T17:45:00.000-05:00"}, {"value": "733", "qualifiers": ["A"], "dateTime": "2008-06-21T18:00:00.000-05:00"}, {"value": "733", "qualifiers": ["A"], "dateTime": "2008-06-21T18:15:00.000-05:00"}, {"value": "731", "qualifiers": ["A"], "dateTime": "2008-06-21T18:30:00.000-05:00"}, {"value": "733", "qualifiers": ["A"], "dateTime": "2008-06-21T18:45:00.000-05:00"}, {"value": "733", "qualifiers": ["A"], "dateTime": "2008-06-21T19:00:00.000-05:00"}, {"value": "728", "qualifiers": ["A"], "dateTime": "2008-06-21T19:15:00.000-05:00"}, {"value": "731", "qualifiers": ["A"], "dateTime": "2008-06-21T19:30:00.000-05:00"}, {"value": "726", "qualifiers": ["A"], "dateTime": "2008-06-21T19:45:00.000-05:00"}, {"value": "731", "qualifiers": ["A"], "dateTime": "2008-06-21T20:00:00.000-05:00"}, {"value": "728", "qualifiers": ["A"], "dateTime": "2008-06-21T20:15:00.000-05:00"}, {"value": "731", "qualifiers": ["A"], "dateTime": "2008-06-21T20:30:00.000-05:00"}, {"value": "728", "qualifiers": ["A"], "dateTime": "2008-06-21T20:45:00.000-05:00"}, {"value": "726", "qualifiers": ["A"], "dateTime": "2008-06-21T21:00:00.000-05:00"}, {"value": "728", "qualifiers": ["A"], "dateTime": "2008-06-21T21:15:00.000-05:00"}, {"value": "722", "qualifiers": ["A"], "dateTime": "2008-06-21T21:30:00.000-05:00"}, {"value": "724", "qualifiers": ["A"], "dateTime": "2008-06-21T21:45:00.000-05:00"}, {"value": "726", "qualifiers": ["A"], "dateTime": "2008-06-21T22:00:00.000-05:00"}, {"value": "719", "qualifiers": ["A"], "dateTime": "2008-06-21T22:15:00.000-05:00"}, {"value": "722", "qualifiers": ["A"], "dateTime": "2008-06-21T22:30:00.000-05:00"}, {"value": "719", "qualifiers": ["A"], "dateTime": "2008-06-21T22:45:00.000-05:00"}, {"value": "719", "qualifiers": ["A"], "dateTime": "2008-06-21T23:00:00.000-05:00"}, {"value": "717", "qualifiers": ["A"], "dateTime": "2008-06-21T23:15:00.000-05:00"}, {"value": "717", "qualifiers": ["A"], "dateTime": "2008-06-21T23:30:00.000-05:00"}, {"value": "717", "qualifiers": ["A"], "dateTime": "2008-06-21T23:45:00.000-05:00"}, {"value": "722", "qualifiers": ["A"], "dateTime": "2008-06-22T00:00:00.000-05:00"}, {"value": "712", "qualifiers": ["A"], "dateTime": "2008-06-22T00:15:00.000-05:00"}, {"value": "715", "qualifiers": ["A"], "dateTime": "2008-06-22T00:30:00.000-05:00"}, {"value": "715", "qualifiers": ["A"], "dateTime": "2008-06-22T00:45:00.000-05:00"}, {"value": "717", "qualifiers": ["A"], "dateTime": "2008-06-22T01:00:00.000-05:00"}, {"value": "715", "qualifiers": ["A"], "dateTime": "2008-06-22T01:15:00.000-05:00"}, {"value": "710", "qualifiers": ["A"], "dateTime": "2008-06-22T01:30:00.000-05:00"}, {"value": "715", "qualifiers": ["A"], "dateTime": "2008-06-22T01:45:00.000-05:00"}, {"value": "712", "qualifiers": ["A"], "dateTime": "2008-06-22T02:00:00.000-05:00"}, {"value": "712", "qualifiers": ["A"], "dateTime": "2008-06-22T02:15:00.000-05:00"}, {"value": "710", "qualifiers": ["A"], "dateTime": "2008-06-22T02:30:00.000-05:00"}, {"value": "710", "qualifiers": ["A"], "dateTime": "2008-06-22T02:45:00.000-05:00"}, {"value": "710", "qualifiers": ["A"], "dateTime": "2008-06-22T03:00:00.000-05:00"}, {"value": "708", "qualifiers": ["A"], "dateTime": "2008-06-22T03:15:00.000-05:00"}, {"value": "712", "qualifiers": ["A"], "dateTime": "2008-06-22T03:30:00.000-05:00"}, {"value": "710", "qualifiers": ["A"], "dateTime": "2008-06-22T03:45:00.000-05:00"}, {"value": "715", "qualifiers": ["A"], "dateTime": "2008-06-22T04:00:00.000-05:00"}, {"value": "715", "qualifiers": ["A"], "dateTime": "2008-06-22T04:15:00.000-05:00"}, {"value": "706", "qualifiers": ["A"], "dateTime": "2008-06-22T04:30:00.000-05:00"}, {"value": "703", "qualifiers": ["A"], "dateTime": "2008-06-22T04:45:00.000-05:00"}, {"value": "710", "qualifiers": ["A"], "dateTime": "2008-06-22T05:00:00.000-05:00"}, {"value": "708", "qualifiers": ["A"], "dateTime": "2008-06-22T05:15:00.000-05:00"}, {"value": "710", "qualifiers": ["A"], "dateTime": "2008-06-22T05:30:00.000-05:00"}, {"value": "706", "qualifiers": ["A"], "dateTime": "2008-06-22T05:45:00.000-05:00"}, {"value": "706", "qualifiers": ["A"], "dateTime": "2008-06-22T06:00:00.000-05:00"}, {"value": "706", "qualifiers": ["A"], "dateTime": "2008-06-22T06:15:00.000-05:00"}, {"value": "708", "qualifiers": ["A"], "dateTime": "2008-06-22T06:30:00.000-05:00"}, {"value": "701", "qualifiers": ["A"], "dateTime": "2008-06-22T06:45:00.000-05:00"}, {"value": "701", "qualifiers": ["A"], "dateTime": "2008-06-22T07:00:00.000-05:00"}, {"value": "706", "qualifiers": ["A"], "dateTime": "2008-06-22T07:15:00.000-05:00"}, {"value": "699", "qualifiers": ["A"], "dateTime": "2008-06-22T07:30:00.000-05:00"}, {"value": "708", "qualifiers": ["A"], "dateTime": "2008-06-22T07:45:00.000-05:00"}, {"value": "699", "qualifiers": ["A"], "dateTime": "2008-06-22T08:00:00.000-05:00"}, {"value": "701", "qualifiers": ["A"], "dateTime": "2008-06-22T08:15:00.000-05:00"}, {"value": "706", "qualifiers": ["A"], "dateTime": "2008-06-22T08:30:00.000-05:00"}, {"value": "703", "qualifiers": ["A"], "dateTime": "2008-06-22T08:45:00.000-05:00"}, {"value": "701", "qualifiers": ["A"], "dateTime": "2008-06-22T09:00:00.000-05:00"}, {"value": "699", "qualifiers": ["A"], "dateTime": "2008-06-22T09:15:00.000-05:00"}, {"value": "703", "qualifiers": ["A"], "dateTime": "2008-06-22T09:30:00.000-05:00"}, {"value": "703", "qualifiers": ["A"], "dateTime": "2008-06-22T09:45:00.000-05:00"}, {"value": "697", "qualifiers": ["A"], "dateTime": "2008-06-22T10:00:00.000-05:00"}, {"value": "697", "qualifiers": ["A"], "dateTime": "2008-06-22T10:15:00.000-05:00"}, {"value": "694", "qualifiers": ["A"], "dateTime": "2008-06-22T10:30:00.000-05:00"}, {"value": "694", "qualifiers": ["A"], "dateTime": "2008-06-22T10:45:00.000-05:00"}, {"value": "699", "qualifiers": ["A"], "dateTime": "2008-06-22T11:00:00.000-05:00"}, {"value": "699", "qualifiers": ["A"], "dateTime": "2008-06-22T11:15:00.000-05:00"}, {"value": "692", "qualifiers": ["A"], "dateTime": "2008-06-22T11:30:00.000-05:00"}, {"value": "692", "qualifiers": ["A"], "dateTime": "2008-06-22T11:45:00.000-05:00"}, {"value": "694", "qualifiers": ["A"], "dateTime": "2008-06-22T12:00:00.000-05:00"}, {"value": "694", "qualifiers": ["A"], "dateTime": "2008-06-22T12:15:00.000-05:00"}, {"value": "690", "qualifiers": ["A"], "dateTime": "2008-06-22T12:30:00.000-05:00"}, {"value": "694", "qualifiers": ["A"], "dateTime": "2008-06-22T12:45:00.000-05:00"}, {"value": "688", "qualifiers": ["A"], "dateTime": "2008-06-22T13:00:00.000-05:00"}, {"value": "697", "qualifiers": ["A"], "dateTime": "2008-06-22T13:15:00.000-05:00"}, {"value": "686", "qualifiers": ["A"], "dateTime": "2008-06-22T13:30:00.000-05:00"}, {"value": "690", "qualifiers": ["A"], "dateTime": "2008-06-22T13:45:00.000-05:00"}, {"value": "688", "qualifiers": ["A"], "dateTime": "2008-06-22T14:00:00.000-05:00"}, {"value": "686", "qualifiers": ["A"], "dateTime": "2008-06-22T14:15:00.000-05:00"}, {"value": "683", "qualifiers": ["A"], "dateTime": "2008-06-22T14:30:00.000-05:00"}, {"value": "686", "qualifiers": ["A"], "dateTime": "2008-06-22T14:45:00.000-05:00"}, {"value": "688", "qualifiers": ["A"], "dateTime": "2008-06-22T15:00:00.000-05:00"}, {"value": "692", "qualifiers": ["A"], "dateTime": "2008-06-22T15:15:00.000-05:00"}, {"value": "683", "qualifiers": ["A"], "dateTime": "2008-06-22T15:30:00.000-05:00"}, {"value": "683", "qualifiers": ["A"], "dateTime": "2008-06-22T15:45:00.000-05:00"}, {"value": "686", "qualifiers": ["A"], "dateTime": "2008-06-22T16:00:00.000-05:00"}, {"value": "681", "qualifiers": ["A"], "dateTime": "2008-06-22T16:15:00.000-05:00"}, {"value": "686", "qualifiers": ["A"], "dateTime": "2008-06-22T16:30:00.000-05:00"}, {"value": "681", "qualifiers": ["A"], "dateTime": "2008-06-22T16:45:00.000-05:00"}, {"value": "681", "qualifiers": ["A"], "dateTime": "2008-06-22T17:00:00.000-05:00"}, {"value": "688", "qualifiers": ["A"], "dateTime": "2008-06-22T17:15:00.000-05:00"}, {"value": "677", "qualifiers": ["A"], "dateTime": "2008-06-22T17:30:00.000-05:00"}, {"value": "679", "qualifiers": ["A"], "dateTime": "2008-06-22T17:45:00.000-05:00"}, {"value": "677", "qualifiers": ["A"], "dateTime": "2008-06-22T18:00:00.000-05:00"}, {"value": "683", "qualifiers": ["A"], "dateTime": "2008-06-22T18:15:00.000-05:00"}, {"value": "681", "qualifiers": ["A"], "dateTime": "2008-06-22T18:30:00.000-05:00"}, {"value": "679", "qualifiers": ["A"], "dateTime": "2008-06-22T18:45:00.000-05:00"}, {"value": "677", "qualifiers": ["A"], "dateTime": "2008-06-22T19:00:00.000-05:00"}, {"value": "679", "qualifiers": ["A"], "dateTime": "2008-06-22T19:15:00.000-05:00"}, {"value": "679", "qualifiers": ["A"], "dateTime": "2008-06-22T19:30:00.000-05:00"}, {"value": "670", "qualifiers": ["A"], "dateTime": "2008-06-22T19:45:00.000-05:00"}, {"value": "674", "qualifiers": ["A"], "dateTime": "2008-06-22T20:00:00.000-05:00"}, {"value": "674", "qualifiers": ["A"], "dateTime": "2008-06-22T20:15:00.000-05:00"}, {"value": "670", "qualifiers": ["A"], "dateTime": "2008-06-22T20:30:00.000-05:00"}, {"value": "677", "qualifiers": ["A"], "dateTime": "2008-06-22T20:45:00.000-05:00"}, {"value": "677", "qualifiers": ["A"], "dateTime": "2008-06-22T21:00:00.000-05:00"}, {"value": "677", "qualifiers": ["A"], "dateTime": "2008-06-22T21:15:00.000-05:00"}, {"value": "674", "qualifiers": ["A"], "dateTime": "2008-06-22T21:30:00.000-05:00"}, {"value": "668", "qualifiers": ["A"], "dateTime": "2008-06-22T21:45:00.000-05:00"}, {"value": "668", "qualifiers": ["A"], "dateTime": "2008-06-22T22:00:00.000-05:00"}, {"value": "668", "qualifiers": ["A"], "dateTime": "2008-06-22T22:15:00.000-05:00"}, {"value": "674", "qualifiers": ["A"], "dateTime": "2008-06-22T22:30:00.000-05:00"}, {"value": "672", "qualifiers": ["A"], "dateTime": "2008-06-22T22:45:00.000-05:00"}, {"value": "674", "qualifiers": ["A"], "dateTime": "2008-06-22T23:00:00.000-05:00"}, {"value": "670", "qualifiers": ["A"], "dateTime": "2008-06-22T23:15:00.000-05:00"}, {"value": "670", "qualifiers": ["A"], "dateTime": "2008-06-22T23:30:00.000-05:00"}, {"value": "666", "qualifiers": ["A"], "dateTime": "2008-06-22T23:45:00.000-05:00"}, {"value": "668", "qualifiers": ["A"], "dateTime": "2008-06-23T00:00:00.000-05:00"}, {"value": "661", "qualifiers": ["A"], "dateTime": "2008-06-23T00:15:00.000-05:00"}, {"value": "668", "qualifiers": ["A"], "dateTime": "2008-06-23T00:30:00.000-05:00"}, {"value": "663", "qualifiers": ["A"], "dateTime": "2008-06-23T00:45:00.000-05:00"}, {"value": "666", "qualifiers": ["A"], "dateTime": "2008-06-23T01:00:00.000-05:00"}, {"value": "663", "qualifiers": ["A"], "dateTime": "2008-06-23T01:15:00.000-05:00"}, {"value": "666", "qualifiers": ["A"], "dateTime": "2008-06-23T01:30:00.000-05:00"}, {"value": "663", "qualifiers": ["A"], "dateTime": "2008-06-23T01:45:00.000-05:00"}, {"value": "666", "qualifiers": ["A"], "dateTime": "2008-06-23T02:00:00.000-05:00"}, {"value": "668", "qualifiers": ["A"], "dateTime": "2008-06-23T02:15:00.000-05:00"}, {"value": "661", "qualifiers": ["A"], "dateTime": "2008-06-23T02:30:00.000-05:00"}, {"value": "661", "qualifiers": ["A"], "dateTime": "2008-06-23T02:45:00.000-05:00"}, {"value": "657", "qualifiers": ["A"], "dateTime": "2008-06-23T03:00:00.000-05:00"}, {"value": "663", "qualifiers": ["A"], "dateTime": "2008-06-23T03:15:00.000-05:00"}, {"value": "663", "qualifiers": ["A"], "dateTime": "2008-06-23T03:30:00.000-05:00"}, {"value": "661", "qualifiers": ["A"], "dateTime": "2008-06-23T03:45:00.000-05:00"}, {"value": "659", "qualifiers": ["A"], "dateTime": "2008-06-23T04:00:00.000-05:00"}, {"value": "657", "qualifiers": ["A"], "dateTime": "2008-06-23T04:15:00.000-05:00"}, {"value": "661", "qualifiers": ["A"], "dateTime": "2008-06-23T04:30:00.000-05:00"}, {"value": "657", "qualifiers": ["A"], "dateTime": "2008-06-23T04:45:00.000-05:00"}, {"value": "659", "qualifiers": ["A"], "dateTime": "2008-06-23T05:00:00.000-05:00"}, {"value": "653", "qualifiers": ["A"], "dateTime": "2008-06-23T05:15:00.000-05:00"}, {"value": "650", "qualifiers": ["A"], "dateTime": "2008-06-23T05:30:00.000-05:00"}, {"value": "655", "qualifiers": ["A"], "dateTime": "2008-06-23T05:45:00.000-05:00"}, {"value": "659", "qualifiers": ["A"], "dateTime": "2008-06-23T06:00:00.000-05:00"}, {"value": "653", "qualifiers": ["A"], "dateTime": "2008-06-23T06:15:00.000-05:00"}, {"value": "653", "qualifiers": ["A"], "dateTime": "2008-06-23T06:30:00.000-05:00"}, {"value": "657", "qualifiers": ["A"], "dateTime": "2008-06-23T06:45:00.000-05:00"}, {"value": "655", "qualifiers": ["A"], "dateTime": "2008-06-23T07:00:00.000-05:00"}, {"value": "653", "qualifiers": ["A"], "dateTime": "2008-06-23T07:15:00.000-05:00"}, {"value": "655", "qualifiers": ["A"], "dateTime": "2008-06-23T07:30:00.000-05:00"}, {"value": "648", "qualifiers": ["A"], "dateTime": "2008-06-23T07:45:00.000-05:00"}, {"value": "648", "qualifiers": ["A"], "dateTime": "2008-06-23T08:00:00.000-05:00"}, {"value": "650", "qualifiers": ["A"], "dateTime": "2008-06-23T08:15:00.000-05:00"}, {"value": "646", "qualifiers": ["A"], "dateTime": "2008-06-23T08:30:00.000-05:00"}, {"value": "648", "qualifiers": ["A"], "dateTime": "2008-06-23T08:45:00.000-05:00"}, {"value": "646", "qualifiers": ["A"], "dateTime": "2008-06-23T09:00:00.000-05:00"}, {"value": "642", "qualifiers": ["A"], "dateTime": "2008-06-23T09:15:00.000-05:00"}, {"value": "648", "qualifiers": ["A"], "dateTime": "2008-06-23T09:30:00.000-05:00"}, {"value": "648", "qualifiers": ["A"], "dateTime": "2008-06-23T09:45:00.000-05:00"}, {"value": "644", "qualifiers": ["A"], "dateTime": "2008-06-23T10:00:00.000-05:00"}, {"value": "644", "qualifiers": ["A"], "dateTime": "2008-06-23T10:15:00.000-05:00"}, {"value": "640", "qualifiers": ["A"], "dateTime": "2008-06-23T10:30:00.000-05:00"}, {"value": "646", "qualifiers": ["A"], "dateTime": "2008-06-23T10:45:00.000-05:00"}, {"value": "648", "qualifiers": ["A"], "dateTime": "2008-06-23T11:00:00.000-05:00"}, {"value": "644", "qualifiers": ["A"], "dateTime": "2008-06-23T11:15:00.000-05:00"}, {"value": "644", "qualifiers": ["A"], "dateTime": "2008-06-23T11:30:00.000-05:00"}, {"value": "642", "qualifiers": ["A"], "dateTime": "2008-06-23T11:45:00.000-05:00"}, {"value": "637", "qualifiers": ["A"], "dateTime": "2008-06-23T12:00:00.000-05:00"}, {"value": "646", "qualifiers": ["A"], "dateTime": "2008-06-23T12:15:00.000-05:00"}, {"value": "640", "qualifiers": ["A"], "dateTime": "2008-06-23T12:30:00.000-05:00"}, {"value": "642", "qualifiers": ["A"], "dateTime": "2008-06-23T12:45:00.000-05:00"}, {"value": "637", "qualifiers": ["A"], "dateTime": "2008-06-23T13:00:00.000-05:00"}, {"value": "644", "qualifiers": ["A"], "dateTime": "2008-06-23T13:15:00.000-05:00"}, {"value": "644", "qualifiers": ["A"], "dateTime": "2008-06-23T13:30:00.000-05:00"}, {"value": "635", "qualifiers": ["A"], "dateTime": "2008-06-23T13:45:00.000-05:00"}, {"value": "644", "qualifiers": ["A"], "dateTime": "2008-06-23T14:00:00.000-05:00"}, {"value": "640", "qualifiers": ["A"], "dateTime": "2008-06-23T14:15:00.000-05:00"}, {"value": "640", "qualifiers": ["A"], "dateTime": "2008-06-23T14:30:00.000-05:00"}, {"value": "644", "qualifiers": ["A"], "dateTime": "2008-06-23T14:45:00.000-05:00"}, {"value": "633", "qualifiers": ["A"], "dateTime": "2008-06-23T15:00:00.000-05:00"}, {"value": "633", "qualifiers": ["A"], "dateTime": "2008-06-23T15:15:00.000-05:00"}, {"value": "637", "qualifiers": ["A"], "dateTime": "2008-06-23T15:30:00.000-05:00"}, {"value": "640", "qualifiers": ["A"], "dateTime": "2008-06-23T15:45:00.000-05:00"}, {"value": "633", "qualifiers": ["A"], "dateTime": "2008-06-23T16:00:00.000-05:00"}, {"value": "637", "qualifiers": ["A"], "dateTime": "2008-06-23T16:15:00.000-05:00"}, {"value": "631", "qualifiers": ["A"], "dateTime": "2008-06-23T16:30:00.000-05:00"}, {"value": "631", "qualifiers": ["A"], "dateTime": "2008-06-23T16:45:00.000-05:00"}, {"value": "633", "qualifiers": ["A"], "dateTime": "2008-06-23T17:00:00.000-05:00"}, {"value": "629", "qualifiers": ["A"], "dateTime": "2008-06-23T17:15:00.000-05:00"}, {"value": "627", "qualifiers": ["A"], "dateTime": "2008-06-23T17:30:00.000-05:00"}, {"value": "635", "qualifiers": ["A"], "dateTime": "2008-06-23T17:45:00.000-05:00"}, {"value": "631", "qualifiers": ["A"], "dateTime": "2008-06-23T18:00:00.000-05:00"}, {"value": "631", "qualifiers": ["A"], "dateTime": "2008-06-23T18:15:00.000-05:00"}, {"value": "627", "qualifiers": ["A"], "dateTime": "2008-06-23T18:30:00.000-05:00"}, {"value": "629", "qualifiers": ["A"], "dateTime": "2008-06-23T18:45:00.000-05:00"}, {"value": "627", "qualifiers": ["A"], "dateTime": "2008-06-23T19:00:00.000-05:00"}, {"value": "633", "qualifiers": ["A"], "dateTime": "2008-06-23T19:15:00.000-05:00"}, {"value": "629", "qualifiers": ["A"], "dateTime": "2008-06-23T19:30:00.000-05:00"}, {"value": "624", "qualifiers": ["A"], "dateTime": "2008-06-23T19:45:00.000-05:00"}, {"value": "631", "qualifiers": ["A"], "dateTime": "2008-06-23T20:00:00.000-05:00"}, {"value": "620", "qualifiers": ["A"], "dateTime": "2008-06-23T20:15:00.000-05:00"}, {"value": "622", "qualifiers": ["A"], "dateTime": "2008-06-23T20:30:00.000-05:00"}, {"value": "622", "qualifiers": ["A"], "dateTime": "2008-06-23T20:45:00.000-05:00"}, {"value": "620", "qualifiers": ["A"], "dateTime": "2008-06-23T21:00:00.000-05:00"}, {"value": "620", "qualifiers": ["A"], "dateTime": "2008-06-23T21:15:00.000-05:00"}, {"value": "627", "qualifiers": ["A"], "dateTime": "2008-06-23T21:30:00.000-05:00"}, {"value": "620", "qualifiers": ["A"], "dateTime": "2008-06-23T21:45:00.000-05:00"}, {"value": "622", "qualifiers": ["A"], "dateTime": "2008-06-23T22:00:00.000-05:00"}, {"value": "627", "qualifiers": ["A"], "dateTime": "2008-06-23T22:15:00.000-05:00"}, {"value": "622", "qualifiers": ["A"], "dateTime": "2008-06-23T22:30:00.000-05:00"}, {"value": "618", "qualifiers": ["A"], "dateTime": "2008-06-23T22:45:00.000-05:00"}, {"value": "618", "qualifiers": ["A"], "dateTime": "2008-06-23T23:00:00.000-05:00"}, {"value": "620", "qualifiers": ["A"], "dateTime": "2008-06-23T23:15:00.000-05:00"}, {"value": "620", "qualifiers": ["A"], "dateTime": "2008-06-23T23:30:00.000-05:00"}, {"value": "624", "qualifiers": ["A"], "dateTime": "2008-06-23T23:45:00.000-05:00"}, {"value": "622", "qualifiers": ["A"], "dateTime": "2008-06-24T00:00:00.000-05:00"}, {"value": "618", "qualifiers": ["A"], "dateTime": "2008-06-24T00:15:00.000-05:00"}, {"value": "616", "qualifiers": ["A"], "dateTime": "2008-06-24T00:30:00.000-05:00"}, {"value": "618", "qualifiers": ["A"], "dateTime": "2008-06-24T00:45:00.000-05:00"}, {"value": "622", "qualifiers": ["A"], "dateTime": "2008-06-24T01:00:00.000-05:00"}, {"value": "616", "qualifiers": ["A"], "dateTime": "2008-06-24T01:15:00.000-05:00"}, {"value": "616", "qualifiers": ["A"], "dateTime": "2008-06-24T01:30:00.000-05:00"}, {"value": "624", "qualifiers": ["A"], "dateTime": "2008-06-24T01:45:00.000-05:00"}, {"value": "622", "qualifiers": ["A"], "dateTime": "2008-06-24T02:00:00.000-05:00"}, {"value": "616", "qualifiers": ["A"], "dateTime": "2008-06-24T02:15:00.000-05:00"}, {"value": "622", "qualifiers": ["A"], "dateTime": "2008-06-24T02:30:00.000-05:00"}, {"value": "614", "qualifiers": ["A"], "dateTime": "2008-06-24T02:45:00.000-05:00"}, {"value": "618", "qualifiers": ["A"], "dateTime": "2008-06-24T03:00:00.000-05:00"}, {"value": "610", "qualifiers": ["A"], "dateTime": "2008-06-24T03:15:00.000-05:00"}, {"value": "620", "qualifiers": ["A"], "dateTime": "2008-06-24T03:30:00.000-05:00"}, {"value": "612", "qualifiers": ["A"], "dateTime": "2008-06-24T03:45:00.000-05:00"}, {"value": "610", "qualifiers": ["A"], "dateTime": "2008-06-24T04:00:00.000-05:00"}, {"value": "618", "qualifiers": ["A"], "dateTime": "2008-06-24T04:15:00.000-05:00"}, {"value": "610", "qualifiers": ["A"], "dateTime": "2008-06-24T04:30:00.000-05:00"}, {"value": "616", "qualifiers": ["A"], "dateTime": "2008-06-24T04:45:00.000-05:00"}, {"value": "612", "qualifiers": ["A"], "dateTime": "2008-06-24T05:00:00.000-05:00"}, {"value": "610", "qualifiers": ["A"], "dateTime": "2008-06-24T05:15:00.000-05:00"}, {"value": "614", "qualifiers": ["A"], "dateTime": "2008-06-24T05:30:00.000-05:00"}, {"value": "612", "qualifiers": ["A"], "dateTime": "2008-06-24T05:45:00.000-05:00"}, {"value": "616", "qualifiers": ["A"], "dateTime": "2008-06-24T06:00:00.000-05:00"}, {"value": "616", "qualifiers": ["A"], "dateTime": "2008-06-24T06:15:00.000-05:00"}, {"value": "616", "qualifiers": ["A"], "dateTime": "2008-06-24T06:30:00.000-05:00"}, {"value": "610", "qualifiers": ["A"], "dateTime": "2008-06-24T06:45:00.000-05:00"}, {"value": "608", "qualifiers": ["A"], "dateTime": "2008-06-24T07:00:00.000-05:00"}, {"value": "610", "qualifiers": ["A"], "dateTime": "2008-06-24T07:15:00.000-05:00"}, {"value": "610", "qualifiers": ["A"], "dateTime": "2008-06-24T07:30:00.000-05:00"}, {"value": "608", "qualifiers": ["A"], "dateTime": "2008-06-24T07:45:00.000-05:00"}, {"value": "612", "qualifiers": ["A"], "dateTime": "2008-06-24T08:00:00.000-05:00"}, {"value": "608", "qualifiers": ["A"], "dateTime": "2008-06-24T08:15:00.000-05:00"}, {"value": "608", "qualifiers": ["A"], "dateTime": "2008-06-24T08:30:00.000-05:00"}, {"value": "616", "qualifiers": ["A"], "dateTime": "2008-06-24T08:45:00.000-05:00"}, {"value": "610", "qualifiers": ["A"], "dateTime": "2008-06-24T09:00:00.000-05:00"}, {"value": "608", "qualifiers": ["A"], "dateTime": "2008-06-24T09:15:00.000-05:00"}, {"value": "608", "qualifiers": ["A"], "dateTime": "2008-06-24T09:30:00.000-05:00"}, {"value": "610", "qualifiers": ["A"], "dateTime": "2008-06-24T09:45:00.000-05:00"}, {"value": "605", "qualifiers": ["A"], "dateTime": "2008-06-24T10:00:00.000-05:00"}, {"value": "610", "qualifiers": ["A"], "dateTime": "2008-06-24T10:15:00.000-05:00"}, {"value": "612", "qualifiers": ["A"], "dateTime": "2008-06-24T10:30:00.000-05:00"}, {"value": "612", "qualifiers": ["A"], "dateTime": "2008-06-24T10:45:00.000-05:00"}, {"value": "601", "qualifiers": ["A"], "dateTime": "2008-06-24T11:00:00.000-05:00"}, {"value": "605", "qualifiers": ["A"], "dateTime": "2008-06-24T11:15:00.000-05:00"}, {"value": "605", "qualifiers": ["A"], "dateTime": "2008-06-24T11:30:00.000-05:00"}, {"value": "608", "qualifiers": ["A"], "dateTime": "2008-06-24T11:45:00.000-05:00"}, {"value": "612", "qualifiers": ["A"], "dateTime": "2008-06-24T12:00:00.000-05:00"}, {"value": "610", "qualifiers": ["A"], "dateTime": "2008-06-24T12:15:00.000-05:00"}, {"value": "608", "qualifiers": ["A"], "dateTime": "2008-06-24T12:30:00.000-05:00"}, {"value": "603", "qualifiers": ["A"], "dateTime": "2008-06-24T12:45:00.000-05:00"}, {"value": "605", "qualifiers": ["A"], "dateTime": "2008-06-24T13:00:00.000-05:00"}, {"value": "608", "qualifiers": ["A"], "dateTime": "2008-06-24T13:15:00.000-05:00"}, {"value": "601", "qualifiers": ["A"], "dateTime": "2008-06-24T13:30:00.000-05:00"}, {"value": "597", "qualifiers": ["A"], "dateTime": "2008-06-24T13:45:00.000-05:00"}, {"value": "599", "qualifiers": ["A"], "dateTime": "2008-06-24T14:00:00.000-05:00"}, {"value": "605", "qualifiers": ["A"], "dateTime": "2008-06-24T14:15:00.000-05:00"}, {"value": "605", "qualifiers": ["A"], "dateTime": "2008-06-24T14:30:00.000-05:00"}, {"value": "603", "qualifiers": ["A"], "dateTime": "2008-06-24T14:45:00.000-05:00"}, {"value": "597", "qualifiers": ["A"], "dateTime": "2008-06-24T15:00:00.000-05:00"}, {"value": "601", "qualifiers": ["A"], "dateTime": "2008-06-24T15:15:00.000-05:00"}, {"value": "593", "qualifiers": ["A"], "dateTime": "2008-06-24T15:30:00.000-05:00"}, {"value": "599", "qualifiers": ["A"], "dateTime": "2008-06-24T15:45:00.000-05:00"}, {"value": "601", "qualifiers": ["A"], "dateTime": "2008-06-24T16:00:00.000-05:00"}, {"value": "599", "qualifiers": ["A"], "dateTime": "2008-06-24T16:15:00.000-05:00"}, {"value": "599", "qualifiers": ["A"], "dateTime": "2008-06-24T16:30:00.000-05:00"}, {"value": "599", "qualifiers": ["A"], "dateTime": "2008-06-24T16:45:00.000-05:00"}, {"value": "591", "qualifiers": ["A"], "dateTime": "2008-06-24T17:00:00.000-05:00"}, {"value": "595", "qualifiers": ["A"], "dateTime": "2008-06-24T17:15:00.000-05:00"}, {"value": "599", "qualifiers": ["A"], "dateTime": "2008-06-24T17:30:00.000-05:00"}, {"value": "597", "qualifiers": ["A"], "dateTime": "2008-06-24T17:45:00.000-05:00"}, {"value": "593", "qualifiers": ["A"], "dateTime": "2008-06-24T18:00:00.000-05:00"}, {"value": "597", "qualifiers": ["A"], "dateTime": "2008-06-24T18:15:00.000-05:00"}, {"value": "595", "qualifiers": ["A"], "dateTime": "2008-06-24T18:30:00.000-05:00"}, {"value": "599", "qualifiers": ["A"], "dateTime": "2008-06-24T18:45:00.000-05:00"}, {"value": "589", "qualifiers": ["A"], "dateTime": "2008-06-24T19:00:00.000-05:00"}, {"value": "587", "qualifiers": ["A"], "dateTime": "2008-06-24T19:15:00.000-05:00"}, {"value": "591", "qualifiers": ["A"], "dateTime": "2008-06-24T19:30:00.000-05:00"}, {"value": "595", "qualifiers": ["A"], "dateTime": "2008-06-24T19:45:00.000-05:00"}, {"value": "593", "qualifiers": ["A"], "dateTime": "2008-06-24T20:00:00.000-05:00"}, {"value": "589", "qualifiers": ["A"], "dateTime": "2008-06-24T20:15:00.000-05:00"}, {"value": "593", "qualifiers": ["A"], "dateTime": "2008-06-24T20:30:00.000-05:00"}, {"value": "593", "qualifiers": ["A"], "dateTime": "2008-06-24T20:45:00.000-05:00"}, {"value": "589", "qualifiers": ["A"], "dateTime": "2008-06-24T21:00:00.000-05:00"}, {"value": "585", "qualifiers": ["A"], "dateTime": "2008-06-24T21:15:00.000-05:00"}, {"value": "589", "qualifiers": ["A"], "dateTime": "2008-06-24T21:30:00.000-05:00"}, {"value": "587", "qualifiers": ["A"], "dateTime": "2008-06-24T21:45:00.000-05:00"}, {"value": "589", "qualifiers": ["A"], "dateTime": "2008-06-24T22:00:00.000-05:00"}, {"value": "589", "qualifiers": ["A"], "dateTime": "2008-06-24T22:15:00.000-05:00"}, {"value": "583", "qualifiers": ["A"], "dateTime": "2008-06-24T22:30:00.000-05:00"}, {"value": "585", "qualifiers": ["A"], "dateTime": "2008-06-24T22:45:00.000-05:00"}, {"value": "585", "qualifiers": ["A"], "dateTime": "2008-06-24T23:00:00.000-05:00"}, {"value": "589", "qualifiers": ["A"], "dateTime": "2008-06-24T23:15:00.000-05:00"}, {"value": "583", "qualifiers": ["A"], "dateTime": "2008-06-24T23:30:00.000-05:00"}, {"value": "583", "qualifiers": ["A"], "dateTime": "2008-06-24T23:45:00.000-05:00"}, {"value": "585", "qualifiers": ["A"], "dateTime": "2008-06-25T00:00:00.000-05:00"}, {"value": "585", "qualifiers": ["A"], "dateTime": "2008-06-25T00:15:00.000-05:00"}, {"value": "580", "qualifiers": ["A"], "dateTime": "2008-06-25T00:30:00.000-05:00"}, {"value": "578", "qualifiers": ["A"], "dateTime": "2008-06-25T00:45:00.000-05:00"}, {"value": "587", "qualifiers": ["A"], "dateTime": "2008-06-25T01:00:00.000-05:00"}, {"value": "587", "qualifiers": ["A"], "dateTime": "2008-06-25T01:15:00.000-05:00"}, {"value": "587", "qualifiers": ["A"], "dateTime": "2008-06-25T01:30:00.000-05:00"}, {"value": "583", "qualifiers": ["A"], "dateTime": "2008-06-25T01:45:00.000-05:00"}, {"value": "578", "qualifiers": ["A"], "dateTime": "2008-06-25T02:00:00.000-05:00"}, {"value": "578", "qualifiers": ["A"], "dateTime": "2008-06-25T02:15:00.000-05:00"}, {"value": "583", "qualifiers": ["A"], "dateTime": "2008-06-25T02:30:00.000-05:00"}, {"value": "576", "qualifiers": ["A"], "dateTime": "2008-06-25T02:45:00.000-05:00"}, {"value": "576", "qualifiers": ["A"], "dateTime": "2008-06-25T03:00:00.000-05:00"}, {"value": "578", "qualifiers": ["A"], "dateTime": "2008-06-25T03:15:00.000-05:00"}, {"value": "578", "qualifiers": ["A"], "dateTime": "2008-06-25T03:30:00.000-05:00"}, {"value": "585", "qualifiers": ["A"], "dateTime": "2008-06-25T03:45:00.000-05:00"}, {"value": "583", "qualifiers": ["A"], "dateTime": "2008-06-25T04:00:00.000-05:00"}, {"value": "576", "qualifiers": ["A"], "dateTime": "2008-06-25T04:15:00.000-05:00"}, {"value": "580", "qualifiers": ["A"], "dateTime": "2008-06-25T04:30:00.000-05:00"}, {"value": "580", "qualifiers": ["A"], "dateTime": "2008-06-25T04:45:00.000-05:00"}, {"value": "580", "qualifiers": ["A"], "dateTime": "2008-06-25T05:00:00.000-05:00"}, {"value": "574", "qualifiers": ["A"], "dateTime": "2008-06-25T05:15:00.000-05:00"}, {"value": "576", "qualifiers": ["A"], "dateTime": "2008-06-25T05:30:00.000-05:00"}, {"value": "580", "qualifiers": ["A"], "dateTime": "2008-06-25T05:45:00.000-05:00"}, {"value": "574", "qualifiers": ["A"], "dateTime": "2008-06-25T06:00:00.000-05:00"}, {"value": "580", "qualifiers": ["A"], "dateTime": "2008-06-25T06:15:00.000-05:00"}, {"value": "578", "qualifiers": ["A"], "dateTime": "2008-06-25T06:30:00.000-05:00"}, {"value": "574", "qualifiers": ["A"], "dateTime": "2008-06-25T06:45:00.000-05:00"}, {"value": "572", "qualifiers": ["A"], "dateTime": "2008-06-25T07:00:00.000-05:00"}, {"value": "580", "qualifiers": ["A"], "dateTime": "2008-06-25T07:15:00.000-05:00"}, {"value": "576", "qualifiers": ["A"], "dateTime": "2008-06-25T07:30:00.000-05:00"}, {"value": "574", "qualifiers": ["A"], "dateTime": "2008-06-25T07:45:00.000-05:00"}, {"value": "578", "qualifiers": ["A"], "dateTime": "2008-06-25T08:00:00.000-05:00"}, {"value": "576", "qualifiers": ["A"], "dateTime": "2008-06-25T08:15:00.000-05:00"}, {"value": "574", "qualifiers": ["A"], "dateTime": "2008-06-25T08:30:00.000-05:00"}, {"value": "572", "qualifiers": ["A"], "dateTime": "2008-06-25T08:45:00.000-05:00"}, {"value": "572", "qualifiers": ["A"], "dateTime": "2008-06-25T09:00:00.000-05:00"}, {"value": "568", "qualifiers": ["A"], "dateTime": "2008-06-25T09:15:00.000-05:00"}, {"value": "574", "qualifiers": ["A"], "dateTime": "2008-06-25T09:30:00.000-05:00"}, {"value": "570", "qualifiers": ["A"], "dateTime": "2008-06-25T09:45:00.000-05:00"}, {"value": "568", "qualifiers": ["A"], "dateTime": "2008-06-25T10:00:00.000-05:00"}, {"value": "572", "qualifiers": ["A"], "dateTime": "2008-06-25T10:15:00.000-05:00"}, {"value": "576", "qualifiers": ["A"], "dateTime": "2008-06-25T10:30:00.000-05:00"}, {"value": "566", "qualifiers": ["A"], "dateTime": "2008-06-25T10:45:00.000-05:00"}, {"value": "568", "qualifiers": ["A"], "dateTime": "2008-06-25T11:00:00.000-05:00"}, {"value": "568", "qualifiers": ["A"], "dateTime": "2008-06-25T11:15:00.000-05:00"}, {"value": "576", "qualifiers": ["A"], "dateTime": "2008-06-25T11:30:00.000-05:00"}, {"value": "572", "qualifiers": ["A"], "dateTime": "2008-06-25T11:45:00.000-05:00"}, {"value": "568", "qualifiers": ["A"], "dateTime": "2008-06-25T12:00:00.000-05:00"}, {"value": "572", "qualifiers": ["A"], "dateTime": "2008-06-25T12:15:00.000-05:00"}, {"value": "570", "qualifiers": ["A"], "dateTime": "2008-06-25T12:30:00.000-05:00"}, {"value": "570", "qualifiers": ["A"], "dateTime": "2008-06-25T12:45:00.000-05:00"}, {"value": "566", "qualifiers": ["A"], "dateTime": "2008-06-25T13:00:00.000-05:00"}, {"value": "570", "qualifiers": ["A"], "dateTime": "2008-06-25T13:15:00.000-05:00"}, {"value": "564", "qualifiers": ["A"], "dateTime": "2008-06-25T13:30:00.000-05:00"}, {"value": "572", "qualifiers": ["A"], "dateTime": "2008-06-25T13:45:00.000-05:00"}, {"value": "564", "qualifiers": ["A"], "dateTime": "2008-06-25T14:00:00.000-05:00"}, {"value": "562", "qualifiers": ["A"], "dateTime": "2008-06-25T14:15:00.000-05:00"}, {"value": "568", "qualifiers": ["A"], "dateTime": "2008-06-25T14:30:00.000-05:00"}, {"value": "570", "qualifiers": ["A"], "dateTime": "2008-06-25T14:45:00.000-05:00"}, {"value": "562", "qualifiers": ["A"], "dateTime": "2008-06-25T15:00:00.000-05:00"}, {"value": "566", "qualifiers": ["A"], "dateTime": "2008-06-25T15:15:00.000-05:00"}, {"value": "568", "qualifiers": ["A"], "dateTime": "2008-06-25T15:30:00.000-05:00"}, {"value": "560", "qualifiers": ["A"], "dateTime": "2008-06-25T15:45:00.000-05:00"}, {"value": "568", "qualifiers": ["A"], "dateTime": "2008-06-25T16:00:00.000-05:00"}, {"value": "568", "qualifiers": ["A"], "dateTime": "2008-06-25T16:15:00.000-05:00"}, {"value": "560", "qualifiers": ["A"], "dateTime": "2008-06-25T16:30:00.000-05:00"}, {"value": "562", "qualifiers": ["A"], "dateTime": "2008-06-25T16:45:00.000-05:00"}, {"value": "562", "qualifiers": ["A"], "dateTime": "2008-06-25T17:00:00.000-05:00"}, {"value": "564", "qualifiers": ["A"], "dateTime": "2008-06-25T17:15:00.000-05:00"}, {"value": "564", "qualifiers": ["A"], "dateTime": "2008-06-25T17:30:00.000-05:00"}, {"value": "560", "qualifiers": ["A"], "dateTime": "2008-06-25T17:45:00.000-05:00"}, {"value": "562", "qualifiers": ["A"], "dateTime": "2008-06-25T18:00:00.000-05:00"}, {"value": "560", "qualifiers": ["A"], "dateTime": "2008-06-25T18:15:00.000-05:00"}, {"value": "558", "qualifiers": ["A"], "dateTime": "2008-06-25T18:30:00.000-05:00"}, {"value": "562", "qualifiers": ["A"], "dateTime": "2008-06-25T18:45:00.000-05:00"}, {"value": "558", "qualifiers": ["A"], "dateTime": "2008-06-25T19:00:00.000-05:00"}, {"value": "562", "qualifiers": ["A"], "dateTime": "2008-06-25T19:15:00.000-05:00"}, {"value": "558", "qualifiers": ["A"], "dateTime": "2008-06-25T19:30:00.000-05:00"}, {"value": "560", "qualifiers": ["A"], "dateTime": "2008-06-25T19:45:00.000-05:00"}, {"value": "562", "qualifiers": ["A"], "dateTime": "2008-06-25T20:00:00.000-05:00"}, {"value": "554", "qualifiers": ["A"], "dateTime": "2008-06-25T20:15:00.000-05:00"}, {"value": "556", "qualifiers": ["A"], "dateTime": "2008-06-25T20:30:00.000-05:00"}, {"value": "560", "qualifiers": ["A"], "dateTime": "2008-06-25T20:45:00.000-05:00"}, {"value": "554", "qualifiers": ["A"], "dateTime": "2008-06-25T21:00:00.000-05:00"}, {"value": "560", "qualifiers": ["A"], "dateTime": "2008-06-25T21:15:00.000-05:00"}, {"value": "558", "qualifiers": ["A"], "dateTime": "2008-06-25T21:30:00.000-05:00"}, {"value": "558", "qualifiers": ["A"], "dateTime": "2008-06-25T21:45:00.000-05:00"}, {"value": "554", "qualifiers": ["A"], "dateTime": "2008-06-25T22:00:00.000-05:00"}, {"value": "560", "qualifiers": ["A"], "dateTime": "2008-06-25T22:15:00.000-05:00"}, {"value": "556", "qualifiers": ["A"], "dateTime": "2008-06-25T22:30:00.000-05:00"}, {"value": "554", "qualifiers": ["A"], "dateTime": "2008-06-25T22:45:00.000-05:00"}, {"value": "558", "qualifiers": ["A"], "dateTime": "2008-06-25T23:00:00.000-05:00"}, {"value": "550", "qualifiers": ["A"], "dateTime": "2008-06-25T23:15:00.000-05:00"}, {"value": "552", "qualifiers": ["A"], "dateTime": "2008-06-25T23:30:00.000-05:00"}, {"value": "550", "qualifiers": ["A"], "dateTime": "2008-06-25T23:45:00.000-05:00"}, {"value": "554", "qualifiers": ["A"], "dateTime": "2008-06-26T00:00:00.000-05:00"}, {"value": "556", "qualifiers": ["A"], "dateTime": "2008-06-26T00:15:00.000-05:00"}, {"value": "548", "qualifiers": ["A"], "dateTime": "2008-06-26T00:30:00.000-05:00"}, {"value": "548", "qualifiers": ["A"], "dateTime": "2008-06-26T00:45:00.000-05:00"}, {"value": "556", "qualifiers": ["A"], "dateTime": "2008-06-26T01:00:00.000-05:00"}, {"value": "554", "qualifiers": ["A"], "dateTime": "2008-06-26T01:15:00.000-05:00"}, {"value": "556", "qualifiers": ["A"], "dateTime": "2008-06-26T01:30:00.000-05:00"}, {"value": "556", "qualifiers": ["A"], "dateTime": "2008-06-26T01:45:00.000-05:00"}, {"value": "548", "qualifiers": ["A"], "dateTime": "2008-06-26T02:00:00.000-05:00"}, {"value": "550", "qualifiers": ["A"], "dateTime": "2008-06-26T02:15:00.000-05:00"}, {"value": "550", "qualifiers": ["A"], "dateTime": "2008-06-26T02:30:00.000-05:00"}, {"value": "548", "qualifiers": ["A"], "dateTime": "2008-06-26T02:45:00.000-05:00"}, {"value": "554", "qualifiers": ["A"], "dateTime": "2008-06-26T03:00:00.000-05:00"}, {"value": "548", "qualifiers": ["A"], "dateTime": "2008-06-26T03:15:00.000-05:00"}, {"value": "548", "qualifiers": ["A"], "dateTime": "2008-06-26T03:30:00.000-05:00"}, {"value": "546", "qualifiers": ["A"], "dateTime": "2008-06-26T03:45:00.000-05:00"}, {"value": "554", "qualifiers": ["A"], "dateTime": "2008-06-26T04:00:00.000-05:00"}, {"value": "548", "qualifiers": ["A"], "dateTime": "2008-06-26T04:15:00.000-05:00"}, {"value": "550", "qualifiers": ["A"], "dateTime": "2008-06-26T04:30:00.000-05:00"}, {"value": "552", "qualifiers": ["A"], "dateTime": "2008-06-26T04:45:00.000-05:00"}, {"value": "554", "qualifiers": ["A"], "dateTime": "2008-06-26T05:00:00.000-05:00"}, {"value": "552", "qualifiers": ["A"], "dateTime": "2008-06-26T05:15:00.000-05:00"}, {"value": "554", "qualifiers": ["A"], "dateTime": "2008-06-26T05:30:00.000-05:00"}, {"value": "556", "qualifiers": ["A"], "dateTime": "2008-06-26T05:45:00.000-05:00"}, {"value": "548", "qualifiers": ["A"], "dateTime": "2008-06-26T06:00:00.000-05:00"}, {"value": "548", "qualifiers": ["A"], "dateTime": "2008-06-26T06:15:00.000-05:00"}, {"value": "552", "qualifiers": ["A"], "dateTime": "2008-06-26T06:30:00.000-05:00"}, {"value": "552", "qualifiers": ["A"], "dateTime": "2008-06-26T06:45:00.000-05:00"}, {"value": "548", "qualifiers": ["A"], "dateTime": "2008-06-26T07:00:00.000-05:00"}, {"value": "550", "qualifiers": ["A"], "dateTime": "2008-06-26T07:15:00.000-05:00"}, {"value": "550", "qualifiers": ["A"], "dateTime": "2008-06-26T07:30:00.000-05:00"}, {"value": "546", "qualifiers": ["A"], "dateTime": "2008-06-26T07:45:00.000-05:00"}, {"value": "554", "qualifiers": ["A"], "dateTime": "2008-06-26T08:00:00.000-05:00"}, {"value": "552", "qualifiers": ["A"], "dateTime": "2008-06-26T08:15:00.000-05:00"}, {"value": "550", "qualifiers": ["A"], "dateTime": "2008-06-26T08:30:00.000-05:00"}, {"value": "546", "qualifiers": ["A"], "dateTime": "2008-06-26T08:45:00.000-05:00"}, {"value": "550", "qualifiers": ["A"], "dateTime": "2008-06-26T09:00:00.000-05:00"}, {"value": "548", "qualifiers": ["A"], "dateTime": "2008-06-26T09:15:00.000-05:00"}, {"value": "550", "qualifiers": ["A"], "dateTime": "2008-06-26T09:30:00.000-05:00"}, {"value": "554", "qualifiers": ["A"], "dateTime": "2008-06-26T09:45:00.000-05:00"}, {"value": "548", "qualifiers": ["A"], "dateTime": "2008-06-26T10:00:00.000-05:00"}, {"value": "546", "qualifiers": ["A"], "dateTime": "2008-06-26T10:15:00.000-05:00"}, {"value": "544", "qualifiers": ["A"], "dateTime": "2008-06-26T10:30:00.000-05:00"}, {"value": "552", "qualifiers": ["A"], "dateTime": "2008-06-26T10:45:00.000-05:00"}, {"value": "552", "qualifiers": ["A"], "dateTime": "2008-06-26T11:00:00.000-05:00"}, {"value": "548", "qualifiers": ["A"], "dateTime": "2008-06-26T11:15:00.000-05:00"}, {"value": "546", "qualifiers": ["A"], "dateTime": "2008-06-26T11:30:00.000-05:00"}, {"value": "550", "qualifiers": ["A"], "dateTime": "2008-06-26T11:45:00.000-05:00"}, {"value": "546", "qualifiers": ["A"], "dateTime": "2008-06-26T12:00:00.000-05:00"}, {"value": "548", "qualifiers": ["A"], "dateTime": "2008-06-26T12:15:00.000-05:00"}, {"value": "552", "qualifiers": ["A"], "dateTime": "2008-06-26T12:30:00.000-05:00"}, {"value": "544", "qualifiers": ["A"], "dateTime": "2008-06-26T12:45:00.000-05:00"}, {"value": "542", "qualifiers": ["A"], "dateTime": "2008-06-26T13:00:00.000-05:00"}, {"value": "546", "qualifiers": ["A"], "dateTime": "2008-06-26T13:15:00.000-05:00"}, {"value": "550", "qualifiers": ["A"], "dateTime": "2008-06-26T13:30:00.000-05:00"}, {"value": "544", "qualifiers": ["A"], "dateTime": "2008-06-26T13:45:00.000-05:00"}, {"value": "550", "qualifiers": ["A"], "dateTime": "2008-06-26T14:00:00.000-05:00"}, {"value": "544", "qualifiers": ["A"], "dateTime": "2008-06-26T14:15:00.000-05:00"}, {"value": "544", "qualifiers": ["A"], "dateTime": "2008-06-26T14:30:00.000-05:00"}, {"value": "550", "qualifiers": ["A"], "dateTime": "2008-06-26T14:45:00.000-05:00"}, {"value": "538", "qualifiers": ["A"], "dateTime": "2008-06-26T15:00:00.000-05:00"}, {"value": "538", "qualifiers": ["A"], "dateTime": "2008-06-26T15:15:00.000-05:00"}, {"value": "538", "qualifiers": ["A"], "dateTime": "2008-06-26T15:30:00.000-05:00"}, {"value": "536", "qualifiers": ["A"], "dateTime": "2008-06-26T15:45:00.000-05:00"}, {"value": "536", "qualifiers": ["A"], "dateTime": "2008-06-26T16:00:00.000-05:00"}, {"value": "534", "qualifiers": ["A"], "dateTime": "2008-06-26T16:15:00.000-05:00"}, {"value": "536", "qualifiers": ["A"], "dateTime": "2008-06-26T16:30:00.000-05:00"}, {"value": "536", "qualifiers": ["A"], "dateTime": "2008-06-26T16:45:00.000-05:00"}, {"value": "536", "qualifiers": ["A"], "dateTime": "2008-06-26T17:00:00.000-05:00"}, {"value": "536", "qualifiers": ["A"], "dateTime": "2008-06-26T17:15:00.000-05:00"}, {"value": "536", "qualifiers": ["A"], "dateTime": "2008-06-26T17:30:00.000-05:00"}, {"value": "536", "qualifiers": ["A"], "dateTime": "2008-06-26T17:45:00.000-05:00"}, {"value": "536", "qualifiers": ["A"], "dateTime": "2008-06-26T18:00:00.000-05:00"}, {"value": "534", "qualifiers": ["A"], "dateTime": "2008-06-26T18:15:00.000-05:00"}, {"value": "534", "qualifiers": ["A"], "dateTime": "2008-06-26T18:30:00.000-05:00"}, {"value": "536", "qualifiers": ["A"], "dateTime": "2008-06-26T18:45:00.000-05:00"}, {"value": "536", "qualifiers": ["A"], "dateTime": "2008-06-26T19:00:00.000-05:00"}, {"value": "536", "qualifiers": ["A"], "dateTime": "2008-06-26T19:15:00.000-05:00"}, {"value": "536", "qualifiers": ["A"], "dateTime": "2008-06-26T19:30:00.000-05:00"}, {"value": "534", "qualifiers": ["A"], "dateTime": "2008-06-26T19:45:00.000-05:00"}, {"value": "534", "qualifiers": ["A"], "dateTime": "2008-06-26T20:00:00.000-05:00"}, {"value": "534", "qualifiers": ["A"], "dateTime": "2008-06-26T20:15:00.000-05:00"}, {"value": "534", "qualifiers": ["A"], "dateTime": "2008-06-26T20:30:00.000-05:00"}, {"value": "534", "qualifiers": ["A"], "dateTime": "2008-06-26T20:45:00.000-05:00"}, {"value": "534", "qualifiers": ["A"], "dateTime": "2008-06-26T21:00:00.000-05:00"}, {"value": "534", "qualifiers": ["A"], "dateTime": "2008-06-26T21:15:00.000-05:00"}, {"value": "534", "qualifiers": ["A"], "dateTime": "2008-06-26T21:30:00.000-05:00"}, {"value": "534", "qualifiers": ["A"], "dateTime": "2008-06-26T21:45:00.000-05:00"}, {"value": "534", "qualifiers": ["A"], "dateTime": "2008-06-26T22:00:00.000-05:00"}, {"value": "534", "qualifiers": ["A"], "dateTime": "2008-06-26T22:15:00.000-05:00"}, {"value": "532", "qualifiers": ["A"], "dateTime": "2008-06-26T22:30:00.000-05:00"}, {"value": "532", "qualifiers": ["A"], "dateTime": "2008-06-26T22:45:00.000-05:00"}, {"value": "532", "qualifiers": ["A"], "dateTime": "2008-06-26T23:00:00.000-05:00"}, {"value": "532", "qualifiers": ["A"], "dateTime": "2008-06-26T23:15:00.000-05:00"}, {"value": "530", "qualifiers": ["A"], "dateTime": "2008-06-26T23:30:00.000-05:00"}, {"value": "532", "qualifiers": ["A"], "dateTime": "2008-06-26T23:45:00.000-05:00"}, {"value": "530", "qualifiers": ["A"], "dateTime": "2008-06-27T00:00:00.000-05:00"}, {"value": "532", "qualifiers": ["A"], "dateTime": "2008-06-27T00:15:00.000-05:00"}, {"value": "530", "qualifiers": ["A"], "dateTime": "2008-06-27T00:30:00.000-05:00"}, {"value": "532", "qualifiers": ["A"], "dateTime": "2008-06-27T00:45:00.000-05:00"}, {"value": "530", "qualifiers": ["A"], "dateTime": "2008-06-27T01:00:00.000-05:00"}, {"value": "530", "qualifiers": ["A"], "dateTime": "2008-06-27T01:15:00.000-05:00"}, {"value": "530", "qualifiers": ["A"], "dateTime": "2008-06-27T01:30:00.000-05:00"}, {"value": "528", "qualifiers": ["A"], "dateTime": "2008-06-27T01:45:00.000-05:00"}, {"value": "530", "qualifiers": ["A"], "dateTime": "2008-06-27T02:00:00.000-05:00"}, {"value": "530", "qualifiers": ["A"], "dateTime": "2008-06-27T02:15:00.000-05:00"}, {"value": "528", "qualifiers": ["A"], "dateTime": "2008-06-27T02:30:00.000-05:00"}, {"value": "528", "qualifiers": ["A"], "dateTime": "2008-06-27T02:45:00.000-05:00"}, {"value": "528", "qualifiers": ["A"], "dateTime": "2008-06-27T03:00:00.000-05:00"}, {"value": "528", "qualifiers": ["A"], "dateTime": "2008-06-27T03:15:00.000-05:00"}, {"value": "528", "qualifiers": ["A"], "dateTime": "2008-06-27T03:30:00.000-05:00"}, {"value": "528", "qualifiers": ["A"], "dateTime": "2008-06-27T03:45:00.000-05:00"}, {"value": "528", "qualifiers": ["A"], "dateTime": "2008-06-27T04:00:00.000-05:00"}, {"value": "528", "qualifiers": ["A"], "dateTime": "2008-06-27T04:15:00.000-05:00"}, {"value": "528", "qualifiers": ["A"], "dateTime": "2008-06-27T04:30:00.000-05:00"}, {"value": "528", "qualifiers": ["A"], "dateTime": "2008-06-27T04:45:00.000-05:00"}, {"value": "528", "qualifiers": ["A"], "dateTime": "2008-06-27T05:00:00.000-05:00"}, {"value": "528", "qualifiers": ["A"], "dateTime": "2008-06-27T05:15:00.000-05:00"}, {"value": "528", "qualifiers": ["A"], "dateTime": "2008-06-27T05:30:00.000-05:00"}, {"value": "528", "qualifiers": ["A"], "dateTime": "2008-06-27T05:45:00.000-05:00"}, {"value": "528", "qualifiers": ["A"], "dateTime": "2008-06-27T06:00:00.000-05:00"}, {"value": "528", "qualifiers": ["A"], "dateTime": "2008-06-27T06:15:00.000-05:00"}, {"value": "528", "qualifiers": ["A"], "dateTime": "2008-06-27T06:30:00.000-05:00"}, {"value": "526", "qualifiers": ["A"], "dateTime": "2008-06-27T06:45:00.000-05:00"}, {"value": "526", "qualifiers": ["A"], "dateTime": "2008-06-27T07:00:00.000-05:00"}, {"value": "528", "qualifiers": ["A"], "dateTime": "2008-06-27T07:15:00.000-05:00"}, {"value": "526", "qualifiers": ["A"], "dateTime": "2008-06-27T07:30:00.000-05:00"}, {"value": "524", "qualifiers": ["A"], "dateTime": "2008-06-27T07:45:00.000-05:00"}, {"value": "526", "qualifiers": ["A"], "dateTime": "2008-06-27T08:00:00.000-05:00"}, {"value": "526", "qualifiers": ["A"], "dateTime": "2008-06-27T08:15:00.000-05:00"}, {"value": "524", "qualifiers": ["A"], "dateTime": "2008-06-27T08:30:00.000-05:00"}, {"value": "526", "qualifiers": ["A"], "dateTime": "2008-06-27T08:45:00.000-05:00"}, {"value": "524", "qualifiers": ["A"], "dateTime": "2008-06-27T09:00:00.000-05:00"}, {"value": "524", "qualifiers": ["A"], "dateTime": "2008-06-27T09:15:00.000-05:00"}, {"value": "524", "qualifiers": ["A"], "dateTime": "2008-06-27T09:30:00.000-05:00"}, {"value": "524", "qualifiers": ["A"], "dateTime": "2008-06-27T09:45:00.000-05:00"}, {"value": "524", "qualifiers": ["A"], "dateTime": "2008-06-27T10:00:00.000-05:00"}, {"value": "524", "qualifiers": ["A"], "dateTime": "2008-06-27T10:15:00.000-05:00"}, {"value": "524", "qualifiers": ["A"], "dateTime": "2008-06-27T10:30:00.000-05:00"}, {"value": "524", "qualifiers": ["A"], "dateTime": "2008-06-27T10:45:00.000-05:00"}, {"value": "522", "qualifiers": ["A"], "dateTime": "2008-06-27T11:00:00.000-05:00"}, {"value": "518", "qualifiers": ["A"], "dateTime": "2008-06-27T11:15:00.000-05:00"}, {"value": "522", "qualifiers": ["A"], "dateTime": "2008-06-27T11:30:00.000-05:00"}, {"value": "522", "qualifiers": ["A"], "dateTime": "2008-06-27T11:45:00.000-05:00"}, {"value": "520", "qualifiers": ["A"], "dateTime": "2008-06-27T12:00:00.000-05:00"}, {"value": "520", "qualifiers": ["A"], "dateTime": "2008-06-27T12:15:00.000-05:00"}, {"value": "518", "qualifiers": ["A"], "dateTime": "2008-06-27T12:30:00.000-05:00"}, {"value": "520", "qualifiers": ["A"], "dateTime": "2008-06-27T12:45:00.000-05:00"}, {"value": "518", "qualifiers": ["A"], "dateTime": "2008-06-27T13:00:00.000-05:00"}, {"value": "518", "qualifiers": ["A"], "dateTime": "2008-06-27T13:15:00.000-05:00"}, {"value": "518", "qualifiers": ["A"], "dateTime": "2008-06-27T13:30:00.000-05:00"}, {"value": "516", "qualifiers": ["A"], "dateTime": "2008-06-27T13:45:00.000-05:00"}, {"value": "516", "qualifiers": ["A"], "dateTime": "2008-06-27T14:00:00.000-05:00"}, {"value": "518", "qualifiers": ["A"], "dateTime": "2008-06-27T14:15:00.000-05:00"}, {"value": "518", "qualifiers": ["A"], "dateTime": "2008-06-27T14:30:00.000-05:00"}, {"value": "518", "qualifiers": ["A"], "dateTime": "2008-06-27T14:45:00.000-05:00"}, {"value": "516", "qualifiers": ["A"], "dateTime": "2008-06-27T15:00:00.000-05:00"}, {"value": "516", "qualifiers": ["A"], "dateTime": "2008-06-27T15:15:00.000-05:00"}, {"value": "516", "qualifiers": ["A"], "dateTime": "2008-06-27T15:30:00.000-05:00"}, {"value": "516", "qualifiers": ["A"], "dateTime": "2008-06-27T15:45:00.000-05:00"}, {"value": "516", "qualifiers": ["A"], "dateTime": "2008-06-27T16:00:00.000-05:00"}, {"value": "514", "qualifiers": ["A"], "dateTime": "2008-06-27T16:15:00.000-05:00"}, {"value": "514", "qualifiers": ["A"], "dateTime": "2008-06-27T16:30:00.000-05:00"}, {"value": "514", "qualifiers": ["A"], "dateTime": "2008-06-27T16:45:00.000-05:00"}, {"value": "512", "qualifiers": ["A"], "dateTime": "2008-06-27T17:00:00.000-05:00"}, {"value": "512", "qualifiers": ["A"], "dateTime": "2008-06-27T17:15:00.000-05:00"}, {"value": "512", "qualifiers": ["A"], "dateTime": "2008-06-27T17:30:00.000-05:00"}, {"value": "512", "qualifiers": ["A"], "dateTime": "2008-06-27T17:45:00.000-05:00"}, {"value": "512", "qualifiers": ["A"], "dateTime": "2008-06-27T18:00:00.000-05:00"}, {"value": "512", "qualifiers": ["A"], "dateTime": "2008-06-27T18:15:00.000-05:00"}, {"value": "512", "qualifiers": ["A"], "dateTime": "2008-06-27T18:30:00.000-05:00"}, {"value": "510", "qualifiers": ["A"], "dateTime": "2008-06-27T18:45:00.000-05:00"}, {"value": "510", "qualifiers": ["A"], "dateTime": "2008-06-27T19:00:00.000-05:00"}, {"value": "508", "qualifiers": ["A"], "dateTime": "2008-06-27T19:15:00.000-05:00"}, {"value": "510", "qualifiers": ["A"], "dateTime": "2008-06-27T19:30:00.000-05:00"}, {"value": "508", "qualifiers": ["A"], "dateTime": "2008-06-27T19:45:00.000-05:00"}, {"value": "508", "qualifiers": ["A"], "dateTime": "2008-06-27T20:00:00.000-05:00"}, {"value": "508", "qualifiers": ["A"], "dateTime": "2008-06-27T20:15:00.000-05:00"}, {"value": "508", "qualifiers": ["A"], "dateTime": "2008-06-27T20:30:00.000-05:00"}, {"value": "508", "qualifiers": ["A"], "dateTime": "2008-06-27T20:45:00.000-05:00"}, {"value": "508", "qualifiers": ["A"], "dateTime": "2008-06-27T21:00:00.000-05:00"}, {"value": "506", "qualifiers": ["A"], "dateTime": "2008-06-27T21:15:00.000-05:00"}, {"value": "506", "qualifiers": ["A"], "dateTime": "2008-06-27T21:30:00.000-05:00"}, {"value": "506", "qualifiers": ["A"], "dateTime": "2008-06-27T21:45:00.000-05:00"}, {"value": "506", "qualifiers": ["A"], "dateTime": "2008-06-27T22:00:00.000-05:00"}, {"value": "506", "qualifiers": ["A"], "dateTime": "2008-06-27T22:15:00.000-05:00"}, {"value": "504", "qualifiers": ["A"], "dateTime": "2008-06-27T22:30:00.000-05:00"}, {"value": "504", "qualifiers": ["A"], "dateTime": "2008-06-27T22:45:00.000-05:00"}, {"value": "503", "qualifiers": ["A"], "dateTime": "2008-06-27T23:00:00.000-05:00"}, {"value": "503", "qualifiers": ["A"], "dateTime": "2008-06-27T23:15:00.000-05:00"}, {"value": "503", "qualifiers": ["A"], "dateTime": "2008-06-27T23:30:00.000-05:00"}, {"value": "503", "qualifiers": ["A"], "dateTime": "2008-06-27T23:45:00.000-05:00"}, {"value": "503", "qualifiers": ["A"], "dateTime": "2008-06-28T00:00:00.000-05:00"}, {"value": "503", "qualifiers": ["A"], "dateTime": "2008-06-28T00:15:00.000-05:00"}, {"value": "501", "qualifiers": ["A"], "dateTime": "2008-06-28T00:30:00.000-05:00"}, {"value": "503", "qualifiers": ["A"], "dateTime": "2008-06-28T00:45:00.000-05:00"}, {"value": "501", "qualifiers": ["A"], "dateTime": "2008-06-28T01:00:00.000-05:00"}, {"value": "501", "qualifiers": ["A"], "dateTime": "2008-06-28T01:15:00.000-05:00"}, {"value": "499", "qualifiers": ["A"], "dateTime": "2008-06-28T01:30:00.000-05:00"}, {"value": "499", "qualifiers": ["A"], "dateTime": "2008-06-28T01:45:00.000-05:00"}, {"value": "499", "qualifiers": ["A"], "dateTime": "2008-06-28T02:00:00.000-05:00"}, {"value": "499", "qualifiers": ["A"], "dateTime": "2008-06-28T02:15:00.000-05:00"}, {"value": "499", "qualifiers": ["A"], "dateTime": "2008-06-28T02:30:00.000-05:00"}, {"value": "499", "qualifiers": ["A"], "dateTime": "2008-06-28T02:45:00.000-05:00"}, {"value": "497", "qualifiers": ["A"], "dateTime": "2008-06-28T03:00:00.000-05:00"}, {"value": "497", "qualifiers": ["A"], "dateTime": "2008-06-28T03:15:00.000-05:00"}, {"value": "497", "qualifiers": ["A"], "dateTime": "2008-06-28T03:30:00.000-05:00"}, {"value": "497", "qualifiers": ["A"], "dateTime": "2008-06-28T03:45:00.000-05:00"}, {"value": "497", "qualifiers": ["A"], "dateTime": "2008-06-28T04:00:00.000-05:00"}, {"value": "497", "qualifiers": ["A"], "dateTime": "2008-06-28T04:15:00.000-05:00"}, {"value": "497", "qualifiers": ["A"], "dateTime": "2008-06-28T04:30:00.000-05:00"}, {"value": "495", "qualifiers": ["A"], "dateTime": "2008-06-28T04:45:00.000-05:00"}, {"value": "495", "qualifiers": ["A"], "dateTime": "2008-06-28T05:00:00.000-05:00"}, {"value": "495", "qualifiers": ["A"], "dateTime": "2008-06-28T05:15:00.000-05:00"}, {"value": "495", "qualifiers": ["A"], "dateTime": "2008-06-28T05:30:00.000-05:00"}, {"value": "495", "qualifiers": ["A"], "dateTime": "2008-06-28T05:45:00.000-05:00"}, {"value": "491", "qualifiers": ["A"], "dateTime": "2008-06-28T06:00:00.000-05:00"}, {"value": "493", "qualifiers": ["A"], "dateTime": "2008-06-28T06:15:00.000-05:00"}, {"value": "493", "qualifiers": ["A"], "dateTime": "2008-06-28T06:30:00.000-05:00"}, {"value": "493", "qualifiers": ["A"], "dateTime": "2008-06-28T06:45:00.000-05:00"}, {"value": "493", "qualifiers": ["A"], "dateTime": "2008-06-28T07:00:00.000-05:00"}, {"value": "491", "qualifiers": ["A"], "dateTime": "2008-06-28T07:15:00.000-05:00"}, {"value": "491", "qualifiers": ["A"], "dateTime": "2008-06-28T07:30:00.000-05:00"}, {"value": "491", "qualifiers": ["A"], "dateTime": "2008-06-28T07:45:00.000-05:00"}, {"value": "491", "qualifiers": ["A"], "dateTime": "2008-06-28T08:00:00.000-05:00"}, {"value": "489", "qualifiers": ["A"], "dateTime": "2008-06-28T08:15:00.000-05:00"}, {"value": "489", "qualifiers": ["A"], "dateTime": "2008-06-28T08:30:00.000-05:00"}, {"value": "487", "qualifiers": ["A"], "dateTime": "2008-06-28T08:45:00.000-05:00"}, {"value": "489", "qualifiers": ["A"], "dateTime": "2008-06-28T09:00:00.000-05:00"}, {"value": "487", "qualifiers": ["A"], "dateTime": "2008-06-28T09:15:00.000-05:00"}, {"value": "487", "qualifiers": ["A"], "dateTime": "2008-06-28T09:30:00.000-05:00"}, {"value": "487", "qualifiers": ["A"], "dateTime": "2008-06-28T09:45:00.000-05:00"}, {"value": "487", "qualifiers": ["A"], "dateTime": "2008-06-28T10:00:00.000-05:00"}, {"value": "485", "qualifiers": ["A"], "dateTime": "2008-06-28T10:15:00.000-05:00"}, {"value": "485", "qualifiers": ["A"], "dateTime": "2008-06-28T10:30:00.000-05:00"}, {"value": "487", "qualifiers": ["A"], "dateTime": "2008-06-28T10:45:00.000-05:00"}, {"value": "485", "qualifiers": ["A"], "dateTime": "2008-06-28T11:00:00.000-05:00"}, {"value": "487", "qualifiers": ["A"], "dateTime": "2008-06-28T11:15:00.000-05:00"}, {"value": "483", "qualifiers": ["A"], "dateTime": "2008-06-28T11:30:00.000-05:00"}, {"value": "483", "qualifiers": ["A"], "dateTime": "2008-06-28T11:45:00.000-05:00"}, {"value": "485", "qualifiers": ["A"], "dateTime": "2008-06-28T12:00:00.000-05:00"}, {"value": "481", "qualifiers": ["A"], "dateTime": "2008-06-28T12:15:00.000-05:00"}, {"value": "483", "qualifiers": ["A"], "dateTime": "2008-06-28T12:30:00.000-05:00"}, {"value": "481", "qualifiers": ["A"], "dateTime": "2008-06-28T12:45:00.000-05:00"}, {"value": "483", "qualifiers": ["A"], "dateTime": "2008-06-28T13:00:00.000-05:00"}, {"value": "483", "qualifiers": ["A"], "dateTime": "2008-06-28T13:15:00.000-05:00"}, {"value": "481", "qualifiers": ["A"], "dateTime": "2008-06-28T13:30:00.000-05:00"}, {"value": "481", "qualifiers": ["A"], "dateTime": "2008-06-28T13:45:00.000-05:00"}, {"value": "485", "qualifiers": ["A"], "dateTime": "2008-06-28T14:00:00.000-05:00"}, {"value": "481", "qualifiers": ["A"], "dateTime": "2008-06-28T14:15:00.000-05:00"}, {"value": "481", "qualifiers": ["A"], "dateTime": "2008-06-28T14:30:00.000-05:00"}, {"value": "481", "qualifiers": ["A"], "dateTime": "2008-06-28T14:45:00.000-05:00"}, {"value": "481", "qualifiers": ["A"], "dateTime": "2008-06-28T15:00:00.000-05:00"}, {"value": "480", "qualifiers": ["A"], "dateTime": "2008-06-28T15:15:00.000-05:00"}, {"value": "478", "qualifiers": ["A"], "dateTime": "2008-06-28T15:30:00.000-05:00"}, {"value": "478", "qualifiers": ["A"], "dateTime": "2008-06-28T15:45:00.000-05:00"}, {"value": "478", "qualifiers": ["A"], "dateTime": "2008-06-28T16:00:00.000-05:00"}, {"value": "480", "qualifiers": ["A"], "dateTime": "2008-06-28T16:15:00.000-05:00"}, {"value": "476", "qualifiers": ["A"], "dateTime": "2008-06-28T16:30:00.000-05:00"}, {"value": "478", "qualifiers": ["A"], "dateTime": "2008-06-28T16:45:00.000-05:00"}, {"value": "478", "qualifiers": ["A"], "dateTime": "2008-06-28T17:00:00.000-05:00"}, {"value": "476", "qualifiers": ["A"], "dateTime": "2008-06-28T17:15:00.000-05:00"}, {"value": "476", "qualifiers": ["A"], "dateTime": "2008-06-28T17:30:00.000-05:00"}, {"value": "476", "qualifiers": ["A"], "dateTime": "2008-06-28T17:45:00.000-05:00"}, {"value": "474", "qualifiers": ["A"], "dateTime": "2008-06-28T18:00:00.000-05:00"}, {"value": "476", "qualifiers": ["A"], "dateTime": "2008-06-28T18:15:00.000-05:00"}, {"value": "474", "qualifiers": ["A"], "dateTime": "2008-06-28T18:30:00.000-05:00"}, {"value": "472", "qualifiers": ["A"], "dateTime": "2008-06-28T18:45:00.000-05:00"}, {"value": "476", "qualifiers": ["A"], "dateTime": "2008-06-28T19:00:00.000-05:00"}, {"value": "474", "qualifiers": ["A"], "dateTime": "2008-06-28T19:15:00.000-05:00"}, {"value": "474", "qualifiers": ["A"], "dateTime": "2008-06-28T19:30:00.000-05:00"}, {"value": "474", "qualifiers": ["A"], "dateTime": "2008-06-28T19:45:00.000-05:00"}, {"value": "472", "qualifiers": ["A"], "dateTime": "2008-06-28T20:00:00.000-05:00"}, {"value": "472", "qualifiers": ["A"], "dateTime": "2008-06-28T20:15:00.000-05:00"}, {"value": "472", "qualifiers": ["A"], "dateTime": "2008-06-28T20:30:00.000-05:00"}, {"value": "472", "qualifiers": ["A"], "dateTime": "2008-06-28T20:45:00.000-05:00"}, {"value": "472", "qualifiers": ["A"], "dateTime": "2008-06-28T21:00:00.000-05:00"}, {"value": "472", "qualifiers": ["A"], "dateTime": "2008-06-28T21:15:00.000-05:00"}, {"value": "472", "qualifiers": ["A"], "dateTime": "2008-06-28T21:30:00.000-05:00"}, {"value": "470", "qualifiers": ["A"], "dateTime": "2008-06-28T21:45:00.000-05:00"}, {"value": "472", "qualifiers": ["A"], "dateTime": "2008-06-28T22:00:00.000-05:00"}, {"value": "470", "qualifiers": ["A"], "dateTime": "2008-06-28T22:15:00.000-05:00"}, {"value": "470", "qualifiers": ["A"], "dateTime": "2008-06-28T22:30:00.000-05:00"}, {"value": "470", "qualifiers": ["A"], "dateTime": "2008-06-28T22:45:00.000-05:00"}, {"value": "470", "qualifiers": ["A"], "dateTime": "2008-06-28T23:00:00.000-05:00"}, {"value": "470", "qualifiers": ["A"], "dateTime": "2008-06-28T23:15:00.000-05:00"}, {"value": "470", "qualifiers": ["A"], "dateTime": "2008-06-28T23:30:00.000-05:00"}, {"value": "470", "qualifiers": ["A"], "dateTime": "2008-06-28T23:45:00.000-05:00"}, {"value": "470", "qualifiers": ["A"], "dateTime": "2008-06-29T00:00:00.000-05:00"}, {"value": "468", "qualifiers": ["A"], "dateTime": "2008-06-29T00:15:00.000-05:00"}, {"value": "468", "qualifiers": ["A"], "dateTime": "2008-06-29T00:30:00.000-05:00"}, {"value": "468", "qualifiers": ["A"], "dateTime": "2008-06-29T00:45:00.000-05:00"}, {"value": "468", "qualifiers": ["A"], "dateTime": "2008-06-29T01:00:00.000-05:00"}, {"value": "466", "qualifiers": ["A"], "dateTime": "2008-06-29T01:15:00.000-05:00"}, {"value": "466", "qualifiers": ["A"], "dateTime": "2008-06-29T01:30:00.000-05:00"}, {"value": "466", "qualifiers": ["A"], "dateTime": "2008-06-29T01:45:00.000-05:00"}, {"value": "468", "qualifiers": ["A"], "dateTime": "2008-06-29T02:00:00.000-05:00"}, {"value": "466", "qualifiers": ["A"], "dateTime": "2008-06-29T02:15:00.000-05:00"}, {"value": "466", "qualifiers": ["A"], "dateTime": "2008-06-29T02:30:00.000-05:00"}, {"value": "466", "qualifiers": ["A"], "dateTime": "2008-06-29T02:45:00.000-05:00"}, {"value": "466", "qualifiers": ["A"], "dateTime": "2008-06-29T03:00:00.000-05:00"}, {"value": "466", "qualifiers": ["A"], "dateTime": "2008-06-29T03:15:00.000-05:00"}, {"value": "466", "qualifiers": ["A"], "dateTime": "2008-06-29T03:30:00.000-05:00"}, {"value": "466", "qualifiers": ["A"], "dateTime": "2008-06-29T03:45:00.000-05:00"}, {"value": "466", "qualifiers": ["A"], "dateTime": "2008-06-29T04:00:00.000-05:00"}, {"value": "465", "qualifiers": ["A"], "dateTime": "2008-06-29T04:15:00.000-05:00"}, {"value": "466", "qualifiers": ["A"], "dateTime": "2008-06-29T04:30:00.000-05:00"}, {"value": "466", "qualifiers": ["A"], "dateTime": "2008-06-29T04:45:00.000-05:00"}, {"value": "465", "qualifiers": ["A"], "dateTime": "2008-06-29T05:00:00.000-05:00"}, {"value": "465", "qualifiers": ["A"], "dateTime": "2008-06-29T05:15:00.000-05:00"}, {"value": "465", "qualifiers": ["A"], "dateTime": "2008-06-29T05:30:00.000-05:00"}, {"value": "465", "qualifiers": ["A"], "dateTime": "2008-06-29T05:45:00.000-05:00"}, {"value": "465", "qualifiers": ["A"], "dateTime": "2008-06-29T06:00:00.000-05:00"}, {"value": "465", "qualifiers": ["A"], "dateTime": "2008-06-29T06:15:00.000-05:00"}, {"value": "465", "qualifiers": ["A"], "dateTime": "2008-06-29T06:30:00.000-05:00"}, {"value": "463", "qualifiers": ["A"], "dateTime": "2008-06-29T06:45:00.000-05:00"}, {"value": "465", "qualifiers": ["A"], "dateTime": "2008-06-29T07:00:00.000-05:00"}, {"value": "461", "qualifiers": ["A"], "dateTime": "2008-06-29T07:15:00.000-05:00"}, {"value": "463", "qualifiers": ["A"], "dateTime": "2008-06-29T07:30:00.000-05:00"}, {"value": "461", "qualifiers": ["A"], "dateTime": "2008-06-29T07:45:00.000-05:00"}, {"value": "463", "qualifiers": ["A"], "dateTime": "2008-06-29T08:00:00.000-05:00"}, {"value": "463", "qualifiers": ["A"], "dateTime": "2008-06-29T08:15:00.000-05:00"}, {"value": "461", "qualifiers": ["A"], "dateTime": "2008-06-29T08:30:00.000-05:00"}, {"value": "463", "qualifiers": ["A"], "dateTime": "2008-06-29T08:45:00.000-05:00"}, {"value": "459", "qualifiers": ["A"], "dateTime": "2008-06-29T09:00:00.000-05:00"}, {"value": "463", "qualifiers": ["A"], "dateTime": "2008-06-29T09:15:00.000-05:00"}, {"value": "461", "qualifiers": ["A"], "dateTime": "2008-06-29T09:30:00.000-05:00"}, {"value": "461", "qualifiers": ["A"], "dateTime": "2008-06-29T09:45:00.000-05:00"}, {"value": "459", "qualifiers": ["A"], "dateTime": "2008-06-29T10:00:00.000-05:00"}, {"value": "459", "qualifiers": ["A"], "dateTime": "2008-06-29T10:15:00.000-05:00"}, {"value": "459", "qualifiers": ["A"], "dateTime": "2008-06-29T10:30:00.000-05:00"}, {"value": "459", "qualifiers": ["A"], "dateTime": "2008-06-29T10:45:00.000-05:00"}, {"value": "459", "qualifiers": ["A"], "dateTime": "2008-06-29T11:00:00.000-05:00"}, {"value": "459", "qualifiers": ["A"], "dateTime": "2008-06-29T11:15:00.000-05:00"}, {"value": "459", "qualifiers": ["A"], "dateTime": "2008-06-29T11:30:00.000-05:00"}, {"value": "459", "qualifiers": ["A"], "dateTime": "2008-06-29T11:45:00.000-05:00"}, {"value": "457", "qualifiers": ["A"], "dateTime": "2008-06-29T12:00:00.000-05:00"}, {"value": "457", "qualifiers": ["A"], "dateTime": "2008-06-29T12:15:00.000-05:00"}, {"value": "457", "qualifiers": ["A"], "dateTime": "2008-06-29T12:30:00.000-05:00"}, {"value": "457", "qualifiers": ["A"], "dateTime": "2008-06-29T12:45:00.000-05:00"}, {"value": "457", "qualifiers": ["A"], "dateTime": "2008-06-29T13:00:00.000-05:00"}, {"value": "455", "qualifiers": ["A"], "dateTime": "2008-06-29T13:15:00.000-05:00"}, {"value": "455", "qualifiers": ["A"], "dateTime": "2008-06-29T13:30:00.000-05:00"}, {"value": "453", "qualifiers": ["A"], "dateTime": "2008-06-29T13:45:00.000-05:00"}, {"value": "455", "qualifiers": ["A"], "dateTime": "2008-06-29T14:00:00.000-05:00"}, {"value": "453", "qualifiers": ["A"], "dateTime": "2008-06-29T14:15:00.000-05:00"}, {"value": "455", "qualifiers": ["A"], "dateTime": "2008-06-29T14:30:00.000-05:00"}, {"value": "455", "qualifiers": ["A"], "dateTime": "2008-06-29T14:45:00.000-05:00"}, {"value": "453", "qualifiers": ["A"], "dateTime": "2008-06-29T15:00:00.000-05:00"}, {"value": "453", "qualifiers": ["A"], "dateTime": "2008-06-29T15:15:00.000-05:00"}, {"value": "453", "qualifiers": ["A"], "dateTime": "2008-06-29T15:30:00.000-05:00"}, {"value": "453", "qualifiers": ["A"], "dateTime": "2008-06-29T15:45:00.000-05:00"}, {"value": "452", "qualifiers": ["A"], "dateTime": "2008-06-29T16:00:00.000-05:00"}, {"value": "453", "qualifiers": ["A"], "dateTime": "2008-06-29T16:15:00.000-05:00"}, {"value": "453", "qualifiers": ["A"], "dateTime": "2008-06-29T16:30:00.000-05:00"}, {"value": "452", "qualifiers": ["A"], "dateTime": "2008-06-29T16:45:00.000-05:00"}, {"value": "452", "qualifiers": ["A"], "dateTime": "2008-06-29T17:00:00.000-05:00"}, {"value": "450", "qualifiers": ["A"], "dateTime": "2008-06-29T17:15:00.000-05:00"}, {"value": "450", "qualifiers": ["A"], "dateTime": "2008-06-29T17:30:00.000-05:00"}, {"value": "452", "qualifiers": ["A"], "dateTime": "2008-06-29T17:45:00.000-05:00"}, {"value": "450", "qualifiers": ["A"], "dateTime": "2008-06-29T18:00:00.000-05:00"}, {"value": "450", "qualifiers": ["A"], "dateTime": "2008-06-29T18:15:00.000-05:00"}, {"value": "450", "qualifiers": ["A"], "dateTime": "2008-06-29T18:30:00.000-05:00"}, {"value": "450", "qualifiers": ["A"], "dateTime": "2008-06-29T18:45:00.000-05:00"}, {"value": "448", "qualifiers": ["A"], "dateTime": "2008-06-29T19:00:00.000-05:00"}, {"value": "448", "qualifiers": ["A"], "dateTime": "2008-06-29T19:15:00.000-05:00"}, {"value": "448", "qualifiers": ["A"], "dateTime": "2008-06-29T19:30:00.000-05:00"}, {"value": "448", "qualifiers": ["A"], "dateTime": "2008-06-29T19:45:00.000-05:00"}, {"value": "446", "qualifiers": ["A"], "dateTime": "2008-06-29T20:00:00.000-05:00"}, {"value": "446", "qualifiers": ["A"], "dateTime": "2008-06-29T20:15:00.000-05:00"}, {"value": "446", "qualifiers": ["A"], "dateTime": "2008-06-29T20:30:00.000-05:00"}, {"value": "444", "qualifiers": ["A"], "dateTime": "2008-06-29T20:45:00.000-05:00"}, {"value": "444", "qualifiers": ["A"], "dateTime": "2008-06-29T21:00:00.000-05:00"}, {"value": "444", "qualifiers": ["A"], "dateTime": "2008-06-29T21:15:00.000-05:00"}, {"value": "442", "qualifiers": ["A"], "dateTime": "2008-06-29T21:30:00.000-05:00"}, {"value": "444", "qualifiers": ["A"], "dateTime": "2008-06-29T21:45:00.000-05:00"}, {"value": "444", "qualifiers": ["A"], "dateTime": "2008-06-29T22:00:00.000-05:00"}, {"value": "442", "qualifiers": ["A"], "dateTime": "2008-06-29T22:15:00.000-05:00"}, {"value": "442", "qualifiers": ["A"], "dateTime": "2008-06-29T22:30:00.000-05:00"}, {"value": "441", "qualifiers": ["A"], "dateTime": "2008-06-29T22:45:00.000-05:00"}, {"value": "441", "qualifiers": ["A"], "dateTime": "2008-06-29T23:00:00.000-05:00"}, {"value": "442", "qualifiers": ["A"], "dateTime": "2008-06-29T23:15:00.000-05:00"}, {"value": "441", "qualifiers": ["A"], "dateTime": "2008-06-29T23:30:00.000-05:00"}, {"value": "441", "qualifiers": ["A"], "dateTime": "2008-06-29T23:45:00.000-05:00"}, {"value": "439", "qualifiers": ["A"], "dateTime": "2008-06-30T00:00:00.000-05:00"}, {"value": "439", "qualifiers": ["A"], "dateTime": "2008-06-30T00:15:00.000-05:00"}, {"value": "439", "qualifiers": ["A"], "dateTime": "2008-06-30T00:30:00.000-05:00"}, {"value": "437", "qualifiers": ["A"], "dateTime": "2008-06-30T00:45:00.000-05:00"}, {"value": "437", "qualifiers": ["A"], "dateTime": "2008-06-30T01:00:00.000-05:00"}, {"value": "437", "qualifiers": ["A"], "dateTime": "2008-06-30T01:15:00.000-05:00"}, {"value": "437", "qualifiers": ["A"], "dateTime": "2008-06-30T01:30:00.000-05:00"}, {"value": "435", "qualifiers": ["A"], "dateTime": "2008-06-30T01:45:00.000-05:00"}, {"value": "437", "qualifiers": ["A"], "dateTime": "2008-06-30T02:00:00.000-05:00"}, {"value": "435", "qualifiers": ["A"], "dateTime": "2008-06-30T02:15:00.000-05:00"}, {"value": "435", "qualifiers": ["A"], "dateTime": "2008-06-30T02:30:00.000-05:00"}, {"value": "435", "qualifiers": ["A"], "dateTime": "2008-06-30T02:45:00.000-05:00"}, {"value": "433", "qualifiers": ["A"], "dateTime": "2008-06-30T03:00:00.000-05:00"}, {"value": "433", "qualifiers": ["A"], "dateTime": "2008-06-30T03:15:00.000-05:00"}, {"value": "435", "qualifiers": ["A"], "dateTime": "2008-06-30T03:30:00.000-05:00"}, {"value": "433", "qualifiers": ["A"], "dateTime": "2008-06-30T03:45:00.000-05:00"}, {"value": "433", "qualifiers": ["A"], "dateTime": "2008-06-30T04:00:00.000-05:00"}, {"value": "432", "qualifiers": ["A"], "dateTime": "2008-06-30T04:15:00.000-05:00"}, {"value": "432", "qualifiers": ["A"], "dateTime": "2008-06-30T04:30:00.000-05:00"}, {"value": "433", "qualifiers": ["A"], "dateTime": "2008-06-30T04:45:00.000-05:00"}, {"value": "432", "qualifiers": ["A"], "dateTime": "2008-06-30T05:00:00.000-05:00"}, {"value": "433", "qualifiers": ["A"], "dateTime": "2008-06-30T05:15:00.000-05:00"}, {"value": "432", "qualifiers": ["A"], "dateTime": "2008-06-30T05:30:00.000-05:00"}, {"value": "432", "qualifiers": ["A"], "dateTime": "2008-06-30T05:45:00.000-05:00"}, {"value": "432", "qualifiers": ["A"], "dateTime": "2008-06-30T06:00:00.000-05:00"}, {"value": "430", "qualifiers": ["A"], "dateTime": "2008-06-30T06:15:00.000-05:00"}, {"value": "430", "qualifiers": ["A"], "dateTime": "2008-06-30T06:30:00.000-05:00"}, {"value": "430", "qualifiers": ["A"], "dateTime": "2008-06-30T06:45:00.000-05:00"}, {"value": "430", "qualifiers": ["A"], "dateTime": "2008-06-30T07:00:00.000-05:00"}, {"value": "430", "qualifiers": ["A"], "dateTime": "2008-06-30T07:15:00.000-05:00"}, {"value": "430", "qualifiers": ["A"], "dateTime": "2008-06-30T07:30:00.000-05:00"}, {"value": "430", "qualifiers": ["A"], "dateTime": "2008-06-30T07:45:00.000-05:00"}, {"value": "428", "qualifiers": ["A"], "dateTime": "2008-06-30T08:00:00.000-05:00"}, {"value": "428", "qualifiers": ["A"], "dateTime": "2008-06-30T08:15:00.000-05:00"}, {"value": "428", "qualifiers": ["A"], "dateTime": "2008-06-30T08:30:00.000-05:00"}, {"value": "428", "qualifiers": ["A"], "dateTime": "2008-06-30T08:45:00.000-05:00"}, {"value": "428", "qualifiers": ["A"], "dateTime": "2008-06-30T09:00:00.000-05:00"}, {"value": "426", "qualifiers": ["A"], "dateTime": "2008-06-30T09:15:00.000-05:00"}, {"value": "426", "qualifiers": ["A"], "dateTime": "2008-06-30T09:30:00.000-05:00"}, {"value": "426", "qualifiers": ["A"], "dateTime": "2008-06-30T09:45:00.000-05:00"}, {"value": "426", "qualifiers": ["A"], "dateTime": "2008-06-30T10:00:00.000-05:00"}, {"value": "426", "qualifiers": ["A"], "dateTime": "2008-06-30T10:15:00.000-05:00"}, {"value": "426", "qualifiers": ["A"], "dateTime": "2008-06-30T10:30:00.000-05:00"}, {"value": "426", "qualifiers": ["A"], "dateTime": "2008-06-30T10:45:00.000-05:00"}, {"value": "426", "qualifiers": ["A"], "dateTime": "2008-06-30T11:00:00.000-05:00"}, {"value": "424", "qualifiers": ["A"], "dateTime": "2008-06-30T11:15:00.000-05:00"}, {"value": "424", "qualifiers": ["A"], "dateTime": "2008-06-30T11:30:00.000-05:00"}, {"value": "424", "qualifiers": ["A"], "dateTime": "2008-06-30T11:45:00.000-05:00"}, {"value": "424", "qualifiers": ["A"], "dateTime": "2008-06-30T12:00:00.000-05:00"}, {"value": "424", "qualifiers": ["A"], "dateTime": "2008-06-30T12:15:00.000-05:00"}, {"value": "424", "qualifiers": ["A"], "dateTime": "2008-06-30T12:30:00.000-05:00"}, {"value": "424", "qualifiers": ["A"], "dateTime": "2008-06-30T12:45:00.000-05:00"}, {"value": "423", "qualifiers": ["A"], "dateTime": "2008-06-30T13:00:00.000-05:00"}, {"value": "424", "qualifiers": ["A"], "dateTime": "2008-06-30T13:15:00.000-05:00"}, {"value": "423", "qualifiers": ["A"], "dateTime": "2008-06-30T13:30:00.000-05:00"}, {"value": "421", "qualifiers": ["A"], "dateTime": "2008-06-30T13:45:00.000-05:00"}, {"value": "423", "qualifiers": ["A"], "dateTime": "2008-06-30T14:00:00.000-05:00"}, {"value": "423", "qualifiers": ["A"], "dateTime": "2008-06-30T14:15:00.000-05:00"}, {"value": "421", "qualifiers": ["A"], "dateTime": "2008-06-30T14:30:00.000-05:00"}, {"value": "423", "qualifiers": ["A"], "dateTime": "2008-06-30T14:45:00.000-05:00"}, {"value": "421", "qualifiers": ["A"], "dateTime": "2008-06-30T15:00:00.000-05:00"}, {"value": "421", "qualifiers": ["A"], "dateTime": "2008-06-30T15:15:00.000-05:00"}, {"value": "421", "qualifiers": ["A"], "dateTime": "2008-06-30T15:30:00.000-05:00"}, {"value": "419", "qualifiers": ["A"], "dateTime": "2008-06-30T15:45:00.000-05:00"}, {"value": "419", "qualifiers": ["A"], "dateTime": "2008-06-30T16:00:00.000-05:00"}, {"value": "419", "qualifiers": ["A"], "dateTime": "2008-06-30T16:15:00.000-05:00"}, {"value": "419", "qualifiers": ["A"], "dateTime": "2008-06-30T16:30:00.000-05:00"}, {"value": "419", "qualifiers": ["A"], "dateTime": "2008-06-30T16:45:00.000-05:00"}, {"value": "417", "qualifiers": ["A"], "dateTime": "2008-06-30T17:00:00.000-05:00"}, {"value": "417", "qualifiers": ["A"], "dateTime": "2008-06-30T17:15:00.000-05:00"}, {"value": "417", "qualifiers": ["A"], "dateTime": "2008-06-30T17:30:00.000-05:00"}, {"value": "417", "qualifiers": ["A"], "dateTime": "2008-06-30T17:45:00.000-05:00"}, {"value": "417", "qualifiers": ["A"], "dateTime": "2008-06-30T18:00:00.000-05:00"}, {"value": "417", "qualifiers": ["A"], "dateTime": "2008-06-30T18:15:00.000-05:00"}, {"value": "415", "qualifiers": ["A"], "dateTime": "2008-06-30T18:30:00.000-05:00"}, {"value": "415", "qualifiers": ["A"], "dateTime": "2008-06-30T18:45:00.000-05:00"}, {"value": "415", "qualifiers": ["A"], "dateTime": "2008-06-30T19:00:00.000-05:00"}, {"value": "415", "qualifiers": ["A"], "dateTime": "2008-06-30T19:15:00.000-05:00"}, {"value": "415", "qualifiers": ["A"], "dateTime": "2008-06-30T19:30:00.000-05:00"}, {"value": "414", "qualifiers": ["A"], "dateTime": "2008-06-30T19:45:00.000-05:00"}, {"value": "414", "qualifiers": ["A"], "dateTime": "2008-06-30T20:00:00.000-05:00"}, {"value": "414", "qualifiers": ["A"], "dateTime": "2008-06-30T20:15:00.000-05:00"}, {"value": "414", "qualifiers": ["A"], "dateTime": "2008-06-30T20:30:00.000-05:00"}, {"value": "414", "qualifiers": ["A"], "dateTime": "2008-06-30T20:45:00.000-05:00"}, {"value": "414", "qualifiers": ["A"], "dateTime": "2008-06-30T21:00:00.000-05:00"}, {"value": "414", "qualifiers": ["A"], "dateTime": "2008-06-30T21:15:00.000-05:00"}, {"value": "414", "qualifiers": ["A"], "dateTime": "2008-06-30T21:30:00.000-05:00"}, {"value": "412", "qualifiers": ["A"], "dateTime": "2008-06-30T21:45:00.000-05:00"}, {"value": "412", "qualifiers": ["A"], "dateTime": "2008-06-30T22:00:00.000-05:00"}, {"value": "410", "qualifiers": ["A"], "dateTime": "2008-06-30T22:15:00.000-05:00"}, {"value": "412", "qualifiers": ["A"], "dateTime": "2008-06-30T22:30:00.000-05:00"}, {"value": "412", "qualifiers": ["A"], "dateTime": "2008-06-30T22:45:00.000-05:00"}, {"value": "412", "qualifiers": ["A"], "dateTime": "2008-06-30T23:00:00.000-05:00"}, {"value": "412", "qualifiers": ["A"], "dateTime": "2008-06-30T23:15:00.000-05:00"}, {"value": "410", "qualifiers": ["A"], "dateTime": "2008-06-30T23:30:00.000-05:00"}, {"value": "410", "qualifiers": ["A"], "dateTime": "2008-06-30T23:45:00.000-05:00"}, {"value": "410", "qualifiers": ["A"], "dateTime": "2008-07-01T00:00:00.000-05:00"}, {"value": "410", "qualifiers": ["A"], "dateTime": "2008-07-01T00:15:00.000-05:00"}, {"value": "408", "qualifiers": ["A"], "dateTime": "2008-07-01T00:30:00.000-05:00"}, {"value": "410", "qualifiers": ["A"], "dateTime": "2008-07-01T00:45:00.000-05:00"}, {"value": "408", "qualifiers": ["A"], "dateTime": "2008-07-01T01:00:00.000-05:00"}, {"value": "408", "qualifiers": ["A"], "dateTime": "2008-07-01T01:15:00.000-05:00"}, {"value": "408", "qualifiers": ["A"], "dateTime": "2008-07-01T01:30:00.000-05:00"}, {"value": "408", "qualifiers": ["A"], "dateTime": "2008-07-01T01:45:00.000-05:00"}, {"value": "408", "qualifiers": ["A"], "dateTime": "2008-07-01T02:00:00.000-05:00"}, {"value": "407", "qualifiers": ["A"], "dateTime": "2008-07-01T02:15:00.000-05:00"}, {"value": "408", "qualifiers": ["A"], "dateTime": "2008-07-01T02:30:00.000-05:00"}, {"value": "408", "qualifiers": ["A"], "dateTime": "2008-07-01T02:45:00.000-05:00"}, {"value": "407", "qualifiers": ["A"], "dateTime": "2008-07-01T03:00:00.000-05:00"}, {"value": "407", "qualifiers": ["A"], "dateTime": "2008-07-01T03:15:00.000-05:00"}, {"value": "407", "qualifiers": ["A"], "dateTime": "2008-07-01T03:30:00.000-05:00"}, {"value": "407", "qualifiers": ["A"], "dateTime": "2008-07-01T03:45:00.000-05:00"}, {"value": "407", "qualifiers": ["A"], "dateTime": "2008-07-01T04:00:00.000-05:00"}, {"value": "407", "qualifiers": ["A"], "dateTime": "2008-07-01T04:15:00.000-05:00"}, {"value": "407", "qualifiers": ["A"], "dateTime": "2008-07-01T04:30:00.000-05:00"}, {"value": "407", "qualifiers": ["A"], "dateTime": "2008-07-01T04:45:00.000-05:00"}, {"value": "407", "qualifiers": ["A"], "dateTime": "2008-07-01T05:00:00.000-05:00"}, {"value": "407", "qualifiers": ["A"], "dateTime": "2008-07-01T05:15:00.000-05:00"}, {"value": "407", "qualifiers": ["A"], "dateTime": "2008-07-01T05:30:00.000-05:00"}, {"value": "407", "qualifiers": ["A"], "dateTime": "2008-07-01T05:45:00.000-05:00"}, {"value": "407", "qualifiers": ["A"], "dateTime": "2008-07-01T06:00:00.000-05:00"}, {"value": "407", "qualifiers": ["A"], "dateTime": "2008-07-01T06:15:00.000-05:00"}, {"value": "407", "qualifiers": ["A"], "dateTime": "2008-07-01T06:30:00.000-05:00"}, {"value": "407", "qualifiers": ["A"], "dateTime": "2008-07-01T06:45:00.000-05:00"}, {"value": "407", "qualifiers": ["A"], "dateTime": "2008-07-01T07:00:00.000-05:00"}, {"value": "407", "qualifiers": ["A"], "dateTime": "2008-07-01T07:15:00.000-05:00"}, {"value": "405", "qualifiers": ["A"], "dateTime": "2008-07-01T07:30:00.000-05:00"}, {"value": "407", "qualifiers": ["A"], "dateTime": "2008-07-01T07:45:00.000-05:00"}, {"value": "407", "qualifiers": ["A"], "dateTime": "2008-07-01T08:00:00.000-05:00"}, {"value": "407", "qualifiers": ["A"], "dateTime": "2008-07-01T08:15:00.000-05:00"}, {"value": "405", "qualifiers": ["A"], "dateTime": "2008-07-01T08:30:00.000-05:00"}, {"value": "407", "qualifiers": ["A"], "dateTime": "2008-07-01T08:45:00.000-05:00"}, {"value": "405", "qualifiers": ["A"], "dateTime": "2008-07-01T09:00:00.000-05:00"}, {"value": "405", "qualifiers": ["A"], "dateTime": "2008-07-01T09:15:00.000-05:00"}, {"value": "405", "qualifiers": ["A"], "dateTime": "2008-07-01T09:30:00.000-05:00"}, {"value": "405", "qualifiers": ["A"], "dateTime": "2008-07-01T09:45:00.000-05:00"}, {"value": "403", "qualifiers": ["A"], "dateTime": "2008-07-01T10:00:00.000-05:00"}, {"value": "405", "qualifiers": ["A"], "dateTime": "2008-07-01T10:15:00.000-05:00"}, {"value": "403", "qualifiers": ["A"], "dateTime": "2008-07-01T10:30:00.000-05:00"}, {"value": "403", "qualifiers": ["A"], "dateTime": "2008-07-01T10:45:00.000-05:00"}, {"value": "401", "qualifiers": ["A"], "dateTime": "2008-07-01T11:00:00.000-05:00"}, {"value": "403", "qualifiers": ["A"], "dateTime": "2008-07-01T11:15:00.000-05:00"}, {"value": "403", "qualifiers": ["A"], "dateTime": "2008-07-01T11:30:00.000-05:00"}, {"value": "403", "qualifiers": ["A"], "dateTime": "2008-07-01T11:45:00.000-05:00"}, {"value": "405", "qualifiers": ["A"], "dateTime": "2008-07-01T12:00:00.000-05:00"}, {"value": "403", "qualifiers": ["A"], "dateTime": "2008-07-01T12:15:00.000-05:00"}, {"value": "403", "qualifiers": ["A"], "dateTime": "2008-07-01T12:30:00.000-05:00"}, {"value": "403", "qualifiers": ["A"], "dateTime": "2008-07-01T12:45:00.000-05:00"}, {"value": "403", "qualifiers": ["A"], "dateTime": "2008-07-01T13:00:00.000-05:00"}, {"value": "403", "qualifiers": ["A"], "dateTime": "2008-07-01T13:15:00.000-05:00"}, {"value": "401", "qualifiers": ["A"], "dateTime": "2008-07-01T13:30:00.000-05:00"}, {"value": "401", "qualifiers": ["A"], "dateTime": "2008-07-01T13:45:00.000-05:00"}, {"value": "401", "qualifiers": ["A"], "dateTime": "2008-07-01T14:00:00.000-05:00"}, {"value": "403", "qualifiers": ["A"], "dateTime": "2008-07-01T14:15:00.000-05:00"}, {"value": "401", "qualifiers": ["A"], "dateTime": "2008-07-01T14:30:00.000-05:00"}, {"value": "401", "qualifiers": ["A"], "dateTime": "2008-07-01T14:45:00.000-05:00"}, {"value": "401", "qualifiers": ["A"], "dateTime": "2008-07-01T15:00:00.000-05:00"}, {"value": "401", "qualifiers": ["A"], "dateTime": "2008-07-01T15:15:00.000-05:00"}, {"value": "400", "qualifiers": ["A"], "dateTime": "2008-07-01T15:30:00.000-05:00"}, {"value": "401", "qualifiers": ["A"], "dateTime": "2008-07-01T15:45:00.000-05:00"}, {"value": "400", "qualifiers": ["A"], "dateTime": "2008-07-01T16:00:00.000-05:00"}, {"value": "400", "qualifiers": ["A"], "dateTime": "2008-07-01T16:15:00.000-05:00"}, {"value": "398", "qualifiers": ["A"], "dateTime": "2008-07-01T16:30:00.000-05:00"}, {"value": "400", "qualifiers": ["A"], "dateTime": "2008-07-01T16:45:00.000-05:00"}, {"value": "400", "qualifiers": ["A"], "dateTime": "2008-07-01T17:00:00.000-05:00"}, {"value": "398", "qualifiers": ["A"], "dateTime": "2008-07-01T17:15:00.000-05:00"}, {"value": "398", "qualifiers": ["A"], "dateTime": "2008-07-01T17:30:00.000-05:00"}, {"value": "398", "qualifiers": ["A"], "dateTime": "2008-07-01T17:45:00.000-05:00"}, {"value": "398", "qualifiers": ["A"], "dateTime": "2008-07-01T18:00:00.000-05:00"}, {"value": "396", "qualifiers": ["A"], "dateTime": "2008-07-01T18:15:00.000-05:00"}, {"value": "398", "qualifiers": ["A"], "dateTime": "2008-07-01T18:30:00.000-05:00"}, {"value": "396", "qualifiers": ["A"], "dateTime": "2008-07-01T18:45:00.000-05:00"}, {"value": "396", "qualifiers": ["A"], "dateTime": "2008-07-01T19:00:00.000-05:00"}, {"value": "398", "qualifiers": ["A"], "dateTime": "2008-07-01T19:15:00.000-05:00"}, {"value": "396", "qualifiers": ["A"], "dateTime": "2008-07-01T19:30:00.000-05:00"}, {"value": "396", "qualifiers": ["A"], "dateTime": "2008-07-01T19:45:00.000-05:00"}, {"value": "396", "qualifiers": ["A"], "dateTime": "2008-07-01T20:00:00.000-05:00"}, {"value": "396", "qualifiers": ["A"], "dateTime": "2008-07-01T20:15:00.000-05:00"}, {"value": "396", "qualifiers": ["A"], "dateTime": "2008-07-01T20:30:00.000-05:00"}, {"value": "394", "qualifiers": ["A"], "dateTime": "2008-07-01T20:45:00.000-05:00"}, {"value": "396", "qualifiers": ["A"], "dateTime": "2008-07-01T21:00:00.000-05:00"}, {"value": "394", "qualifiers": ["A"], "dateTime": "2008-07-01T21:15:00.000-05:00"}, {"value": "394", "qualifiers": ["A"], "dateTime": "2008-07-01T21:30:00.000-05:00"}, {"value": "394", "qualifiers": ["A"], "dateTime": "2008-07-01T21:45:00.000-05:00"}, {"value": "394", "qualifiers": ["A"], "dateTime": "2008-07-01T22:00:00.000-05:00"}, {"value": "394", "qualifiers": ["A"], "dateTime": "2008-07-01T22:15:00.000-05:00"}, {"value": "394", "qualifiers": ["A"], "dateTime": "2008-07-01T22:30:00.000-05:00"}, {"value": "393", "qualifiers": ["A"], "dateTime": "2008-07-01T22:45:00.000-05:00"}, {"value": "393", "qualifiers": ["A"], "dateTime": "2008-07-01T23:00:00.000-05:00"}, {"value": "393", "qualifiers": ["A"], "dateTime": "2008-07-01T23:15:00.000-05:00"}, {"value": "393", "qualifiers": ["A"], "dateTime": "2008-07-01T23:30:00.000-05:00"}, {"value": "393", "qualifiers": ["A"], "dateTime": "2008-07-01T23:45:00.000-05:00"}, {"value": "393", "qualifiers": ["A"], "dateTime": "2008-07-02T00:00:00.000-05:00"}, {"value": "391", "qualifiers": ["A"], "dateTime": "2008-07-02T00:15:00.000-05:00"}, {"value": "393", "qualifiers": ["A"], "dateTime": "2008-07-02T00:30:00.000-05:00"}, {"value": "391", "qualifiers": ["A"], "dateTime": "2008-07-02T00:45:00.000-05:00"}, {"value": "391", "qualifiers": ["A"], "dateTime": "2008-07-02T01:00:00.000-05:00"}, {"value": "391", "qualifiers": ["A"], "dateTime": "2008-07-02T01:15:00.000-05:00"}, {"value": "393", "qualifiers": ["A"], "dateTime": "2008-07-02T01:30:00.000-05:00"}, {"value": "391", "qualifiers": ["A"], "dateTime": "2008-07-02T01:45:00.000-05:00"}, {"value": "391", "qualifiers": ["A"], "dateTime": "2008-07-02T02:00:00.000-05:00"}, {"value": "391", "qualifiers": ["A"], "dateTime": "2008-07-02T02:15:00.000-05:00"}, {"value": "391", "qualifiers": ["A"], "dateTime": "2008-07-02T02:30:00.000-05:00"}, {"value": "391", "qualifiers": ["A"], "dateTime": "2008-07-02T02:45:00.000-05:00"}, {"value": "391", "qualifiers": ["A"], "dateTime": "2008-07-02T03:00:00.000-05:00"}, {"value": "391", "qualifiers": ["A"], "dateTime": "2008-07-02T03:15:00.000-05:00"}, {"value": "391", "qualifiers": ["A"], "dateTime": "2008-07-02T03:30:00.000-05:00"}, {"value": "391", "qualifiers": ["A"], "dateTime": "2008-07-02T03:45:00.000-05:00"}, {"value": "391", "qualifiers": ["A"], "dateTime": "2008-07-02T04:00:00.000-05:00"}, {"value": "391", "qualifiers": ["A"], "dateTime": "2008-07-02T04:15:00.000-05:00"}, {"value": "389", "qualifiers": ["A"], "dateTime": "2008-07-02T04:30:00.000-05:00"}, {"value": "391", "qualifiers": ["A"], "dateTime": "2008-07-02T04:45:00.000-05:00"}, {"value": "391", "qualifiers": ["A"], "dateTime": "2008-07-02T05:00:00.000-05:00"}, {"value": "391", "qualifiers": ["A"], "dateTime": "2008-07-02T05:15:00.000-05:00"}, {"value": "391", "qualifiers": ["A"], "dateTime": "2008-07-02T05:30:00.000-05:00"}, {"value": "391", "qualifiers": ["A"], "dateTime": "2008-07-02T05:45:00.000-05:00"}, {"value": "391", "qualifiers": ["A"], "dateTime": "2008-07-02T06:00:00.000-05:00"}, {"value": "391", "qualifiers": ["A"], "dateTime": "2008-07-02T06:15:00.000-05:00"}, {"value": "391", "qualifiers": ["A"], "dateTime": "2008-07-02T06:30:00.000-05:00"}, {"value": "391", "qualifiers": ["A"], "dateTime": "2008-07-02T06:45:00.000-05:00"}, {"value": "391", "qualifiers": ["A"], "dateTime": "2008-07-02T07:00:00.000-05:00"}, {"value": "389", "qualifiers": ["A"], "dateTime": "2008-07-02T07:15:00.000-05:00"}, {"value": "389", "qualifiers": ["A"], "dateTime": "2008-07-02T07:30:00.000-05:00"}, {"value": "388", "qualifiers": ["A"], "dateTime": "2008-07-02T07:45:00.000-05:00"}, {"value": "389", "qualifiers": ["A"], "dateTime": "2008-07-02T08:00:00.000-05:00"}, {"value": "389", "qualifiers": ["A"], "dateTime": "2008-07-02T08:15:00.000-05:00"}, {"value": "389", "qualifiers": ["A"], "dateTime": "2008-07-02T08:30:00.000-05:00"}, {"value": "389", "qualifiers": ["A"], "dateTime": "2008-07-02T08:45:00.000-05:00"}, {"value": "389", "qualifiers": ["A"], "dateTime": "2008-07-02T09:00:00.000-05:00"}, {"value": "389", "qualifiers": ["A"], "dateTime": "2008-07-02T09:15:00.000-05:00"}, {"value": "388", "qualifiers": ["A"], "dateTime": "2008-07-02T09:30:00.000-05:00"}, {"value": "389", "qualifiers": ["A"], "dateTime": "2008-07-02T09:45:00.000-05:00"}, {"value": "389", "qualifiers": ["A"], "dateTime": "2008-07-02T10:00:00.000-05:00"}, {"value": "389", "qualifiers": ["A"], "dateTime": "2008-07-02T10:15:00.000-05:00"}, {"value": "389", "qualifiers": ["A"], "dateTime": "2008-07-02T10:30:00.000-05:00"}, {"value": "388", "qualifiers": ["A"], "dateTime": "2008-07-02T10:45:00.000-05:00"}, {"value": "389", "qualifiers": ["A"], "dateTime": "2008-07-02T11:00:00.000-05:00"}, {"value": "389", "qualifiers": ["A"], "dateTime": "2008-07-02T11:15:00.000-05:00"}, {"value": "389", "qualifiers": ["A"], "dateTime": "2008-07-02T11:30:00.000-05:00"}, {"value": "389", "qualifiers": ["A"], "dateTime": "2008-07-02T11:45:00.000-05:00"}, {"value": "388", "qualifiers": ["A"], "dateTime": "2008-07-02T12:00:00.000-05:00"}, {"value": "389", "qualifiers": ["A"], "dateTime": "2008-07-02T12:15:00.000-05:00"}, {"value": "388", "qualifiers": ["A"], "dateTime": "2008-07-02T12:30:00.000-05:00"}, {"value": "388", "qualifiers": ["A"], "dateTime": "2008-07-02T12:45:00.000-05:00"}, {"value": "389", "qualifiers": ["A"], "dateTime": "2008-07-02T13:00:00.000-05:00"}, {"value": "388", "qualifiers": ["A"], "dateTime": "2008-07-02T13:15:00.000-05:00"}, {"value": "388", "qualifiers": ["A"], "dateTime": "2008-07-02T13:30:00.000-05:00"}, {"value": "388", "qualifiers": ["A"], "dateTime": "2008-07-02T13:45:00.000-05:00"}, {"value": "388", "qualifiers": ["A"], "dateTime": "2008-07-02T14:00:00.000-05:00"}, {"value": "388", "qualifiers": ["A"], "dateTime": "2008-07-02T14:15:00.000-05:00"}, {"value": "388", "qualifiers": ["A"], "dateTime": "2008-07-02T14:30:00.000-05:00"}, {"value": "386", "qualifiers": ["A"], "dateTime": "2008-07-02T14:45:00.000-05:00"}, {"value": "388", "qualifiers": ["A"], "dateTime": "2008-07-02T15:00:00.000-05:00"}, {"value": "388", "qualifiers": ["A"], "dateTime": "2008-07-02T15:15:00.000-05:00"}, {"value": "388", "qualifiers": ["A"], "dateTime": "2008-07-02T15:30:00.000-05:00"}, {"value": "386", "qualifiers": ["A"], "dateTime": "2008-07-02T15:45:00.000-05:00"}, {"value": "388", "qualifiers": ["A"], "dateTime": "2008-07-02T16:00:00.000-05:00"}, {"value": "388", "qualifiers": ["A"], "dateTime": "2008-07-02T16:15:00.000-05:00"}, {"value": "388", "qualifiers": ["A"], "dateTime": "2008-07-02T16:30:00.000-05:00"}, {"value": "388", "qualifiers": ["A"], "dateTime": "2008-07-02T16:45:00.000-05:00"}, {"value": "388", "qualifiers": ["A"], "dateTime": "2008-07-02T17:00:00.000-05:00"}, {"value": "388", "qualifiers": ["A"], "dateTime": "2008-07-02T17:15:00.000-05:00"}, {"value": "386", "qualifiers": ["A"], "dateTime": "2008-07-02T17:30:00.000-05:00"}, {"value": "386", "qualifiers": ["A"], "dateTime": "2008-07-02T17:45:00.000-05:00"}, {"value": "386", "qualifiers": ["A"], "dateTime": "2008-07-02T18:00:00.000-05:00"}, {"value": "386", "qualifiers": ["A"], "dateTime": "2008-07-02T18:15:00.000-05:00"}, {"value": "384", "qualifiers": ["A"], "dateTime": "2008-07-02T18:30:00.000-05:00"}, {"value": "384", "qualifiers": ["A"], "dateTime": "2008-07-02T18:45:00.000-05:00"}, {"value": "384", "qualifiers": ["A"], "dateTime": "2008-07-02T19:00:00.000-05:00"}, {"value": "384", "qualifiers": ["A"], "dateTime": "2008-07-02T19:15:00.000-05:00"}, {"value": "382", "qualifiers": ["A"], "dateTime": "2008-07-02T19:30:00.000-05:00"}, {"value": "382", "qualifiers": ["A"], "dateTime": "2008-07-02T19:45:00.000-05:00"}, {"value": "382", "qualifiers": ["A"], "dateTime": "2008-07-02T20:00:00.000-05:00"}, {"value": "382", "qualifiers": ["A"], "dateTime": "2008-07-02T20:15:00.000-05:00"}, {"value": "382", "qualifiers": ["A"], "dateTime": "2008-07-02T20:30:00.000-05:00"}, {"value": "382", "qualifiers": ["A"], "dateTime": "2008-07-02T20:45:00.000-05:00"}, {"value": "381", "qualifiers": ["A"], "dateTime": "2008-07-02T21:00:00.000-05:00"}, {"value": "382", "qualifiers": ["A"], "dateTime": "2008-07-02T21:15:00.000-05:00"}, {"value": "381", "qualifiers": ["A"], "dateTime": "2008-07-02T21:30:00.000-05:00"}, {"value": "382", "qualifiers": ["A"], "dateTime": "2008-07-02T21:45:00.000-05:00"}, {"value": "381", "qualifiers": ["A"], "dateTime": "2008-07-02T22:00:00.000-05:00"}, {"value": "381", "qualifiers": ["A"], "dateTime": "2008-07-02T22:15:00.000-05:00"}, {"value": "381", "qualifiers": ["A"], "dateTime": "2008-07-02T22:30:00.000-05:00"}, {"value": "379", "qualifiers": ["A"], "dateTime": "2008-07-02T22:45:00.000-05:00"}, {"value": "379", "qualifiers": ["A"], "dateTime": "2008-07-02T23:00:00.000-05:00"}, {"value": "379", "qualifiers": ["A"], "dateTime": "2008-07-02T23:15:00.000-05:00"}, {"value": "377", "qualifiers": ["A"], "dateTime": "2008-07-02T23:30:00.000-05:00"}, {"value": "377", "qualifiers": ["A"], "dateTime": "2008-07-02T23:45:00.000-05:00"}, {"value": "377", "qualifiers": ["A"], "dateTime": "2008-07-03T00:00:00.000-05:00"}, {"value": "377", "qualifiers": ["A"], "dateTime": "2008-07-03T00:15:00.000-05:00"}, {"value": "376", "qualifiers": ["A"], "dateTime": "2008-07-03T00:30:00.000-05:00"}, {"value": "377", "qualifiers": ["A"], "dateTime": "2008-07-03T00:45:00.000-05:00"}, {"value": "376", "qualifiers": ["A"], "dateTime": "2008-07-03T01:00:00.000-05:00"}, {"value": "376", "qualifiers": ["A"], "dateTime": "2008-07-03T01:15:00.000-05:00"}, {"value": "376", "qualifiers": ["A"], "dateTime": "2008-07-03T01:30:00.000-05:00"}, {"value": "374", "qualifiers": ["A"], "dateTime": "2008-07-03T01:45:00.000-05:00"}, {"value": "376", "qualifiers": ["A"], "dateTime": "2008-07-03T02:00:00.000-05:00"}, {"value": "374", "qualifiers": ["A"], "dateTime": "2008-07-03T02:15:00.000-05:00"}, {"value": "376", "qualifiers": ["A"], "dateTime": "2008-07-03T02:30:00.000-05:00"}, {"value": "376", "qualifiers": ["A"], "dateTime": "2008-07-03T02:45:00.000-05:00"}, {"value": "374", "qualifiers": ["A"], "dateTime": "2008-07-03T03:00:00.000-05:00"}, {"value": "374", "qualifiers": ["A"], "dateTime": "2008-07-03T03:15:00.000-05:00"}, {"value": "374", "qualifiers": ["A"], "dateTime": "2008-07-03T03:30:00.000-05:00"}, {"value": "374", "qualifiers": ["A"], "dateTime": "2008-07-03T03:45:00.000-05:00"}, {"value": "374", "qualifiers": ["A"], "dateTime": "2008-07-03T04:00:00.000-05:00"}, {"value": "374", "qualifiers": ["A"], "dateTime": "2008-07-03T04:15:00.000-05:00"}, {"value": "374", "qualifiers": ["A"], "dateTime": "2008-07-03T04:30:00.000-05:00"}, {"value": "372", "qualifiers": ["A"], "dateTime": "2008-07-03T04:45:00.000-05:00"}, {"value": "374", "qualifiers": ["A"], "dateTime": "2008-07-03T05:00:00.000-05:00"}, {"value": "374", "qualifiers": ["A"], "dateTime": "2008-07-03T05:15:00.000-05:00"}, {"value": "372", "qualifiers": ["A"], "dateTime": "2008-07-03T05:30:00.000-05:00"}, {"value": "374", "qualifiers": ["A"], "dateTime": "2008-07-03T05:45:00.000-05:00"}, {"value": "374", "qualifiers": ["A"], "dateTime": "2008-07-03T06:00:00.000-05:00"}, {"value": "372", "qualifiers": ["A"], "dateTime": "2008-07-03T06:15:00.000-05:00"}, {"value": "372", "qualifiers": ["A"], "dateTime": "2008-07-03T06:30:00.000-05:00"}, {"value": "371", "qualifiers": ["A"], "dateTime": "2008-07-03T06:45:00.000-05:00"}, {"value": "371", "qualifiers": ["A"], "dateTime": "2008-07-03T07:00:00.000-05:00"}, {"value": "372", "qualifiers": ["A"], "dateTime": "2008-07-03T07:15:00.000-05:00"}, {"value": "371", "qualifiers": ["A"], "dateTime": "2008-07-03T07:30:00.000-05:00"}, {"value": "371", "qualifiers": ["A"], "dateTime": "2008-07-03T07:45:00.000-05:00"}, {"value": "371", "qualifiers": ["A"], "dateTime": "2008-07-03T08:00:00.000-05:00"}, {"value": "371", "qualifiers": ["A"], "dateTime": "2008-07-03T08:15:00.000-05:00"}, {"value": "369", "qualifiers": ["A"], "dateTime": "2008-07-03T08:30:00.000-05:00"}, {"value": "371", "qualifiers": ["A"], "dateTime": "2008-07-03T08:45:00.000-05:00"}, {"value": "371", "qualifiers": ["A"], "dateTime": "2008-07-03T09:00:00.000-05:00"}, {"value": "371", "qualifiers": ["A"], "dateTime": "2008-07-03T09:15:00.000-05:00"}, {"value": "369", "qualifiers": ["A"], "dateTime": "2008-07-03T09:30:00.000-05:00"}, {"value": "369", "qualifiers": ["A"], "dateTime": "2008-07-03T09:45:00.000-05:00"}, {"value": "369", "qualifiers": ["A"], "dateTime": "2008-07-03T10:00:00.000-05:00"}, {"value": "369", "qualifiers": ["A"], "dateTime": "2008-07-03T10:15:00.000-05:00"}, {"value": "369", "qualifiers": ["A"], "dateTime": "2008-07-03T10:30:00.000-05:00"}, {"value": "367", "qualifiers": ["A"], "dateTime": "2008-07-03T10:45:00.000-05:00"}, {"value": "367", "qualifiers": ["A"], "dateTime": "2008-07-03T11:00:00.000-05:00"}, {"value": "367", "qualifiers": ["A"], "dateTime": "2008-07-03T11:15:00.000-05:00"}, {"value": "367", "qualifiers": ["A"], "dateTime": "2008-07-03T11:30:00.000-05:00"}, {"value": "367", "qualifiers": ["A"], "dateTime": "2008-07-03T11:45:00.000-05:00"}, {"value": "367", "qualifiers": ["A"], "dateTime": "2008-07-03T12:00:00.000-05:00"}, {"value": "364", "qualifiers": ["A"], "dateTime": "2008-07-03T12:15:00.000-05:00"}, {"value": "366", "qualifiers": ["A"], "dateTime": "2008-07-03T12:30:00.000-05:00"}, {"value": "366", "qualifiers": ["A"], "dateTime": "2008-07-03T12:45:00.000-05:00"}, {"value": "366", "qualifiers": ["A"], "dateTime": "2008-07-03T13:00:00.000-05:00"}, {"value": "364", "qualifiers": ["A"], "dateTime": "2008-07-03T13:15:00.000-05:00"}, {"value": "364", "qualifiers": ["A"], "dateTime": "2008-07-03T13:30:00.000-05:00"}, {"value": "364", "qualifiers": ["A"], "dateTime": "2008-07-03T13:45:00.000-05:00"}, {"value": "364", "qualifiers": ["A"], "dateTime": "2008-07-03T14:00:00.000-05:00"}, {"value": "364", "qualifiers": ["A"], "dateTime": "2008-07-03T14:15:00.000-05:00"}, {"value": "364", "qualifiers": ["A"], "dateTime": "2008-07-03T14:30:00.000-05:00"}, {"value": "362", "qualifiers": ["A"], "dateTime": "2008-07-03T14:45:00.000-05:00"}, {"value": "362", "qualifiers": ["A"], "dateTime": "2008-07-03T15:00:00.000-05:00"}, {"value": "362", "qualifiers": ["A"], "dateTime": "2008-07-03T15:15:00.000-05:00"}, {"value": "362", "qualifiers": ["A"], "dateTime": "2008-07-03T15:30:00.000-05:00"}, {"value": "362", "qualifiers": ["A"], "dateTime": "2008-07-03T15:45:00.000-05:00"}, {"value": "362", "qualifiers": ["A"], "dateTime": "2008-07-03T16:00:00.000-05:00"}, {"value": "362", "qualifiers": ["A"], "dateTime": "2008-07-03T16:15:00.000-05:00"}, {"value": "362", "qualifiers": ["A"], "dateTime": "2008-07-03T16:30:00.000-05:00"}, {"value": "361", "qualifiers": ["A"], "dateTime": "2008-07-03T16:45:00.000-05:00"}, {"value": "361", "qualifiers": ["A"], "dateTime": "2008-07-03T17:00:00.000-05:00"}, {"value": "361", "qualifiers": ["A"], "dateTime": "2008-07-03T17:15:00.000-05:00"}, {"value": "361", "qualifiers": ["A"], "dateTime": "2008-07-03T17:30:00.000-05:00"}, {"value": "361", "qualifiers": ["A"], "dateTime": "2008-07-03T17:45:00.000-05:00"}, {"value": "361", "qualifiers": ["A"], "dateTime": "2008-07-03T18:00:00.000-05:00"}, {"value": "361", "qualifiers": ["A"], "dateTime": "2008-07-03T18:15:00.000-05:00"}, {"value": "359", "qualifiers": ["A"], "dateTime": "2008-07-03T18:30:00.000-05:00"}, {"value": "361", "qualifiers": ["A"], "dateTime": "2008-07-03T18:45:00.000-05:00"}, {"value": "361", "qualifiers": ["A"], "dateTime": "2008-07-03T19:00:00.000-05:00"}, {"value": "359", "qualifiers": ["A"], "dateTime": "2008-07-03T19:15:00.000-05:00"}, {"value": "359", "qualifiers": ["A"], "dateTime": "2008-07-03T19:30:00.000-05:00"}, {"value": "359", "qualifiers": ["A"], "dateTime": "2008-07-03T19:45:00.000-05:00"}, {"value": "359", "qualifiers": ["A"], "dateTime": "2008-07-03T20:00:00.000-05:00"}, {"value": "357", "qualifiers": ["A"], "dateTime": "2008-07-03T20:15:00.000-05:00"}, {"value": "359", "qualifiers": ["A"], "dateTime": "2008-07-03T20:30:00.000-05:00"}, {"value": "357", "qualifiers": ["A"], "dateTime": "2008-07-03T20:45:00.000-05:00"}, {"value": "357", "qualifiers": ["A"], "dateTime": "2008-07-03T21:00:00.000-05:00"}, {"value": "356", "qualifiers": ["A"], "dateTime": "2008-07-03T21:15:00.000-05:00"}, {"value": "357", "qualifiers": ["A"], "dateTime": "2008-07-03T21:30:00.000-05:00"}, {"value": "356", "qualifiers": ["A"], "dateTime": "2008-07-03T21:45:00.000-05:00"}, {"value": "356", "qualifiers": ["A"], "dateTime": "2008-07-03T22:00:00.000-05:00"}, {"value": "356", "qualifiers": ["A"], "dateTime": "2008-07-03T22:15:00.000-05:00"}, {"value": "356", "qualifiers": ["A"], "dateTime": "2008-07-03T22:30:00.000-05:00"}, {"value": "354", "qualifiers": ["A"], "dateTime": "2008-07-03T22:45:00.000-05:00"}, {"value": "356", "qualifiers": ["A"], "dateTime": "2008-07-03T23:00:00.000-05:00"}, {"value": "354", "qualifiers": ["A"], "dateTime": "2008-07-03T23:15:00.000-05:00"}, {"value": "354", "qualifiers": ["A"], "dateTime": "2008-07-03T23:30:00.000-05:00"}, {"value": "354", "qualifiers": ["A"], "dateTime": "2008-07-03T23:45:00.000-05:00"}, {"value": "354", "qualifiers": ["A"], "dateTime": "2008-07-04T00:00:00.000-05:00"}, {"value": "354", "qualifiers": ["A"], "dateTime": "2008-07-04T00:15:00.000-05:00"}, {"value": "354", "qualifiers": ["A"], "dateTime": "2008-07-04T00:30:00.000-05:00"}, {"value": "354", "qualifiers": ["A"], "dateTime": "2008-07-04T00:45:00.000-05:00"}, {"value": "354", "qualifiers": ["A"], "dateTime": "2008-07-04T01:00:00.000-05:00"}, {"value": "352", "qualifiers": ["A"], "dateTime": "2008-07-04T01:15:00.000-05:00"}, {"value": "352", "qualifiers": ["A"], "dateTime": "2008-07-04T01:30:00.000-05:00"}, {"value": "352", "qualifiers": ["A"], "dateTime": "2008-07-04T01:45:00.000-05:00"}, {"value": "352", "qualifiers": ["A"], "dateTime": "2008-07-04T02:00:00.000-05:00"}, {"value": "352", "qualifiers": ["A"], "dateTime": "2008-07-04T02:15:00.000-05:00"}, {"value": "352", "qualifiers": ["A"], "dateTime": "2008-07-04T02:30:00.000-05:00"}, {"value": "351", "qualifiers": ["A"], "dateTime": "2008-07-04T02:45:00.000-05:00"}, {"value": "351", "qualifiers": ["A"], "dateTime": "2008-07-04T03:00:00.000-05:00"}, {"value": "351", "qualifiers": ["A"], "dateTime": "2008-07-04T03:15:00.000-05:00"}, {"value": "351", "qualifiers": ["A"], "dateTime": "2008-07-04T03:30:00.000-05:00"}, {"value": "352", "qualifiers": ["A"], "dateTime": "2008-07-04T03:45:00.000-05:00"}, {"value": "351", "qualifiers": ["A"], "dateTime": "2008-07-04T04:00:00.000-05:00"}, {"value": "351", "qualifiers": ["A"], "dateTime": "2008-07-04T04:15:00.000-05:00"}, {"value": "351", "qualifiers": ["A"], "dateTime": "2008-07-04T04:30:00.000-05:00"}, {"value": "351", "qualifiers": ["A"], "dateTime": "2008-07-04T04:45:00.000-05:00"}, {"value": "351", "qualifiers": ["A"], "dateTime": "2008-07-04T05:00:00.000-05:00"}, {"value": "351", "qualifiers": ["A"], "dateTime": "2008-07-04T05:15:00.000-05:00"}, {"value": "349", "qualifiers": ["A"], "dateTime": "2008-07-04T05:30:00.000-05:00"}, {"value": "351", "qualifiers": ["A"], "dateTime": "2008-07-04T05:45:00.000-05:00"}, {"value": "349", "qualifiers": ["A"], "dateTime": "2008-07-04T06:00:00.000-05:00"}, {"value": "349", "qualifiers": ["A"], "dateTime": "2008-07-04T06:15:00.000-05:00"}, {"value": "349", "qualifiers": ["A"], "dateTime": "2008-07-04T06:30:00.000-05:00"}, {"value": "349", "qualifiers": ["A"], "dateTime": "2008-07-04T06:45:00.000-05:00"}, {"value": "349", "qualifiers": ["A"], "dateTime": "2008-07-04T07:00:00.000-05:00"}, {"value": "349", "qualifiers": ["A"], "dateTime": "2008-07-04T07:15:00.000-05:00"}, {"value": "349", "qualifiers": ["A"], "dateTime": "2008-07-04T07:30:00.000-05:00"}, {"value": "347", "qualifiers": ["A"], "dateTime": "2008-07-04T07:45:00.000-05:00"}, {"value": "349", "qualifiers": ["A"], "dateTime": "2008-07-04T08:00:00.000-05:00"}, {"value": "347", "qualifiers": ["A"], "dateTime": "2008-07-04T08:15:00.000-05:00"}, {"value": "349", "qualifiers": ["A"], "dateTime": "2008-07-04T08:30:00.000-05:00"}, {"value": "347", "qualifiers": ["A"], "dateTime": "2008-07-04T08:45:00.000-05:00"}, {"value": "347", "qualifiers": ["A"], "dateTime": "2008-07-04T09:00:00.000-05:00"}, {"value": "349", "qualifiers": ["A"], "dateTime": "2008-07-04T09:15:00.000-05:00"}, {"value": "349", "qualifiers": ["A"], "dateTime": "2008-07-04T09:30:00.000-05:00"}, {"value": "347", "qualifiers": ["A"], "dateTime": "2008-07-04T09:45:00.000-05:00"}, {"value": "347", "qualifiers": ["A"], "dateTime": "2008-07-04T10:00:00.000-05:00"}, {"value": "349", "qualifiers": ["A"], "dateTime": "2008-07-04T10:15:00.000-05:00"}, {"value": "347", "qualifiers": ["A"], "dateTime": "2008-07-04T10:30:00.000-05:00"}, {"value": "347", "qualifiers": ["A"], "dateTime": "2008-07-04T10:45:00.000-05:00"}, {"value": "347", "qualifiers": ["A"], "dateTime": "2008-07-04T11:00:00.000-05:00"}, {"value": "347", "qualifiers": ["A"], "dateTime": "2008-07-04T11:15:00.000-05:00"}, {"value": "347", "qualifiers": ["A"], "dateTime": "2008-07-04T11:30:00.000-05:00"}, {"value": "347", "qualifiers": ["A"], "dateTime": "2008-07-04T11:45:00.000-05:00"}, {"value": "347", "qualifiers": ["A"], "dateTime": "2008-07-04T12:00:00.000-05:00"}, {"value": "347", "qualifiers": ["A"], "dateTime": "2008-07-04T12:15:00.000-05:00"}, {"value": "346", "qualifiers": ["A"], "dateTime": "2008-07-04T12:30:00.000-05:00"}, {"value": "347", "qualifiers": ["A"], "dateTime": "2008-07-04T12:45:00.000-05:00"}, {"value": "347", "qualifiers": ["A"], "dateTime": "2008-07-04T13:00:00.000-05:00"}, {"value": "346", "qualifiers": ["A"], "dateTime": "2008-07-04T13:15:00.000-05:00"}, {"value": "346", "qualifiers": ["A"], "dateTime": "2008-07-04T13:30:00.000-05:00"}, {"value": "346", "qualifiers": ["A"], "dateTime": "2008-07-04T13:45:00.000-05:00"}, {"value": "347", "qualifiers": ["A"], "dateTime": "2008-07-04T14:00:00.000-05:00"}, {"value": "346", "qualifiers": ["A"], "dateTime": "2008-07-04T14:15:00.000-05:00"}, {"value": "346", "qualifiers": ["A"], "dateTime": "2008-07-04T14:30:00.000-05:00"}, {"value": "346", "qualifiers": ["A"], "dateTime": "2008-07-04T14:45:00.000-05:00"}, {"value": "346", "qualifiers": ["A"], "dateTime": "2008-07-04T15:00:00.000-05:00"}, {"value": "346", "qualifiers": ["A"], "dateTime": "2008-07-04T15:15:00.000-05:00"}, {"value": "346", "qualifiers": ["A"], "dateTime": "2008-07-04T15:30:00.000-05:00"}, {"value": "346", "qualifiers": ["A"], "dateTime": "2008-07-04T15:45:00.000-05:00"}, {"value": "346", "qualifiers": ["A"], "dateTime": "2008-07-04T16:00:00.000-05:00"}, {"value": "346", "qualifiers": ["A"], "dateTime": "2008-07-04T16:15:00.000-05:00"}, {"value": "346", "qualifiers": ["A"], "dateTime": "2008-07-04T16:30:00.000-05:00"}, {"value": "346", "qualifiers": ["A"], "dateTime": "2008-07-04T16:45:00.000-05:00"}, {"value": "346", "qualifiers": ["A"], "dateTime": "2008-07-04T17:00:00.000-05:00"}, {"value": "346", "qualifiers": ["A"], "dateTime": "2008-07-04T17:15:00.000-05:00"}, {"value": "346", "qualifiers": ["A"], "dateTime": "2008-07-04T17:30:00.000-05:00"}, {"value": "344", "qualifiers": ["A"], "dateTime": "2008-07-04T17:45:00.000-05:00"}, {"value": "344", "qualifiers": ["A"], "dateTime": "2008-07-04T18:00:00.000-05:00"}, {"value": "344", "qualifiers": ["A"], "dateTime": "2008-07-04T18:15:00.000-05:00"}, {"value": "344", "qualifiers": ["A"], "dateTime": "2008-07-04T18:30:00.000-05:00"}, {"value": "344", "qualifiers": ["A"], "dateTime": "2008-07-04T18:45:00.000-05:00"}, {"value": "344", "qualifiers": ["A"], "dateTime": "2008-07-04T19:00:00.000-05:00"}, {"value": "343", "qualifiers": ["A"], "dateTime": "2008-07-04T19:15:00.000-05:00"}, {"value": "344", "qualifiers": ["A"], "dateTime": "2008-07-04T19:30:00.000-05:00"}, {"value": "344", "qualifiers": ["A"], "dateTime": "2008-07-04T19:45:00.000-05:00"}, {"value": "343", "qualifiers": ["A"], "dateTime": "2008-07-04T20:00:00.000-05:00"}, {"value": "343", "qualifiers": ["A"], "dateTime": "2008-07-04T20:15:00.000-05:00"}, {"value": "344", "qualifiers": ["A"], "dateTime": "2008-07-04T20:30:00.000-05:00"}, {"value": "343", "qualifiers": ["A"], "dateTime": "2008-07-04T20:45:00.000-05:00"}, {"value": "343", "qualifiers": ["A"], "dateTime": "2008-07-04T21:00:00.000-05:00"}, {"value": "343", "qualifiers": ["A"], "dateTime": "2008-07-04T21:15:00.000-05:00"}, {"value": "343", "qualifiers": ["A"], "dateTime": "2008-07-04T21:30:00.000-05:00"}, {"value": "341", "qualifiers": ["A"], "dateTime": "2008-07-04T21:45:00.000-05:00"}, {"value": "341", "qualifiers": ["A"], "dateTime": "2008-07-04T22:00:00.000-05:00"}, {"value": "341", "qualifiers": ["A"], "dateTime": "2008-07-04T22:15:00.000-05:00"}, {"value": "343", "qualifiers": ["A"], "dateTime": "2008-07-04T22:30:00.000-05:00"}, {"value": "343", "qualifiers": ["A"], "dateTime": "2008-07-04T22:45:00.000-05:00"}, {"value": "341", "qualifiers": ["A"], "dateTime": "2008-07-04T23:00:00.000-05:00"}, {"value": "341", "qualifiers": ["A"], "dateTime": "2008-07-04T23:15:00.000-05:00"}, {"value": "341", "qualifiers": ["A"], "dateTime": "2008-07-04T23:30:00.000-05:00"}, {"value": "341", "qualifiers": ["A"], "dateTime": "2008-07-04T23:45:00.000-05:00"}, {"value": "339", "qualifiers": ["A"], "dateTime": "2008-07-05T00:00:00.000-05:00"}, {"value": "339", "qualifiers": ["A"], "dateTime": "2008-07-05T00:15:00.000-05:00"}, {"value": "339", "qualifiers": ["A"], "dateTime": "2008-07-05T00:30:00.000-05:00"}, {"value": "339", "qualifiers": ["A"], "dateTime": "2008-07-05T00:45:00.000-05:00"}, {"value": "339", "qualifiers": ["A"], "dateTime": "2008-07-05T01:00:00.000-05:00"}, {"value": "338", "qualifiers": ["A"], "dateTime": "2008-07-05T01:15:00.000-05:00"}, {"value": "339", "qualifiers": ["A"], "dateTime": "2008-07-05T01:30:00.000-05:00"}, {"value": "339", "qualifiers": ["A"], "dateTime": "2008-07-05T01:45:00.000-05:00"}, {"value": "339", "qualifiers": ["A"], "dateTime": "2008-07-05T02:00:00.000-05:00"}, {"value": "339", "qualifiers": ["A"], "dateTime": "2008-07-05T02:15:00.000-05:00"}, {"value": "339", "qualifiers": ["A"], "dateTime": "2008-07-05T02:30:00.000-05:00"}, {"value": "339", "qualifiers": ["A"], "dateTime": "2008-07-05T02:45:00.000-05:00"}, {"value": "338", "qualifiers": ["A"], "dateTime": "2008-07-05T03:00:00.000-05:00"}, {"value": "339", "qualifiers": ["A"], "dateTime": "2008-07-05T03:15:00.000-05:00"}, {"value": "339", "qualifiers": ["A"], "dateTime": "2008-07-05T03:30:00.000-05:00"}, {"value": "338", "qualifiers": ["A"], "dateTime": "2008-07-05T03:45:00.000-05:00"}, {"value": "338", "qualifiers": ["A"], "dateTime": "2008-07-05T04:00:00.000-05:00"}, {"value": "336", "qualifiers": ["A"], "dateTime": "2008-07-05T04:15:00.000-05:00"}, {"value": "338", "qualifiers": ["A"], "dateTime": "2008-07-05T04:30:00.000-05:00"}, {"value": "338", "qualifiers": ["A"], "dateTime": "2008-07-05T04:45:00.000-05:00"}, {"value": "338", "qualifiers": ["A"], "dateTime": "2008-07-05T05:00:00.000-05:00"}, {"value": "336", "qualifiers": ["A"], "dateTime": "2008-07-05T05:15:00.000-05:00"}, {"value": "336", "qualifiers": ["A"], "dateTime": "2008-07-05T05:30:00.000-05:00"}, {"value": "336", "qualifiers": ["A"], "dateTime": "2008-07-05T05:45:00.000-05:00"}, {"value": "334", "qualifiers": ["A"], "dateTime": "2008-07-05T06:00:00.000-05:00"}, {"value": "336", "qualifiers": ["A"], "dateTime": "2008-07-05T06:15:00.000-05:00"}, {"value": "336", "qualifiers": ["A"], "dateTime": "2008-07-05T06:30:00.000-05:00"}, {"value": "336", "qualifiers": ["A"], "dateTime": "2008-07-05T06:45:00.000-05:00"}, {"value": "336", "qualifiers": ["A"], "dateTime": "2008-07-05T07:00:00.000-05:00"}, {"value": "336", "qualifiers": ["A"], "dateTime": "2008-07-05T07:15:00.000-05:00"}, {"value": "336", "qualifiers": ["A"], "dateTime": "2008-07-05T07:30:00.000-05:00"}, {"value": "336", "qualifiers": ["A"], "dateTime": "2008-07-05T07:45:00.000-05:00"}, {"value": "336", "qualifiers": ["A"], "dateTime": "2008-07-05T08:00:00.000-05:00"}, {"value": "334", "qualifiers": ["A"], "dateTime": "2008-07-05T08:15:00.000-05:00"}, {"value": "336", "qualifiers": ["A"], "dateTime": "2008-07-05T08:30:00.000-05:00"}, {"value": "334", "qualifiers": ["A"], "dateTime": "2008-07-05T08:45:00.000-05:00"}, {"value": "336", "qualifiers": ["A"], "dateTime": "2008-07-05T09:00:00.000-05:00"}, {"value": "334", "qualifiers": ["A"], "dateTime": "2008-07-05T09:15:00.000-05:00"}, {"value": "334", "qualifiers": ["A"], "dateTime": "2008-07-05T09:30:00.000-05:00"}, {"value": "333", "qualifiers": ["A"], "dateTime": "2008-07-05T09:45:00.000-05:00"}, {"value": "334", "qualifiers": ["A"], "dateTime": "2008-07-05T10:00:00.000-05:00"}, {"value": "333", "qualifiers": ["A"], "dateTime": "2008-07-05T10:15:00.000-05:00"}, {"value": "334", "qualifiers": ["A"], "dateTime": "2008-07-05T10:30:00.000-05:00"}, {"value": "334", "qualifiers": ["A"], "dateTime": "2008-07-05T10:45:00.000-05:00"}, {"value": "334", "qualifiers": ["A"], "dateTime": "2008-07-05T11:00:00.000-05:00"}, {"value": "334", "qualifiers": ["A"], "dateTime": "2008-07-05T11:15:00.000-05:00"}, {"value": "334", "qualifiers": ["A"], "dateTime": "2008-07-05T11:30:00.000-05:00"}, {"value": "334", "qualifiers": ["A"], "dateTime": "2008-07-05T11:45:00.000-05:00"}, {"value": "333", "qualifiers": ["A"], "dateTime": "2008-07-05T12:00:00.000-05:00"}, {"value": "334", "qualifiers": ["A"], "dateTime": "2008-07-05T12:15:00.000-05:00"}, {"value": "333", "qualifiers": ["A"], "dateTime": "2008-07-05T12:30:00.000-05:00"}, {"value": "333", "qualifiers": ["A"], "dateTime": "2008-07-05T12:45:00.000-05:00"}, {"value": "331", "qualifiers": ["A"], "dateTime": "2008-07-05T13:00:00.000-05:00"}, {"value": "331", "qualifiers": ["A"], "dateTime": "2008-07-05T13:15:00.000-05:00"}, {"value": "334", "qualifiers": ["A"], "dateTime": "2008-07-05T13:30:00.000-05:00"}, {"value": "334", "qualifiers": ["A"], "dateTime": "2008-07-05T13:45:00.000-05:00"}, {"value": "333", "qualifiers": ["A"], "dateTime": "2008-07-05T14:00:00.000-05:00"}, {"value": "333", "qualifiers": ["A"], "dateTime": "2008-07-05T14:15:00.000-05:00"}, {"value": "333", "qualifiers": ["A"], "dateTime": "2008-07-05T14:30:00.000-05:00"}, {"value": "331", "qualifiers": ["A"], "dateTime": "2008-07-05T14:45:00.000-05:00"}, {"value": "333", "qualifiers": ["A"], "dateTime": "2008-07-05T15:00:00.000-05:00"}, {"value": "333", "qualifiers": ["A"], "dateTime": "2008-07-05T15:15:00.000-05:00"}, {"value": "331", "qualifiers": ["A"], "dateTime": "2008-07-05T15:30:00.000-05:00"}, {"value": "331", "qualifiers": ["A"], "dateTime": "2008-07-05T15:45:00.000-05:00"}, {"value": "331", "qualifiers": ["A"], "dateTime": "2008-07-05T16:00:00.000-05:00"}, {"value": "331", "qualifiers": ["A"], "dateTime": "2008-07-05T16:15:00.000-05:00"}, {"value": "331", "qualifiers": ["A"], "dateTime": "2008-07-05T16:30:00.000-05:00"}, {"value": "331", "qualifiers": ["A"], "dateTime": "2008-07-05T16:45:00.000-05:00"}, {"value": "331", "qualifiers": ["A"], "dateTime": "2008-07-05T17:00:00.000-05:00"}, {"value": "331", "qualifiers": ["A"], "dateTime": "2008-07-05T17:15:00.000-05:00"}, {"value": "331", "qualifiers": ["A"], "dateTime": "2008-07-05T17:30:00.000-05:00"}, {"value": "331", "qualifiers": ["A"], "dateTime": "2008-07-05T17:45:00.000-05:00"}, {"value": "331", "qualifiers": ["A"], "dateTime": "2008-07-05T18:00:00.000-05:00"}, {"value": "330", "qualifiers": ["A"], "dateTime": "2008-07-05T18:15:00.000-05:00"}, {"value": "330", "qualifiers": ["A"], "dateTime": "2008-07-05T18:30:00.000-05:00"}, {"value": "330", "qualifiers": ["A"], "dateTime": "2008-07-05T18:45:00.000-05:00"}, {"value": "330", "qualifiers": ["A"], "dateTime": "2008-07-05T19:00:00.000-05:00"}, {"value": "330", "qualifiers": ["A"], "dateTime": "2008-07-05T19:15:00.000-05:00"}, {"value": "330", "qualifiers": ["A"], "dateTime": "2008-07-05T19:30:00.000-05:00"}, {"value": "330", "qualifiers": ["A"], "dateTime": "2008-07-05T19:45:00.000-05:00"}, {"value": "330", "qualifiers": ["A"], "dateTime": "2008-07-05T20:00:00.000-05:00"}, {"value": "330", "qualifiers": ["A"], "dateTime": "2008-07-05T20:15:00.000-05:00"}, {"value": "330", "qualifiers": ["A"], "dateTime": "2008-07-05T20:30:00.000-05:00"}, {"value": "330", "qualifiers": ["A"], "dateTime": "2008-07-05T20:45:00.000-05:00"}, {"value": "330", "qualifiers": ["A"], "dateTime": "2008-07-05T21:00:00.000-05:00"}, {"value": "330", "qualifiers": ["A"], "dateTime": "2008-07-05T21:15:00.000-05:00"}, {"value": "330", "qualifiers": ["A"], "dateTime": "2008-07-05T21:30:00.000-05:00"}, {"value": "328", "qualifiers": ["A"], "dateTime": "2008-07-05T21:45:00.000-05:00"}, {"value": "330", "qualifiers": ["A"], "dateTime": "2008-07-05T22:00:00.000-05:00"}, {"value": "328", "qualifiers": ["A"], "dateTime": "2008-07-05T22:15:00.000-05:00"}, {"value": "328", "qualifiers": ["A"], "dateTime": "2008-07-05T22:30:00.000-05:00"}, {"value": "328", "qualifiers": ["A"], "dateTime": "2008-07-05T22:45:00.000-05:00"}, {"value": "328", "qualifiers": ["A"], "dateTime": "2008-07-05T23:00:00.000-05:00"}, {"value": "328", "qualifiers": ["A"], "dateTime": "2008-07-05T23:15:00.000-05:00"}, {"value": "328", "qualifiers": ["A"], "dateTime": "2008-07-05T23:30:00.000-05:00"}, {"value": "328", "qualifiers": ["A"], "dateTime": "2008-07-05T23:45:00.000-05:00"}], "qualifier": [{"qualifierCode": "A", "qualifierDescription": "Approved for publication -- Processing and review completed.", "qualifierID": 0, "network": "NWIS", "vocabulary": "uv_rmk_cd"}], "qualityControlLevel": [], "method": [{"methodDescription": "", "methodID": 43974}], "source": [], "offset": [], "sample": [], "censorCode": []}], "name": "USGS:05463500:00060:00000"}]
-    self.url = 'https://waterservices.usgs.gov/nwis/iv/?site=05463500&startDT=2008-05-20&endDT=2008-07-05&parameterCD=00060&format=json' 
-    self.prefix = 'https://waterservices.usgs.gov/nwis/'
-    self.H_START_DT = '2008-05-20'
-    self.H_END_DT = '2008-07-05'
-    self.sites = ['05463500', '05471050', '05420680', '05479000', '05484000', '05481000', '05486000', '05421000', '05485500',
-              '05455100', '05470500', '05451500', '05458000', '05471000', '05462000', '05457700', '05458500', '05470000',
-              '05484500', '05481300', '05464220', '05458900', '05485605', '05463000', '05471200', '05476750', '05411850',
-              '05454220', '05481950', '05416900', '05464500', '05487470']
-
-  def test_status_code(self):
-    with requests_mock.Mocker() as m:
-      m.get(self.url, status_code=404)
-      self.assertEqual(req_hydrodata(['05463500'], self.H_START_DT, self.H_END_DT, self.prefix), None)
-
-
-
-  def test_bad_list(self):
-    self.assertEqual(req_hydrodata(['x'], self.H_START_DT, self.H_END_DT, self.prefix), None)
-    self.assertEqual(req_hydrodata([123456, 100], self.H_START_DT, self.H_END_DT, self.prefix), None)
-    self.assertEqual(req_hydrodata([], self.H_START_DT, self.H_END_DT, self.prefix), None)
-
-  def test_bad_url(self):
-    self.assertEqual(req_hydrodata(self.sites, self.H_START_DT, self.H_END_DT, 'badurl.noob' ), None)
-    self.assertEqual(req_hydrodata(self.sites, self.H_START_DT, self.H_END_DT, ''), None)
-
-  def test_valid_data(self):
-    self.assertEqual(req_hydrodata(['05463500'], self.H_START_DT, self.H_END_DT, self.prefix), self.valid_data)
- 
-
-class TestParseHydroData(unittest.TestCase):
-
-  def setUp(self):
-
-    self.mock_data = [{"sourceInfo": {
-   "siteName": "Black Hawk Creek at Hudson, IA",
-   "siteCode": [{"value": "05463500","network": "NWIS","agencyCode": "USGS"}],
-   "timeZoneInfo": {"defaultTimeZone": {"zoneOffset": "-06:00","zoneAbbreviation": "CST"},"daylightSavingsTimeZone": {
-    "zoneOffset": "-05:00","zoneAbbreviation": "CDT"},"siteUsesDaylightSavingsTime": True},"geoLocation": {
-    "geogLocation": {"srs": "EPSG:4326","latitude": 42.4077639,"longitude": -92.4632451},"localSiteXY": []},
-   "note": [],"siteType": [],},"variable": {"variableCode": [],"variableName": "Streamflow, ft&#179;/s",
-   "variableDescription": "Discharge, cubic feet per second","valueType": "Derived Value","unit": {"unitCode": "ft3/s"},
-   "note": [],},"values": [{"value": [{
-      "value": "446","qualifiers": ["A"],"dateTime": "2008-05-20T00:00:00.000-05:00"},{
-      "value": "446","qualifiers": ["A"],"dateTime": "2008-05-20T00:15:00.000-05:00"},{
-      "value": "446","qualifiers": ["A"],"dateTime": "2008-05-20T00:30:00.000-05:00"},{
-      "value": "444","qualifiers": ["A"],"dateTime": "2008-05-20T00:45:00.000-05:00"},{
-      "value": "446","qualifiers": ["A"],"dateTime": "2008-05-20T01:00:00.000-05:00"},{
-      "value": "446","qualifiers": ["A"],"dateTime": "2008-05-20T01:15:00.000-05:00"}]}]}]
-
-    self.mock_parsed_data = [{'key': 'Black Hawk Creek at Hudson, IA', 'values': [
-     {'date': '2008-05-20', 'time': 0, 'value': 0}, 
-     {'date': '2008-05-20', 'time': '00:00:00', 'value': '446'},
-     {'date': '2008-05-20', 'time': '00:15:00', 'value': '446'},
-     {'date': '2008-05-20', 'time': '00:30:00', 'value': '446'},
-     {'date': '2008-05-20', 'time': '00:45:00', 'value': '444'}, 
-     {'date': '2008-05-20', 'time': '01:00:00', 'value': '446'},
-     {'date': '2008-05-20', 'time': '01:15:00', 'value': '446'}]}] 
-
-
-  def test_empty_data(self):
-    self.assertEqual(parse_hydrodata([]), [])
-    self.assertEqual(parse_hydrodata(None), [])
-
-  def test_data_not_list(self):
-    self.assertEqual(parse_hydrodata({}), [])
-
-  def test_valid_data(self):
-    # Delete 'time_mili' key before comparing due to python time conversion inconsistencies.     
-    ret = parse_hydrodata(self.mock_data)[0]
-    for k1, v1 in ret.items():
-      if k1 is 'values':
-        for item in ret[k1]:
-          if item['time_mili'] is not None:
-            del item['time_mili']
-    self.assertEqual([ret], self.mock_parsed_data)
